Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-03-03   13.19219   9.222351  10.326633         NA         NA         NA
#> 2020-04-23   13.05305  11.367966   8.900458   8.977792         NA         NA
#> 2020-05-18   12.23271  10.276382  10.419803   9.651772   8.345286         NA
#> 2021-03-04   12.15595   8.909430  10.259338   9.605484   9.045405   9.088301
#> 2021-04-30   11.94392   9.270128  10.434061   9.137313   8.992241  10.815723
#> 2021-12-01   12.03563   9.222598  10.266913   9.549466   8.226533  13.120533
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-03-03         NA         NA         NA         NA         NA         NA
#> 2020-04-23         NA         NA         NA         NA         NA         NA
#> 2020-05-18         NA         NA         NA         NA         NA         NA
#> 2021-03-04    8.51937   5.677529   11.67753   5.962036   6.472324   2.636981
#> 2021-04-30   10.22893   8.526011   10.97551   7.620775   4.416657   5.756089
#> 2021-12-01   11.11699   8.572881   11.23256   8.808101   5.176483   6.331243
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-03-03         NA         NA         NA         NA         NA         NA
#> 2020-04-23         NA         NA         NA         NA         NA         NA
#> 2020-05-18         NA         NA         NA         NA         NA         NA
#> 2021-03-04   4.016348   3.450894   9.431155         NA         NA         NA
#> 2021-04-30   6.438128   7.136296   9.790140   7.231323         NA         NA
#> 2021-12-01   5.825506   7.922060   9.879795   8.778133   11.56936   13.61273
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-03-03         NA         NA         NA         NA         NA         NA
#> 2020-04-23         NA         NA         NA         NA         NA         NA
#> 2020-05-18         NA         NA         NA         NA         NA         NA
#> 2021-03-04         NA         NA         NA         NA         NA         NA
#> 2021-04-30         NA         NA         NA         NA         NA         NA
#> 2021-12-01    14.7243   9.654179   10.76991   11.27963  0.1115221   1.376587