Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-05-31 2020-01-01  8.55651591
#> 2   2020-05-31 2020-02-01 14.35856607
#> 3   2020-05-31 2020-03-01 10.41101091
#> 4   2020-05-31 2020-04-01  4.82432591
#> 5   2020-05-31 2020-05-01  1.84461025
#> 6   2020-05-31 2020-06-01          NA
#> 7   2020-05-31 2020-07-01          NA
#> 8   2020-05-31 2020-08-01          NA
#> 9   2020-05-31 2020-09-01          NA
#> 10  2020-05-31 2020-10-01          NA
#> 11  2020-05-31 2020-11-01          NA
#> 12  2020-05-31 2020-12-01          NA
#> 13  2020-05-31 2021-01-01          NA
#> 14  2020-05-31 2021-02-01          NA
#> 15  2020-05-31 2021-03-01          NA
#> 16  2020-05-31 2021-04-01          NA
#> 17  2020-05-31 2021-05-01          NA
#> 18  2020-05-31 2021-06-01          NA
#> 19  2020-05-31 2021-07-01          NA
#> 20  2020-05-31 2021-08-01          NA
#> 21  2020-05-31 2021-09-01          NA
#> 22  2020-05-31 2021-10-01          NA
#> 23  2020-05-31 2021-11-01          NA
#> 24  2020-05-31 2021-12-01          NA
#> 25  2020-07-30 2020-01-01 12.05910313
#> 26  2020-07-30 2020-02-01 10.14370012
#> 27  2020-07-30 2020-03-01  4.57151005
#> 28  2020-07-30 2020-04-01  3.31228966
#> 29  2020-07-30 2020-05-01  1.65058502
#> 30  2020-07-30 2020-06-01 -0.04574678
#> 31  2020-07-30 2020-07-01  1.62381094
#> 32  2020-07-30 2020-08-01          NA
#> 33  2020-07-30 2020-09-01          NA
#> 34  2020-07-30 2020-10-01          NA
#> 35  2020-07-30 2020-11-01          NA
#> 36  2020-07-30 2020-12-01          NA
#> 37  2020-07-30 2021-01-01          NA
#> 38  2020-07-30 2021-02-01          NA
#> 39  2020-07-30 2021-03-01          NA
#> 40  2020-07-30 2021-04-01          NA
#> 41  2020-07-30 2021-05-01          NA
#> 42  2020-07-30 2021-06-01          NA
#> 43  2020-07-30 2021-07-01          NA
#> 44  2020-07-30 2021-08-01          NA
#> 45  2020-07-30 2021-09-01          NA
#> 46  2020-07-30 2021-10-01          NA
#> 47  2020-07-30 2021-11-01          NA
#> 48  2020-07-30 2021-12-01          NA
#> 49  2020-11-20 2020-01-01 11.84359256
#> 50  2020-11-20 2020-02-01 12.14212725
#> 51  2020-11-20 2020-03-01  9.08144869
#> 52  2020-11-20 2020-04-01  4.11781652
#> 53  2020-11-20 2020-05-01  2.25412518
#> 54  2020-11-20 2020-06-01  4.24378641
#> 55  2020-11-20 2020-07-01 -1.53950843
#> 56  2020-11-20 2020-08-01 -6.55474391
#> 57  2020-11-20 2020-09-01 -9.55986253
#> 58  2020-11-20 2020-10-01 -9.74528794
#> 59  2020-11-20 2020-11-01 -7.07063445
#> 60  2020-11-20 2020-12-01          NA
#> 61  2020-11-20 2021-01-01          NA
#> 62  2020-11-20 2021-02-01          NA
#> 63  2020-11-20 2021-03-01          NA
#> 64  2020-11-20 2021-04-01          NA
#> 65  2020-11-20 2021-05-01          NA
#> 66  2020-11-20 2021-06-01          NA
#> 67  2020-11-20 2021-07-01          NA
#> 68  2020-11-20 2021-08-01          NA
#> 69  2020-11-20 2021-09-01          NA
#> 70  2020-11-20 2021-10-01          NA
#> 71  2020-11-20 2021-11-01          NA
#> 72  2020-11-20 2021-12-01          NA
#> 73  2020-12-12 2020-01-01 12.24215355
#> 74  2020-12-12 2020-02-01 11.83624194
#> 75  2020-12-12 2020-03-01  8.57627698
#> 76  2020-12-12 2020-04-01  3.17022450
#> 77  2020-12-12 2020-05-01  2.68133143
#> 78  2020-12-12 2020-06-01  2.07116675
#> 79  2020-12-12 2020-07-01 -1.86928629
#> 80  2020-12-12 2020-08-01 -8.35276321
#> 81  2020-12-12 2020-09-01 -8.92320718
#> 82  2020-12-12 2020-10-01 -8.87007141
#> 83  2020-12-12 2020-11-01 -5.29742568
#> 84  2020-12-12 2020-12-01 -6.38034044
#> 85  2020-12-12 2021-01-01          NA
#> 86  2020-12-12 2021-02-01          NA
#> 87  2020-12-12 2021-03-01          NA
#> 88  2020-12-12 2021-04-01          NA
#> 89  2020-12-12 2021-05-01          NA
#> 90  2020-12-12 2021-06-01          NA
#> 91  2020-12-12 2021-07-01          NA
#> 92  2020-12-12 2021-08-01          NA
#> 93  2020-12-12 2021-09-01          NA
#> 94  2020-12-12 2021-10-01          NA
#> 95  2020-12-12 2021-11-01          NA
#> 96  2020-12-12 2021-12-01          NA
#> 97  2021-04-26 2020-01-01 11.67135186
#> 98  2021-04-26 2020-02-01 11.96957170
#> 99  2021-04-26 2020-03-01  7.95004509
#> 100 2021-04-26 2020-04-01  3.69659563
#> 101 2021-04-26 2020-05-01  2.38905558
#> 102 2021-04-26 2020-06-01  2.65269490
#> 103 2021-04-26 2020-07-01 -1.08051825
#> 104 2021-04-26 2020-08-01 -8.92343439
#> 105 2021-04-26 2020-09-01 -9.57842486
#> 106 2021-04-26 2020-10-01 -9.60296823
#> 107 2021-04-26 2020-11-01 -6.13187148
#> 108 2021-04-26 2020-12-01 -5.07955211
#> 109 2021-04-26 2021-01-01 -3.06543416
#> 110 2021-04-26 2021-02-01  1.51825708
#> 111 2021-04-26 2021-03-01 -2.39512545
#> 112 2021-04-26 2021-04-01 -3.77433700
#> 113 2021-04-26 2021-05-01          NA
#> 114 2021-04-26 2021-06-01          NA
#> 115 2021-04-26 2021-07-01          NA
#> 116 2021-04-26 2021-08-01          NA
#> 117 2021-04-26 2021-09-01          NA
#> 118 2021-04-26 2021-10-01          NA
#> 119 2021-04-26 2021-11-01          NA
#> 120 2021-04-26 2021-12-01          NA
#> 121 2021-05-04 2020-01-01 11.91311866
#> 122 2021-05-04 2020-02-01 11.95935347
#> 123 2021-05-04 2020-03-01  8.23293601
#> 124 2021-05-04 2020-04-01  3.56092922
#> 125 2021-05-04 2020-05-01  2.42806573
#> 126 2021-05-04 2020-06-01  2.66784833
#> 127 2021-05-04 2020-07-01 -1.44164615
#> 128 2021-05-04 2020-08-01 -9.27525534
#> 129 2021-05-04 2020-09-01 -9.25154349
#> 130 2021-05-04 2020-10-01 -8.61345764
#> 131 2021-05-04 2020-11-01 -5.83895470
#> 132 2021-05-04 2020-12-01 -5.00364090
#> 133 2021-05-04 2021-01-01 -0.80177867
#> 134 2021-05-04 2021-02-01  3.54095091
#> 135 2021-05-04 2021-03-01 -2.87906447
#> 136 2021-05-04 2021-04-01 -2.84708788
#> 137 2021-05-04 2021-05-01 -0.09864977
#> 138 2021-05-04 2021-06-01          NA
#> 139 2021-05-04 2021-07-01          NA
#> 140 2021-05-04 2021-08-01          NA
#> 141 2021-05-04 2021-09-01          NA
#> 142 2021-05-04 2021-10-01          NA
#> 143 2021-05-04 2021-11-01          NA
#> 144 2021-05-04 2021-12-01          NA