Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   13.57496  14.731799 14.6385710 15.0474608 15.4173718 15.1836039
#> Feb 2020         NA  11.544216 10.4846136 10.6622525 11.3296947 10.9833770
#> Mar 2020         NA   3.746606  6.6029445  6.5071298  6.3866627  6.3504787
#> Apr 2020         NA   9.558742  4.8038953  5.7207196  5.8320110  5.7796335
#> May 2020         NA  -2.444644  1.7926482  1.9666557  1.9124798  2.1452263
#> Jun 2020         NA   8.537736  4.9299376  4.6037148  4.1431702  4.3467233
#> Jul 2020         NA  -3.305657 -0.9041741 -1.3063128 -1.3758856 -1.3263248
#> Aug 2020         NA         NA -1.6257191  0.4751147  0.9649832  0.1613438
#> Sep 2020         NA         NA -2.4371532 -2.3107249 -2.5937448 -1.9687991
#> Oct 2020         NA         NA  6.9647916  2.2409524  3.0479650  3.9751713
#> Nov 2020         NA         NA         NA  4.4368788  5.3008912  3.0105518
#> Dec 2020         NA         NA         NA  2.4981172  5.7041622  5.2178334
#> Jan 2021         NA         NA         NA  8.5128124  7.7156230  8.4241054
#> Feb 2021         NA         NA         NA  7.7766345  5.7129833  3.0340107
#> Mar 2021         NA         NA         NA  0.3301653  0.2405502  1.6895646
#> Apr 2021         NA         NA         NA -0.2425346  3.7895039  3.4172270
#> May 2021         NA         NA         NA  3.4565926  0.8403430  0.5690962
#> Jun 2021         NA         NA         NA  2.9730892  3.2260683  2.7474914
#> Jul 2021         NA         NA         NA         NA  4.8690941  1.9860552
#> Aug 2021         NA         NA         NA         NA         NA  5.0031323
#> Sep 2021         NA         NA         NA         NA         NA -4.9104531
#> Oct 2021         NA         NA         NA         NA         NA  2.2757330
#> Nov 2021         NA         NA         NA         NA         NA  1.0769931
#> Dec 2021         NA         NA         NA         NA         NA  4.2447522