Efficiency Model 1

Linear regression model of the revisions (R) on a preliminary vintage (P)

Usage

efficiencyModel1(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
efficiencyModel1(vintages[["diagonal_view"]])
#>                            N          R2        F intercept.estimate
#> [Release[2]]-[Release[1]] 22 0.158447470 3.765599         2.23898122
#> [Release[3]]-[Release[2]] 19 0.260185498 5.978733         1.65223776
#> [Release[4]]-[Release[3]] 17 0.134074024 2.322497         0.77852549
#> [Release[5]]-[Release[4]] 11 0.001546383 0.013939         0.03095329
#>                           intercept.stderr intercept.pvalue slope.estimate
#> [Release[2]]-[Release[1]]        1.4975099        0.1504913   -0.086412552
#> [Release[3]]-[Release[2]]        0.8052680        0.0559229   -0.062475962
#> [Release[4]]-[Release[3]]        0.4874296        0.1310678   -0.024293219
#> [Release[5]]-[Release[4]]        0.3825488        0.9372817   -0.001848279
#>                           slope.stderr slope.pvalue   skewness   kurtosis
#> [Release[2]]-[Release[1]]   0.04453073   0.06654607  0.2394377 -0.2527084
#> [Release[3]]-[Release[2]]   0.02555103   0.02566541 -0.4587218  0.1127482
#> [Release[4]]-[Release[3]]   0.01594070   0.14831862  0.9674335  1.0542605
#> [Release[5]]-[Release[4]]   0.01565495   0.90861029  0.1143025 -0.5018355
#>                           JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> [Release[2]]-[Release[1]]        0.2565353         0.8796179    2.862038e-03
#> [Release[3]]-[Release[2]]        0.6408113         0.7258546    7.862083e-05
#> [Release[4]]-[Release[3]]        3.2367837         0.1982172    2.163821e-02
#> [Release[5]]-[Release[4]]        0.1267079         0.9386112    6.176018e-03
#>                           BreuschPagan.value BreuschPagan.pvalue   White.R2
#> [Release[2]]-[Release[1]]        0.057405052           0.8130838 0.02007552
#> [Release[3]]-[Release[2]]        0.001336659           0.9712614 0.01537706
#> [Release[4]]-[Release[3]]        0.331751635           0.5731705 0.04013462
#> [Release[5]]-[Release[4]]        0.055929581           0.8183438 0.29843582
#>                           White.value White.pvalue     arch.R2 arch.value
#> [Release[2]]-[Release[1]]   0.4416615    0.8018524 0.004667785 0.09802348
#> [Release[3]]-[Release[2]]   0.2921642    0.8640868 0.084654972 1.52378949
#> [Release[4]]-[Release[3]]   0.6822886    0.7109563 0.056080236 0.89728378
#> [Release[5]]-[Release[4]]   3.2827940    0.1937092 0.143568115 1.43568115
#>                           arch.pvalue
#> [Release[2]]-[Release[1]]   0.7542145
#> [Release[3]]-[Release[2]]   0.2170469
#> [Release[4]]-[Release[3]]   0.3435111
#> [Release[5]]-[Release[4]]   0.2308395