Efficiency Model 2

Linear regression model of R_v on R_{v-1}

Usage

efficiencyModel2(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
efficiencyModel2(vintages$diagonal_view)
#>                            N          R2           F intercept.estimate
#> [Release[3]]-[Release[2]] 24 0.344031034 11.53817200        0.109860001
#> [Release[4]]-[Release[3]] 14 0.001302023  0.01564465        0.000632632
#> [Release[5]]-[Release[4]]  9 0.151834592  1.25310716        0.033831766
#>                           intercept.stderr intercept.pvalue slope.estimate
#> [Release[3]]-[Release[2]]       0.20213049        0.5922439    -0.30115024
#> [Release[4]]-[Release[3]]       0.16074738        0.9969245     0.01919427
#> [Release[5]]-[Release[4]]       0.06573168        0.6226141    -0.11829608
#>                           slope.stderr slope.pvalue    skewness    kurtosis
#> [Release[3]]-[Release[2]]   0.08865734  0.002591094 -0.04615157 -0.36199959
#> [Release[4]]-[Release[3]]   0.15345772  0.902532324 -1.03395405  2.29739182
#> [Release[5]]-[Release[4]]   0.10567597  0.299894962  0.48140272 -0.01239274
#>                           JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> [Release[3]]-[Release[2]]        0.1337484        0.93531285      0.03935263
#> [Release[4]]-[Release[3]]        5.1752204        0.07519954      0.11395178
#> [Release[5]]-[Release[4]]        0.3090493        0.85682237      0.28960014
#>                           BreuschPagan.value BreuschPagan.pvalue   White.R2
#> [Release[3]]-[Release[2]]          0.9012233           0.3527668 0.03974011
#> [Release[4]]-[Release[3]]          1.5432809           0.2378601 0.17483605
#> [Release[5]]-[Release[4]]          2.8536055           0.1350193 0.33025472
#>                           White.value White.pvalue     arch.R2 arch.value
#> [Release[3]]-[Release[2]]   0.9537627    0.6207162 0.007372263  0.1695620
#> [Release[4]]-[Release[3]]   2.4477047    0.2940950 0.098329364  1.2782817
#> [Release[5]]-[Release[4]]   2.9722925    0.2262429 0.013632058  0.1090565
#>                           arch.pvalue
#> [Release[3]]-[Release[2]]   0.6805013
#> [Release[4]]-[Release[3]]   0.2582188
#> [Release[5]]-[Release[4]]   0.7412209