Signal VS Noise

Linear regression models to determine whether revisions are ‘news’ or ‘noise’. For 'noise': R (revisions) on P (preliminary estimate). For 'news': R on L (latter estimate).

Usage

signalnoise(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the `create_vintages()` output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
period_range <- seq(as.Date('2011-01-01'),as.Date('2020-10-01'),by='quarter')
qtr <- (as.numeric(substr(period_range,6,7))+2)/3
time_period <- rep(paste0(format(period_range, "%Y"), "Q", qtr),5)
np <- length(period_range)
rev_date <- c(rep("2021-06-30",np), rep("2021-12-31",np), rep("2022-06-30",np),
            rep("2022-12-31",np), rep("2023-06-30",np))
set.seed(1)
xt <- cumsum(sample(rnorm(1000,0,1), np, TRUE))
rev <- rnorm(np*4,0,.1)
obs_values <- xt
for(i in 1:4) {
  xt <- xt+rev[(1+(i-1)*np):(i*np)]
  obs_values <- c(obs_values,xt)
}
df <- data.frame(rev_date, time_period, obs_values)

## Create vintage and test
vintages <- create_vintages(df, periodicity = 4)
signalnoise(vintages$diagonal_view)
#>                               News.R2    News.F News.pvalue    Noise.R2
#> [Release[2]]-[Release[1]] 0.054010910 2.1604364   0.1292262 0.064342830
#> [Release[3]]-[Release[2]] 0.058006796 2.3202718   0.1120099 0.071263987
#> [Release[4]]-[Release[3]] 0.020451313 0.8180525   0.4489106 0.014328986
#> [Release[5]]-[Release[4]] 0.005068516 0.2027406   0.8173674 0.009136627
#>                             Noise.F Noise.pvalue
#> [Release[2]]-[Release[1]] 2.5737132   0.08948398
#> [Release[3]]-[Release[2]] 2.8505595   0.07023115
#> [Release[4]]-[Release[3]] 0.5731594   0.56853954
#> [Release[5]]-[Release[4]] 0.3654651   0.69628610