Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages[["diagonal_view"]])
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9640344  723.718          0.3924057        0.4250440
#> Release[3] 26 0.9899999 2375.982         -0.3992916        0.2441183
#> Release[4] 17 0.9888331 1328.253          0.7652153        0.3483762
#> Release[5] 14 0.9936468 1876.801         -0.2453058        0.3378674
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue
#> Release[2]       0.36407400      0.9829721   0.03653899   0.64493837
#> Release[3]       0.11496237      1.0073033   0.02066515   0.72686637
#> Release[4]       0.04418801      0.9434990   0.02588815   0.04538415
#> Release[5]       0.48173478      1.0159378   0.02345082   0.50963784
#>               skewness   kurtosis JarqueBera.value JarqueBera.pvalue
#> Release[2] -0.44989446  0.9998122        2.1107852         0.3480557
#> Release[3] -0.72616032 -0.3552829        2.3286054         0.3121402
#> Release[4]  0.02034036 -1.3438358        1.2050330         0.5474323
#> Release[5] -0.43189364 -0.1098506        0.4106893         0.8143666
#>            BreuschPagan.R2 BreuschPagan.value BreuschPagan.pvalue   White.R2
#> Release[2]     0.017672389         0.48573867           0.4917920 0.05235191
#> Release[3]     0.015663978         0.38191783           0.5423992 0.08451522
#> Release[4]     0.001314213         0.01973913           0.8901374 0.02692496
#> Release[5]     0.020526302         0.25147753           0.6251096 0.06374769
#>            White.value White.pvalue    arch.R2 arch.value arch.pvalue
#> Release[2]   1.5182054    0.4680863 0.05799203  1.6237767   0.2025660
#> Release[3]   2.1973958    0.3333048 0.03477148  0.8692870   0.3511528
#> Release[4]   0.4577244    0.7954382 0.01208739  0.1933983   0.6601032
#> Release[5]   0.8924677    0.6400341 0.03421708  0.4448220   0.5048042