Theil's Inequality Coefficient U1

Theil's Inequality Coefficient U1

Usage

theil(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
period_range <- seq(as.Date('2011-01-01'),as.Date('2020-10-01'),by='quarter')
qtr <- (as.numeric(substr(period_range,6,7))+2)/3
time_period <- rep(paste0(format(period_range, "%Y"), "Q", qtr),5)
np <- length(period_range)
rev_date <- c(rep("2021-06-30",np), rep("2021-12-31",np), rep("2022-06-30",np),
            rep("2022-12-31",np), rep("2023-06-30",np))
set.seed(1)
xt <- cumsum(sample(rnorm(1000,0,1), np, TRUE))
rev <- rnorm(np*4,0,.1)
obs_values <- xt
for(i in 1:4) {
  xt <- xt+rev[(1+(i-1)*np):(i*np)]
  obs_values <- c(obs_values,xt)
}
df <- data.frame(rev_date, time_period, obs_values)

## Create vintage and test
vintages <- create_vintages(df, periodicity = 4)
theil(vintages$diagonal_view)
#> [1] 0.009198273 0.011021424 0.009661489 0.009876197