Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   12.71810         NA         NA         NA         NA         NA
#> 2020-05-21    9.01126   8.505849  3.7186866   3.896725   3.948151         NA
#> 2021-04-10   10.66301   5.589661  0.1568806   5.131178   3.117730   4.216198
#> 2021-04-28   11.13696   7.193090 -0.1890166   3.657626   3.375544   3.985718
#> 2021-08-19   10.49150   6.589002 -0.8189697   2.893905   2.823452   2.937796
#> 2021-09-05   10.17340   6.539015 -0.4965269   3.053112   3.312315   3.239869
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10  2.3811212   8.433663   7.215127   1.082369  0.1120284 -0.9374421
#> 2021-04-28  3.4212761   6.054820   7.829253   4.737181  3.1573842  1.4566869
#> 2021-08-19  0.8551097   5.413561   7.149878   4.085611  2.0869891  1.1236689
#> 2021-09-05  1.4817415   5.283859   6.315872   3.553175  1.5032903  1.4991140
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   7.285937   8.730369 -3.8286417 -3.4880655         NA         NA
#> 2021-04-28   9.885812   9.744573 -1.0615448 -0.7528198         NA         NA
#> 2021-08-19   8.621628  10.143243  0.3023545 -1.3033004 -0.8868021  -3.508885
#> 2021-09-05   8.397417   8.926508  0.3598866 -2.1502034 -2.3035197  -3.156892
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19  -11.36901  -9.627540         NA         NA         NA         NA
#> 2021-09-05  -11.68628  -9.869348  -12.81855         NA         NA         NA