Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   9.410619         NA         NA         NA         NA         NA
#> 2020-05-21   5.703780   15.41285   21.18476   24.70244   25.61469         NA
#> 2021-04-10   7.355534   12.49666   17.62296   25.93690   24.78427   25.83047
#> 2021-04-28   7.829477   14.10009   17.27706   24.46334   25.04208   25.59999
#> 2021-08-19   7.184023   13.49601   16.64711   23.69962   24.48999   24.55207
#> 2021-09-05   6.865919   13.44602   16.96955   23.85883   24.97885   24.85415
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   34.64289   37.56037   35.03977   30.03895   34.23803   36.26111
#> 2021-04-28   35.68305   35.18153   35.65390   33.69376   37.28338   38.65524
#> 2021-08-19   33.11688   34.54027   34.97452   33.04219   36.21299   38.32222
#> 2021-09-05   33.74351   34.41057   34.14051   32.50975   35.62929   38.69766
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   28.02089   34.11010   30.55312   35.65559         NA         NA
#> 2021-04-28   30.62076   35.12431   33.32022   38.39084         NA         NA
#> 2021-08-19   29.35658   35.52298   34.68412   37.84036   40.65345   37.47974
#> 2021-09-05   29.13237   34.30624   34.74165   36.99346   39.23674   37.83173
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19   50.29250   54.29774         NA         NA         NA         NA
#> 2021-09-05   49.97523   54.05593   52.41335         NA         NA         NA