Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-14   1.228329         NA         NA         NA         NA         NA
#> 2020-03-19   3.168025   1.159140  0.8422779         NA         NA         NA
#> 2020-06-17   3.449851   3.848333  4.1335493   5.246063   2.935414   4.411188
#> 2020-11-07   2.499796   2.506128  4.4495568   4.971830   7.845300   5.549872
#> 2020-11-27   2.364098   2.461363  4.3312198   7.490447   7.601374   5.524189
#> 2021-11-19   2.526368   2.869088  4.2370482   6.064658   7.098654   4.427190
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-14         NA         NA         NA         NA         NA         NA
#> 2020-03-19         NA         NA         NA         NA         NA         NA
#> 2020-06-17         NA         NA         NA         NA         NA         NA
#> 2020-11-07  11.428804   17.84338   17.91107   17.30095   15.23346         NA
#> 2020-11-27   9.015822   16.25286   19.26956   20.09258   15.87181         NA
#> 2021-11-19   8.650138   15.95207   18.64125   18.41621   15.53770   7.659438
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-14         NA         NA         NA         NA         NA         NA
#> 2020-03-19         NA         NA         NA         NA         NA         NA
#> 2020-06-17         NA         NA         NA         NA         NA         NA
#> 2020-11-07         NA         NA         NA         NA         NA         NA
#> 2020-11-27         NA         NA         NA         NA         NA         NA
#> 2021-11-19   4.685434   9.269384   5.862266   8.191876   12.86778   11.77307
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-14         NA         NA         NA         NA         NA         NA
#> 2020-03-19         NA         NA         NA         NA         NA         NA
#> 2020-06-17         NA         NA         NA         NA         NA         NA
#> 2020-11-07         NA         NA         NA         NA         NA         NA
#> 2020-11-27         NA         NA         NA         NA         NA         NA
#> 2021-11-19   14.95749   16.65671   20.91565   15.93503   18.72645         NA