Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   17.19023         NA         NA         NA         NA         NA
#> 2020-05-21   13.48339   22.02819   28.25133   28.83196   28.29340         NA
#> 2021-04-10   15.13515   19.11200   24.68952   30.06641   27.46298   25.49485
#> 2021-04-28   15.60909   20.71543   24.34363   28.59286   27.72079   25.26437
#> 2021-08-19   14.96364   20.11134   23.71367   27.82914   27.16870   24.21645
#> 2021-09-05   14.64553   20.06136   24.03612   27.98834   27.65756   24.51852
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   27.91946   24.69196   21.26507   15.30894   13.83979   16.22827
#> 2021-04-28   28.95961   22.31312   21.87919   18.96375   16.88515   18.62240
#> 2021-08-19   26.39345   21.67186   21.19982   18.31218   15.81475   18.28938
#> 2021-09-05   27.02008   21.54215   20.36581   17.77974   15.23105   18.66482
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   22.83617   32.65598   33.40746   39.91000         NA         NA
#> 2021-04-28   25.43604   33.67018   36.17455   42.64525         NA         NA
#> 2021-08-19   24.17186   34.06885   37.53845   42.09477   38.12715   32.25271
#> 2021-09-05   23.94765   32.85212   37.59599   41.24786   36.71043   32.60470
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19   30.61148   26.66935         NA         NA         NA         NA
#> 2021-09-05   30.29422   26.42754   26.52756         NA         NA         NA