Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   7.204358         NA         NA         NA         NA         NA
#> 2020-05-21   3.497519   6.753144  10.857578   12.08743   16.83585         NA
#> 2021-04-10   5.149273   3.836956   7.295772   13.32188   16.00543   18.80101
#> 2021-04-28   5.623216   5.440386   6.949875   11.84833   16.26324   18.57053
#> 2021-08-19   4.977762   4.836297   6.319922   11.08461   15.71115   17.52261
#> 2021-09-05   4.659658   4.786311   6.642365   11.24381   16.20001   17.82468
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   22.17226   19.24543   10.10620   1.545001   2.515776 -0.6871231
#> 2021-04-28   23.21242   16.86658   10.72032   5.199813   5.561132  1.7070059
#> 2021-08-19   20.64625   16.22532   10.04095   4.548243   4.490737  1.3739880
#> 2021-09-05   21.27288   16.09562    9.20694   4.015807   3.907038  1.7494330
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10 -3.0959005  -2.903753 -2.2635668   4.196987         NA         NA
#> 2021-04-28 -0.4960251  -1.889549  0.5035301   6.932233         NA         NA
#> 2021-08-19 -1.7602091  -1.490879  1.8674294   6.381752  10.242147   14.34462
#> 2021-09-05 -1.9844203  -2.707615  1.9249615   5.534849   8.825429   14.69662
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19   13.00790   6.326846         NA         NA         NA         NA
#> 2021-09-05   12.69063   6.085038  -2.705249         NA         NA         NA