Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   6.894417         NA         NA         NA         NA         NA
#> 2020-05-21   3.187578  0.1418382  4.3879401   5.021847   7.386103         NA
#> 2021-04-10   4.839333 -2.7743499  0.8261341   6.256300   6.555682   3.253217
#> 2021-04-28   5.313275 -1.1709204  0.4802369   4.782748   6.813496   3.022737
#> 2021-08-19   4.667822 -1.7750089 -0.1497161   4.019028   6.261403   1.974815
#> 2021-09-05   4.349718 -1.8249952  0.1727267   4.178235   6.750267   2.276888
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10  -3.783351  -1.129453  -2.165798  -8.035776  -8.056904  -7.618795
#> 2021-04-28  -2.743196  -3.508297  -1.551673  -4.380965  -5.011548  -5.224666
#> 2021-08-19  -5.309363  -4.149556  -2.231048  -5.032534  -6.081944  -5.557684
#> 2021-09-05  -4.682731  -4.279258  -3.065053  -5.564970  -6.665642  -5.182239
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10  -13.53348  -16.97182  -15.13190  -6.793055         NA         NA
#> 2021-04-28  -10.93360  -15.95761  -12.36480  -4.057810         NA         NA
#> 2021-08-19  -12.19779  -15.55894  -11.00090  -4.608290 -0.5690297  -4.944698
#> 2021-09-05  -12.42200  -16.77568  -10.94337  -5.455193 -1.9857474  -4.592705
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19  -11.22665  -10.72161         NA         NA         NA         NA
#> 2021-09-05  -11.54392  -10.96342   -14.3967         NA         NA         NA