Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   9.409973         NA         NA         NA         NA         NA
#> 2020-05-21   5.703134   8.823859   14.99416   20.13711   25.60973         NA
#> 2021-04-10   7.354888   5.907671   11.43235   21.37156   24.77931   26.53311
#> 2021-04-28   7.828831   7.511100   11.08645   19.89801   25.03712   26.30263
#> 2021-08-19   7.183378   6.907012   10.45650   19.13429   24.48503   25.25471
#> 2021-09-05   6.865273   6.857025   10.77894   19.29350   24.97389   25.55679
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   24.49935   32.21730   33.05462   29.29812   29.91792   34.03770
#> 2021-04-28   25.53950   29.83845   33.66875   32.95293   32.96328   36.43183
#> 2021-08-19   22.97334   29.19720   32.98937   32.30136   31.89288   36.09881
#> 2021-09-05   23.59997   29.06749   32.15537   31.76893   31.30918   36.47426
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   33.64873   36.47515   37.12595   44.52493         NA         NA
#> 2021-04-28   36.24861   37.48935   39.89305   47.26018         NA         NA
#> 2021-08-19   34.98442   37.88802   41.25695   46.70970   59.49478   58.31480
#> 2021-09-05   34.76021   36.67129   41.31448   45.86279   58.07806   58.66679
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19   54.47223   59.10329         NA         NA         NA         NA
#> 2021-09-05   54.15497   58.86148   58.82443         NA         NA         NA