Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-14   5.678589         NA         NA         NA         NA         NA
#> 2020-03-19   7.618284   2.104870  -1.603694         NA         NA         NA
#> 2020-06-17   7.900110   4.794062   1.687577   1.188256  -6.231966  -5.591241
#> 2020-11-07   6.950055   3.451857   2.003585   0.914023  -1.322080  -4.452557
#> 2020-11-27   6.814358   3.407093   1.885248   3.432641  -1.566007  -4.478241
#> 2021-11-19   6.976627   3.814818   1.791076   2.006851  -2.068726  -5.575239
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-14         NA         NA         NA         NA         NA         NA
#> 2020-03-19         NA         NA         NA         NA         NA         NA
#> 2020-06-17         NA         NA         NA         NA         NA         NA
#> 2020-11-07 -0.8599041 1.95788819  0.3339097   1.696043  -2.086700         NA
#> 2020-11-27 -3.2728865 0.36737360  1.6924024   4.487675  -1.448360         NA
#> 2021-11-19 -3.6385700 0.06657616  1.0640928   2.811306  -1.782464  -7.731952
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-14         NA         NA         NA         NA         NA         NA
#> 2020-03-19         NA         NA         NA         NA         NA         NA
#> 2020-06-17         NA         NA         NA         NA         NA         NA
#> 2020-11-07         NA         NA         NA         NA         NA         NA
#> 2020-11-27         NA         NA         NA         NA         NA         NA
#> 2021-11-19  -11.79328  -6.483175  -8.754999  -11.30776  -15.84788  -23.16336
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-14         NA         NA         NA         NA         NA         NA
#> 2020-03-19         NA         NA         NA         NA         NA         NA
#> 2020-06-17         NA         NA         NA         NA         NA         NA
#> 2020-11-07         NA         NA         NA         NA         NA         NA
#> 2020-11-27         NA         NA         NA         NA         NA         NA
#> 2021-11-19  -20.69307  -17.30487  -15.04897  -15.08514  -22.58212         NA