Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   18.52483         NA         NA         NA         NA         NA
#> 2020-05-21   14.81799   18.81228   19.17803   14.13386   11.77313         NA
#> 2021-04-10   16.46974   15.89609   15.61622   15.36832   10.94271  10.806241
#> 2021-04-28   16.94368   17.49952   15.27032   13.89476   11.20053  10.575760
#> 2021-08-19   16.29823   16.89543   14.64037   13.13104   10.64843   9.527839
#> 2021-09-05   15.98013   16.84544   14.96281   13.29025   11.13730   9.829912
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   8.164964   9.801090   11.35726   14.81498   18.10557   12.16541
#> 2021-04-28   9.205119   7.422247   11.97139   18.46979   21.15093   14.55954
#> 2021-08-19   6.638952   6.780988   11.29201   17.81822   20.08053   14.22652
#> 2021-09-05   7.265584   6.651286   10.45801   17.28579   19.49683   14.60197
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   18.16082   17.06835   11.27961   11.00388         NA         NA
#> 2021-04-28   20.76070   18.08255   14.04671   13.73913         NA         NA
#> 2021-08-19   19.49651   18.48122   15.41060   13.18865   8.923356   3.683117
#> 2021-09-05   19.27230   17.26449   15.46814   12.34175   7.506638   4.035109
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19   2.943554   2.137215         NA         NA         NA         NA
#> 2021-09-05   2.626289   1.895407   5.367617         NA         NA         NA