Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   17.05047         NA         NA         NA         NA         NA
#> 2020-05-21   13.34363   23.13980   31.71240   33.91245   34.24156         NA
#> 2021-04-10   14.99538   20.22361   28.15059   35.14690   33.41114   31.43248
#> 2021-04-28   15.46932   21.82704   27.80470   33.67335   33.66895   31.20200
#> 2021-08-19   14.82387   21.22295   27.17474   32.90963   33.11686   30.15408
#> 2021-09-05   14.50577   21.17296   27.49719   33.06884   33.60572   30.45616
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   28.19703   25.20745   20.86279   13.75720   16.04172   7.566225
#> 2021-04-28   29.23718   22.82861   21.47691   17.41202   19.08708   9.960354
#> 2021-08-19   26.67101   22.18735   20.79754   16.76045   18.01668   9.627336
#> 2021-09-05   27.29765   22.05764   19.96353   16.22801   17.43299  10.002781
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   11.54808   18.47522   21.06900   26.84877         NA         NA
#> 2021-04-28   14.14795   19.48943   23.83609   29.58401         NA         NA
#> 2021-08-19   12.88377   19.88810   25.19999   29.03353   30.95087   26.29962
#> 2021-09-05   12.65956   18.67136   25.25753   28.18663   29.53415   26.65161
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19   22.52011   17.27620         NA         NA         NA         NA
#> 2021-09-05   22.20285   17.03439   15.20561         NA         NA         NA