Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01  2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-14   3.114074         NA          NA         NA         NA         NA
#> 2020-03-19   5.053770   2.118482 -3.52936303         NA         NA         NA
#> 2020-06-17   5.335596   4.807675 -0.23809162   1.363782  -3.407154   1.501551
#> 2020-11-07   4.385541   3.465470  0.07791594   1.089549   1.502732   2.640235
#> 2020-11-27   4.249843   3.420706 -0.04042111   3.608166   1.258805   2.614552
#> 2021-11-19   4.412113   3.828431 -0.13459265   2.182377   0.756086   1.517553
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-14         NA         NA         NA         NA         NA         NA
#> 2020-03-19         NA         NA         NA         NA         NA         NA
#> 2020-06-17         NA         NA         NA         NA         NA         NA
#> 2020-11-07  3.0876460  10.496955   3.781625   2.124699   2.790015         NA
#> 2020-11-27  0.6746636   8.906440   5.140118   4.916331   3.428355         NA
#> 2021-11-19  0.3089802   8.605643   4.511809   3.239961   3.094251  -8.924347
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-14         NA         NA         NA         NA         NA         NA
#> 2020-03-19         NA         NA         NA         NA         NA         NA
#> 2020-06-17         NA         NA         NA         NA         NA         NA
#> 2020-11-07         NA         NA         NA         NA         NA         NA
#> 2020-11-27         NA         NA         NA         NA         NA         NA
#> 2021-11-19  -21.93089  -17.76445  -24.99353  -21.88155  -18.37024   -16.5332
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-14         NA         NA         NA         NA         NA         NA
#> 2020-03-19         NA         NA         NA         NA         NA         NA
#> 2020-06-17         NA         NA         NA         NA         NA         NA
#> 2020-11-07         NA         NA         NA         NA         NA         NA
#> 2020-11-27         NA         NA         NA         NA         NA         NA
#> 2021-11-19  -19.45811  -13.86197  -7.308099  -11.61078  -16.72984         NA