Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   15.45913         NA         NA         NA         NA         NA
#> 2020-05-21   11.75229   16.83551   17.59213   11.96891  10.451192         NA
#> 2021-04-10   13.40405   13.91932   14.03032   13.20336   9.620771  10.203380
#> 2021-04-28   13.87799   15.52275   13.68442   11.72981   9.878585   9.972900
#> 2021-08-19   13.23254   14.91866   13.05447   10.96609   9.326493   8.924978
#> 2021-09-05   12.91443   14.86868   13.37691   11.12530   9.815356   9.227051
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   13.64633   19.46951   21.50892   19.79096   29.01067   26.41829
#> 2021-04-28   14.68649   17.09067   22.12305   23.44577   32.05603   28.81242
#> 2021-08-19   12.12032   16.44941   21.44368   22.79420   30.98563   28.47940
#> 2021-09-05   12.74695   16.31971   20.60967   22.26176   30.40193   28.85485
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   32.66309   32.44684   29.26364   30.74342         NA         NA
#> 2021-04-28   35.26297   33.46105   32.03073   33.47867         NA         NA
#> 2021-08-19   33.99878   33.85972   33.39463   32.92819   30.59584    29.1423
#> 2021-09-05   33.77457   32.64298   33.45217   32.08129   29.17912    29.4943
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19   32.49268   39.56595         NA         NA         NA         NA
#> 2021-09-05   32.17541   39.32414   45.37951         NA         NA         NA