Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   5.675727         NA         NA         NA         NA         NA
#> 2020-05-21   1.968888  0.9848330  3.1323501   3.505630   3.081891         NA
#> 2021-04-10   3.620642 -1.9313551 -0.4294559   4.740083   2.251471   4.772574
#> 2021-04-28   4.094585 -0.3279256 -0.7753531   3.266531   2.509284   4.542094
#> 2021-08-19   3.449132 -0.9320141 -1.4053062   2.502811   1.957192   3.494172
#> 2021-09-05   3.131027 -0.9820003 -1.0828633   2.662018   2.446055   3.796245
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   8.575993   13.55622   11.13294   5.598133 -1.2380377 -2.6375491
#> 2021-04-28   9.616148   11.17738   11.74707   9.252945  1.8073180 -0.2434201
#> 2021-08-19   7.049982   10.53612   11.06769   8.601375  0.7369229 -0.5764381
#> 2021-09-05   7.676614   10.40642   10.23368   8.068939  0.1532241 -0.2009930
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10  -5.892003  -14.55634  -19.98337  -15.78816         NA         NA
#> 2021-04-28  -3.292127  -13.54213  -17.21627  -13.05292         NA         NA
#> 2021-08-19  -4.556311  -13.14346  -15.85238  -13.60340  -19.55936  -26.42923
#> 2021-09-05  -4.780522  -14.36020  -15.79484  -14.45030  -20.97608  -26.07724
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19  -28.22546  -24.52587         NA         NA         NA         NA
#> 2021-09-05  -28.54273  -24.76768   -30.1845         NA         NA         NA