Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   8.644119         NA         NA         NA         NA         NA
#> 2020-05-21   4.937280   9.925405  11.882590   4.127022  -1.041611         NA
#> 2021-04-10   6.589035   7.009217   8.320784   5.361475  -1.872032 -0.3591832
#> 2021-04-28   7.062977   8.612646   7.974887   3.887923  -1.614218 -0.5896635
#> 2021-08-19   6.417524   8.008558   7.344934   3.124202  -2.166310 -1.6375850
#> 2021-09-05   6.099419   7.958572   7.667377   3.283409  -1.677447 -1.3355120
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01  2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA          NA         NA
#> 2020-05-21         NA         NA         NA         NA          NA         NA
#> 2021-04-10  0.5250960   4.710672   6.059496  -1.044787 -3.11504466  -4.352041
#> 2021-04-28  1.5652509   2.331828   6.673621   2.610025 -0.06968891  -1.957912
#> 2021-08-19 -1.0009155   1.690569   5.994246   1.958455 -1.14008405  -2.290930
#> 2021-09-05 -0.3742837   1.560867   5.160240   1.426019 -1.72378284  -1.915485
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10 -10.318020  -9.176063  -14.67805  -16.30381         NA         NA
#> 2021-04-28  -7.718145  -8.161859  -11.91096  -13.56856         NA         NA
#> 2021-08-19  -8.982329  -7.763189  -10.54706  -14.11904  -18.33230  -18.42585
#> 2021-09-05  -9.206540  -8.979925  -10.48953  -14.96594  -19.74902  -18.07386
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19  -15.55657  -12.98305         NA         NA         NA         NA
#> 2021-09-05  -15.87384  -13.22485  -11.93426         NA         NA         NA