Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29   8.537683         NA         NA         NA         NA         NA
#> 2020-05-21   4.830843   13.82324   19.85749   21.43131   24.45750         NA
#> 2021-04-10   6.482598   10.90705   16.29568   22.66576   23.62708   24.79078
#> 2021-04-28   6.956541   12.51048   15.94978   21.19221   23.88490   24.56030
#> 2021-08-19   6.311087   11.90640   15.31983   20.42849   23.33281   23.51237
#> 2021-09-05   5.992983   11.85641   15.64227   20.58770   23.82167   23.81445
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   23.38781   18.18816   15.04381   10.87166   14.44443   19.39686
#> 2021-04-28   24.42796   15.80931   15.65794   14.52647   17.48979   21.79099
#> 2021-08-19   21.86180   15.16805   14.97856   13.87490   16.41939   21.45797
#> 2021-09-05   22.48843   15.03835   14.14456   13.34246   15.83569   21.83341
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   20.71480   28.08575   25.91802   27.83210         NA         NA
#> 2021-04-28   23.31468   29.09996   28.68512   30.56735         NA         NA
#> 2021-08-19   22.05049   29.49863   30.04902   30.01687   30.75368   31.31420
#> 2021-09-05   21.82628   28.28189   30.10655   29.16996   29.33696   31.66619
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19   30.76956   28.28402         NA         NA         NA         NA
#> 2021-09-05   30.45229   28.04221   29.39814         NA         NA         NA