Check horizontal format

Check horizontal format

Usage

check_horizontal(x, ...)

# S3 method for class 'data.frame'
check_horizontal(x, ...)

# S3 method for class 'matrix'
check_horizontal(x, date_format = "%Y-%m-%d")

# Default S3 method
check_horizontal(x, ...)

Arguments

x

a formatted data.frame containing the input in the horizontal format

...

Arguments to be passed to check_horizontal according to the class of the object x

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
horizontal_format <- rjd3revisions:::from_long_to_horizontal(long_format)
check_horizontal(horizontal_format)
#>            2020-01-01 2020-02-01 2020-03-01 2020-04-01 2020-05-01 2020-06-01
#> 2020-01-29  13.025406         NA         NA         NA         NA         NA
#> 2020-05-21   9.318567   14.69338   15.85610   9.696389   8.233216         NA
#> 2021-04-10  10.970321   11.77719   12.29430  10.930842   7.402796   12.29928
#> 2021-04-28  11.444264   13.38062   11.94840   9.457289   7.660609   12.06880
#> 2021-08-19  10.798811   12.77653   11.31845   8.693569   7.108517   11.02088
#> 2021-09-05  10.480706   12.72654   11.64089   8.852776   7.597380   11.32295
#>            2020-07-01 2020-08-01 2020-09-01 2020-10-01 2020-11-01 2020-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   15.15822   19.24775   22.38855   19.78578   17.87088   20.84461
#> 2021-04-28   16.19838   16.86891   23.00267   23.44059   20.91624   23.23874
#> 2021-08-19   13.63221   16.22765   22.32330   22.78902   19.84584   22.90572
#> 2021-09-05   14.25884   16.09795   21.48929   22.25659   19.26214   23.28116
#>            2021-01-01 2021-02-01 2021-03-01 2021-04-01 2021-05-01 2021-06-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10   20.72742   22.47572   21.88352   22.52482         NA         NA
#> 2021-04-28   23.32730   23.48993   24.65061   25.26007         NA         NA
#> 2021-08-19   22.06311   23.88860   26.01451   24.70959   25.10640    33.5032
#> 2021-09-05   21.83890   22.67186   26.07205   23.86269   23.68969    33.8552
#>            2021-07-01 2021-08-01 2021-09-01 2021-10-01 2021-11-01 2021-12-01
#> 2020-01-29         NA         NA         NA         NA         NA         NA
#> 2020-05-21         NA         NA         NA         NA         NA         NA
#> 2021-04-10         NA         NA         NA         NA         NA         NA
#> 2021-04-28         NA         NA         NA         NA         NA         NA
#> 2021-08-19   35.62340   39.31682         NA         NA         NA         NA
#> 2021-09-05   35.30614   39.07501   42.55708         NA         NA         NA