Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01   6.2238136
#> 2   2020-02-17 2020-02-01   5.6169089
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01   5.5115349
#> 26  2020-06-05 2020-02-01   2.3440241
#> 27  2020-06-05 2020-03-01   1.2469576
#> 28  2020-06-05 2020-04-01   1.8116074
#> 29  2020-06-05 2020-05-01   5.1974121
#> 30  2020-06-05 2020-06-01   1.3702018
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01   5.5096052
#> 50  2021-04-17 2020-02-01   2.0841627
#> 51  2021-04-17 2020-03-01   1.9063093
#> 52  2021-04-17 2020-04-01  -0.7684392
#> 53  2021-04-17 2020-05-01   2.5112258
#> 54  2021-04-17 2020-06-01   1.5051984
#> 55  2021-04-17 2020-07-01  -4.8166220
#> 56  2021-04-17 2020-08-01  -1.8211575
#> 57  2021-04-17 2020-09-01  -0.7776238
#> 58  2021-04-17 2020-10-01  -1.4234916
#> 59  2021-04-17 2020-11-01  -3.8592875
#> 60  2021-04-17 2020-12-01 -10.7907365
#> 61  2021-04-17 2021-01-01 -20.1460694
#> 62  2021-04-17 2021-02-01 -27.1693235
#> 63  2021-04-17 2021-03-01 -33.8564101
#> 64  2021-04-17 2021-04-01 -38.6138540
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01   5.1272716
#> 74  2021-09-08 2020-02-01   2.7613394
#> 75  2021-09-08 2020-03-01   2.6755150
#> 76  2021-09-08 2020-04-01  -0.8890723
#> 77  2021-09-08 2020-05-01   1.9403644
#> 78  2021-09-08 2020-06-01   2.5006268
#> 79  2021-09-08 2020-07-01  -3.5707730
#> 80  2021-09-08 2020-08-01  -3.1465479
#> 81  2021-09-08 2020-09-01  -4.3592510
#> 82  2021-09-08 2020-10-01   2.3234180
#> 83  2021-09-08 2020-11-01  -4.4179701
#> 84  2021-09-08 2020-12-01 -11.6125696
#> 85  2021-09-08 2021-01-01 -18.6580925
#> 86  2021-09-08 2021-02-01 -28.0526880
#> 87  2021-09-08 2021-03-01 -34.9951381
#> 88  2021-09-08 2021-04-01 -38.7672852
#> 89  2021-09-08 2021-05-01 -31.4632590
#> 90  2021-09-08 2021-06-01 -32.3519825
#> 91  2021-09-08 2021-07-01 -38.1691379
#> 92  2021-09-08 2021-08-01 -42.9688152
#> 93  2021-09-08 2021-09-01 -43.2396049
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01   5.0745978
#> 98  2021-11-12 2020-02-01   2.8032361
#> 99  2021-11-12 2020-03-01   1.5561076
#> 100 2021-11-12 2020-04-01  -0.6002551
#> 101 2021-11-12 2020-05-01   2.6093003
#> 102 2021-11-12 2020-06-01   2.6548513
#> 103 2021-11-12 2020-07-01  -2.3096831
#> 104 2021-11-12 2020-08-01  -2.5660520
#> 105 2021-11-12 2020-09-01  -2.6888632
#> 106 2021-11-12 2020-10-01   0.9678170
#> 107 2021-11-12 2020-11-01  -4.5229505
#> 108 2021-11-12 2020-12-01  -9.6697106
#> 109 2021-11-12 2021-01-01 -19.3748483
#> 110 2021-11-12 2021-02-01 -28.6431238
#> 111 2021-11-12 2021-03-01 -34.0544584
#> 112 2021-11-12 2021-04-01 -37.4447502
#> 113 2021-11-12 2021-05-01 -30.1271047
#> 114 2021-11-12 2021-06-01 -31.2504405
#> 115 2021-11-12 2021-07-01 -36.5468385
#> 116 2021-11-12 2021-08-01 -41.2224789
#> 117 2021-11-12 2021-09-01 -39.8076147
#> 118 2021-11-12 2021-10-01 -37.2927228
#> 119 2021-11-12 2021-11-01 -38.4773229
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01   5.4298121
#> 122 2021-11-17 2020-02-01   2.6795064
#> 123 2021-11-17 2020-03-01   2.1357044
#> 124 2021-11-17 2020-04-01  -0.5404497
#> 125 2021-11-17 2020-05-01   2.4653176
#> 126 2021-11-17 2020-06-01   2.5389352
#> 127 2021-11-17 2020-07-01  -2.5473395
#> 128 2021-11-17 2020-08-01  -2.7672060
#> 129 2021-11-17 2020-09-01  -2.6256001
#> 130 2021-11-17 2020-10-01  -0.2677358
#> 131 2021-11-17 2020-11-01  -4.6237129
#> 132 2021-11-17 2020-12-01  -9.2777324
#> 133 2021-11-17 2021-01-01 -18.9274552
#> 134 2021-11-17 2021-02-01 -27.8310390
#> 135 2021-11-17 2021-03-01 -33.6872975
#> 136 2021-11-17 2021-04-01 -36.9570601
#> 137 2021-11-17 2021-05-01 -31.9402690
#> 138 2021-11-17 2021-06-01 -30.0592764
#> 139 2021-11-17 2021-07-01 -35.3641128
#> 140 2021-11-17 2021-08-01 -40.6112088
#> 141 2021-11-17 2021-09-01 -41.0371659
#> 142 2021-11-17 2021-10-01 -37.3646158
#> 143 2021-11-17 2021-11-01 -35.8986964
#> 144 2021-11-17 2021-12-01          NA