Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01  10.375949
#> 2   2020-02-17 2020-02-01  17.972254
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01   9.663671
#> 26  2020-06-05 2020-02-01  14.699369
#> 27  2020-06-05 2020-03-01  14.082121
#> 28  2020-06-05 2020-04-01  18.118544
#> 29  2020-06-05 2020-05-01  18.647447
#> 30  2020-06-05 2020-06-01  21.397988
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01   9.661741
#> 50  2021-04-17 2020-02-01  14.439508
#> 51  2021-04-17 2020-03-01  14.741473
#> 52  2021-04-17 2020-04-01  15.538498
#> 53  2021-04-17 2020-05-01  15.961261
#> 54  2021-04-17 2020-06-01  21.532985
#> 55  2021-04-17 2020-07-01  21.418355
#> 56  2021-04-17 2020-08-01  28.494171
#> 57  2021-04-17 2020-09-01  39.464741
#> 58  2021-04-17 2020-10-01  41.446740
#> 59  2021-04-17 2020-11-01  47.381363
#> 60  2021-04-17 2020-12-01  51.124022
#> 61  2021-04-17 2021-01-01  53.164436
#> 62  2021-04-17 2021-02-01  56.670506
#> 63  2021-04-17 2021-03-01  48.185474
#> 64  2021-04-17 2021-04-01  46.814239
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01   9.279407
#> 74  2021-09-08 2020-02-01  15.116684
#> 75  2021-09-08 2020-03-01  15.510679
#> 76  2021-09-08 2020-04-01  15.417865
#> 77  2021-09-08 2020-05-01  15.390399
#> 78  2021-09-08 2020-06-01  22.528413
#> 79  2021-09-08 2020-07-01  22.664204
#> 80  2021-09-08 2020-08-01  27.168781
#> 81  2021-09-08 2020-09-01  35.883113
#> 82  2021-09-08 2020-10-01  45.193650
#> 83  2021-09-08 2020-11-01  46.822680
#> 84  2021-09-08 2020-12-01  50.302189
#> 85  2021-09-08 2021-01-01  54.652413
#> 86  2021-09-08 2021-02-01  55.787142
#> 87  2021-09-08 2021-03-01  47.046746
#> 88  2021-09-08 2021-04-01  46.660808
#> 89  2021-09-08 2021-05-01  53.358909
#> 90  2021-09-08 2021-06-01  58.656953
#> 91  2021-09-08 2021-07-01  55.100089
#> 92  2021-09-08 2021-08-01  55.254267
#> 93  2021-09-08 2021-09-01  61.544510
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01   9.226734
#> 98  2021-11-12 2020-02-01  15.158581
#> 99  2021-11-12 2020-03-01  14.391271
#> 100 2021-11-12 2020-04-01  15.706682
#> 101 2021-11-12 2020-05-01  16.059335
#> 102 2021-11-12 2020-06-01  22.682638
#> 103 2021-11-12 2020-07-01  23.925294
#> 104 2021-11-12 2020-08-01  27.749277
#> 105 2021-11-12 2020-09-01  37.553501
#> 106 2021-11-12 2020-10-01  43.838049
#> 107 2021-11-12 2020-11-01  46.717700
#> 108 2021-11-12 2020-12-01  52.245048
#> 109 2021-11-12 2021-01-01  53.935657
#> 110 2021-11-12 2021-02-01  55.196706
#> 111 2021-11-12 2021-03-01  47.987425
#> 112 2021-11-12 2021-04-01  47.983343
#> 113 2021-11-12 2021-05-01  54.695063
#> 114 2021-11-12 2021-06-01  59.758495
#> 115 2021-11-12 2021-07-01  56.722388
#> 116 2021-11-12 2021-08-01  57.000603
#> 117 2021-11-12 2021-09-01  64.976501
#> 118 2021-11-12 2021-10-01  70.208218
#> 119 2021-11-12 2021-11-01  71.955216
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01   9.581948
#> 122 2021-11-17 2020-02-01  15.034852
#> 123 2021-11-17 2020-03-01  14.970868
#> 124 2021-11-17 2020-04-01  15.766487
#> 125 2021-11-17 2020-05-01  15.915352
#> 126 2021-11-17 2020-06-01  22.566722
#> 127 2021-11-17 2020-07-01  23.687638
#> 128 2021-11-17 2020-08-01  27.548123
#> 129 2021-11-17 2020-09-01  37.616764
#> 130 2021-11-17 2020-10-01  42.602496
#> 131 2021-11-17 2020-11-01  46.616937
#> 132 2021-11-17 2020-12-01  52.637026
#> 133 2021-11-17 2021-01-01  54.383050
#> 134 2021-11-17 2021-02-01  56.008791
#> 135 2021-11-17 2021-03-01  48.354586
#> 136 2021-11-17 2021-04-01  48.471033
#> 137 2021-11-17 2021-05-01  52.881899
#> 138 2021-11-17 2021-06-01  60.949659
#> 139 2021-11-17 2021-07-01  57.905114
#> 140 2021-11-17 2021-08-01  57.611874
#> 141 2021-11-17 2021-09-01  63.746949
#> 142 2021-11-17 2021-10-01  70.136325
#> 143 2021-11-17 2021-11-01  74.533843
#> 144 2021-11-17 2021-12-01         NA