Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01  14.5251304
#> 2   2020-02-17 2020-02-01  13.5475466
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01  13.8128517
#> 26  2020-06-05 2020-02-01  10.2746617
#> 27  2020-06-05 2020-03-01   4.8461764
#> 28  2020-06-05 2020-04-01   7.4402322
#> 29  2020-06-05 2020-05-01   9.4906647
#> 30  2020-06-05 2020-06-01   8.7685011
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01  13.8109220
#> 50  2021-04-17 2020-02-01  10.0148004
#> 51  2021-04-17 2020-03-01   5.5055282
#> 52  2021-04-17 2020-04-01   4.8601857
#> 53  2021-04-17 2020-05-01   6.8044784
#> 54  2021-04-17 2020-06-01   8.9034978
#> 55  2021-04-17 2020-07-01   3.6254892
#> 56  2021-04-17 2020-08-01  11.8119036
#> 57  2021-04-17 2020-09-01  10.6460585
#> 58  2021-04-17 2020-10-01   7.5854478
#> 59  2021-04-17 2020-11-01   6.5040711
#> 60  2021-04-17 2020-12-01   2.2200491
#> 61  2021-04-17 2021-01-01   5.6970556
#> 62  2021-04-17 2021-02-01   2.0633927
#> 63  2021-04-17 2021-03-01  -6.7193730
#> 64  2021-04-17 2021-04-01 -17.4672515
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01  13.4285884
#> 74  2021-09-08 2020-02-01  10.6919770
#> 75  2021-09-08 2020-03-01   6.2747339
#> 76  2021-09-08 2020-04-01   4.7395525
#> 77  2021-09-08 2020-05-01   6.2336170
#> 78  2021-09-08 2020-06-01   9.8989262
#> 79  2021-09-08 2020-07-01   4.8713382
#> 80  2021-09-08 2020-08-01  10.4865131
#> 81  2021-09-08 2020-09-01   7.0644313
#> 82  2021-09-08 2020-10-01  11.3323574
#> 83  2021-09-08 2020-11-01   5.9453885
#> 84  2021-09-08 2020-12-01   1.3982160
#> 85  2021-09-08 2021-01-01   7.1850326
#> 86  2021-09-08 2021-02-01   1.1800282
#> 87  2021-09-08 2021-03-01  -7.8581010
#> 88  2021-09-08 2021-04-01 -17.6206827
#> 89  2021-09-08 2021-05-01 -18.5806520
#> 90  2021-09-08 2021-06-01 -18.9493911
#> 91  2021-09-08 2021-07-01 -23.1814823
#> 92  2021-09-08 2021-08-01 -17.1078346
#> 93  2021-09-08 2021-09-01 -20.7624485
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01  13.3759146
#> 98  2021-11-12 2020-02-01  10.7338737
#> 99  2021-11-12 2020-03-01   5.1553264
#> 100 2021-11-12 2020-04-01   5.0283697
#> 101 2021-11-12 2020-05-01   6.9025529
#> 102 2021-11-12 2020-06-01  10.0531506
#> 103 2021-11-12 2020-07-01   6.1324281
#> 104 2021-11-12 2020-08-01  11.0670090
#> 105 2021-11-12 2020-09-01   8.7348191
#> 106 2021-11-12 2020-10-01   9.9767563
#> 107 2021-11-12 2020-11-01   5.8404081
#> 108 2021-11-12 2020-12-01   3.3410750
#> 109 2021-11-12 2021-01-01   6.4682768
#> 110 2021-11-12 2021-02-01   0.5895924
#> 111 2021-11-12 2021-03-01  -6.9174213
#> 112 2021-11-12 2021-04-01 -16.2981477
#> 113 2021-11-12 2021-05-01 -17.2444976
#> 114 2021-11-12 2021-06-01 -17.8478492
#> 115 2021-11-12 2021-07-01 -21.5591830
#> 116 2021-11-12 2021-08-01 -15.3614983
#> 117 2021-11-12 2021-09-01 -17.3304583
#> 118 2021-11-12 2021-10-01 -22.9985265
#> 119 2021-11-12 2021-11-01 -28.2446429
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01  13.7311289
#> 122 2021-11-17 2020-02-01  10.6101441
#> 123 2021-11-17 2020-03-01   5.7349233
#> 124 2021-11-17 2020-04-01   5.0881751
#> 125 2021-11-17 2020-05-01   6.7585702
#> 126 2021-11-17 2020-06-01   9.9372346
#> 127 2021-11-17 2020-07-01   5.8947716
#> 128 2021-11-17 2020-08-01  10.8658550
#> 129 2021-11-17 2020-09-01   8.7980822
#> 130 2021-11-17 2020-10-01   8.7412036
#> 131 2021-11-17 2020-11-01   5.7396457
#> 132 2021-11-17 2020-12-01   3.7330532
#> 133 2021-11-17 2021-01-01   6.9156699
#> 134 2021-11-17 2021-02-01   1.4016772
#> 135 2021-11-17 2021-03-01  -6.5502604
#> 136 2021-11-17 2021-04-01 -15.8104576
#> 137 2021-11-17 2021-05-01 -19.0576619
#> 138 2021-11-17 2021-06-01 -16.6566850
#> 139 2021-11-17 2021-07-01 -20.3764572
#> 140 2021-11-17 2021-08-01 -14.7502282
#> 141 2021-11-17 2021-09-01 -18.5600095
#> 142 2021-11-17 2021-10-01 -23.0704195
#> 143 2021-11-17 2021-11-01 -25.6660164
#> 144 2021-11-17 2021-12-01          NA