Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time    obs_values
#> 1   2020-02-17 2020-01-01   7.679233475
#> 2   2020-02-17 2020-02-01   2.559868107
#> 3   2020-02-17 2020-03-01            NA
#> 4   2020-02-17 2020-04-01            NA
#> 5   2020-02-17 2020-05-01            NA
#> 6   2020-02-17 2020-06-01            NA
#> 7   2020-02-17 2020-07-01            NA
#> 8   2020-02-17 2020-08-01            NA
#> 9   2020-02-17 2020-09-01            NA
#> 10  2020-02-17 2020-10-01            NA
#> 11  2020-02-17 2020-11-01            NA
#> 12  2020-02-17 2020-12-01            NA
#> 13  2020-02-17 2021-01-01            NA
#> 14  2020-02-17 2021-02-01            NA
#> 15  2020-02-17 2021-03-01            NA
#> 16  2020-02-17 2021-04-01            NA
#> 17  2020-02-17 2021-05-01            NA
#> 18  2020-02-17 2021-06-01            NA
#> 19  2020-02-17 2021-07-01            NA
#> 20  2020-02-17 2021-08-01            NA
#> 21  2020-02-17 2021-09-01            NA
#> 22  2020-02-17 2021-10-01            NA
#> 23  2020-02-17 2021-11-01            NA
#> 24  2020-02-17 2021-12-01            NA
#> 25  2020-06-05 2020-01-01   6.966954834
#> 26  2020-06-05 2020-02-01  -0.713016760
#> 27  2020-06-05 2020-03-01  -5.234416334
#> 28  2020-06-05 2020-04-01   2.133566861
#> 29  2020-06-05 2020-05-01   1.604798575
#> 30  2020-06-05 2020-06-01  -4.388836679
#> 31  2020-06-05 2020-07-01            NA
#> 32  2020-06-05 2020-08-01            NA
#> 33  2020-06-05 2020-09-01            NA
#> 34  2020-06-05 2020-10-01            NA
#> 35  2020-06-05 2020-11-01            NA
#> 36  2020-06-05 2020-12-01            NA
#> 37  2020-06-05 2021-01-01            NA
#> 38  2020-06-05 2021-02-01            NA
#> 39  2020-06-05 2021-03-01            NA
#> 40  2020-06-05 2021-04-01            NA
#> 41  2020-06-05 2021-05-01            NA
#> 42  2020-06-05 2021-06-01            NA
#> 43  2020-06-05 2021-07-01            NA
#> 44  2020-06-05 2021-08-01            NA
#> 45  2020-06-05 2021-09-01            NA
#> 46  2020-06-05 2021-10-01            NA
#> 47  2020-06-05 2021-11-01            NA
#> 48  2020-06-05 2021-12-01            NA
#> 49  2021-04-17 2020-01-01   6.965025093
#> 50  2021-04-17 2020-02-01  -0.972878108
#> 51  2021-04-17 2020-03-01  -4.575064595
#> 52  2021-04-17 2020-04-01  -0.446479661
#> 53  2021-04-17 2020-05-01  -1.081387692
#> 54  2021-04-17 2020-06-01  -4.253840017
#> 55  2021-04-17 2020-07-01  -9.606064296
#> 56  2021-04-17 2020-08-01  -5.554600770
#> 57  2021-04-17 2020-09-01  -3.763892736
#> 58  2021-04-17 2020-10-01 -10.402033504
#> 59  2021-04-17 2020-11-01  -9.683254627
#> 60  2021-04-17 2020-12-01  -4.745276544
#> 61  2021-04-17 2021-01-01   0.941295605
#> 62  2021-04-17 2021-02-01   0.528646048
#> 63  2021-04-17 2021-03-01  -3.225190698
#> 64  2021-04-17 2021-04-01  -1.313161271
#> 65  2021-04-17 2021-05-01            NA
#> 66  2021-04-17 2021-06-01            NA
#> 67  2021-04-17 2021-07-01            NA
#> 68  2021-04-17 2021-08-01            NA
#> 69  2021-04-17 2021-09-01            NA
#> 70  2021-04-17 2021-10-01            NA
#> 71  2021-04-17 2021-11-01            NA
#> 72  2021-04-17 2021-12-01            NA
#> 73  2021-09-08 2020-01-01   6.582691551
#> 74  2021-09-08 2020-02-01  -0.295701445
#> 75  2021-09-08 2020-03-01  -3.805858865
#> 76  2021-09-08 2020-04-01  -0.567112818
#> 77  2021-09-08 2020-05-01  -1.652249095
#> 78  2021-09-08 2020-06-01  -3.258411609
#> 79  2021-09-08 2020-07-01  -8.360215308
#> 80  2021-09-08 2020-08-01  -6.879991235
#> 81  2021-09-08 2020-09-01  -7.345519934
#> 82  2021-09-08 2020-10-01  -6.655123891
#> 83  2021-09-08 2020-11-01 -10.241937222
#> 84  2021-09-08 2020-12-01  -5.567109645
#> 85  2021-09-08 2021-01-01   2.429272524
#> 86  2021-09-08 2021-02-01  -0.354718442
#> 87  2021-09-08 2021-03-01  -4.363918738
#> 88  2021-09-08 2021-04-01  -1.466592554
#> 89  2021-09-08 2021-05-01  -1.330531465
#> 90  2021-09-08 2021-06-01  -8.714537633
#> 91  2021-09-08 2021-07-01  -7.616013742
#> 92  2021-09-08 2021-08-01  -0.113086262
#> 93  2021-09-08 2021-09-01   0.525826129
#> 94  2021-09-08 2021-10-01            NA
#> 95  2021-09-08 2021-11-01            NA
#> 96  2021-09-08 2021-12-01            NA
#> 97  2021-11-12 2020-01-01   6.530017682
#> 98  2021-11-12 2020-02-01  -0.253804766
#> 99  2021-11-12 2020-03-01  -4.925266312
#> 100 2021-11-12 2020-04-01  -0.278295636
#> 101 2021-11-12 2020-05-01  -0.983313200
#> 102 2021-11-12 2020-06-01  -3.104187127
#> 103 2021-11-12 2020-07-01  -7.099125419
#> 104 2021-11-12 2020-08-01  -6.299495345
#> 105 2021-11-12 2020-09-01  -5.675132116
#> 106 2021-11-12 2020-10-01  -8.010724962
#> 107 2021-11-12 2020-11-01 -10.346917602
#> 108 2021-11-12 2020-12-01  -3.624250587
#> 109 2021-11-12 2021-01-01   1.712516726
#> 110 2021-11-12 2021-02-01  -0.945154205
#> 111 2021-11-12 2021-03-01  -3.423239021
#> 112 2021-11-12 2021-04-01  -0.144057476
#> 113 2021-11-12 2021-05-01   0.005622897
#> 114 2021-11-12 2021-06-01  -7.612995697
#> 115 2021-11-12 2021-07-01  -5.993714366
#> 116 2021-11-12 2021-08-01   1.633249990
#> 117 2021-11-12 2021-09-01   3.957816311
#> 118 2021-11-12 2021-10-01   3.293710083
#> 119 2021-11-12 2021-11-01   2.489122135
#> 120 2021-11-12 2021-12-01            NA
#> 121 2021-11-17 2020-01-01   6.885232004
#> 122 2021-11-17 2020-02-01  -0.377534405
#> 123 2021-11-17 2020-03-01  -4.345669447
#> 124 2021-11-17 2020-04-01  -0.218490246
#> 125 2021-11-17 2020-05-01  -1.127295904
#> 126 2021-11-17 2020-06-01  -3.220103207
#> 127 2021-11-17 2020-07-01  -7.336781873
#> 128 2021-11-17 2020-08-01  -6.500649318
#> 129 2021-11-17 2020-09-01  -5.611868980
#> 130 2021-11-17 2020-10-01  -9.246277731
#> 131 2021-11-17 2020-11-01 -10.447680014
#> 132 2021-11-17 2020-12-01  -3.232272413
#> 133 2021-11-17 2021-01-01   2.159909869
#> 134 2021-11-17 2021-02-01  -0.133069394
#> 135 2021-11-17 2021-03-01  -3.056078173
#> 136 2021-11-17 2021-04-01   0.343632626
#> 137 2021-11-17 2021-05-01  -1.807541418
#> 138 2021-11-17 2021-06-01  -6.421831540
#> 139 2021-11-17 2021-07-01  -4.810988612
#> 140 2021-11-17 2021-08-01   2.244520078
#> 141 2021-11-17 2021-09-01   2.728265127
#> 142 2021-11-17 2021-10-01   3.221817086
#> 143 2021-11-17 2021-11-01   5.067748632
#> 144 2021-11-17 2021-12-01            NA