Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-05-31 2020-01-01  8.25231916
#> 2   2020-05-31 2020-02-01  9.06110139
#> 3   2020-05-31 2020-03-01  6.74560329
#> 4   2020-05-31 2020-04-01  7.67382698
#> 5   2020-05-31 2020-05-01  3.23472649
#> 6   2020-05-31 2020-06-01          NA
#> 7   2020-05-31 2020-07-01          NA
#> 8   2020-05-31 2020-08-01          NA
#> 9   2020-05-31 2020-09-01          NA
#> 10  2020-05-31 2020-10-01          NA
#> 11  2020-05-31 2020-11-01          NA
#> 12  2020-05-31 2020-12-01          NA
#> 13  2020-05-31 2021-01-01          NA
#> 14  2020-05-31 2021-02-01          NA
#> 15  2020-05-31 2021-03-01          NA
#> 16  2020-05-31 2021-04-01          NA
#> 17  2020-05-31 2021-05-01          NA
#> 18  2020-05-31 2021-06-01          NA
#> 19  2020-05-31 2021-07-01          NA
#> 20  2020-05-31 2021-08-01          NA
#> 21  2020-05-31 2021-09-01          NA
#> 22  2020-05-31 2021-10-01          NA
#> 23  2020-05-31 2021-11-01          NA
#> 24  2020-05-31 2021-12-01          NA
#> 25  2020-07-30 2020-01-01 11.75490638
#> 26  2020-07-30 2020-02-01  4.84623544
#> 27  2020-07-30 2020-03-01  0.90610243
#> 28  2020-07-30 2020-04-01  6.16179074
#> 29  2020-07-30 2020-05-01  3.04070126
#> 30  2020-07-30 2020-06-01 -1.71085134
#> 31  2020-07-30 2020-07-01  1.55047397
#> 32  2020-07-30 2020-08-01          NA
#> 33  2020-07-30 2020-09-01          NA
#> 34  2020-07-30 2020-10-01          NA
#> 35  2020-07-30 2020-11-01          NA
#> 36  2020-07-30 2020-12-01          NA
#> 37  2020-07-30 2021-01-01          NA
#> 38  2020-07-30 2021-02-01          NA
#> 39  2020-07-30 2021-03-01          NA
#> 40  2020-07-30 2021-04-01          NA
#> 41  2020-07-30 2021-05-01          NA
#> 42  2020-07-30 2021-06-01          NA
#> 43  2020-07-30 2021-07-01          NA
#> 44  2020-07-30 2021-08-01          NA
#> 45  2020-07-30 2021-09-01          NA
#> 46  2020-07-30 2021-10-01          NA
#> 47  2020-07-30 2021-11-01          NA
#> 48  2020-07-30 2021-12-01          NA
#> 49  2020-11-20 2020-01-01 11.53939580
#> 50  2020-11-20 2020-02-01  6.84466258
#> 51  2020-11-20 2020-03-01  5.41604107
#> 52  2020-11-20 2020-04-01  6.96731759
#> 53  2020-11-20 2020-05-01  3.64424142
#> 54  2020-11-20 2020-06-01  2.57868184
#> 55  2020-11-20 2020-07-01 -1.61284540
#> 56  2020-11-20 2020-08-01 -1.30882093
#> 57  2020-11-20 2020-09-01  3.70148593
#> 58  2020-11-20 2020-10-01  6.95293962
#> 59  2020-11-20 2020-11-01 -0.09860754
#> 60  2020-11-20 2020-12-01          NA
#> 61  2020-11-20 2021-01-01          NA
#> 62  2020-11-20 2021-02-01          NA
#> 63  2020-11-20 2021-03-01          NA
#> 64  2020-11-20 2021-04-01          NA
#> 65  2020-11-20 2021-05-01          NA
#> 66  2020-11-20 2021-06-01          NA
#> 67  2020-11-20 2021-07-01          NA
#> 68  2020-11-20 2021-08-01          NA
#> 69  2020-11-20 2021-09-01          NA
#> 70  2020-11-20 2021-10-01          NA
#> 71  2020-11-20 2021-11-01          NA
#> 72  2020-11-20 2021-12-01          NA
#> 73  2020-12-12 2020-01-01 11.93795679
#> 74  2020-12-12 2020-02-01  6.53877727
#> 75  2020-12-12 2020-03-01  4.91086936
#> 76  2020-12-12 2020-04-01  6.01972557
#> 77  2020-12-12 2020-05-01  4.07144767
#> 78  2020-12-12 2020-06-01  0.40606218
#> 79  2020-12-12 2020-07-01 -1.94262327
#> 80  2020-12-12 2020-08-01 -3.10684023
#> 81  2020-12-12 2020-09-01  4.33814127
#> 82  2020-12-12 2020-10-01  7.82815615
#> 83  2020-12-12 2020-11-01  1.67460122
#> 84  2020-12-12 2020-12-01  4.89786537
#> 85  2020-12-12 2021-01-01          NA
#> 86  2020-12-12 2021-02-01          NA
#> 87  2020-12-12 2021-03-01          NA
#> 88  2020-12-12 2021-04-01          NA
#> 89  2020-12-12 2021-05-01          NA
#> 90  2020-12-12 2021-06-01          NA
#> 91  2020-12-12 2021-07-01          NA
#> 92  2020-12-12 2021-08-01          NA
#> 93  2020-12-12 2021-09-01          NA
#> 94  2020-12-12 2021-10-01          NA
#> 95  2020-12-12 2021-11-01          NA
#> 96  2020-12-12 2021-12-01          NA
#> 97  2021-04-26 2020-01-01 11.36715510
#> 98  2021-04-26 2020-02-01  6.67210702
#> 99  2021-04-26 2020-03-01  4.28463746
#> 100 2021-04-26 2020-04-01  6.54609671
#> 101 2021-04-26 2020-05-01  3.77917182
#> 102 2021-04-26 2020-06-01  0.98759033
#> 103 2021-04-26 2020-07-01 -1.15385522
#> 104 2021-04-26 2020-08-01 -3.67751142
#> 105 2021-04-26 2020-09-01  3.68292359
#> 106 2021-04-26 2020-10-01  7.09525933
#> 107 2021-04-26 2020-11-01  0.84015542
#> 108 2021-04-26 2020-12-01  6.19865370
#> 109 2021-04-26 2021-01-01  9.11713549
#> 110 2021-04-26 2021-02-01  3.21053707
#> 111 2021-04-26 2021-03-01 -1.18919474
#> 112 2021-04-26 2021-04-01 -0.39832461
#> 113 2021-04-26 2021-05-01          NA
#> 114 2021-04-26 2021-06-01          NA
#> 115 2021-04-26 2021-07-01          NA
#> 116 2021-04-26 2021-08-01          NA
#> 117 2021-04-26 2021-09-01          NA
#> 118 2021-04-26 2021-10-01          NA
#> 119 2021-04-26 2021-11-01          NA
#> 120 2021-04-26 2021-12-01          NA
#> 121 2021-05-04 2020-01-01 11.60892190
#> 122 2021-05-04 2020-02-01  6.66188879
#> 123 2021-05-04 2020-03-01  4.56752839
#> 124 2021-05-04 2020-04-01  6.41043029
#> 125 2021-05-04 2020-05-01  3.81818197
#> 126 2021-05-04 2020-06-01  1.00274376
#> 127 2021-05-04 2020-07-01 -1.51498313
#> 128 2021-05-04 2020-08-01 -4.02933236
#> 129 2021-05-04 2020-09-01  4.00980496
#> 130 2021-05-04 2020-10-01  8.08476992
#> 131 2021-05-04 2020-11-01  1.13307221
#> 132 2021-05-04 2020-12-01  6.27456491
#> 133 2021-05-04 2021-01-01 11.38079099
#> 134 2021-05-04 2021-02-01  5.23323090
#> 135 2021-05-04 2021-03-01 -1.67313376
#> 136 2021-05-04 2021-04-01  0.52892451
#> 137 2021-05-04 2021-05-01  1.27791467
#> 138 2021-05-04 2021-06-01          NA
#> 139 2021-05-04 2021-07-01          NA
#> 140 2021-05-04 2021-08-01          NA
#> 141 2021-05-04 2021-09-01          NA
#> 142 2021-05-04 2021-10-01          NA
#> 143 2021-05-04 2021-11-01          NA
#> 144 2021-05-04 2021-12-01          NA