Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time   obs_values
#> 1   2020-02-17 2020-01-01   7.96357543
#> 2   2020-02-17 2020-02-01  13.49459010
#> 3   2020-02-17 2020-03-01           NA
#> 4   2020-02-17 2020-04-01           NA
#> 5   2020-02-17 2020-05-01           NA
#> 6   2020-02-17 2020-06-01           NA
#> 7   2020-02-17 2020-07-01           NA
#> 8   2020-02-17 2020-08-01           NA
#> 9   2020-02-17 2020-09-01           NA
#> 10  2020-02-17 2020-10-01           NA
#> 11  2020-02-17 2020-11-01           NA
#> 12  2020-02-17 2020-12-01           NA
#> 13  2020-02-17 2021-01-01           NA
#> 14  2020-02-17 2021-02-01           NA
#> 15  2020-02-17 2021-03-01           NA
#> 16  2020-02-17 2021-04-01           NA
#> 17  2020-02-17 2021-05-01           NA
#> 18  2020-02-17 2021-06-01           NA
#> 19  2020-02-17 2021-07-01           NA
#> 20  2020-02-17 2021-08-01           NA
#> 21  2020-02-17 2021-09-01           NA
#> 22  2020-02-17 2021-10-01           NA
#> 23  2020-02-17 2021-11-01           NA
#> 24  2020-02-17 2021-12-01           NA
#> 25  2020-06-05 2020-01-01   7.25129679
#> 26  2020-06-05 2020-02-01  10.22170524
#> 27  2020-06-05 2020-03-01  13.71298943
#> 28  2020-06-05 2020-04-01   8.84300140
#> 29  2020-06-05 2020-05-01   2.72075497
#> 30  2020-06-05 2020-06-01  -5.16153452
#> 31  2020-06-05 2020-07-01           NA
#> 32  2020-06-05 2020-08-01           NA
#> 33  2020-06-05 2020-09-01           NA
#> 34  2020-06-05 2020-10-01           NA
#> 35  2020-06-05 2020-11-01           NA
#> 36  2020-06-05 2020-12-01           NA
#> 37  2020-06-05 2021-01-01           NA
#> 38  2020-06-05 2021-02-01           NA
#> 39  2020-06-05 2021-03-01           NA
#> 40  2020-06-05 2021-04-01           NA
#> 41  2020-06-05 2021-05-01           NA
#> 42  2020-06-05 2021-06-01           NA
#> 43  2020-06-05 2021-07-01           NA
#> 44  2020-06-05 2021-08-01           NA
#> 45  2020-06-05 2021-09-01           NA
#> 46  2020-06-05 2021-10-01           NA
#> 47  2020-06-05 2021-11-01           NA
#> 48  2020-06-05 2021-12-01           NA
#> 49  2021-04-17 2020-01-01   7.24936705
#> 50  2021-04-17 2020-02-01   9.96184389
#> 51  2021-04-17 2020-03-01  14.37234117
#> 52  2021-04-17 2020-04-01   6.26295488
#> 53  2021-04-17 2020-05-01   0.03456870
#> 54  2021-04-17 2020-06-01  -5.02653786
#> 55  2021-04-17 2020-07-01 -11.95511942
#> 56  2021-04-17 2020-08-01  -6.61106620
#> 57  2021-04-17 2020-09-01  -6.34971180
#> 58  2021-04-17 2020-10-01   0.76312858
#> 59  2021-04-17 2020-11-01   6.18277254
#> 60  2021-04-17 2020-12-01   2.52437407
#> 61  2021-04-17 2021-01-01  -0.75890491
#> 62  2021-04-17 2021-02-01   7.80725803
#> 63  2021-04-17 2021-03-01   8.73821526
#> 64  2021-04-17 2021-04-01   4.96417252
#> 65  2021-04-17 2021-05-01           NA
#> 66  2021-04-17 2021-06-01           NA
#> 67  2021-04-17 2021-07-01           NA
#> 68  2021-04-17 2021-08-01           NA
#> 69  2021-04-17 2021-09-01           NA
#> 70  2021-04-17 2021-10-01           NA
#> 71  2021-04-17 2021-11-01           NA
#> 72  2021-04-17 2021-12-01           NA
#> 73  2021-09-08 2020-01-01   6.86703350
#> 74  2021-09-08 2020-02-01  10.63902055
#> 75  2021-09-08 2020-03-01  15.14154690
#> 76  2021-09-08 2020-04-01   6.14232172
#> 77  2021-09-08 2020-05-01  -0.53629270
#> 78  2021-09-08 2020-06-01  -4.03110945
#> 79  2021-09-08 2020-07-01 -10.70927044
#> 80  2021-09-08 2020-08-01  -7.93645667
#> 81  2021-09-08 2020-09-01  -9.93133900
#> 82  2021-09-08 2020-10-01   4.51003819
#> 83  2021-09-08 2020-11-01   5.62408994
#> 84  2021-09-08 2020-12-01   1.70254097
#> 85  2021-09-08 2021-01-01   0.72907201
#> 86  2021-09-08 2021-02-01   6.92389354
#> 87  2021-09-08 2021-03-01   7.59948722
#> 88  2021-09-08 2021-04-01   4.81074124
#> 89  2021-09-08 2021-05-01   1.98666557
#> 90  2021-09-08 2021-06-01  -5.28097689
#> 91  2021-09-08 2021-07-01 -13.61361480
#> 92  2021-09-08 2021-08-01  -9.35952207
#> 93  2021-09-08 2021-09-01  -8.82585964
#> 94  2021-09-08 2021-10-01           NA
#> 95  2021-09-08 2021-11-01           NA
#> 96  2021-09-08 2021-12-01           NA
#> 97  2021-11-12 2020-01-01   6.81435964
#> 98  2021-11-12 2020-02-01  10.68091723
#> 99  2021-11-12 2020-03-01  14.02213946
#> 100 2021-11-12 2020-04-01   6.43113890
#> 101 2021-11-12 2020-05-01   0.13264319
#> 102 2021-11-12 2020-06-01  -3.87688497
#> 103 2021-11-12 2020-07-01  -9.44818055
#> 104 2021-11-12 2020-08-01  -7.35596078
#> 105 2021-11-12 2020-09-01  -8.26095118
#> 106 2021-11-12 2020-10-01   3.15443712
#> 107 2021-11-12 2020-11-01   5.51910956
#> 108 2021-11-12 2020-12-01   3.64540002
#> 109 2021-11-12 2021-01-01   0.01231621
#> 110 2021-11-12 2021-02-01   6.33345778
#> 111 2021-11-12 2021-03-01   8.54016693
#> 112 2021-11-12 2021-04-01   6.13327632
#> 113 2021-11-12 2021-05-01   3.32281993
#> 114 2021-11-12 2021-06-01  -4.17943495
#> 115 2021-11-12 2021-07-01 -11.99131542
#> 116 2021-11-12 2021-08-01  -7.61318582
#> 117 2021-11-12 2021-09-01  -5.39386946
#> 118 2021-11-12 2021-10-01  -0.01409892
#> 119 2021-11-12 2021-11-01   1.59275517
#> 120 2021-11-12 2021-12-01           NA
#> 121 2021-11-17 2020-01-01   7.16957396
#> 122 2021-11-17 2020-02-01  10.55718759
#> 123 2021-11-17 2020-03-01  14.60173632
#> 124 2021-11-17 2020-04-01   6.49094429
#> 125 2021-11-17 2020-05-01  -0.01133951
#> 126 2021-11-17 2020-06-01  -3.99280105
#> 127 2021-11-17 2020-07-01  -9.68583700
#> 128 2021-11-17 2020-08-01  -7.55711475
#> 129 2021-11-17 2020-09-01  -8.19768805
#> 130 2021-11-17 2020-10-01   1.91888435
#> 131 2021-11-17 2020-11-01   5.41834715
#> 132 2021-11-17 2020-12-01   4.03737820
#> 133 2021-11-17 2021-01-01   0.45970935
#> 134 2021-11-17 2021-02-01   7.14554259
#> 135 2021-11-17 2021-03-01   8.90732778
#> 136 2021-11-17 2021-04-01   6.62096642
#> 137 2021-11-17 2021-05-01   1.50965562
#> 138 2021-11-17 2021-06-01  -2.98827080
#> 139 2021-11-17 2021-07-01 -10.80858967
#> 140 2021-11-17 2021-08-01  -7.00191573
#> 141 2021-11-17 2021-09-01  -6.62342064
#> 142 2021-11-17 2021-10-01  -0.08599192
#> 143 2021-11-17 2021-11-01   4.17138167
#> 144 2021-11-17 2021-12-01           NA