Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01   3.3691888
#> 2   2020-02-17 2020-02-01  -0.9196899
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01   2.6569101
#> 26  2020-06-05 2020-02-01  -4.1925747
#> 27  2020-06-05 2020-03-01  -8.7119588
#> 28  2020-06-05 2020-04-01  -4.3307094
#> 29  2020-06-05 2020-05-01  -4.0665584
#> 30  2020-06-05 2020-06-01 -10.9508849
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01   2.6549804
#> 50  2021-04-17 2020-02-01  -4.4524361
#> 51  2021-04-17 2020-03-01  -8.0526071
#> 52  2021-04-17 2020-04-01  -6.9107559
#> 53  2021-04-17 2020-05-01  -6.7527447
#> 54  2021-04-17 2020-06-01 -10.8158883
#> 55  2021-04-17 2020-07-01 -17.2721542
#> 56  2021-04-17 2020-08-01 -13.6571444
#> 57  2021-04-17 2020-09-01 -16.7245668
#> 58  2021-04-17 2020-10-01 -27.1599964
#> 59  2021-04-17 2020-11-01 -26.4098781
#> 60  2021-04-17 2020-12-01 -30.4635422
#> 61  2021-04-17 2021-01-01 -38.4500736
#> 62  2021-04-17 2021-02-01 -40.7931065
#> 63  2021-04-17 2021-03-01 -44.3937012
#> 64  2021-04-17 2021-04-01 -46.1101923
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01   2.2726469
#> 74  2021-09-08 2020-02-01  -3.7752594
#> 75  2021-09-08 2020-03-01  -7.2834013
#> 76  2021-09-08 2020-04-01  -7.0313890
#> 77  2021-09-08 2020-05-01  -7.3236061
#> 78  2021-09-08 2020-06-01  -9.8204598
#> 79  2021-09-08 2020-07-01 -16.0263053
#> 80  2021-09-08 2020-08-01 -14.9825348
#> 81  2021-09-08 2020-09-01 -20.3061940
#> 82  2021-09-08 2020-10-01 -23.4130868
#> 83  2021-09-08 2020-11-01 -26.9685607
#> 84  2021-09-08 2020-12-01 -31.2853753
#> 85  2021-09-08 2021-01-01 -36.9620966
#> 86  2021-09-08 2021-02-01 -41.6764709
#> 87  2021-09-08 2021-03-01 -45.5324293
#> 88  2021-09-08 2021-04-01 -46.2636236
#> 89  2021-09-08 2021-05-01 -44.8669849
#> 90  2021-09-08 2021-06-01 -49.3320806
#> 91  2021-09-08 2021-07-01 -55.8679983
#> 92  2021-09-08 2021-08-01 -58.3119806
#> 93  2021-09-08 2021-09-01 -58.4505900
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01   2.2199730
#> 98  2021-11-12 2020-02-01  -3.7333627
#> 99  2021-11-12 2020-03-01  -8.4028088
#> 100 2021-11-12 2020-04-01  -6.7425719
#> 101 2021-11-12 2020-05-01  -6.6546702
#> 102 2021-11-12 2020-06-01  -9.6662354
#> 103 2021-11-12 2020-07-01 -14.7652154
#> 104 2021-11-12 2020-08-01 -14.4020389
#> 105 2021-11-12 2020-09-01 -18.6358062
#> 106 2021-11-12 2020-10-01 -24.7686879
#> 107 2021-11-12 2020-11-01 -27.0735411
#> 108 2021-11-12 2020-12-01 -29.3425162
#> 109 2021-11-12 2021-01-01 -37.6788524
#> 110 2021-11-12 2021-02-01 -42.2669067
#> 111 2021-11-12 2021-03-01 -44.5917496
#> 112 2021-11-12 2021-04-01 -44.9410885
#> 113 2021-11-12 2021-05-01 -43.5308306
#> 114 2021-11-12 2021-06-01 -48.2305386
#> 115 2021-11-12 2021-07-01 -54.2456990
#> 116 2021-11-12 2021-08-01 -56.5656444
#> 117 2021-11-12 2021-09-01 -55.0185998
#> 118 2021-11-12 2021-10-01 -58.0939643
#> 119 2021-11-12 2021-11-01 -53.9708727
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01   2.5751873
#> 122 2021-11-17 2020-02-01  -3.8570924
#> 123 2021-11-17 2020-03-01  -7.8232119
#> 124 2021-11-17 2020-04-01  -6.6827665
#> 125 2021-11-17 2020-05-01  -6.7986529
#> 126 2021-11-17 2020-06-01  -9.7821514
#> 127 2021-11-17 2020-07-01 -15.0028718
#> 128 2021-11-17 2020-08-01 -14.6031929
#> 129 2021-11-17 2020-09-01 -18.5725430
#> 130 2021-11-17 2020-10-01 -26.0042407
#> 131 2021-11-17 2020-11-01 -27.1743035
#> 132 2021-11-17 2020-12-01 -28.9505380
#> 133 2021-11-17 2021-01-01 -37.2314593
#> 134 2021-11-17 2021-02-01 -41.4548219
#> 135 2021-11-17 2021-03-01 -44.2245887
#> 136 2021-11-17 2021-04-01 -44.4533984
#> 137 2021-11-17 2021-05-01 -45.3439949
#> 138 2021-11-17 2021-06-01 -47.0393745
#> 139 2021-11-17 2021-07-01 -53.0629732
#> 140 2021-11-17 2021-08-01 -55.9543743
#> 141 2021-11-17 2021-09-01 -56.2481510
#> 142 2021-11-17 2021-10-01 -58.1658573
#> 143 2021-11-17 2021-11-01 -51.3922462
#> 144 2021-11-17 2021-12-01          NA