Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01  17.763610
#> 2   2020-02-17 2020-02-01  16.353238
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01  17.051331
#> 26  2020-06-05 2020-02-01  13.080353
#> 27  2020-06-05 2020-03-01  12.511325
#> 28  2020-06-05 2020-04-01  14.399986
#> 29  2020-06-05 2020-05-01  12.138711
#> 30  2020-06-05 2020-06-01   6.739590
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01  17.049401
#> 50  2021-04-17 2020-02-01  12.820492
#> 51  2021-04-17 2020-03-01  13.170677
#> 52  2021-04-17 2020-04-01  11.819939
#> 53  2021-04-17 2020-05-01   9.452524
#> 54  2021-04-17 2020-06-01   6.874586
#> 55  2021-04-17 2020-07-01  11.372622
#> 56  2021-04-17 2020-08-01  11.750973
#> 57  2021-04-17 2020-09-01   8.968892
#> 58  2021-04-17 2020-10-01   3.950371
#> 59  2021-04-17 2020-11-01   6.561165
#> 60  2021-04-17 2020-12-01   2.031547
#> 61  2021-04-17 2021-01-01   9.240621
#> 62  2021-04-17 2021-02-01   4.895435
#> 63  2021-04-17 2021-03-01   2.915738
#> 64  2021-04-17 2021-04-01   4.854520
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01  16.667068
#> 74  2021-09-08 2020-02-01  13.497668
#> 75  2021-09-08 2020-03-01  13.939883
#> 76  2021-09-08 2020-04-01  11.699306
#> 77  2021-09-08 2020-05-01   8.881663
#> 78  2021-09-08 2020-06-01   7.870015
#> 79  2021-09-08 2020-07-01  12.618471
#> 80  2021-09-08 2020-08-01  10.425582
#> 81  2021-09-08 2020-09-01   5.387265
#> 82  2021-09-08 2020-10-01   7.697281
#> 83  2021-09-08 2020-11-01   6.002483
#> 84  2021-09-08 2020-12-01   1.209714
#> 85  2021-09-08 2021-01-01  10.728598
#> 86  2021-09-08 2021-02-01   4.012070
#> 87  2021-09-08 2021-03-01   1.777010
#> 88  2021-09-08 2021-04-01   4.701089
#> 89  2021-09-08 2021-05-01   5.586323
#> 90  2021-09-08 2021-06-01   3.313528
#> 91  2021-09-08 2021-07-01   7.067200
#> 92  2021-09-08 2021-08-01   5.602528
#> 93  2021-09-08 2021-09-01   1.540432
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01  16.614394
#> 98  2021-11-12 2020-02-01  13.539565
#> 99  2021-11-12 2020-03-01  12.820475
#> 100 2021-11-12 2020-04-01  11.988123
#> 101 2021-11-12 2020-05-01   9.550599
#> 102 2021-11-12 2020-06-01   8.024239
#> 103 2021-11-12 2020-07-01  13.879561
#> 104 2021-11-12 2020-08-01  11.006078
#> 105 2021-11-12 2020-09-01   7.057653
#> 106 2021-11-12 2020-10-01   6.341680
#> 107 2021-11-12 2020-11-01   5.897502
#> 108 2021-11-12 2020-12-01   3.152573
#> 109 2021-11-12 2021-01-01  10.011842
#> 110 2021-11-12 2021-02-01   3.421635
#> 111 2021-11-12 2021-03-01   2.717690
#> 112 2021-11-12 2021-04-01   6.023624
#> 113 2021-11-12 2021-05-01   6.922477
#> 114 2021-11-12 2021-06-01   4.415070
#> 115 2021-11-12 2021-07-01   8.689500
#> 116 2021-11-12 2021-08-01   7.348864
#> 117 2021-11-12 2021-09-01   4.972422
#> 118 2021-11-12 2021-10-01   4.612016
#> 119 2021-11-12 2021-11-01   5.186863
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01  16.969608
#> 122 2021-11-17 2020-02-01  13.415835
#> 123 2021-11-17 2020-03-01  13.400072
#> 124 2021-11-17 2020-04-01  12.047928
#> 125 2021-11-17 2020-05-01   9.406616
#> 126 2021-11-17 2020-06-01   7.908323
#> 127 2021-11-17 2020-07-01  13.641904
#> 128 2021-11-17 2020-08-01  10.804924
#> 129 2021-11-17 2020-09-01   7.120916
#> 130 2021-11-17 2020-10-01   5.106127
#> 131 2021-11-17 2020-11-01   5.796740
#> 132 2021-11-17 2020-12-01   3.544551
#> 133 2021-11-17 2021-01-01  10.459236
#> 134 2021-11-17 2021-02-01   4.233719
#> 135 2021-11-17 2021-03-01   3.084850
#> 136 2021-11-17 2021-04-01   6.511314
#> 137 2021-11-17 2021-05-01   5.109313
#> 138 2021-11-17 2021-06-01   5.606234
#> 139 2021-11-17 2021-07-01   9.872225
#> 140 2021-11-17 2021-08-01   7.960134
#> 141 2021-11-17 2021-09-01   3.742871
#> 142 2021-11-17 2021-10-01   4.540123
#> 143 2021-11-17 2021-11-01   7.765490
#> 144 2021-11-17 2021-12-01         NA