Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01  4.0479907
#> 2   2020-02-17 2020-02-01  2.3131548
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01  3.3357120
#> 26  2020-06-05 2020-02-01 -0.9597300
#> 27  2020-06-05 2020-03-01  1.2297348
#> 28  2020-06-05 2020-04-01  9.3237792
#> 29  2020-06-05 2020-05-01  5.8601999
#> 30  2020-06-05 2020-06-01 -3.8235603
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01  3.3337823
#> 50  2021-04-17 2020-02-01 -1.2195914
#> 51  2021-04-17 2020-03-01  1.8890866
#> 52  2021-04-17 2020-04-01  6.7437326
#> 53  2021-04-17 2020-05-01  3.1740136
#> 54  2021-04-17 2020-06-01 -3.6885636
#> 55  2021-04-17 2020-07-01 -2.6479434
#> 56  2021-04-17 2020-08-01 -3.4029193
#> 57  2021-04-17 2020-09-01  1.3973335
#> 58  2021-04-17 2020-10-01 -0.5806657
#> 59  2021-04-17 2020-11-01  1.5704449
#> 60  2021-04-17 2020-12-01  2.6108496
#> 61  2021-04-17 2021-01-01  2.5099056
#> 62  2021-04-17 2021-02-01  3.8587305
#> 63  2021-04-17 2021-03-01 11.6477949
#> 64  2021-04-17 2021-04-01 23.4846559
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01  2.9514488
#> 74  2021-09-08 2020-02-01 -0.5424147
#> 75  2021-09-08 2020-03-01  2.6582923
#> 76  2021-09-08 2020-04-01  6.6230995
#> 77  2021-09-08 2020-05-01  2.6031522
#> 78  2021-09-08 2020-06-01 -2.6931352
#> 79  2021-09-08 2020-07-01 -1.4020944
#> 80  2021-09-08 2020-08-01 -4.7283098
#> 81  2021-09-08 2020-09-01 -2.1842937
#> 82  2021-09-08 2020-10-01  3.1662439
#> 83  2021-09-08 2020-11-01  1.0117623
#> 84  2021-09-08 2020-12-01  1.7890165
#> 85  2021-09-08 2021-01-01  3.9978825
#> 86  2021-09-08 2021-02-01  2.9753660
#> 87  2021-09-08 2021-03-01 10.5090669
#> 88  2021-09-08 2021-04-01 23.3312247
#> 89  2021-09-08 2021-05-01 30.0146795
#> 90  2021-09-08 2021-06-01 25.8295967
#> 91  2021-09-08 2021-07-01 27.0964553
#> 92  2021-09-08 2021-08-01 26.8167809
#> 93  2021-09-08 2021-09-01 29.3431667
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01  2.8987749
#> 98  2021-11-12 2020-02-01 -0.5005180
#> 99  2021-11-12 2020-03-01  1.5388848
#> 100 2021-11-12 2020-04-01  6.9119167
#> 101 2021-11-12 2020-05-01  3.2720881
#> 102 2021-11-12 2020-06-01 -2.5389107
#> 103 2021-11-12 2020-07-01 -0.1410045
#> 104 2021-11-12 2020-08-01 -4.1478139
#> 105 2021-11-12 2020-09-01 -0.5139058
#> 106 2021-11-12 2020-10-01  1.8106429
#> 107 2021-11-12 2020-11-01  0.9067819
#> 108 2021-11-12 2020-12-01  3.7318756
#> 109 2021-11-12 2021-01-01  3.2811267
#> 110 2021-11-12 2021-02-01  2.3849303
#> 111 2021-11-12 2021-03-01 11.4497466
#> 112 2021-11-12 2021-04-01 24.6537597
#> 113 2021-11-12 2021-05-01 31.3508338
#> 114 2021-11-12 2021-06-01 26.9311386
#> 115 2021-11-12 2021-07-01 28.7187547
#> 116 2021-11-12 2021-08-01 28.5631171
#> 117 2021-11-12 2021-09-01 32.7751569
#> 118 2021-11-12 2021-10-01 32.5526548
#> 119 2021-11-12 2021-11-01 31.2217537
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01  3.2539892
#> 122 2021-11-17 2020-02-01 -0.6242477
#> 123 2021-11-17 2020-03-01  2.1184817
#> 124 2021-11-17 2020-04-01  6.9717220
#> 125 2021-11-17 2020-05-01  3.1281054
#> 126 2021-11-17 2020-06-01 -2.6548268
#> 127 2021-11-17 2020-07-01 -0.3786609
#> 128 2021-11-17 2020-08-01 -4.3489679
#> 129 2021-11-17 2020-09-01 -0.4506427
#> 130 2021-11-17 2020-10-01  0.5750901
#> 131 2021-11-17 2020-11-01  0.8060195
#> 132 2021-11-17 2020-12-01  4.1238537
#> 133 2021-11-17 2021-01-01  3.7285198
#> 134 2021-11-17 2021-02-01  3.1970151
#> 135 2021-11-17 2021-03-01 11.8169075
#> 136 2021-11-17 2021-04-01 25.1414498
#> 137 2021-11-17 2021-05-01 29.5376695
#> 138 2021-11-17 2021-06-01 28.1223027
#> 139 2021-11-17 2021-07-01 29.9014805
#> 140 2021-11-17 2021-08-01 29.1743872
#> 141 2021-11-17 2021-09-01 31.5456057
#> 142 2021-11-17 2021-10-01 32.4807618
#> 143 2021-11-17 2021-11-01 33.8003802
#> 144 2021-11-17 2021-12-01         NA