Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01   9.791538
#> 2   2020-02-17 2020-02-01   9.619741
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01   9.079259
#> 26  2020-06-05 2020-02-01   6.346856
#> 27  2020-06-05 2020-03-01   8.795301
#> 28  2020-06-05 2020-04-01  15.371163
#> 29  2020-06-05 2020-05-01  19.708920
#> 30  2020-06-05 2020-06-01  16.302862
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01   9.077329
#> 50  2021-04-17 2020-02-01   6.086995
#> 51  2021-04-17 2020-03-01   9.454653
#> 52  2021-04-17 2020-04-01  12.791117
#> 53  2021-04-17 2020-05-01  17.022733
#> 54  2021-04-17 2020-06-01  16.437858
#> 55  2021-04-17 2020-07-01  16.360515
#> 56  2021-04-17 2020-08-01  16.189685
#> 57  2021-04-17 2020-09-01  15.431916
#> 58  2021-04-17 2020-10-01  11.088358
#> 59  2021-04-17 2020-11-01  12.871906
#> 60  2021-04-17 2020-12-01  12.759391
#> 61  2021-04-17 2021-01-01  11.058914
#> 62  2021-04-17 2021-02-01   9.858087
#> 63  2021-04-17 2021-03-01  10.234741
#> 64  2021-04-17 2021-04-01  13.121626
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01   8.694996
#> 74  2021-09-08 2020-02-01   6.764172
#> 75  2021-09-08 2020-03-01  10.223859
#> 76  2021-09-08 2020-04-01  12.670484
#> 77  2021-09-08 2020-05-01  16.451872
#> 78  2021-09-08 2020-06-01  17.433287
#> 79  2021-09-08 2020-07-01  17.606364
#> 80  2021-09-08 2020-08-01  14.864295
#> 81  2021-09-08 2020-09-01  11.850289
#> 82  2021-09-08 2020-10-01  14.835268
#> 83  2021-09-08 2020-11-01  12.313224
#> 84  2021-09-08 2020-12-01  11.937558
#> 85  2021-09-08 2021-01-01  12.546891
#> 86  2021-09-08 2021-02-01   8.974723
#> 87  2021-09-08 2021-03-01   9.096013
#> 88  2021-09-08 2021-04-01  12.968194
#> 89  2021-09-08 2021-05-01  19.203417
#> 90  2021-09-08 2021-06-01  20.423785
#> 91  2021-09-08 2021-07-01  22.439321
#> 92  2021-09-08 2021-08-01  17.451990
#> 93  2021-09-08 2021-09-01  12.700808
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01   8.642322
#> 98  2021-11-12 2020-02-01   6.806068
#> 99  2021-11-12 2020-03-01   9.104451
#> 100 2021-11-12 2020-04-01  12.959301
#> 101 2021-11-12 2020-05-01  17.120808
#> 102 2021-11-12 2020-06-01  17.587511
#> 103 2021-11-12 2020-07-01  18.867453
#> 104 2021-11-12 2020-08-01  15.444791
#> 105 2021-11-12 2020-09-01  13.520676
#> 106 2021-11-12 2020-10-01  13.479667
#> 107 2021-11-12 2020-11-01  12.208243
#> 108 2021-11-12 2020-12-01  13.880417
#> 109 2021-11-12 2021-01-01  11.830135
#> 110 2021-11-12 2021-02-01   8.384287
#> 111 2021-11-12 2021-03-01  10.036693
#> 112 2021-11-12 2021-04-01  14.290729
#> 113 2021-11-12 2021-05-01  20.539572
#> 114 2021-11-12 2021-06-01  21.525327
#> 115 2021-11-12 2021-07-01  24.061620
#> 116 2021-11-12 2021-08-01  19.198326
#> 117 2021-11-12 2021-09-01  16.132798
#> 118 2021-11-12 2021-10-01  12.827899
#> 119 2021-11-12 2021-11-01  11.780988
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01   8.997536
#> 122 2021-11-17 2020-02-01   6.682339
#> 123 2021-11-17 2020-03-01   9.684048
#> 124 2021-11-17 2020-04-01  13.019106
#> 125 2021-11-17 2020-05-01  16.976825
#> 126 2021-11-17 2020-06-01  17.471595
#> 127 2021-11-17 2020-07-01  18.629797
#> 128 2021-11-17 2020-08-01  15.243637
#> 129 2021-11-17 2020-09-01  13.583940
#> 130 2021-11-17 2020-10-01  12.244114
#> 131 2021-11-17 2020-11-01  12.107481
#> 132 2021-11-17 2020-12-01  14.272395
#> 133 2021-11-17 2021-01-01  12.277528
#> 134 2021-11-17 2021-02-01   9.196372
#> 135 2021-11-17 2021-03-01  10.403854
#> 136 2021-11-17 2021-04-01  14.778420
#> 137 2021-11-17 2021-05-01  18.726407
#> 138 2021-11-17 2021-06-01  22.716491
#> 139 2021-11-17 2021-07-01  25.244346
#> 140 2021-11-17 2021-08-01  19.809596
#> 141 2021-11-17 2021-09-01  14.903247
#> 142 2021-11-17 2021-10-01  12.756006
#> 143 2021-11-17 2021-11-01  14.359614
#> 144 2021-11-17 2021-12-01         NA