Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time   obs_values
#> 1   2020-02-17 2020-01-01   5.39716868
#> 2   2020-02-17 2020-02-01   6.70800739
#> 3   2020-02-17 2020-03-01           NA
#> 4   2020-02-17 2020-04-01           NA
#> 5   2020-02-17 2020-05-01           NA
#> 6   2020-02-17 2020-06-01           NA
#> 7   2020-02-17 2020-07-01           NA
#> 8   2020-02-17 2020-08-01           NA
#> 9   2020-02-17 2020-09-01           NA
#> 10  2020-02-17 2020-10-01           NA
#> 11  2020-02-17 2020-11-01           NA
#> 12  2020-02-17 2020-12-01           NA
#> 13  2020-02-17 2021-01-01           NA
#> 14  2020-02-17 2021-02-01           NA
#> 15  2020-02-17 2021-03-01           NA
#> 16  2020-02-17 2021-04-01           NA
#> 17  2020-02-17 2021-05-01           NA
#> 18  2020-02-17 2021-06-01           NA
#> 19  2020-02-17 2021-07-01           NA
#> 20  2020-02-17 2021-08-01           NA
#> 21  2020-02-17 2021-09-01           NA
#> 22  2020-02-17 2021-10-01           NA
#> 23  2020-02-17 2021-11-01           NA
#> 24  2020-02-17 2021-12-01           NA
#> 25  2020-06-05 2020-01-01   4.68489004
#> 26  2020-06-05 2020-02-01   3.43512253
#> 27  2020-06-05 2020-03-01   0.68962867
#> 28  2020-06-05 2020-04-01   0.38460776
#> 29  2020-06-05 2020-05-01  -3.30396449
#> 30  2020-06-05 2020-06-01  -4.32182951
#> 31  2020-06-05 2020-07-01           NA
#> 32  2020-06-05 2020-08-01           NA
#> 33  2020-06-05 2020-09-01           NA
#> 34  2020-06-05 2020-10-01           NA
#> 35  2020-06-05 2020-11-01           NA
#> 36  2020-06-05 2020-12-01           NA
#> 37  2020-06-05 2021-01-01           NA
#> 38  2020-06-05 2021-02-01           NA
#> 39  2020-06-05 2021-03-01           NA
#> 40  2020-06-05 2021-04-01           NA
#> 41  2020-06-05 2021-05-01           NA
#> 42  2020-06-05 2021-06-01           NA
#> 43  2020-06-05 2021-07-01           NA
#> 44  2020-06-05 2021-08-01           NA
#> 45  2020-06-05 2021-09-01           NA
#> 46  2020-06-05 2021-10-01           NA
#> 47  2020-06-05 2021-11-01           NA
#> 48  2020-06-05 2021-12-01           NA
#> 49  2021-04-17 2020-01-01   4.68296030
#> 50  2021-04-17 2020-02-01   3.17526118
#> 51  2021-04-17 2020-03-01   1.34898041
#> 52  2021-04-17 2020-04-01  -2.19543876
#> 53  2021-04-17 2020-05-01  -5.99015076
#> 54  2021-04-17 2020-06-01  -4.18683284
#> 55  2021-04-17 2020-07-01  -2.32465239
#> 56  2021-04-17 2020-08-01   2.58525079
#> 57  2021-04-17 2020-09-01   4.99157230
#> 58  2021-04-17 2020-10-01   1.64390177
#> 59  2021-04-17 2020-11-01   3.61847042
#> 60  2021-04-17 2020-12-01  -1.14146505
#> 61  2021-04-17 2021-01-01  -8.56030307
#> 62  2021-04-17 2021-02-01  -7.95334055
#> 63  2021-04-17 2021-03-01  -8.57289788
#> 64  2021-04-17 2021-04-01 -14.49944598
#> 65  2021-04-17 2021-05-01           NA
#> 66  2021-04-17 2021-06-01           NA
#> 67  2021-04-17 2021-07-01           NA
#> 68  2021-04-17 2021-08-01           NA
#> 69  2021-04-17 2021-09-01           NA
#> 70  2021-04-17 2021-10-01           NA
#> 71  2021-04-17 2021-11-01           NA
#> 72  2021-04-17 2021-12-01           NA
#> 73  2021-09-08 2020-01-01   4.30062676
#> 74  2021-09-08 2020-02-01   3.85243784
#> 75  2021-09-08 2020-03-01   2.11818614
#> 76  2021-09-08 2020-04-01  -2.31607192
#> 77  2021-09-08 2020-05-01  -6.56101216
#> 78  2021-09-08 2020-06-01  -3.19140444
#> 79  2021-09-08 2020-07-01  -1.07880340
#> 80  2021-09-08 2020-08-01   1.25986033
#> 81  2021-09-08 2020-09-01   1.40994510
#> 82  2021-09-08 2020-10-01   5.39081138
#> 83  2021-09-08 2020-11-01   3.05978782
#> 84  2021-09-08 2020-12-01  -1.96329815
#> 85  2021-09-08 2021-01-01  -7.07232615
#> 86  2021-09-08 2021-02-01  -8.83670504
#> 87  2021-09-08 2021-03-01  -9.71162592
#> 88  2021-09-08 2021-04-01 -14.65287726
#> 89  2021-09-08 2021-05-01 -13.94608034
#> 90  2021-09-08 2021-06-01 -10.26657454
#> 91  2021-09-08 2021-07-01  -6.70269983
#> 92  2021-09-08 2021-08-01  -5.56701776
#> 93  2021-09-08 2021-09-01  -5.94417471
#> 94  2021-09-08 2021-10-01           NA
#> 95  2021-09-08 2021-11-01           NA
#> 96  2021-09-08 2021-12-01           NA
#> 97  2021-11-12 2020-01-01   4.24795289
#> 98  2021-11-12 2020-02-01   3.89433452
#> 99  2021-11-12 2020-03-01   0.99877869
#> 100 2021-11-12 2020-04-01  -2.02725474
#> 101 2021-11-12 2020-05-01  -5.89207626
#> 102 2021-11-12 2020-06-01  -3.03717995
#> 103 2021-11-12 2020-07-01   0.18228649
#> 104 2021-11-12 2020-08-01   1.84035622
#> 105 2021-11-12 2020-09-01   3.08033292
#> 106 2021-11-12 2020-10-01   4.03521031
#> 107 2021-11-12 2020-11-01   2.95480744
#> 108 2021-11-12 2020-12-01  -0.02043909
#> 109 2021-11-12 2021-01-01  -7.78908195
#> 110 2021-11-12 2021-02-01  -9.42714080
#> 111 2021-11-12 2021-03-01  -8.77094621
#> 112 2021-11-12 2021-04-01 -13.33034219
#> 113 2021-11-12 2021-05-01 -12.60992597
#> 114 2021-11-12 2021-06-01  -9.16503260
#> 115 2021-11-12 2021-07-01  -5.08040046
#> 116 2021-11-12 2021-08-01  -3.82068151
#> 117 2021-11-12 2021-09-01  -2.51218453
#> 118 2021-11-12 2021-10-01  -0.45018365
#> 119 2021-11-12 2021-11-01   4.08415975
#> 120 2021-11-12 2021-12-01           NA
#> 121 2021-11-17 2020-01-01   4.60316721
#> 122 2021-11-17 2020-02-01   3.77060488
#> 123 2021-11-17 2020-03-01   1.57837555
#> 124 2021-11-17 2020-04-01  -1.96744935
#> 125 2021-11-17 2020-05-01  -6.03605897
#> 126 2021-11-17 2020-06-01  -3.15309603
#> 127 2021-11-17 2020-07-01  -0.05536996
#> 128 2021-11-17 2020-08-01   1.63920224
#> 129 2021-11-17 2020-09-01   3.14359606
#> 130 2021-11-17 2020-10-01   2.79965754
#> 131 2021-11-17 2020-11-01   2.85404503
#> 132 2021-11-17 2020-12-01   0.37153909
#> 133 2021-11-17 2021-01-01  -7.34168881
#> 134 2021-11-17 2021-02-01  -8.61505599
#> 135 2021-11-17 2021-03-01  -8.40378536
#> 136 2021-11-17 2021-04-01 -12.84265208
#> 137 2021-11-17 2021-05-01 -14.42309029
#> 138 2021-11-17 2021-06-01  -7.97386844
#> 139 2021-11-17 2021-07-01  -3.89767470
#> 140 2021-11-17 2021-08-01  -3.20941142
#> 141 2021-11-17 2021-09-01  -3.74173571
#> 142 2021-11-17 2021-10-01  -0.52207664
#> 143 2021-11-17 2021-11-01   6.66278625
#> 144 2021-11-17 2021-12-01           NA