Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-05-19 2020-01-01   7.4483440
#> 2   2020-05-19 2020-02-01  13.6484059
#> 3   2020-05-19 2020-03-01   4.1018045
#> 4   2020-05-19 2020-04-01   7.5425523
#> 5   2020-05-19 2020-05-01   5.1088680
#> 6   2020-05-19 2020-06-01          NA
#> 7   2020-05-19 2020-07-01          NA
#> 8   2020-05-19 2020-08-01          NA
#> 9   2020-05-19 2020-09-01          NA
#> 10  2020-05-19 2020-10-01          NA
#> 11  2020-05-19 2020-11-01          NA
#> 12  2020-05-19 2020-12-01          NA
#> 13  2020-05-19 2021-01-01          NA
#> 14  2020-05-19 2021-02-01          NA
#> 15  2020-05-19 2021-03-01          NA
#> 16  2020-05-19 2021-04-01          NA
#> 17  2020-05-19 2021-05-01          NA
#> 18  2020-05-19 2021-06-01          NA
#> 19  2020-05-19 2021-07-01          NA
#> 20  2020-05-19 2021-08-01          NA
#> 21  2020-05-19 2021-09-01          NA
#> 22  2020-05-19 2021-10-01          NA
#> 23  2020-05-19 2021-11-01          NA
#> 24  2020-05-19 2021-12-01          NA
#> 25  2020-07-30 2020-01-01   8.0660644
#> 26  2020-07-30 2020-02-01   9.5815884
#> 27  2020-07-30 2020-03-01   7.0233632
#> 28  2020-07-30 2020-04-01   5.8889903
#> 29  2020-07-30 2020-05-01   3.8793858
#> 30  2020-07-30 2020-06-01   3.4633294
#> 31  2020-07-30 2020-07-01   5.5083122
#> 32  2020-07-30 2020-08-01          NA
#> 33  2020-07-30 2020-09-01          NA
#> 34  2020-07-30 2020-10-01          NA
#> 35  2020-07-30 2020-11-01          NA
#> 36  2020-07-30 2020-12-01          NA
#> 37  2020-07-30 2021-01-01          NA
#> 38  2020-07-30 2021-02-01          NA
#> 39  2020-07-30 2021-03-01          NA
#> 40  2020-07-30 2021-04-01          NA
#> 41  2020-07-30 2021-05-01          NA
#> 42  2020-07-30 2021-06-01          NA
#> 43  2020-07-30 2021-07-01          NA
#> 44  2020-07-30 2021-08-01          NA
#> 45  2020-07-30 2021-09-01          NA
#> 46  2020-07-30 2021-10-01          NA
#> 47  2020-07-30 2021-11-01          NA
#> 48  2020-07-30 2021-12-01          NA
#> 49  2020-11-20 2020-01-01  10.5043611
#> 50  2020-11-20 2020-02-01   9.0909366
#> 51  2020-11-20 2020-03-01   5.3260699
#> 52  2020-11-20 2020-04-01   6.9986011
#> 53  2020-11-20 2020-05-01   4.6613056
#> 54  2020-11-20 2020-06-01   4.9309677
#> 55  2020-11-20 2020-07-01   4.8066898
#> 56  2020-11-20 2020-08-01   6.0311963
#> 57  2020-11-20 2020-09-01   6.1018909
#> 58  2020-11-20 2020-10-01   1.3572720
#> 59  2020-11-20 2020-11-01   2.7940947
#> 60  2020-11-20 2020-12-01          NA
#> 61  2020-11-20 2021-01-01          NA
#> 62  2020-11-20 2021-02-01          NA
#> 63  2020-11-20 2021-03-01          NA
#> 64  2020-11-20 2021-04-01          NA
#> 65  2020-11-20 2021-05-01          NA
#> 66  2020-11-20 2021-06-01          NA
#> 67  2020-11-20 2021-07-01          NA
#> 68  2020-11-20 2021-08-01          NA
#> 69  2020-11-20 2021-09-01          NA
#> 70  2020-11-20 2021-10-01          NA
#> 71  2020-11-20 2021-11-01          NA
#> 72  2020-11-20 2021-12-01          NA
#> 73  2021-01-16 2020-01-01   9.8799381
#> 74  2021-01-16 2020-02-01   8.8005489
#> 75  2021-01-16 2020-03-01   5.3536986
#> 76  2021-01-16 2020-04-01   6.3006560
#> 77  2021-01-16 2020-05-01   4.4619811
#> 78  2021-01-16 2020-06-01   3.5093486
#> 79  2021-01-16 2020-07-01   3.6651335
#> 80  2021-01-16 2020-08-01   6.2225492
#> 81  2021-01-16 2020-09-01   3.9755353
#> 82  2021-01-16 2020-10-01   4.6526147
#> 83  2021-01-16 2020-11-01   5.9901850
#> 84  2021-01-16 2020-12-01   0.3171446
#> 85  2021-01-16 2021-01-01 -11.1926160
#> 86  2021-01-16 2021-02-01          NA
#> 87  2021-01-16 2021-03-01          NA
#> 88  2021-01-16 2021-04-01          NA
#> 89  2021-01-16 2021-05-01          NA
#> 90  2021-01-16 2021-06-01          NA
#> 91  2021-01-16 2021-07-01          NA
#> 92  2021-01-16 2021-08-01          NA
#> 93  2021-01-16 2021-09-01          NA
#> 94  2021-01-16 2021-10-01          NA
#> 95  2021-01-16 2021-11-01          NA
#> 96  2021-01-16 2021-12-01          NA
#> 97  2021-04-26 2020-01-01   9.9905305
#> 98  2021-04-26 2020-02-01   8.4699640
#> 99  2021-04-26 2020-03-01   5.4019122
#> 100 2021-04-26 2020-04-01   6.2843512
#> 101 2021-04-26 2020-05-01   5.3090112
#> 102 2021-04-26 2020-06-01   4.2801405
#> 103 2021-04-26 2020-07-01   4.6265849
#> 104 2021-04-26 2020-08-01   6.2444954
#> 105 2021-04-26 2020-09-01   3.3987596
#> 106 2021-04-26 2020-10-01   3.7587823
#> 107 2021-04-26 2020-11-01   5.5993642
#> 108 2021-04-26 2020-12-01   0.1200310
#> 109 2021-04-26 2021-01-01  -8.0693524
#> 110 2021-04-26 2021-02-01 -11.5527353
#> 111 2021-04-26 2021-03-01 -12.5860314
#> 112 2021-04-26 2021-04-01 -18.2876905
#> 113 2021-04-26 2021-05-01          NA
#> 114 2021-04-26 2021-06-01          NA
#> 115 2021-04-26 2021-07-01          NA
#> 116 2021-04-26 2021-08-01          NA
#> 117 2021-04-26 2021-09-01          NA
#> 118 2021-04-26 2021-10-01          NA
#> 119 2021-04-26 2021-11-01          NA
#> 120 2021-04-26 2021-12-01          NA
#> 121 2021-11-14 2020-01-01   9.7317592
#> 122 2021-11-14 2020-02-01   8.4681354
#> 123 2021-11-14 2020-03-01   5.3371527
#> 124 2021-11-14 2020-04-01   6.1654529
#> 125 2021-11-14 2020-05-01   4.9058073
#> 126 2021-11-14 2020-06-01   4.1539934
#> 127 2021-11-14 2020-07-01   4.3489164
#> 128 2021-11-14 2020-08-01   5.9327367
#> 129 2021-11-14 2020-09-01   3.8122991
#> 130 2021-11-14 2020-10-01   4.6019465
#> 131 2021-11-14 2020-11-01   5.7078327
#> 132 2021-11-14 2020-12-01  -1.8765972
#> 133 2021-11-14 2021-01-01  -8.8368913
#> 134 2021-11-14 2021-02-01 -15.3915418
#> 135 2021-11-14 2021-03-01 -15.3487272
#> 136 2021-11-14 2021-04-01 -17.6904886
#> 137 2021-11-14 2021-05-01 -20.9386737
#> 138 2021-11-14 2021-06-01 -29.4164986
#> 139 2021-11-14 2021-07-01 -33.7900663
#> 140 2021-11-14 2021-08-01 -31.4756765
#> 141 2021-11-14 2021-09-01 -29.1490352
#> 142 2021-11-14 2021-10-01 -25.8862476
#> 143 2021-11-14 2021-11-01 -28.8929123
#> 144 2021-11-14 2021-12-01          NA