Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01  12.2325587
#> 2   2020-02-17 2020-02-01  12.0867959
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01  11.5202801
#> 26  2020-06-05 2020-02-01   8.8139110
#> 27  2020-06-05 2020-03-01  10.3612690
#> 28  2020-06-05 2020-04-01  16.7713033
#> 29  2020-06-05 2020-05-01  12.8043676
#> 30  2020-06-05 2020-06-01   2.4801194
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01  11.5183503
#> 50  2021-04-17 2020-02-01   8.5540497
#> 51  2021-04-17 2020-03-01  11.0206207
#> 52  2021-04-17 2020-04-01  14.1912568
#> 53  2021-04-17 2020-05-01  10.1181813
#> 54  2021-04-17 2020-06-01   2.6151161
#> 55  2021-04-17 2020-07-01   3.0202642
#> 56  2021-04-17 2020-08-01   7.8624062
#> 57  2021-04-17 2020-09-01   4.5343421
#> 58  2021-04-17 2020-10-01   0.6682246
#> 59  2021-04-17 2020-11-01   3.4896036
#> 60  2021-04-17 2020-12-01   1.6957748
#> 61  2021-04-17 2021-01-01   2.8013740
#> 62  2021-04-17 2021-02-01   4.3356370
#> 63  2021-04-17 2021-03-01   0.7304715
#> 64  2021-04-17 2021-04-01   3.4137786
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01  11.1360168
#> 74  2021-09-08 2020-02-01   9.2312264
#> 75  2021-09-08 2020-03-01  11.7898265
#> 76  2021-09-08 2020-04-01  14.0706237
#> 77  2021-09-08 2020-05-01   9.5473199
#> 78  2021-09-08 2020-06-01   3.6105445
#> 79  2021-09-08 2020-07-01   4.2661132
#> 80  2021-09-08 2020-08-01   6.5370157
#> 81  2021-09-08 2020-09-01   0.9527149
#> 82  2021-09-08 2020-10-01   4.4151342
#> 83  2021-09-08 2020-11-01   2.9309210
#> 84  2021-09-08 2020-12-01   0.8739417
#> 85  2021-09-08 2021-01-01   4.2893509
#> 86  2021-09-08 2021-02-01   3.4522725
#> 87  2021-09-08 2021-03-01  -0.4082566
#> 88  2021-09-08 2021-04-01   3.2603473
#> 89  2021-09-08 2021-05-01   3.0617739
#> 90  2021-09-08 2021-06-01  -3.0195805
#> 91  2021-09-08 2021-07-01  -4.7568160
#> 92  2021-09-08 2021-08-01  -8.5595925
#> 93  2021-09-08 2021-09-01 -17.6343327
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01  11.0833429
#> 98  2021-11-12 2020-02-01   9.2731230
#> 99  2021-11-12 2020-03-01  10.6704190
#> 100 2021-11-12 2020-04-01  14.3594408
#> 101 2021-11-12 2020-05-01  10.2162558
#> 102 2021-11-12 2020-06-01   3.7647690
#> 103 2021-11-12 2020-07-01   5.5272031
#> 104 2021-11-12 2020-08-01   7.1175116
#> 105 2021-11-12 2020-09-01   2.6231027
#> 106 2021-11-12 2020-10-01   3.0595331
#> 107 2021-11-12 2020-11-01   2.8259406
#> 108 2021-11-12 2020-12-01   2.8168008
#> 109 2021-11-12 2021-01-01   3.5725951
#> 110 2021-11-12 2021-02-01   2.8618368
#> 111 2021-11-12 2021-03-01   0.5324231
#> 112 2021-11-12 2021-04-01   4.5828824
#> 113 2021-11-12 2021-05-01   4.3979283
#> 114 2021-11-12 2021-06-01  -1.9180386
#> 115 2021-11-12 2021-07-01  -3.1345167
#> 116 2021-11-12 2021-08-01  -6.8132562
#> 117 2021-11-12 2021-09-01 -14.2023425
#> 118 2021-11-12 2021-10-01 -18.7459550
#> 119 2021-11-12 2021-11-01 -24.5711794
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01  11.4385573
#> 122 2021-11-17 2020-02-01   9.1493934
#> 123 2021-11-17 2020-03-01  11.2500159
#> 124 2021-11-17 2020-04-01  14.4192462
#> 125 2021-11-17 2020-05-01  10.0722731
#> 126 2021-11-17 2020-06-01   3.6488529
#> 127 2021-11-17 2020-07-01   5.2895466
#> 128 2021-11-17 2020-08-01   6.9163576
#> 129 2021-11-17 2020-09-01   2.6863658
#> 130 2021-11-17 2020-10-01   1.8239804
#> 131 2021-11-17 2020-11-01   2.7251782
#> 132 2021-11-17 2020-12-01   3.2087789
#> 133 2021-11-17 2021-01-01   4.0199883
#> 134 2021-11-17 2021-02-01   3.6739216
#> 135 2021-11-17 2021-03-01   0.8995840
#> 136 2021-11-17 2021-04-01   5.0705725
#> 137 2021-11-17 2021-05-01   2.5847640
#> 138 2021-11-17 2021-06-01  -0.7268744
#> 139 2021-11-17 2021-07-01  -1.9517909
#> 140 2021-11-17 2021-08-01  -6.2019861
#> 141 2021-11-17 2021-09-01 -15.4318937
#> 142 2021-11-17 2021-10-01 -18.8178480
#> 143 2021-11-17 2021-11-01 -21.9925529
#> 144 2021-11-17 2021-12-01          NA