Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01   9.553487
#> 2   2020-02-17 2020-02-01   7.861273
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01   8.841208
#> 26  2020-06-05 2020-02-01   4.588388
#> 27  2020-06-05 2020-03-01   2.624364
#> 28  2020-06-05 2020-04-01   3.995132
#> 29  2020-06-05 2020-05-01   8.236425
#> 30  2020-06-05 2020-06-01   8.251431
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01   8.839278
#> 50  2021-04-17 2020-02-01   4.328527
#> 51  2021-04-17 2020-03-01   3.283716
#> 52  2021-04-17 2020-04-01   1.415086
#> 53  2021-04-17 2020-05-01   5.550239
#> 54  2021-04-17 2020-06-01   8.386427
#> 55  2021-04-17 2020-07-01   2.443234
#> 56  2021-04-17 2020-08-01   5.002036
#> 57  2021-04-17 2020-09-01   9.617757
#> 58  2021-04-17 2020-10-01   9.438229
#> 59  2021-04-17 2020-11-01  15.661168
#> 60  2021-04-17 2020-12-01  14.766249
#> 61  2021-04-17 2021-01-01  12.944977
#> 62  2021-04-17 2021-02-01  18.257595
#> 63  2021-04-17 2021-03-01  23.794918
#> 64  2021-04-17 2021-04-01  24.853760
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01   8.456945
#> 74  2021-09-08 2020-02-01   5.005703
#> 75  2021-09-08 2020-03-01   4.052922
#> 76  2021-09-08 2020-04-01   1.294453
#> 77  2021-09-08 2020-05-01   4.979378
#> 78  2021-09-08 2020-06-01   9.381856
#> 79  2021-09-08 2020-07-01   3.689083
#> 80  2021-09-08 2020-08-01   3.676646
#> 81  2021-09-08 2020-09-01   6.036129
#> 82  2021-09-08 2020-10-01  13.185138
#> 83  2021-09-08 2020-11-01  15.102485
#> 84  2021-09-08 2020-12-01  13.944415
#> 85  2021-09-08 2021-01-01  14.432954
#> 86  2021-09-08 2021-02-01  17.374231
#> 87  2021-09-08 2021-03-01  22.656190
#> 88  2021-09-08 2021-04-01  24.700329
#> 89  2021-09-08 2021-05-01  23.929585
#> 90  2021-09-08 2021-06-01  21.660294
#> 91  2021-09-08 2021-07-01  14.594744
#> 92  2021-09-08 2021-08-01  10.225824
#> 93  2021-09-08 2021-09-01  10.801322
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01   8.404271
#> 98  2021-11-12 2020-02-01   5.047600
#> 99  2021-11-12 2020-03-01   2.933514
#> 100 2021-11-12 2020-04-01   1.583270
#> 101 2021-11-12 2020-05-01   5.648314
#> 102 2021-11-12 2020-06-01   9.536080
#> 103 2021-11-12 2020-07-01   4.950173
#> 104 2021-11-12 2020-08-01   4.257142
#> 105 2021-11-12 2020-09-01   7.706517
#> 106 2021-11-12 2020-10-01  11.829537
#> 107 2021-11-12 2020-11-01  14.997505
#> 108 2021-11-12 2020-12-01  15.887274
#> 109 2021-11-12 2021-01-01  13.716198
#> 110 2021-11-12 2021-02-01  16.783795
#> 111 2021-11-12 2021-03-01  23.596869
#> 112 2021-11-12 2021-04-01  26.022864
#> 113 2021-11-12 2021-05-01  25.265739
#> 114 2021-11-12 2021-06-01  22.761836
#> 115 2021-11-12 2021-07-01  16.217043
#> 116 2021-11-12 2021-08-01  11.972160
#> 117 2021-11-12 2021-09-01  14.233312
#> 118 2021-11-12 2021-10-01  15.370180
#> 119 2021-11-12 2021-11-01  16.771732
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01   8.759485
#> 122 2021-11-17 2020-02-01   4.923871
#> 123 2021-11-17 2020-03-01   3.513111
#> 124 2021-11-17 2020-04-01   1.643075
#> 125 2021-11-17 2020-05-01   5.504331
#> 126 2021-11-17 2020-06-01   9.420164
#> 127 2021-11-17 2020-07-01   4.712517
#> 128 2021-11-17 2020-08-01   4.055988
#> 129 2021-11-17 2020-09-01   7.769780
#> 130 2021-11-17 2020-10-01  10.593984
#> 131 2021-11-17 2020-11-01  14.896742
#> 132 2021-11-17 2020-12-01  16.279253
#> 133 2021-11-17 2021-01-01  14.163592
#> 134 2021-11-17 2021-02-01  17.595880
#> 135 2021-11-17 2021-03-01  23.964030
#> 136 2021-11-17 2021-04-01  26.510554
#> 137 2021-11-17 2021-05-01  23.452575
#> 138 2021-11-17 2021-06-01  23.953000
#> 139 2021-11-17 2021-07-01  17.399769
#> 140 2021-11-17 2021-08-01  12.583430
#> 141 2021-11-17 2021-09-01  13.003761
#> 142 2021-11-17 2021-10-01  15.298287
#> 143 2021-11-17 2021-11-01  19.350359
#> 144 2021-11-17 2021-12-01         NA