Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-05-31 2020-01-01  4.2383271
#> 2   2020-05-31 2020-02-01 10.8386894
#> 3   2020-05-31 2020-03-01 13.6519095
#> 4   2020-05-31 2020-04-01 10.8775101
#> 5   2020-05-31 2020-05-01  7.2052777
#> 6   2020-05-31 2020-06-01         NA
#> 7   2020-05-31 2020-07-01         NA
#> 8   2020-05-31 2020-08-01         NA
#> 9   2020-05-31 2020-09-01         NA
#> 10  2020-05-31 2020-10-01         NA
#> 11  2020-05-31 2020-11-01         NA
#> 12  2020-05-31 2020-12-01         NA
#> 13  2020-05-31 2021-01-01         NA
#> 14  2020-05-31 2021-02-01         NA
#> 15  2020-05-31 2021-03-01         NA
#> 16  2020-05-31 2021-04-01         NA
#> 17  2020-05-31 2021-05-01         NA
#> 18  2020-05-31 2021-06-01         NA
#> 19  2020-05-31 2021-07-01         NA
#> 20  2020-05-31 2021-08-01         NA
#> 21  2020-05-31 2021-09-01         NA
#> 22  2020-05-31 2021-10-01         NA
#> 23  2020-05-31 2021-11-01         NA
#> 24  2020-05-31 2021-12-01         NA
#> 25  2020-07-30 2020-01-01  7.7409143
#> 26  2020-07-30 2020-02-01  6.6238234
#> 27  2020-07-30 2020-03-01  7.8124086
#> 28  2020-07-30 2020-04-01  9.3654739
#> 29  2020-07-30 2020-05-01  7.0112525
#> 30  2020-07-30 2020-06-01  8.4159424
#> 31  2020-07-30 2020-07-01  9.3306558
#> 32  2020-07-30 2020-08-01         NA
#> 33  2020-07-30 2020-09-01         NA
#> 34  2020-07-30 2020-10-01         NA
#> 35  2020-07-30 2020-11-01         NA
#> 36  2020-07-30 2020-12-01         NA
#> 37  2020-07-30 2021-01-01         NA
#> 38  2020-07-30 2021-02-01         NA
#> 39  2020-07-30 2021-03-01         NA
#> 40  2020-07-30 2021-04-01         NA
#> 41  2020-07-30 2021-05-01         NA
#> 42  2020-07-30 2021-06-01         NA
#> 43  2020-07-30 2021-07-01         NA
#> 44  2020-07-30 2021-08-01         NA
#> 45  2020-07-30 2021-09-01         NA
#> 46  2020-07-30 2021-10-01         NA
#> 47  2020-07-30 2021-11-01         NA
#> 48  2020-07-30 2021-12-01         NA
#> 49  2020-11-20 2020-01-01  7.5254037
#> 50  2020-11-20 2020-02-01  8.6222506
#> 51  2020-11-20 2020-03-01 12.3223473
#> 52  2020-11-20 2020-04-01 10.1710007
#> 53  2020-11-20 2020-05-01  7.6147926
#> 54  2020-11-20 2020-06-01 12.7054755
#> 55  2020-11-20 2020-07-01  6.1673365
#> 56  2020-11-20 2020-08-01  4.1705366
#> 57  2020-11-20 2020-09-01  3.1192849
#> 58  2020-11-20 2020-10-01  0.9930179
#> 59  2020-11-20 2020-11-01  2.4617791
#> 60  2020-11-20 2020-12-01         NA
#> 61  2020-11-20 2021-01-01         NA
#> 62  2020-11-20 2021-02-01         NA
#> 63  2020-11-20 2021-03-01         NA
#> 64  2020-11-20 2021-04-01         NA
#> 65  2020-11-20 2021-05-01         NA
#> 66  2020-11-20 2021-06-01         NA
#> 67  2020-11-20 2021-07-01         NA
#> 68  2020-11-20 2021-08-01         NA
#> 69  2020-11-20 2021-09-01         NA
#> 70  2020-11-20 2021-10-01         NA
#> 71  2020-11-20 2021-11-01         NA
#> 72  2020-11-20 2021-12-01         NA
#> 73  2020-12-12 2020-01-01  7.9239647
#> 74  2020-12-12 2020-02-01  8.3163652
#> 75  2020-12-12 2020-03-01 11.8171756
#> 76  2020-12-12 2020-04-01  9.2234087
#> 77  2020-12-12 2020-05-01  8.0419989
#> 78  2020-12-12 2020-06-01 10.5328559
#> 79  2020-12-12 2020-07-01  5.8375586
#> 80  2020-12-12 2020-08-01  2.3725173
#> 81  2020-12-12 2020-09-01  3.7559403
#> 82  2020-12-12 2020-10-01  1.8682344
#> 83  2020-12-12 2020-11-01  4.2349879
#> 84  2020-12-12 2020-12-01  8.2848433
#> 85  2020-12-12 2021-01-01         NA
#> 86  2020-12-12 2021-02-01         NA
#> 87  2020-12-12 2021-03-01         NA
#> 88  2020-12-12 2021-04-01         NA
#> 89  2020-12-12 2021-05-01         NA
#> 90  2020-12-12 2021-06-01         NA
#> 91  2020-12-12 2021-07-01         NA
#> 92  2020-12-12 2021-08-01         NA
#> 93  2020-12-12 2021-09-01         NA
#> 94  2020-12-12 2021-10-01         NA
#> 95  2020-12-12 2021-11-01         NA
#> 96  2020-12-12 2021-12-01         NA
#> 97  2021-04-26 2020-01-01  7.3531630
#> 98  2021-04-26 2020-02-01  8.4496950
#> 99  2021-04-26 2020-03-01 11.1909437
#> 100 2021-04-26 2020-04-01  9.7497798
#> 101 2021-04-26 2020-05-01  7.7497230
#> 102 2021-04-26 2020-06-01 11.1143840
#> 103 2021-04-26 2020-07-01  6.6263266
#> 104 2021-04-26 2020-08-01  1.8018461
#> 105 2021-04-26 2020-09-01  3.1007226
#> 106 2021-04-26 2020-10-01  1.1353376
#> 107 2021-04-26 2020-11-01  3.4005421
#> 108 2021-04-26 2020-12-01  9.5856316
#> 109 2021-04-26 2021-01-01  5.9006570
#> 110 2021-04-26 2021-02-01  3.7077541
#> 111 2021-04-26 2021-03-01  2.8568082
#> 112 2021-04-26 2021-04-01 -0.3836730
#> 113 2021-04-26 2021-05-01         NA
#> 114 2021-04-26 2021-06-01         NA
#> 115 2021-04-26 2021-07-01         NA
#> 116 2021-04-26 2021-08-01         NA
#> 117 2021-04-26 2021-09-01         NA
#> 118 2021-04-26 2021-10-01         NA
#> 119 2021-04-26 2021-11-01         NA
#> 120 2021-04-26 2021-12-01         NA
#> 121 2021-05-04 2020-01-01  7.5949298
#> 122 2021-05-04 2020-02-01  8.4394768
#> 123 2021-05-04 2020-03-01 11.4738346
#> 124 2021-05-04 2020-04-01  9.6141134
#> 125 2021-05-04 2020-05-01  7.7887332
#> 126 2021-05-04 2020-06-01 11.1295375
#> 127 2021-05-04 2020-07-01  6.2651987
#> 128 2021-05-04 2020-08-01  1.4500252
#> 129 2021-05-04 2020-09-01  3.4276040
#> 130 2021-05-04 2020-10-01  2.1248482
#> 131 2021-05-04 2020-11-01  3.6934589
#> 132 2021-05-04 2020-12-01  9.6615428
#> 133 2021-05-04 2021-01-01  8.1643124
#> 134 2021-05-04 2021-02-01  5.7304479
#> 135 2021-05-04 2021-03-01  2.3728692
#> 136 2021-05-04 2021-04-01  0.5435761
#> 137 2021-05-04 2021-05-01  1.2566396
#> 138 2021-05-04 2021-06-01         NA
#> 139 2021-05-04 2021-07-01         NA
#> 140 2021-05-04 2021-08-01         NA
#> 141 2021-05-04 2021-09-01         NA
#> 142 2021-05-04 2021-10-01         NA
#> 143 2021-05-04 2021-11-01         NA
#> 144 2021-05-04 2021-12-01         NA