Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01  4.49917327
#> 2   2020-02-17 2020-02-01  7.17769017
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01  3.78689463
#> 26  2020-06-05 2020-02-01  3.90480530
#> 27  2020-06-05 2020-03-01  7.61462546
#> 28  2020-06-05 2020-04-01  8.56789354
#> 29  2020-06-05 2020-05-01 10.04350965
#> 30  2020-06-05 2020-06-01  2.36982396
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01  3.78496489
#> 50  2021-04-17 2020-02-01  3.64494395
#> 51  2021-04-17 2020-03-01  8.27397720
#> 52  2021-04-17 2020-04-01  5.98784702
#> 53  2021-04-17 2020-05-01  7.35732338
#> 54  2021-04-17 2020-06-01  2.50482062
#> 55  2021-04-17 2020-07-01  1.14896879
#> 56  2021-04-17 2020-08-01  2.90400592
#> 57  2021-04-17 2020-09-01  4.14266107
#> 58  2021-04-17 2020-10-01  1.25381699
#> 59  2021-04-17 2020-11-01  7.70722638
#> 60  2021-04-17 2020-12-01  4.91649781
#> 61  2021-04-17 2021-01-01  0.81706175
#> 62  2021-04-17 2021-02-01 -1.55289471
#> 63  2021-04-17 2021-03-01  2.46522677
#> 64  2021-04-17 2021-04-01 -2.31585724
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01  3.40263135
#> 74  2021-09-08 2020-02-01  4.32212061
#> 75  2021-09-08 2020-03-01  9.04318293
#> 76  2021-09-08 2020-04-01  5.86721386
#> 77  2021-09-08 2020-05-01  6.78646198
#> 78  2021-09-08 2020-06-01  3.50024903
#> 79  2021-09-08 2020-07-01  2.39481778
#> 80  2021-09-08 2020-08-01  1.57861546
#> 81  2021-09-08 2020-09-01  0.56103387
#> 82  2021-09-08 2020-10-01  5.00072660
#> 83  2021-09-08 2020-11-01  7.14854378
#> 84  2021-09-08 2020-12-01  4.09466471
#> 85  2021-09-08 2021-01-01  2.30503867
#> 86  2021-09-08 2021-02-01 -2.43625920
#> 87  2021-09-08 2021-03-01  1.32649873
#> 88  2021-09-08 2021-04-01 -2.46928852
#> 89  2021-09-08 2021-05-01 -0.67281200
#> 90  2021-09-08 2021-06-01 -2.23546234
#> 91  2021-09-08 2021-07-01 -0.14320556
#> 92  2021-09-08 2021-08-01 -3.13339165
#> 93  2021-09-08 2021-09-01 -7.67988456
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01  3.34995748
#> 98  2021-11-12 2020-02-01  4.36401729
#> 99  2021-11-12 2020-03-01  7.92377548
#> 100 2021-11-12 2020-04-01  6.15603104
#> 101 2021-11-12 2020-05-01  7.45539787
#> 102 2021-11-12 2020-06-01  3.65447351
#> 103 2021-11-12 2020-07-01  3.65590767
#> 104 2021-11-12 2020-08-01  2.15911135
#> 105 2021-11-12 2020-09-01  2.23142169
#> 106 2021-11-12 2020-10-01  3.64512553
#> 107 2021-11-12 2020-11-01  7.04356340
#> 108 2021-11-12 2020-12-01  6.03752377
#> 109 2021-11-12 2021-01-01  1.58828287
#> 110 2021-11-12 2021-02-01 -3.02669496
#> 111 2021-11-12 2021-03-01  2.26717845
#> 112 2021-11-12 2021-04-01 -1.14675344
#> 113 2021-11-12 2021-05-01  0.66334237
#> 114 2021-11-12 2021-06-01 -1.13392041
#> 115 2021-11-12 2021-07-01  1.47909381
#> 116 2021-11-12 2021-08-01 -1.38705540
#> 117 2021-11-12 2021-09-01 -4.24789438
#> 118 2021-11-12 2021-10-01 -4.87377165
#> 119 2021-11-12 2021-11-01 -6.98458314
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01  3.70517180
#> 122 2021-11-17 2020-02-01  4.24028765
#> 123 2021-11-17 2020-03-01  8.50337235
#> 124 2021-11-17 2020-04-01  6.21583643
#> 125 2021-11-17 2020-05-01  7.31141517
#> 126 2021-11-17 2020-06-01  3.53855743
#> 127 2021-11-17 2020-07-01  3.41825121
#> 128 2021-11-17 2020-08-01  1.95795737
#> 129 2021-11-17 2020-09-01  2.29468482
#> 130 2021-11-17 2020-10-01  2.40957276
#> 131 2021-11-17 2020-11-01  6.94280099
#> 132 2021-11-17 2020-12-01  6.42950195
#> 133 2021-11-17 2021-01-01  2.03567602
#> 134 2021-11-17 2021-02-01 -2.21461015
#> 135 2021-11-17 2021-03-01  2.63433930
#> 136 2021-11-17 2021-04-01 -0.65906334
#> 137 2021-11-17 2021-05-01 -1.14982195
#> 138 2021-11-17 2021-06-01  0.05724375
#> 139 2021-11-17 2021-07-01  2.66181957
#> 140 2021-11-17 2021-08-01 -0.77578531
#> 141 2021-11-17 2021-09-01 -5.47744556
#> 142 2021-11-17 2021-10-01 -4.94566465
#> 143 2021-11-17 2021-11-01 -4.40595665
#> 144 2021-11-17 2021-12-01          NA