Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01  12.3845386
#> 2   2020-02-17 2020-02-01  12.1037372
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01  11.6722600
#> 26  2020-06-05 2020-02-01   8.8308524
#> 27  2020-06-05 2020-03-01   4.6019075
#> 28  2020-06-05 2020-04-01   0.5619707
#> 29  2020-06-05 2020-05-01  -0.1832351
#> 30  2020-06-05 2020-06-01  -6.0749975
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01  11.6703303
#> 50  2021-04-17 2020-02-01   8.5709910
#> 51  2021-04-17 2020-03-01   5.2612593
#> 52  2021-04-17 2020-04-01  -2.0180758
#> 53  2021-04-17 2020-05-01  -2.8694214
#> 54  2021-04-17 2020-06-01  -5.9400009
#> 55  2021-04-17 2020-07-01  -7.2552569
#> 56  2021-04-17 2020-08-01  -5.5601321
#> 57  2021-04-17 2020-09-01  -7.9084555
#> 58  2021-04-17 2020-10-01  -8.7977605
#> 59  2021-04-17 2020-11-01 -13.4585992
#> 60  2021-04-17 2020-12-01 -13.4788509
#> 61  2021-04-17 2021-01-01  -7.9521814
#> 62  2021-04-17 2021-02-01  -7.1300898
#> 63  2021-04-17 2021-03-01 -13.0869715
#> 64  2021-04-17 2021-04-01 -22.6221481
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01  11.2879967
#> 74  2021-09-08 2020-02-01   9.2481677
#> 75  2021-09-08 2020-03-01   6.0304650
#> 76  2021-09-08 2020-04-01  -2.1387090
#> 77  2021-09-08 2020-05-01  -3.4402828
#> 78  2021-09-08 2020-06-01  -4.9445725
#> 79  2021-09-08 2020-07-01  -6.0094079
#> 80  2021-09-08 2020-08-01  -6.8855226
#> 81  2021-09-08 2020-09-01 -11.4900827
#> 82  2021-09-08 2020-10-01  -5.0508509
#> 83  2021-09-08 2020-11-01 -14.0172818
#> 84  2021-09-08 2020-12-01 -14.3006840
#> 85  2021-09-08 2021-01-01  -6.4642045
#> 86  2021-09-08 2021-02-01  -8.0134543
#> 87  2021-09-08 2021-03-01 -14.2256995
#> 88  2021-09-08 2021-04-01 -22.7755794
#> 89  2021-09-08 2021-05-01 -25.9314431
#> 90  2021-09-08 2021-06-01 -36.4595050
#> 91  2021-09-08 2021-07-01 -44.0037745
#> 92  2021-09-08 2021-08-01 -47.7990485
#> 93  2021-09-08 2021-09-01 -54.5792234
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01  11.2353228
#> 98  2021-11-12 2020-02-01   9.2900644
#> 99  2021-11-12 2020-03-01   4.9110575
#> 100 2021-11-12 2020-04-01  -1.8498918
#> 101 2021-11-12 2020-05-01  -2.7713469
#> 102 2021-11-12 2020-06-01  -4.7903480
#> 103 2021-11-12 2020-07-01  -4.7483181
#> 104 2021-11-12 2020-08-01  -6.3050267
#> 105 2021-11-12 2020-09-01  -9.8196949
#> 106 2021-11-12 2020-10-01  -6.4064519
#> 107 2021-11-12 2020-11-01 -14.1222622
#> 108 2021-11-12 2020-12-01 -12.3578250
#> 109 2021-11-12 2021-01-01  -7.1809603
#> 110 2021-11-12 2021-02-01  -8.6038901
#> 111 2021-11-12 2021-03-01 -13.2850198
#> 112 2021-11-12 2021-04-01 -21.4530443
#> 113 2021-11-12 2021-05-01 -24.5952888
#> 114 2021-11-12 2021-06-01 -35.3579631
#> 115 2021-11-12 2021-07-01 -42.3814751
#> 116 2021-11-12 2021-08-01 -46.0527123
#> 117 2021-11-12 2021-09-01 -51.1472332
#> 118 2021-11-12 2021-10-01 -51.4281600
#> 119 2021-11-12 2021-11-01 -61.0088083
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01  11.5905372
#> 122 2021-11-17 2020-02-01   9.1663347
#> 123 2021-11-17 2020-03-01   5.4906544
#> 124 2021-11-17 2020-04-01  -1.7900864
#> 125 2021-11-17 2020-05-01  -2.9153296
#> 126 2021-11-17 2020-06-01  -4.9062641
#> 127 2021-11-17 2020-07-01  -4.9859745
#> 128 2021-11-17 2020-08-01  -6.5061806
#> 129 2021-11-17 2020-09-01  -9.7564317
#> 130 2021-11-17 2020-10-01  -7.6420047
#> 131 2021-11-17 2020-11-01 -14.2230246
#> 132 2021-11-17 2020-12-01 -11.9658468
#> 133 2021-11-17 2021-01-01  -6.7335672
#> 134 2021-11-17 2021-02-01  -7.7918052
#> 135 2021-11-17 2021-03-01 -12.9178589
#> 136 2021-11-17 2021-04-01 -20.9653542
#> 137 2021-11-17 2021-05-01 -26.4084531
#> 138 2021-11-17 2021-06-01 -34.1667989
#> 139 2021-11-17 2021-07-01 -41.1987493
#> 140 2021-11-17 2021-08-01 -45.4414422
#> 141 2021-11-17 2021-09-01 -52.3767844
#> 142 2021-11-17 2021-10-01 -51.5000530
#> 143 2021-11-17 2021-11-01 -58.4301818
#> 144 2021-11-17 2021-12-01          NA