Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01   6.862841
#> 2   2020-02-17 2020-02-01   7.232582
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01   6.150562
#> 26  2020-06-05 2020-02-01   3.959697
#> 27  2020-06-05 2020-03-01   5.857064
#> 28  2020-06-05 2020-04-01  13.898452
#> 29  2020-06-05 2020-05-01  15.898313
#> 30  2020-06-05 2020-06-01  10.573273
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01   6.148632
#> 50  2021-04-17 2020-02-01   3.699836
#> 51  2021-04-17 2020-03-01   6.516416
#> 52  2021-04-17 2020-04-01  11.318406
#> 53  2021-04-17 2020-05-01  13.212126
#> 54  2021-04-17 2020-06-01  10.708270
#> 55  2021-04-17 2020-07-01   8.864465
#> 56  2021-04-17 2020-08-01  12.932596
#> 57  2021-04-17 2020-09-01  13.677261
#> 58  2021-04-17 2020-10-01  10.706132
#> 59  2021-04-17 2020-11-01  17.094498
#> 60  2021-04-17 2020-12-01  10.521789
#> 61  2021-04-17 2021-01-01   5.677408
#> 62  2021-04-17 2021-02-01   7.372709
#> 63  2021-04-17 2021-03-01   5.000881
#> 64  2021-04-17 2021-04-01   6.443426
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01   5.766299
#> 74  2021-09-08 2020-02-01   4.377013
#> 75  2021-09-08 2020-03-01   7.285621
#> 76  2021-09-08 2020-04-01  11.197773
#> 77  2021-09-08 2020-05-01  12.641265
#> 78  2021-09-08 2020-06-01  11.703698
#> 79  2021-09-08 2020-07-01  10.110314
#> 80  2021-09-08 2020-08-01  11.607205
#> 81  2021-09-08 2020-09-01  10.095634
#> 82  2021-09-08 2020-10-01  14.453041
#> 83  2021-09-08 2020-11-01  16.535815
#> 84  2021-09-08 2020-12-01   9.699956
#> 85  2021-09-08 2021-01-01   7.165385
#> 86  2021-09-08 2021-02-01   6.489345
#> 87  2021-09-08 2021-03-01   3.862153
#> 88  2021-09-08 2021-04-01   6.289995
#> 89  2021-09-08 2021-05-01  12.485545
#> 90  2021-09-08 2021-06-01  11.491802
#> 91  2021-09-08 2021-07-01  11.065087
#> 92  2021-09-08 2021-08-01  13.192223
#> 93  2021-09-08 2021-09-01  11.197563
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01   5.713625
#> 98  2021-11-12 2020-02-01   4.418909
#> 99  2021-11-12 2020-03-01   6.166214
#> 100 2021-11-12 2020-04-01  11.486590
#> 101 2021-11-12 2020-05-01  13.310201
#> 102 2021-11-12 2020-06-01  11.857923
#> 103 2021-11-12 2020-07-01  11.371404
#> 104 2021-11-12 2020-08-01  12.187701
#> 105 2021-11-12 2020-09-01  11.766021
#> 106 2021-11-12 2020-10-01  13.097440
#> 107 2021-11-12 2020-11-01  16.430835
#> 108 2021-11-12 2020-12-01  11.642815
#> 109 2021-11-12 2021-01-01   6.448629
#> 110 2021-11-12 2021-02-01   5.898909
#> 111 2021-11-12 2021-03-01   4.802832
#> 112 2021-11-12 2021-04-01   7.612530
#> 113 2021-11-12 2021-05-01  13.821700
#> 114 2021-11-12 2021-06-01  12.593344
#> 115 2021-11-12 2021-07-01  12.687386
#> 116 2021-11-12 2021-08-01  14.938559
#> 117 2021-11-12 2021-09-01  14.629553
#> 118 2021-11-12 2021-10-01  11.716778
#> 119 2021-11-12 2021-11-01   9.817066
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01   6.068839
#> 122 2021-11-17 2020-02-01   4.295180
#> 123 2021-11-17 2020-03-01   6.745811
#> 124 2021-11-17 2020-04-01  11.546395
#> 125 2021-11-17 2020-05-01  13.166218
#> 126 2021-11-17 2020-06-01  11.742007
#> 127 2021-11-17 2020-07-01  11.133747
#> 128 2021-11-17 2020-08-01  11.986547
#> 129 2021-11-17 2020-09-01  11.829285
#> 130 2021-11-17 2020-10-01  11.861887
#> 131 2021-11-17 2020-11-01  16.330072
#> 132 2021-11-17 2020-12-01  12.034793
#> 133 2021-11-17 2021-01-01   6.896023
#> 134 2021-11-17 2021-02-01   6.710994
#> 135 2021-11-17 2021-03-01   5.169993
#> 136 2021-11-17 2021-04-01   8.100220
#> 137 2021-11-17 2021-05-01  12.008536
#> 138 2021-11-17 2021-06-01  13.784508
#> 139 2021-11-17 2021-07-01  13.870112
#> 140 2021-11-17 2021-08-01  15.549829
#> 141 2021-11-17 2021-09-01  13.400002
#> 142 2021-11-17 2021-10-01  11.644885
#> 143 2021-11-17 2021-11-01  12.395692
#> 144 2021-11-17 2021-12-01         NA