Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01   9.163820
#> 2   2020-02-17 2020-02-01  10.506109
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01   8.451541
#> 26  2020-06-05 2020-02-01   7.233224
#> 27  2020-06-05 2020-03-01   1.690119
#> 28  2020-06-05 2020-04-01   7.267155
#> 29  2020-06-05 2020-05-01   9.735265
#> 30  2020-06-05 2020-06-01   6.568477
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01   8.449611
#> 50  2021-04-17 2020-02-01   6.973363
#> 51  2021-04-17 2020-03-01   2.349470
#> 52  2021-04-17 2020-04-01   4.687108
#> 53  2021-04-17 2020-05-01   7.049079
#> 54  2021-04-17 2020-06-01   6.703474
#> 55  2021-04-17 2020-07-01   5.634270
#> 56  2021-04-17 2020-08-01  10.616708
#> 57  2021-04-17 2020-09-01  11.837732
#> 58  2021-04-17 2020-10-01  10.026700
#> 59  2021-04-17 2020-11-01  12.823029
#> 60  2021-04-17 2020-12-01  12.286744
#> 61  2021-04-17 2021-01-01  10.130593
#> 62  2021-04-17 2021-02-01  10.608403
#> 63  2021-04-17 2021-03-01   4.637633
#> 64  2021-04-17 2021-04-01   8.362972
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01   8.067278
#> 74  2021-09-08 2020-02-01   7.650540
#> 75  2021-09-08 2020-03-01   3.118676
#> 76  2021-09-08 2020-04-01   4.566475
#> 77  2021-09-08 2020-05-01   6.478218
#> 78  2021-09-08 2020-06-01   7.698903
#> 79  2021-09-08 2020-07-01   6.880119
#> 80  2021-09-08 2020-08-01   9.291317
#> 81  2021-09-08 2020-09-01   8.256105
#> 82  2021-09-08 2020-10-01  13.773610
#> 83  2021-09-08 2020-11-01  12.264346
#> 84  2021-09-08 2020-12-01  11.464911
#> 85  2021-09-08 2021-01-01  11.618570
#> 86  2021-09-08 2021-02-01   9.725038
#> 87  2021-09-08 2021-03-01   3.498905
#> 88  2021-09-08 2021-04-01   8.209541
#> 89  2021-09-08 2021-05-01  18.060039
#> 90  2021-09-08 2021-06-01  23.260342
#> 91  2021-09-08 2021-07-01  28.168241
#> 92  2021-09-08 2021-08-01  32.267640
#> 93  2021-09-08 2021-09-01  32.475053
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01   8.014604
#> 98  2021-11-12 2020-02-01   7.692436
#> 99  2021-11-12 2020-03-01   1.999269
#> 100 2021-11-12 2020-04-01   4.855292
#> 101 2021-11-12 2020-05-01   7.147153
#> 102 2021-11-12 2020-06-01   7.853127
#> 103 2021-11-12 2020-07-01   8.141209
#> 104 2021-11-12 2020-08-01   9.871813
#> 105 2021-11-12 2020-09-01   9.926493
#> 106 2021-11-12 2020-10-01  12.418008
#> 107 2021-11-12 2020-11-01  12.159366
#> 108 2021-11-12 2020-12-01  13.407770
#> 109 2021-11-12 2021-01-01  10.901814
#> 110 2021-11-12 2021-02-01   9.134603
#> 111 2021-11-12 2021-03-01   4.439585
#> 112 2021-11-12 2021-04-01   9.532076
#> 113 2021-11-12 2021-05-01  19.396194
#> 114 2021-11-12 2021-06-01  24.361884
#> 115 2021-11-12 2021-07-01  29.790540
#> 116 2021-11-12 2021-08-01  34.013976
#> 117 2021-11-12 2021-09-01  35.907044
#> 118 2021-11-12 2021-10-01  35.932451
#> 119 2021-11-12 2021-11-01  35.517004
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01   8.369818
#> 122 2021-11-17 2020-02-01   7.568707
#> 123 2021-11-17 2020-03-01   2.578866
#> 124 2021-11-17 2020-04-01   4.915097
#> 125 2021-11-17 2020-05-01   7.003171
#> 126 2021-11-17 2020-06-01   7.737211
#> 127 2021-11-17 2020-07-01   7.903552
#> 128 2021-11-17 2020-08-01   9.670659
#> 129 2021-11-17 2020-09-01   9.989756
#> 130 2021-11-17 2020-10-01  11.182456
#> 131 2021-11-17 2020-11-01  12.058603
#> 132 2021-11-17 2020-12-01  13.799748
#> 133 2021-11-17 2021-01-01  11.349208
#> 134 2021-11-17 2021-02-01   9.946687
#> 135 2021-11-17 2021-03-01   4.806746
#> 136 2021-11-17 2021-04-01  10.019766
#> 137 2021-11-17 2021-05-01  17.583030
#> 138 2021-11-17 2021-06-01  25.553048
#> 139 2021-11-17 2021-07-01  30.973266
#> 140 2021-11-17 2021-08-01  34.625246
#> 141 2021-11-17 2021-09-01  34.677492
#> 142 2021-11-17 2021-10-01  35.860558
#> 143 2021-11-17 2021-11-01  38.095631
#> 144 2021-11-17 2021-12-01         NA