Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01   9.9973697
#> 2   2020-02-17 2020-02-01  11.6072333
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01   9.2850911
#> 26  2020-06-05 2020-02-01   8.3343485
#> 27  2020-06-05 2020-03-01   7.7157202
#> 28  2020-06-05 2020-04-01  12.6309418
#> 29  2020-06-05 2020-05-01  15.5207096
#> 30  2020-06-05 2020-06-01   8.3154168
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01   9.2831613
#> 50  2021-04-17 2020-02-01   8.0744871
#> 51  2021-04-17 2020-03-01   8.3750719
#> 52  2021-04-17 2020-04-01  10.0508953
#> 53  2021-04-17 2020-05-01  12.8345233
#> 54  2021-04-17 2020-06-01   8.4504135
#> 55  2021-04-17 2020-07-01   8.9081701
#> 56  2021-04-17 2020-08-01   8.6086810
#> 57  2021-04-17 2020-09-01   6.0023599
#> 58  2021-04-17 2020-10-01   4.1827292
#> 59  2021-04-17 2020-11-01   3.1620859
#> 60  2021-04-17 2020-12-01   1.5304193
#> 61  2021-04-17 2021-01-01  -2.7484799
#> 62  2021-04-17 2021-02-01  -3.7411174
#> 63  2021-04-17 2021-03-01  -9.6642410
#> 64  2021-04-17 2021-04-01 -11.1080989
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01   8.9008278
#> 74  2021-09-08 2020-02-01   8.7516638
#> 75  2021-09-08 2020-03-01   9.1442776
#> 76  2021-09-08 2020-04-01   9.9302622
#> 77  2021-09-08 2020-05-01  12.2636619
#> 78  2021-09-08 2020-06-01   9.4458419
#> 79  2021-09-08 2020-07-01  10.1540191
#> 80  2021-09-08 2020-08-01   7.2832905
#> 81  2021-09-08 2020-09-01   2.4207327
#> 82  2021-09-08 2020-10-01   7.9296388
#> 83  2021-09-08 2020-11-01   2.6034033
#> 84  2021-09-08 2020-12-01   0.7085862
#> 85  2021-09-08 2021-01-01  -1.2605029
#> 86  2021-09-08 2021-02-01  -4.6244819
#> 87  2021-09-08 2021-03-01 -10.8029690
#> 88  2021-09-08 2021-04-01 -11.2615302
#> 89  2021-09-08 2021-05-01  -8.9408758
#> 90  2021-09-08 2021-06-01  -8.5171934
#> 91  2021-09-08 2021-07-01  -7.4670329
#> 92  2021-09-08 2021-08-01 -14.3730387
#> 93  2021-09-08 2021-09-01 -16.0434302
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01   8.8481539
#> 98  2021-11-12 2020-02-01   8.7935604
#> 99  2021-11-12 2020-03-01   8.0248702
#> 100 2021-11-12 2020-04-01  10.2190794
#> 101 2021-11-12 2020-05-01  12.9325978
#> 102 2021-11-12 2020-06-01   9.6000664
#> 103 2021-11-12 2020-07-01  11.4151090
#> 104 2021-11-12 2020-08-01   7.8637864
#> 105 2021-11-12 2020-09-01   4.0911205
#> 106 2021-11-12 2020-10-01   6.5740378
#> 107 2021-11-12 2020-11-01   2.4984229
#> 108 2021-11-12 2020-12-01   2.6514452
#> 109 2021-11-12 2021-01-01  -1.9772587
#> 110 2021-11-12 2021-02-01  -5.2149176
#> 111 2021-11-12 2021-03-01  -9.8622893
#> 112 2021-11-12 2021-04-01  -9.9389951
#> 113 2021-11-12 2021-05-01  -7.6047214
#> 114 2021-11-12 2021-06-01  -7.4156514
#> 115 2021-11-12 2021-07-01  -5.8447335
#> 116 2021-11-12 2021-08-01 -12.6267024
#> 117 2021-11-12 2021-09-01 -12.6114400
#> 118 2021-11-12 2021-10-01 -14.2229800
#> 119 2021-11-12 2021-11-01 -22.0878489
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01   9.2033683
#> 122 2021-11-17 2020-02-01   8.6698308
#> 123 2021-11-17 2020-03-01   8.6044670
#> 124 2021-11-17 2020-04-01  10.2788847
#> 125 2021-11-17 2020-05-01  12.7886151
#> 126 2021-11-17 2020-06-01   9.4841503
#> 127 2021-11-17 2020-07-01  11.1774525
#> 128 2021-11-17 2020-08-01   7.6626324
#> 129 2021-11-17 2020-09-01   4.1543836
#> 130 2021-11-17 2020-10-01   5.3384850
#> 131 2021-11-17 2020-11-01   2.3976605
#> 132 2021-11-17 2020-12-01   3.0434234
#> 133 2021-11-17 2021-01-01  -1.5298656
#> 134 2021-11-17 2021-02-01  -4.4028328
#> 135 2021-11-17 2021-03-01  -9.4951285
#> 136 2021-11-17 2021-04-01  -9.4513050
#> 137 2021-11-17 2021-05-01  -9.4178857
#> 138 2021-11-17 2021-06-01  -6.2244873
#> 139 2021-11-17 2021-07-01  -4.6620078
#> 140 2021-11-17 2021-08-01 -12.0154323
#> 141 2021-11-17 2021-09-01 -13.8409912
#> 142 2021-11-17 2021-10-01 -14.2948730
#> 143 2021-11-17 2021-11-01 -19.5092224
#> 144 2021-11-17 2021-12-01          NA