Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01   12.45330
#> 2   2020-02-17 2020-02-01   15.74365
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01   11.74102
#> 26  2020-06-05 2020-02-01   12.47077
#> 27  2020-06-05 2020-03-01   14.48243
#> 28  2020-06-05 2020-04-01   16.70133
#> 29  2020-06-05 2020-05-01   13.92960
#> 30  2020-06-05 2020-06-01   12.36019
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01   11.73909
#> 50  2021-04-17 2020-02-01   12.21090
#> 51  2021-04-17 2020-03-01   15.14178
#> 52  2021-04-17 2020-04-01   14.12129
#> 53  2021-04-17 2020-05-01   11.24342
#> 54  2021-04-17 2020-06-01   12.49519
#> 55  2021-04-17 2020-07-01   14.91939
#> 56  2021-04-17 2020-08-01   17.55712
#> 57  2021-04-17 2020-09-01   19.84301
#> 58  2021-04-17 2020-10-01   17.89259
#> 59  2021-04-17 2020-11-01   23.98704
#> 60  2021-04-17 2020-12-01   21.33934
#> 61  2021-04-17 2021-01-01   24.49175
#> 62  2021-04-17 2021-02-01   28.74801
#> 63  2021-04-17 2021-03-01   28.17227
#> 64  2021-04-17 2021-04-01   24.58750
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01   11.35676
#> 74  2021-09-08 2020-02-01   12.88808
#> 75  2021-09-08 2020-03-01   15.91099
#> 76  2021-09-08 2020-04-01   14.00065
#> 77  2021-09-08 2020-05-01   10.67256
#> 78  2021-09-08 2020-06-01   13.49062
#> 79  2021-09-08 2020-07-01   16.16523
#> 80  2021-09-08 2020-08-01   16.23173
#> 81  2021-09-08 2020-09-01   16.26138
#> 82  2021-09-08 2020-10-01   21.63950
#> 83  2021-09-08 2020-11-01   23.42836
#> 84  2021-09-08 2020-12-01   20.51751
#> 85  2021-09-08 2021-01-01   25.97973
#> 86  2021-09-08 2021-02-01   27.86465
#> 87  2021-09-08 2021-03-01   27.03355
#> 88  2021-09-08 2021-04-01   24.43407
#> 89  2021-09-08 2021-05-01   27.80019
#> 90  2021-09-08 2021-06-01   32.46042
#> 91  2021-09-08 2021-07-01   33.11125
#> 92  2021-09-08 2021-08-01   31.30227
#> 93  2021-09-08 2021-09-01   30.13807
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01   11.30408
#> 98  2021-11-12 2020-02-01   12.92998
#> 99  2021-11-12 2020-03-01   14.79158
#> 100 2021-11-12 2020-04-01   14.28947
#> 101 2021-11-12 2020-05-01   11.34149
#> 102 2021-11-12 2020-06-01   13.64484
#> 103 2021-11-12 2020-07-01   17.42632
#> 104 2021-11-12 2020-08-01   16.81222
#> 105 2021-11-12 2020-09-01   17.93177
#> 106 2021-11-12 2020-10-01   20.28390
#> 107 2021-11-12 2020-11-01   23.32338
#> 108 2021-11-12 2020-12-01   22.46036
#> 109 2021-11-12 2021-01-01   25.26297
#> 110 2021-11-12 2021-02-01   27.27421
#> 111 2021-11-12 2021-03-01   27.97423
#> 112 2021-11-12 2021-04-01   25.75660
#> 113 2021-11-12 2021-05-01   29.13634
#> 114 2021-11-12 2021-06-01   33.56196
#> 115 2021-11-12 2021-07-01   34.73355
#> 116 2021-11-12 2021-08-01   33.04860
#> 117 2021-11-12 2021-09-01   33.57006
#> 118 2021-11-12 2021-10-01   29.43633
#> 119 2021-11-12 2021-11-01   36.08496
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01   11.65930
#> 122 2021-11-17 2020-02-01   12.80625
#> 123 2021-11-17 2020-03-01   15.37118
#> 124 2021-11-17 2020-04-01   14.34928
#> 125 2021-11-17 2020-05-01   11.19751
#> 126 2021-11-17 2020-06-01   13.52892
#> 127 2021-11-17 2020-07-01   17.18867
#> 128 2021-11-17 2020-08-01   16.61107
#> 129 2021-11-17 2020-09-01   17.99503
#> 130 2021-11-17 2020-10-01   19.04835
#> 131 2021-11-17 2020-11-01   23.22262
#> 132 2021-11-17 2020-12-01   22.85234
#> 133 2021-11-17 2021-01-01   25.71036
#> 134 2021-11-17 2021-02-01   28.08630
#> 135 2021-11-17 2021-03-01   28.34139
#> 136 2021-11-17 2021-04-01   26.24429
#> 137 2021-11-17 2021-05-01   27.32318
#> 138 2021-11-17 2021-06-01   34.75313
#> 139 2021-11-17 2021-07-01   35.91628
#> 140 2021-11-17 2021-08-01   33.65987
#> 141 2021-11-17 2021-09-01   32.34051
#> 142 2021-11-17 2021-10-01   29.36444
#> 143 2021-11-17 2021-11-01   38.66359
#> 144 2021-11-17 2021-12-01         NA