Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01  7.47659173
#> 2   2020-02-17 2020-02-01  6.36242371
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01  6.76431309
#> 26  2020-06-05 2020-02-01  3.08953885
#> 27  2020-06-05 2020-03-01  4.79687566
#> 28  2020-06-05 2020-04-01  6.46202240
#> 29  2020-06-05 2020-05-01  6.08185003
#> 30  2020-06-05 2020-06-01  2.01593509
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01  6.76238335
#> 50  2021-04-17 2020-02-01  2.82967750
#> 51  2021-04-17 2020-03-01  5.45622740
#> 52  2021-04-17 2020-04-01  3.88197588
#> 53  2021-04-17 2020-05-01  3.39566376
#> 54  2021-04-17 2020-06-01  2.15093176
#> 55  2021-04-17 2020-07-01  1.17662745
#> 56  2021-04-17 2020-08-01  1.04811223
#> 57  2021-04-17 2020-09-01  4.39933204
#> 58  2021-04-17 2020-10-01  0.02330306
#> 59  2021-04-17 2020-11-01  3.31030439
#> 60  2021-04-17 2020-12-01  2.09822172
#> 61  2021-04-17 2021-01-01  0.34202799
#> 62  2021-04-17 2021-02-01  3.70759637
#> 63  2021-04-17 2021-03-01  7.69454601
#> 64  2021-04-17 2021-04-01  8.84001314
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01  6.38004981
#> 74  2021-09-08 2020-02-01  3.50685416
#> 75  2021-09-08 2020-03-01  6.22543313
#> 76  2021-09-08 2020-04-01  3.76134272
#> 77  2021-09-08 2020-05-01  2.82480236
#> 78  2021-09-08 2020-06-01  3.14636016
#> 79  2021-09-08 2020-07-01  2.42247644
#> 80  2021-09-08 2020-08-01 -0.27727824
#> 81  2021-09-08 2020-09-01  0.81770484
#> 82  2021-09-08 2020-10-01  3.77021267
#> 83  2021-09-08 2020-11-01  2.75162179
#> 84  2021-09-08 2020-12-01  1.27638862
#> 85  2021-09-08 2021-01-01  1.83000491
#> 86  2021-09-08 2021-02-01  2.82423188
#> 87  2021-09-08 2021-03-01  6.55581797
#> 88  2021-09-08 2021-04-01  8.68658186
#> 89  2021-09-08 2021-05-01  5.68034650
#> 90  2021-09-08 2021-06-01 -3.34278163
#> 91  2021-09-08 2021-07-01 -6.69318892
#> 92  2021-09-08 2021-08-01 -7.92472629
#> 93  2021-09-08 2021-09-01 -2.53583441
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01  6.32737594
#> 98  2021-11-12 2020-02-01  3.54875084
#> 99  2021-11-12 2020-03-01  5.10602568
#> 100 2021-11-12 2020-04-01  4.05015991
#> 101 2021-11-12 2020-05-01  3.49373825
#> 102 2021-11-12 2020-06-01  3.30058465
#> 103 2021-11-12 2020-07-01  3.68356632
#> 104 2021-11-12 2020-08-01  0.30321766
#> 105 2021-11-12 2020-09-01  2.48809266
#> 106 2021-11-12 2020-10-01  2.41461160
#> 107 2021-11-12 2020-11-01  2.64664141
#> 108 2021-11-12 2020-12-01  3.21924768
#> 109 2021-11-12 2021-01-01  1.11324911
#> 110 2021-11-12 2021-02-01  2.23379612
#> 111 2021-11-12 2021-03-01  7.49649769
#> 112 2021-11-12 2021-04-01 10.00911694
#> 113 2021-11-12 2021-05-01  7.01650086
#> 114 2021-11-12 2021-06-01 -2.24123969
#> 115 2021-11-12 2021-07-01 -5.07088955
#> 116 2021-11-12 2021-08-01 -6.17839004
#> 117 2021-11-12 2021-09-01  0.89615577
#> 118 2021-11-12 2021-10-01  0.11699018
#> 119 2021-11-12 2021-11-01 -1.41305000
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01  6.68259026
#> 122 2021-11-17 2020-02-01  3.42502120
#> 123 2021-11-17 2020-03-01  5.68562255
#> 124 2021-11-17 2020-04-01  4.10996530
#> 125 2021-11-17 2020-05-01  3.34975555
#> 126 2021-11-17 2020-06-01  3.18466857
#> 127 2021-11-17 2020-07-01  3.44590987
#> 128 2021-11-17 2020-08-01  0.10206368
#> 129 2021-11-17 2020-09-01  2.55135579
#> 130 2021-11-17 2020-10-01  1.17905883
#> 131 2021-11-17 2020-11-01  2.54587900
#> 132 2021-11-17 2020-12-01  3.61122586
#> 133 2021-11-17 2021-01-01  1.56064226
#> 134 2021-11-17 2021-02-01  3.04588093
#> 135 2021-11-17 2021-03-01  7.86365853
#> 136 2021-11-17 2021-04-01 10.49680704
#> 137 2021-11-17 2021-05-01  5.20333654
#> 138 2021-11-17 2021-06-01 -1.05007553
#> 139 2021-11-17 2021-07-01 -3.88816379
#> 140 2021-11-17 2021-08-01 -5.56711995
#> 141 2021-11-17 2021-09-01 -0.33339541
#> 142 2021-11-17 2021-10-01  0.04509719
#> 143 2021-11-17 2021-11-01  1.16557650
#> 144 2021-11-17 2021-12-01          NA