Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-05-31 2020-01-01  6.2792970
#> 2   2020-05-31 2020-02-01 10.3384917
#> 3   2020-05-31 2020-03-01  6.5557796
#> 4   2020-05-31 2020-04-01  5.4005181
#> 5   2020-05-31 2020-05-01  3.2595717
#> 6   2020-05-31 2020-06-01         NA
#> 7   2020-05-31 2020-07-01         NA
#> 8   2020-05-31 2020-08-01         NA
#> 9   2020-05-31 2020-09-01         NA
#> 10  2020-05-31 2020-10-01         NA
#> 11  2020-05-31 2020-11-01         NA
#> 12  2020-05-31 2020-12-01         NA
#> 13  2020-05-31 2021-01-01         NA
#> 14  2020-05-31 2021-02-01         NA
#> 15  2020-05-31 2021-03-01         NA
#> 16  2020-05-31 2021-04-01         NA
#> 17  2020-05-31 2021-05-01         NA
#> 18  2020-05-31 2021-06-01         NA
#> 19  2020-05-31 2021-07-01         NA
#> 20  2020-05-31 2021-08-01         NA
#> 21  2020-05-31 2021-09-01         NA
#> 22  2020-05-31 2021-10-01         NA
#> 23  2020-05-31 2021-11-01         NA
#> 24  2020-05-31 2021-12-01         NA
#> 25  2020-07-30 2020-01-01  9.7818842
#> 26  2020-07-30 2020-02-01  6.1236257
#> 27  2020-07-30 2020-03-01  0.7162787
#> 28  2020-07-30 2020-04-01  3.8884818
#> 29  2020-07-30 2020-05-01  3.0655465
#> 30  2020-07-30 2020-06-01 -0.7956640
#> 31  2020-07-30 2020-07-01  3.5937427
#> 32  2020-07-30 2020-08-01         NA
#> 33  2020-07-30 2020-09-01         NA
#> 34  2020-07-30 2020-10-01         NA
#> 35  2020-07-30 2020-11-01         NA
#> 36  2020-07-30 2020-12-01         NA
#> 37  2020-07-30 2021-01-01         NA
#> 38  2020-07-30 2021-02-01         NA
#> 39  2020-07-30 2021-03-01         NA
#> 40  2020-07-30 2021-04-01         NA
#> 41  2020-07-30 2021-05-01         NA
#> 42  2020-07-30 2021-06-01         NA
#> 43  2020-07-30 2021-07-01         NA
#> 44  2020-07-30 2021-08-01         NA
#> 45  2020-07-30 2021-09-01         NA
#> 46  2020-07-30 2021-10-01         NA
#> 47  2020-07-30 2021-11-01         NA
#> 48  2020-07-30 2021-12-01         NA
#> 49  2020-11-20 2020-01-01  9.5663737
#> 50  2020-11-20 2020-02-01  8.1220528
#> 51  2020-11-20 2020-03-01  5.2262174
#> 52  2020-11-20 2020-04-01  4.6940087
#> 53  2020-11-20 2020-05-01  3.6690866
#> 54  2020-11-20 2020-06-01  3.4938692
#> 55  2020-11-20 2020-07-01  0.4304233
#> 56  2020-11-20 2020-08-01  0.3938111
#> 57  2020-11-20 2020-09-01  3.3696885
#> 58  2020-11-20 2020-10-01  6.4004909
#> 59  2020-11-20 2020-11-01 13.1807018
#> 60  2020-11-20 2020-12-01         NA
#> 61  2020-11-20 2021-01-01         NA
#> 62  2020-11-20 2021-02-01         NA
#> 63  2020-11-20 2021-03-01         NA
#> 64  2020-11-20 2021-04-01         NA
#> 65  2020-11-20 2021-05-01         NA
#> 66  2020-11-20 2021-06-01         NA
#> 67  2020-11-20 2021-07-01         NA
#> 68  2020-11-20 2021-08-01         NA
#> 69  2020-11-20 2021-09-01         NA
#> 70  2020-11-20 2021-10-01         NA
#> 71  2020-11-20 2021-11-01         NA
#> 72  2020-11-20 2021-12-01         NA
#> 73  2020-12-12 2020-01-01  9.9649347
#> 74  2020-12-12 2020-02-01  7.8161675
#> 75  2020-12-12 2020-03-01  4.7210457
#> 76  2020-12-12 2020-04-01  3.7464167
#> 77  2020-12-12 2020-05-01  4.0962929
#> 78  2020-12-12 2020-06-01  1.3212495
#> 79  2020-12-12 2020-07-01  0.1006455
#> 80  2020-12-12 2020-08-01 -1.4042082
#> 81  2020-12-12 2020-09-01  4.0063439
#> 82  2020-12-12 2020-10-01  7.2757074
#> 83  2020-12-12 2020-11-01 14.9539105
#> 84  2020-12-12 2020-12-01 15.8396562
#> 85  2020-12-12 2021-01-01         NA
#> 86  2020-12-12 2021-02-01         NA
#> 87  2020-12-12 2021-03-01         NA
#> 88  2020-12-12 2021-04-01         NA
#> 89  2020-12-12 2021-05-01         NA
#> 90  2020-12-12 2021-06-01         NA
#> 91  2020-12-12 2021-07-01         NA
#> 92  2020-12-12 2021-08-01         NA
#> 93  2020-12-12 2021-09-01         NA
#> 94  2020-12-12 2021-10-01         NA
#> 95  2020-12-12 2021-11-01         NA
#> 96  2020-12-12 2021-12-01         NA
#> 97  2021-04-26 2020-01-01  9.3941330
#> 98  2021-04-26 2020-02-01  7.9494973
#> 99  2021-04-26 2020-03-01  4.0948138
#> 100 2021-04-26 2020-04-01  4.2727878
#> 101 2021-04-26 2020-05-01  3.8040170
#> 102 2021-04-26 2020-06-01  1.9027777
#> 103 2021-04-26 2020-07-01  0.8894135
#> 104 2021-04-26 2020-08-01 -1.9748794
#> 105 2021-04-26 2020-09-01  3.3511262
#> 106 2021-04-26 2020-10-01  6.5428106
#> 107 2021-04-26 2020-11-01 14.1194647
#> 108 2021-04-26 2020-12-01 17.1404445
#> 109 2021-04-26 2021-01-01 13.2641314
#> 110 2021-04-26 2021-02-01  5.6649888
#> 111 2021-04-26 2021-03-01  1.0217486
#> 112 2021-04-26 2021-04-01  2.1719702
#> 113 2021-04-26 2021-05-01         NA
#> 114 2021-04-26 2021-06-01         NA
#> 115 2021-04-26 2021-07-01         NA
#> 116 2021-04-26 2021-08-01         NA
#> 117 2021-04-26 2021-09-01         NA
#> 118 2021-04-26 2021-10-01         NA
#> 119 2021-04-26 2021-11-01         NA
#> 120 2021-04-26 2021-12-01         NA
#> 121 2021-05-04 2020-01-01  9.6358998
#> 122 2021-05-04 2020-02-01  7.9392791
#> 123 2021-05-04 2020-03-01  4.3777047
#> 124 2021-05-04 2020-04-01  4.1371214
#> 125 2021-05-04 2020-05-01  3.8430272
#> 126 2021-05-04 2020-06-01  1.9179311
#> 127 2021-05-04 2020-07-01  0.5282856
#> 128 2021-05-04 2020-08-01 -2.3267004
#> 129 2021-05-04 2020-09-01  3.6780076
#> 130 2021-05-04 2020-10-01  7.5323212
#> 131 2021-05-04 2020-11-01 14.4123815
#> 132 2021-05-04 2020-12-01 17.2163557
#> 133 2021-05-04 2021-01-01 15.5277869
#> 134 2021-05-04 2021-02-01  7.6876827
#> 135 2021-05-04 2021-03-01  0.5378096
#> 136 2021-05-04 2021-04-01  3.0992193
#> 137 2021-05-04 2021-05-01  4.2639893
#> 138 2021-05-04 2021-06-01         NA
#> 139 2021-05-04 2021-07-01         NA
#> 140 2021-05-04 2021-08-01         NA
#> 141 2021-05-04 2021-09-01         NA
#> 142 2021-05-04 2021-10-01         NA
#> 143 2021-05-04 2021-11-01         NA
#> 144 2021-05-04 2021-12-01         NA