Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-05-31 2020-01-01  13.183987
#> 2   2020-05-31 2020-02-01  16.956998
#> 3   2020-05-31 2020-03-01  18.351974
#> 4   2020-05-31 2020-04-01  16.712462
#> 5   2020-05-31 2020-05-01   9.097334
#> 6   2020-05-31 2020-06-01         NA
#> 7   2020-05-31 2020-07-01         NA
#> 8   2020-05-31 2020-08-01         NA
#> 9   2020-05-31 2020-09-01         NA
#> 10  2020-05-31 2020-10-01         NA
#> 11  2020-05-31 2020-11-01         NA
#> 12  2020-05-31 2020-12-01         NA
#> 13  2020-05-31 2021-01-01         NA
#> 14  2020-05-31 2021-02-01         NA
#> 15  2020-05-31 2021-03-01         NA
#> 16  2020-05-31 2021-04-01         NA
#> 17  2020-05-31 2021-05-01         NA
#> 18  2020-05-31 2021-06-01         NA
#> 19  2020-05-31 2021-07-01         NA
#> 20  2020-05-31 2021-08-01         NA
#> 21  2020-05-31 2021-09-01         NA
#> 22  2020-05-31 2021-10-01         NA
#> 23  2020-05-31 2021-11-01         NA
#> 24  2020-05-31 2021-12-01         NA
#> 25  2020-07-30 2020-01-01  16.686574
#> 26  2020-07-30 2020-02-01  12.742132
#> 27  2020-07-30 2020-03-01  12.512473
#> 28  2020-07-30 2020-04-01  15.200425
#> 29  2020-07-30 2020-05-01   8.903308
#> 30  2020-07-30 2020-06-01   3.983298
#> 31  2020-07-30 2020-07-01   9.015028
#> 32  2020-07-30 2020-08-01         NA
#> 33  2020-07-30 2020-09-01         NA
#> 34  2020-07-30 2020-10-01         NA
#> 35  2020-07-30 2020-11-01         NA
#> 36  2020-07-30 2020-12-01         NA
#> 37  2020-07-30 2021-01-01         NA
#> 38  2020-07-30 2021-02-01         NA
#> 39  2020-07-30 2021-03-01         NA
#> 40  2020-07-30 2021-04-01         NA
#> 41  2020-07-30 2021-05-01         NA
#> 42  2020-07-30 2021-06-01         NA
#> 43  2020-07-30 2021-07-01         NA
#> 44  2020-07-30 2021-08-01         NA
#> 45  2020-07-30 2021-09-01         NA
#> 46  2020-07-30 2021-10-01         NA
#> 47  2020-07-30 2021-11-01         NA
#> 48  2020-07-30 2021-12-01         NA
#> 49  2020-11-20 2020-01-01  16.471064
#> 50  2020-11-20 2020-02-01  14.740559
#> 51  2020-11-20 2020-03-01  17.022412
#> 52  2020-11-20 2020-04-01  16.005952
#> 53  2020-11-20 2020-05-01   9.506849
#> 54  2020-11-20 2020-06-01   8.272831
#> 55  2020-11-20 2020-07-01   5.851708
#> 56  2020-11-20 2020-08-01   8.242666
#> 57  2020-11-20 2020-09-01   2.702215
#> 58  2020-11-20 2020-10-01   5.952846
#> 59  2020-11-20 2020-11-01   6.372049
#> 60  2020-11-20 2020-12-01         NA
#> 61  2020-11-20 2021-01-01         NA
#> 62  2020-11-20 2021-02-01         NA
#> 63  2020-11-20 2021-03-01         NA
#> 64  2020-11-20 2021-04-01         NA
#> 65  2020-11-20 2021-05-01         NA
#> 66  2020-11-20 2021-06-01         NA
#> 67  2020-11-20 2021-07-01         NA
#> 68  2020-11-20 2021-08-01         NA
#> 69  2020-11-20 2021-09-01         NA
#> 70  2020-11-20 2021-10-01         NA
#> 71  2020-11-20 2021-11-01         NA
#> 72  2020-11-20 2021-12-01         NA
#> 73  2020-12-12 2020-01-01  16.869625
#> 74  2020-12-12 2020-02-01  14.434674
#> 75  2020-12-12 2020-03-01  16.517240
#> 76  2020-12-12 2020-04-01  15.058360
#> 77  2020-12-12 2020-05-01   9.934055
#> 78  2020-12-12 2020-06-01   6.100212
#> 79  2020-12-12 2020-07-01   5.521931
#> 80  2020-12-12 2020-08-01   6.444647
#> 81  2020-12-12 2020-09-01   3.338870
#> 82  2020-12-12 2020-10-01   6.828063
#> 83  2020-12-12 2020-11-01   8.145257
#> 84  2020-12-12 2020-12-01   4.509648
#> 85  2020-12-12 2021-01-01         NA
#> 86  2020-12-12 2021-02-01         NA
#> 87  2020-12-12 2021-03-01         NA
#> 88  2020-12-12 2021-04-01         NA
#> 89  2020-12-12 2021-05-01         NA
#> 90  2020-12-12 2021-06-01         NA
#> 91  2020-12-12 2021-07-01         NA
#> 92  2020-12-12 2021-08-01         NA
#> 93  2020-12-12 2021-09-01         NA
#> 94  2020-12-12 2021-10-01         NA
#> 95  2020-12-12 2021-11-01         NA
#> 96  2020-12-12 2021-12-01         NA
#> 97  2021-04-26 2020-01-01  16.298823
#> 98  2021-04-26 2020-02-01  14.568004
#> 99  2021-04-26 2020-03-01  15.891008
#> 100 2021-04-26 2020-04-01  15.584731
#> 101 2021-04-26 2020-05-01   9.641779
#> 102 2021-04-26 2020-06-01   6.681740
#> 103 2021-04-26 2020-07-01   6.310699
#> 104 2021-04-26 2020-08-01   5.873976
#> 105 2021-04-26 2020-09-01   2.683652
#> 106 2021-04-26 2020-10-01   6.095166
#> 107 2021-04-26 2020-11-01   7.310812
#> 108 2021-04-26 2020-12-01   5.810437
#> 109 2021-04-26 2021-01-01   8.937629
#> 110 2021-04-26 2021-02-01   8.673707
#> 111 2021-04-26 2021-03-01   8.802599
#> 112 2021-04-26 2021-04-01   9.649219
#> 113 2021-04-26 2021-05-01         NA
#> 114 2021-04-26 2021-06-01         NA
#> 115 2021-04-26 2021-07-01         NA
#> 116 2021-04-26 2021-08-01         NA
#> 117 2021-04-26 2021-09-01         NA
#> 118 2021-04-26 2021-10-01         NA
#> 119 2021-04-26 2021-11-01         NA
#> 120 2021-04-26 2021-12-01         NA
#> 121 2021-05-04 2020-01-01  16.540590
#> 122 2021-05-04 2020-02-01  14.557786
#> 123 2021-05-04 2020-03-01  16.173899
#> 124 2021-05-04 2020-04-01  15.449065
#> 125 2021-05-04 2020-05-01   9.680789
#> 126 2021-05-04 2020-06-01   6.696893
#> 127 2021-05-04 2020-07-01   5.949571
#> 128 2021-05-04 2020-08-01   5.522155
#> 129 2021-05-04 2020-09-01   3.010534
#> 130 2021-05-04 2020-10-01   7.084677
#> 131 2021-05-04 2020-11-01   7.603728
#> 132 2021-05-04 2020-12-01   5.886348
#> 133 2021-05-04 2021-01-01  11.201285
#> 134 2021-05-04 2021-02-01  10.696401
#> 135 2021-05-04 2021-03-01   8.318660
#> 136 2021-05-04 2021-04-01  10.576468
#> 137 2021-05-04 2021-05-01   5.620817
#> 138 2021-05-04 2021-06-01         NA
#> 139 2021-05-04 2021-07-01         NA
#> 140 2021-05-04 2021-08-01         NA
#> 141 2021-05-04 2021-09-01         NA
#> 142 2021-05-04 2021-10-01         NA
#> 143 2021-05-04 2021-11-01         NA
#> 144 2021-05-04 2021-12-01         NA