Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-05-31 2020-01-01   5.788272
#> 2   2020-05-31 2020-02-01  13.834402
#> 3   2020-05-31 2020-03-01  16.113498
#> 4   2020-05-31 2020-04-01  19.625079
#> 5   2020-05-31 2020-05-01  18.831466
#> 6   2020-05-31 2020-06-01         NA
#> 7   2020-05-31 2020-07-01         NA
#> 8   2020-05-31 2020-08-01         NA
#> 9   2020-05-31 2020-09-01         NA
#> 10  2020-05-31 2020-10-01         NA
#> 11  2020-05-31 2020-11-01         NA
#> 12  2020-05-31 2020-12-01         NA
#> 13  2020-05-31 2021-01-01         NA
#> 14  2020-05-31 2021-02-01         NA
#> 15  2020-05-31 2021-03-01         NA
#> 16  2020-05-31 2021-04-01         NA
#> 17  2020-05-31 2021-05-01         NA
#> 18  2020-05-31 2021-06-01         NA
#> 19  2020-05-31 2021-07-01         NA
#> 20  2020-05-31 2021-08-01         NA
#> 21  2020-05-31 2021-09-01         NA
#> 22  2020-05-31 2021-10-01         NA
#> 23  2020-05-31 2021-11-01         NA
#> 24  2020-05-31 2021-12-01         NA
#> 25  2020-07-30 2020-01-01   9.290859
#> 26  2020-07-30 2020-02-01   9.619537
#> 27  2020-07-30 2020-03-01  10.273997
#> 28  2020-07-30 2020-04-01  18.113043
#> 29  2020-07-30 2020-05-01  18.637440
#> 30  2020-07-30 2020-06-01  13.793076
#> 31  2020-07-30 2020-07-01  20.119418
#> 32  2020-07-30 2020-08-01         NA
#> 33  2020-07-30 2020-09-01         NA
#> 34  2020-07-30 2020-10-01         NA
#> 35  2020-07-30 2020-11-01         NA
#> 36  2020-07-30 2020-12-01         NA
#> 37  2020-07-30 2021-01-01         NA
#> 38  2020-07-30 2021-02-01         NA
#> 39  2020-07-30 2021-03-01         NA
#> 40  2020-07-30 2021-04-01         NA
#> 41  2020-07-30 2021-05-01         NA
#> 42  2020-07-30 2021-06-01         NA
#> 43  2020-07-30 2021-07-01         NA
#> 44  2020-07-30 2021-08-01         NA
#> 45  2020-07-30 2021-09-01         NA
#> 46  2020-07-30 2021-10-01         NA
#> 47  2020-07-30 2021-11-01         NA
#> 48  2020-07-30 2021-12-01         NA
#> 49  2020-11-20 2020-01-01   9.075349
#> 50  2020-11-20 2020-02-01  11.617964
#> 51  2020-11-20 2020-03-01  14.783936
#> 52  2020-11-20 2020-04-01  18.918570
#> 53  2020-11-20 2020-05-01  19.240981
#> 54  2020-11-20 2020-06-01  18.082609
#> 55  2020-11-20 2020-07-01  16.956098
#> 56  2020-11-20 2020-08-01  19.555925
#> 57  2020-11-20 2020-09-01  19.066703
#> 58  2020-11-20 2020-10-01  17.558521
#> 59  2020-11-20 2020-11-01  23.069802
#> 60  2020-11-20 2020-12-01         NA
#> 61  2020-11-20 2021-01-01         NA
#> 62  2020-11-20 2021-02-01         NA
#> 63  2020-11-20 2021-03-01         NA
#> 64  2020-11-20 2021-04-01         NA
#> 65  2020-11-20 2021-05-01         NA
#> 66  2020-11-20 2021-06-01         NA
#> 67  2020-11-20 2021-07-01         NA
#> 68  2020-11-20 2021-08-01         NA
#> 69  2020-11-20 2021-09-01         NA
#> 70  2020-11-20 2021-10-01         NA
#> 71  2020-11-20 2021-11-01         NA
#> 72  2020-11-20 2021-12-01         NA
#> 73  2020-12-12 2020-01-01   9.473910
#> 74  2020-12-12 2020-02-01  11.312078
#> 75  2020-12-12 2020-03-01  14.278764
#> 76  2020-12-12 2020-04-01  17.970977
#> 77  2020-12-12 2020-05-01  19.668187
#> 78  2020-12-12 2020-06-01  15.909989
#> 79  2020-12-12 2020-07-01  16.626320
#> 80  2020-12-12 2020-08-01  17.757906
#> 81  2020-12-12 2020-09-01  19.703358
#> 82  2020-12-12 2020-10-01  18.433737
#> 83  2020-12-12 2020-11-01  24.843011
#> 84  2020-12-12 2020-12-01  21.170572
#> 85  2020-12-12 2021-01-01         NA
#> 86  2020-12-12 2021-02-01         NA
#> 87  2020-12-12 2021-03-01         NA
#> 88  2020-12-12 2021-04-01         NA
#> 89  2020-12-12 2021-05-01         NA
#> 90  2020-12-12 2021-06-01         NA
#> 91  2020-12-12 2021-07-01         NA
#> 92  2020-12-12 2021-08-01         NA
#> 93  2020-12-12 2021-09-01         NA
#> 94  2020-12-12 2021-10-01         NA
#> 95  2020-12-12 2021-11-01         NA
#> 96  2020-12-12 2021-12-01         NA
#> 97  2021-04-26 2020-01-01   8.903108
#> 98  2021-04-26 2020-02-01  11.445408
#> 99  2021-04-26 2020-03-01  13.652532
#> 100 2021-04-26 2020-04-01  18.497349
#> 101 2021-04-26 2020-05-01  19.375911
#> 102 2021-04-26 2020-06-01  16.491517
#> 103 2021-04-26 2020-07-01  17.415088
#> 104 2021-04-26 2020-08-01  17.187235
#> 105 2021-04-26 2020-09-01  19.048140
#> 106 2021-04-26 2020-10-01  17.700841
#> 107 2021-04-26 2020-11-01  24.008565
#> 108 2021-04-26 2020-12-01  22.471360
#> 109 2021-04-26 2021-01-01  19.556820
#> 110 2021-04-26 2021-02-01  15.827755
#> 111 2021-04-26 2021-03-01  15.362701
#> 112 2021-04-26 2021-04-01  21.339665
#> 113 2021-04-26 2021-05-01         NA
#> 114 2021-04-26 2021-06-01         NA
#> 115 2021-04-26 2021-07-01         NA
#> 116 2021-04-26 2021-08-01         NA
#> 117 2021-04-26 2021-09-01         NA
#> 118 2021-04-26 2021-10-01         NA
#> 119 2021-04-26 2021-11-01         NA
#> 120 2021-04-26 2021-12-01         NA
#> 121 2021-05-04 2020-01-01   9.144875
#> 122 2021-05-04 2020-02-01  11.435190
#> 123 2021-05-04 2020-03-01  13.935423
#> 124 2021-05-04 2020-04-01  18.361682
#> 125 2021-05-04 2020-05-01  19.414921
#> 126 2021-05-04 2020-06-01  16.506671
#> 127 2021-05-04 2020-07-01  17.053961
#> 128 2021-05-04 2020-08-01  16.835414
#> 129 2021-05-04 2020-09-01  19.375022
#> 130 2021-05-04 2020-10-01  18.690351
#> 131 2021-05-04 2020-11-01  24.301482
#> 132 2021-05-04 2020-12-01  22.547272
#> 133 2021-05-04 2021-01-01  21.820475
#> 134 2021-05-04 2021-02-01  17.850449
#> 135 2021-05-04 2021-03-01  14.878762
#> 136 2021-05-04 2021-04-01  22.266914
#> 137 2021-05-04 2021-05-01  23.603039
#> 138 2021-05-04 2021-06-01         NA
#> 139 2021-05-04 2021-07-01         NA
#> 140 2021-05-04 2021-08-01         NA
#> 141 2021-05-04 2021-09-01         NA
#> 142 2021-05-04 2021-10-01         NA
#> 143 2021-05-04 2021-11-01         NA
#> 144 2021-05-04 2021-12-01         NA