Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01 12.5140072
#> 2   2020-02-17 2020-02-01 23.3330381
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01 11.8017286
#> 26  2020-06-05 2020-02-01 20.0601533
#> 27  2020-06-05 2020-03-01 17.0293688
#> 28  2020-06-05 2020-04-01 17.7165508
#> 29  2020-06-05 2020-05-01 11.2553186
#> 30  2020-06-05 2020-06-01  9.2019638
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01 11.7997989
#> 50  2021-04-17 2020-02-01 19.8002919
#> 51  2021-04-17 2020-03-01 17.6887206
#> 52  2021-04-17 2020-04-01 15.1365043
#> 53  2021-04-17 2020-05-01  8.5691323
#> 54  2021-04-17 2020-06-01  9.3369605
#> 55  2021-04-17 2020-07-01  9.8407749
#> 56  2021-04-17 2020-08-01  9.3017163
#> 57  2021-04-17 2020-09-01  8.6322260
#> 58  2021-04-17 2020-10-01  5.2000339
#> 59  2021-04-17 2020-11-01  8.1861688
#> 60  2021-04-17 2020-12-01  1.6182013
#> 61  2021-04-17 2021-01-01  5.2242477
#> 62  2021-04-17 2021-02-01 18.9552842
#> 63  2021-04-17 2021-03-01 12.6209491
#> 64  2021-04-17 2021-04-01  6.6202834
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01 11.4174653
#> 74  2021-09-08 2020-02-01 20.4774686
#> 75  2021-09-08 2020-03-01 18.4579263
#> 76  2021-09-08 2020-04-01 15.0158711
#> 77  2021-09-08 2020-05-01  7.9982709
#> 78  2021-09-08 2020-06-01 10.3323889
#> 79  2021-09-08 2020-07-01 11.0866239
#> 80  2021-09-08 2020-08-01  7.9763258
#> 81  2021-09-08 2020-09-01  5.0505988
#> 82  2021-09-08 2020-10-01  8.9469435
#> 83  2021-09-08 2020-11-01  7.6274863
#> 84  2021-09-08 2020-12-01  0.7963682
#> 85  2021-09-08 2021-01-01  6.7122246
#> 86  2021-09-08 2021-02-01 18.0719197
#> 87  2021-09-08 2021-03-01 11.4822210
#> 88  2021-09-08 2021-04-01  6.4668521
#> 89  2021-09-08 2021-05-01  5.4037673
#> 90  2021-09-08 2021-06-01  3.1577266
#> 91  2021-09-08 2021-07-01  6.5780445
#> 92  2021-09-08 2021-08-01  2.6655951
#> 93  2021-09-08 2021-09-01  0.3051948
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01 11.3647915
#> 98  2021-11-12 2020-02-01 20.5193653
#> 99  2021-11-12 2020-03-01 17.3385189
#> 100 2021-11-12 2020-04-01 15.3046883
#> 101 2021-11-12 2020-05-01  8.6672068
#> 102 2021-11-12 2020-06-01 10.4866134
#> 103 2021-11-12 2020-07-01 12.3477138
#> 104 2021-11-12 2020-08-01  8.5568217
#> 105 2021-11-12 2020-09-01  6.7209866
#> 106 2021-11-12 2020-10-01  7.5913425
#> 107 2021-11-12 2020-11-01  7.5225059
#> 108 2021-11-12 2020-12-01  2.7392273
#> 109 2021-11-12 2021-01-01  5.9954688
#> 110 2021-11-12 2021-02-01 17.4814839
#> 111 2021-11-12 2021-03-01 12.4229007
#> 112 2021-11-12 2021-04-01  7.7893872
#> 113 2021-11-12 2021-05-01  6.7399217
#> 114 2021-11-12 2021-06-01  4.2592686
#> 115 2021-11-12 2021-07-01  8.2003439
#> 116 2021-11-12 2021-08-01  4.4119313
#> 117 2021-11-12 2021-09-01  3.7371850
#> 118 2021-11-12 2021-10-01  5.4984963
#> 119 2021-11-12 2021-11-01  4.9480554
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01 11.7200058
#> 122 2021-11-17 2020-02-01 20.3956356
#> 123 2021-11-17 2020-03-01 17.9181157
#> 124 2021-11-17 2020-04-01 15.3644937
#> 125 2021-11-17 2020-05-01  8.5232241
#> 126 2021-11-17 2020-06-01 10.3706973
#> 127 2021-11-17 2020-07-01 12.1100574
#> 128 2021-11-17 2020-08-01  8.3556677
#> 129 2021-11-17 2020-09-01  6.7842497
#> 130 2021-11-17 2020-10-01  6.3557897
#> 131 2021-11-17 2020-11-01  7.4217435
#> 132 2021-11-17 2020-12-01  3.1312055
#> 133 2021-11-17 2021-01-01  6.4428620
#> 134 2021-11-17 2021-02-01 18.2935687
#> 135 2021-11-17 2021-03-01 12.7900616
#> 136 2021-11-17 2021-04-01  8.2770773
#> 137 2021-11-17 2021-05-01  4.9267574
#> 138 2021-11-17 2021-06-01  5.4504327
#> 139 2021-11-17 2021-07-01  9.3830696
#> 140 2021-11-17 2021-08-01  5.0232014
#> 141 2021-11-17 2021-09-01  2.5076338
#> 142 2021-11-17 2021-10-01  5.4266033
#> 143 2021-11-17 2021-11-01  7.5266819
#> 144 2021-11-17 2021-12-01         NA