Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01 11.6164214
#> 2   2020-02-17 2020-02-01 19.0978716
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01 10.9041428
#> 26  2020-06-05 2020-02-01 15.8249867
#> 27  2020-06-05 2020-03-01 17.4461849
#> 28  2020-06-05 2020-04-01 23.1816415
#> 29  2020-06-05 2020-05-01 22.5137863
#> 30  2020-06-05 2020-06-01 17.4996775
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01 10.9022130
#> 50  2021-04-17 2020-02-01 15.5651254
#> 51  2021-04-17 2020-03-01 18.1055367
#> 52  2021-04-17 2020-04-01 20.6015950
#> 53  2021-04-17 2020-05-01 19.8276000
#> 54  2021-04-17 2020-06-01 17.6346742
#> 55  2021-04-17 2020-07-01 13.3674630
#> 56  2021-04-17 2020-08-01 13.2647758
#> 57  2021-04-17 2020-09-01 12.1753102
#> 58  2021-04-17 2020-10-01  7.8368515
#> 59  2021-04-17 2020-11-01 10.1338505
#> 60  2021-04-17 2020-12-01  2.3807528
#> 61  2021-04-17 2021-01-01  1.2998876
#> 62  2021-04-17 2021-02-01  4.7064384
#> 63  2021-04-17 2021-03-01  1.9484850
#> 64  2021-04-17 2021-04-01  4.0094572
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01 10.5198795
#> 74  2021-09-08 2020-02-01 16.2423020
#> 75  2021-09-08 2020-03-01 18.8747424
#> 76  2021-09-08 2020-04-01 20.4809618
#> 77  2021-09-08 2020-05-01 19.2567386
#> 78  2021-09-08 2020-06-01 18.6301026
#> 79  2021-09-08 2020-07-01 14.6133120
#> 80  2021-09-08 2020-08-01 11.9393854
#> 81  2021-09-08 2020-09-01  8.5936830
#> 82  2021-09-08 2020-10-01 11.5837611
#> 83  2021-09-08 2020-11-01  9.5751679
#> 84  2021-09-08 2020-12-01  1.5589197
#> 85  2021-09-08 2021-01-01  2.7878645
#> 86  2021-09-08 2021-02-01  3.8230739
#> 87  2021-09-08 2021-03-01  0.8097570
#> 88  2021-09-08 2021-04-01  3.8560259
#> 89  2021-09-08 2021-05-01  8.9434727
#> 90  2021-09-08 2021-06-01  5.8264792
#> 91  2021-09-08 2021-07-01  4.5060559
#> 92  2021-09-08 2021-08-01  3.6741334
#> 93  2021-09-08 2021-09-01  1.0694401
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01 10.4672056
#> 98  2021-11-12 2020-02-01 16.2841987
#> 99  2021-11-12 2020-03-01 17.7553350
#> 100 2021-11-12 2020-04-01 20.7697790
#> 101 2021-11-12 2020-05-01 19.9256745
#> 102 2021-11-12 2020-06-01 18.7843271
#> 103 2021-11-12 2020-07-01 15.8744019
#> 104 2021-11-12 2020-08-01 12.5198813
#> 105 2021-11-12 2020-09-01 10.2640709
#> 106 2021-11-12 2020-10-01 10.2281600
#> 107 2021-11-12 2020-11-01  9.4701875
#> 108 2021-11-12 2020-12-01  3.5017788
#> 109 2021-11-12 2021-01-01  2.0711087
#> 110 2021-11-12 2021-02-01  3.2326382
#> 111 2021-11-12 2021-03-01  1.7504367
#> 112 2021-11-12 2021-04-01  5.1785610
#> 113 2021-11-12 2021-05-01 10.2796271
#> 114 2021-11-12 2021-06-01  6.9280211
#> 115 2021-11-12 2021-07-01  6.1283553
#> 116 2021-11-12 2021-08-01  5.4204697
#> 117 2021-11-12 2021-09-01  4.5014303
#> 118 2021-11-12 2021-10-01 -0.2440782
#> 119 2021-11-12 2021-11-01 -3.6293343
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01 10.8224199
#> 122 2021-11-17 2020-02-01 16.1604691
#> 123 2021-11-17 2020-03-01 18.3349318
#> 124 2021-11-17 2020-04-01 20.8295844
#> 125 2021-11-17 2020-05-01 19.7816918
#> 126 2021-11-17 2020-06-01 18.6684110
#> 127 2021-11-17 2020-07-01 15.6367455
#> 128 2021-11-17 2020-08-01 12.3187273
#> 129 2021-11-17 2020-09-01 10.3273340
#> 130 2021-11-17 2020-10-01  8.9926073
#> 131 2021-11-17 2020-11-01  9.3694251
#> 132 2021-11-17 2020-12-01  3.8937570
#> 133 2021-11-17 2021-01-01  2.5185019
#> 134 2021-11-17 2021-02-01  4.0447230
#> 135 2021-11-17 2021-03-01  2.1175975
#> 136 2021-11-17 2021-04-01  5.6662511
#> 137 2021-11-17 2021-05-01  8.4664628
#> 138 2021-11-17 2021-06-01  8.1191853
#> 139 2021-11-17 2021-07-01  7.3110810
#> 140 2021-11-17 2021-08-01  6.0317398
#> 141 2021-11-17 2021-09-01  3.2718791
#> 142 2021-11-17 2021-10-01 -0.3159712
#> 143 2021-11-17 2021-11-01 -1.0507078
#> 144 2021-11-17 2021-12-01         NA