Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01  15.297884
#> 2   2020-02-17 2020-02-01  16.149961
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01  14.585605
#> 26  2020-06-05 2020-02-01  12.877076
#> 27  2020-06-05 2020-03-01  11.038091
#> 28  2020-06-05 2020-04-01  13.246943
#> 29  2020-06-05 2020-05-01  17.759235
#> 30  2020-06-05 2020-06-01  10.653104
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01  14.583675
#> 50  2021-04-17 2020-02-01  12.617215
#> 51  2021-04-17 2020-03-01  11.697443
#> 52  2021-04-17 2020-04-01  10.666897
#> 53  2021-04-17 2020-05-01  15.073049
#> 54  2021-04-17 2020-06-01  10.788101
#> 55  2021-04-17 2020-07-01   8.734695
#> 56  2021-04-17 2020-08-01  15.674593
#> 57  2021-04-17 2020-09-01  14.935069
#> 58  2021-04-17 2020-10-01  12.200787
#> 59  2021-04-17 2020-11-01  12.167995
#> 60  2021-04-17 2020-12-01  11.632293
#> 61  2021-04-17 2021-01-01  12.855811
#> 62  2021-04-17 2021-02-01  11.301306
#> 63  2021-04-17 2021-03-01  10.041851
#> 64  2021-04-17 2021-04-01   9.930563
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01  14.201342
#> 74  2021-09-08 2020-02-01  13.294391
#> 75  2021-09-08 2020-03-01  12.466649
#> 76  2021-09-08 2020-04-01  10.546264
#> 77  2021-09-08 2020-05-01  14.502188
#> 78  2021-09-08 2020-06-01  11.783529
#> 79  2021-09-08 2020-07-01   9.980544
#> 80  2021-09-08 2020-08-01  14.349202
#> 81  2021-09-08 2020-09-01  11.353442
#> 82  2021-09-08 2020-10-01  15.947697
#> 83  2021-09-08 2020-11-01  11.609312
#> 84  2021-09-08 2020-12-01  10.810460
#> 85  2021-09-08 2021-01-01  14.343788
#> 86  2021-09-08 2021-02-01  10.417942
#> 87  2021-09-08 2021-03-01   8.903123
#> 88  2021-09-08 2021-04-01   9.777132
#> 89  2021-09-08 2021-05-01  12.437008
#> 90  2021-09-08 2021-06-01  10.220293
#> 91  2021-09-08 2021-07-01  13.552012
#> 92  2021-09-08 2021-08-01  17.139793
#> 93  2021-09-08 2021-09-01  15.092907
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01  14.148668
#> 98  2021-11-12 2020-02-01  13.336288
#> 99  2021-11-12 2020-03-01  11.347241
#> 100 2021-11-12 2020-04-01  10.835081
#> 101 2021-11-12 2020-05-01  15.171124
#> 102 2021-11-12 2020-06-01  11.937754
#> 103 2021-11-12 2020-07-01  11.241634
#> 104 2021-11-12 2020-08-01  14.929698
#> 105 2021-11-12 2020-09-01  13.023830
#> 106 2021-11-12 2020-10-01  14.592096
#> 107 2021-11-12 2020-11-01  11.504332
#> 108 2021-11-12 2020-12-01  12.753319
#> 109 2021-11-12 2021-01-01  13.627032
#> 110 2021-11-12 2021-02-01   9.827506
#> 111 2021-11-12 2021-03-01   9.843802
#> 112 2021-11-12 2021-04-01  11.099667
#> 113 2021-11-12 2021-05-01  13.773162
#> 114 2021-11-12 2021-06-01  11.321835
#> 115 2021-11-12 2021-07-01  15.174312
#> 116 2021-11-12 2021-08-01  18.886130
#> 117 2021-11-12 2021-09-01  18.524898
#> 118 2021-11-12 2021-10-01  18.250980
#> 119 2021-11-12 2021-11-01  12.962905
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01  14.503882
#> 122 2021-11-17 2020-02-01  13.212558
#> 123 2021-11-17 2020-03-01  11.926838
#> 124 2021-11-17 2020-04-01  10.894886
#> 125 2021-11-17 2020-05-01  15.027141
#> 126 2021-11-17 2020-06-01  11.821838
#> 127 2021-11-17 2020-07-01  11.003977
#> 128 2021-11-17 2020-08-01  14.728544
#> 129 2021-11-17 2020-09-01  13.087093
#> 130 2021-11-17 2020-10-01  13.356543
#> 131 2021-11-17 2020-11-01  11.403569
#> 132 2021-11-17 2020-12-01  13.145297
#> 133 2021-11-17 2021-01-01  14.074425
#> 134 2021-11-17 2021-02-01  10.639591
#> 135 2021-11-17 2021-03-01  10.210963
#> 136 2021-11-17 2021-04-01  11.587357
#> 137 2021-11-17 2021-05-01  11.959998
#> 138 2021-11-17 2021-06-01  12.512999
#> 139 2021-11-17 2021-07-01  16.357037
#> 140 2021-11-17 2021-08-01  19.497400
#> 141 2021-11-17 2021-09-01  17.295346
#> 142 2021-11-17 2021-10-01  18.179087
#> 143 2021-11-17 2021-11-01  15.541531
#> 144 2021-11-17 2021-12-01         NA