Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01  9.36319241
#> 2   2020-02-17 2020-02-01 13.44382806
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01  8.65091377
#> 26  2020-06-05 2020-02-01 10.17094319
#> 27  2020-06-05 2020-03-01 14.62964893
#> 28  2020-06-05 2020-04-01 15.08855612
#> 29  2020-06-05 2020-05-01 15.68542213
#> 30  2020-06-05 2020-06-01  7.52252213
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01  8.64898403
#> 50  2021-04-17 2020-02-01  9.91108184
#> 51  2021-04-17 2020-03-01 15.28900067
#> 52  2021-04-17 2020-04-01 12.50850960
#> 53  2021-04-17 2020-05-01 12.99923587
#> 54  2021-04-17 2020-06-01  7.65751880
#> 55  2021-04-17 2020-07-01  5.64684387
#> 56  2021-04-17 2020-08-01  3.29339476
#> 57  2021-04-17 2020-09-01  2.69044931
#> 58  2021-04-17 2020-10-01 -0.09762135
#> 59  2021-04-17 2020-11-01  3.76462998
#> 60  2021-04-17 2020-12-01  9.32352799
#> 61  2021-04-17 2021-01-01 13.66541138
#> 62  2021-04-17 2021-02-01 18.03285863
#> 63  2021-04-17 2021-03-01 18.62982667
#> 64  2021-04-17 2021-04-01 18.02407933
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01  8.26665049
#> 74  2021-09-08 2020-02-01 10.58825851
#> 75  2021-09-08 2020-03-01 16.05820640
#> 76  2021-09-08 2020-04-01 12.38787645
#> 77  2021-09-08 2020-05-01 12.42837446
#> 78  2021-09-08 2020-06-01  8.65294721
#> 79  2021-09-08 2020-07-01  6.89269286
#> 80  2021-09-08 2020-08-01  1.96800430
#> 81  2021-09-08 2020-09-01 -0.89117789
#> 82  2021-09-08 2020-10-01  3.64928826
#> 83  2021-09-08 2020-11-01  3.20594738
#> 84  2021-09-08 2020-12-01  8.50169489
#> 85  2021-09-08 2021-01-01 15.15338830
#> 86  2021-09-08 2021-02-01 17.14949414
#> 87  2021-09-08 2021-03-01 17.49109863
#> 88  2021-09-08 2021-04-01 17.87064805
#> 89  2021-09-08 2021-05-01 22.70347071
#> 90  2021-09-08 2021-06-01 21.01324618
#> 91  2021-09-08 2021-07-01 15.32703000
#> 92  2021-09-08 2021-08-01  9.43940116
#> 93  2021-09-08 2021-09-01  7.78031293
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01  8.21397662
#> 98  2021-11-12 2020-02-01 10.63015519
#> 99  2021-11-12 2020-03-01 14.93879895
#> 100 2021-11-12 2020-04-01 12.67669363
#> 101 2021-11-12 2020-05-01 13.09731036
#> 102 2021-11-12 2020-06-01  8.80717169
#> 103 2021-11-12 2020-07-01  8.15378275
#> 104 2021-11-12 2020-08-01  2.54850019
#> 105 2021-11-12 2020-09-01  0.77920993
#> 106 2021-11-12 2020-10-01  2.29368719
#> 107 2021-11-12 2020-11-01  3.10096700
#> 108 2021-11-12 2020-12-01 10.44455395
#> 109 2021-11-12 2021-01-01 14.43663250
#> 110 2021-11-12 2021-02-01 16.55905838
#> 111 2021-11-12 2021-03-01 18.43177835
#> 112 2021-11-12 2021-04-01 19.19318312
#> 113 2021-11-12 2021-05-01 24.03962507
#> 114 2021-11-12 2021-06-01 22.11478812
#> 115 2021-11-12 2021-07-01 16.94932938
#> 116 2021-11-12 2021-08-01 11.18573741
#> 117 2021-11-12 2021-09-01 11.21230311
#> 118 2021-11-12 2021-10-01  7.46659071
#> 119 2021-11-12 2021-11-01 11.88436907
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01  8.56919094
#> 122 2021-11-17 2020-02-01 10.50642555
#> 123 2021-11-17 2020-03-01 15.51839582
#> 124 2021-11-17 2020-04-01 12.73649902
#> 125 2021-11-17 2020-05-01 12.95332765
#> 126 2021-11-17 2020-06-01  8.69125561
#> 127 2021-11-17 2020-07-01  7.91612630
#> 128 2021-11-17 2020-08-01  2.34734621
#> 129 2021-11-17 2020-09-01  0.84247306
#> 130 2021-11-17 2020-10-01  1.05813442
#> 131 2021-11-17 2020-11-01  3.00020459
#> 132 2021-11-17 2020-12-01 10.83653212
#> 133 2021-11-17 2021-01-01 14.88402565
#> 134 2021-11-17 2021-02-01 17.37114319
#> 135 2021-11-17 2021-03-01 18.79893920
#> 136 2021-11-17 2021-04-01 19.68087323
#> 137 2021-11-17 2021-05-01 22.22646075
#> 138 2021-11-17 2021-06-01 23.30595228
#> 139 2021-11-17 2021-07-01 18.13205514
#> 140 2021-11-17 2021-08-01 11.79700750
#> 141 2021-11-17 2021-09-01  9.98275193
#> 142 2021-11-17 2021-10-01  7.39469771
#> 143 2021-11-17 2021-11-01 14.46299557
#> 144 2021-11-17 2021-12-01          NA