Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time   obs_values
#> 1   2020-02-17 2020-01-01   6.42181515
#> 2   2020-02-17 2020-02-01   4.80008806
#> 3   2020-02-17 2020-03-01           NA
#> 4   2020-02-17 2020-04-01           NA
#> 5   2020-02-17 2020-05-01           NA
#> 6   2020-02-17 2020-06-01           NA
#> 7   2020-02-17 2020-07-01           NA
#> 8   2020-02-17 2020-08-01           NA
#> 9   2020-02-17 2020-09-01           NA
#> 10  2020-02-17 2020-10-01           NA
#> 11  2020-02-17 2020-11-01           NA
#> 12  2020-02-17 2020-12-01           NA
#> 13  2020-02-17 2021-01-01           NA
#> 14  2020-02-17 2021-02-01           NA
#> 15  2020-02-17 2021-03-01           NA
#> 16  2020-02-17 2021-04-01           NA
#> 17  2020-02-17 2021-05-01           NA
#> 18  2020-02-17 2021-06-01           NA
#> 19  2020-02-17 2021-07-01           NA
#> 20  2020-02-17 2021-08-01           NA
#> 21  2020-02-17 2021-09-01           NA
#> 22  2020-02-17 2021-10-01           NA
#> 23  2020-02-17 2021-11-01           NA
#> 24  2020-02-17 2021-12-01           NA
#> 25  2020-06-05 2020-01-01   5.70953651
#> 26  2020-06-05 2020-02-01   1.52720319
#> 27  2020-06-05 2020-03-01  -0.38840728
#> 28  2020-06-05 2020-04-01   0.15324691
#> 29  2020-06-05 2020-05-01   1.53549032
#> 30  2020-06-05 2020-06-01   0.27387270
#> 31  2020-06-05 2020-07-01           NA
#> 32  2020-06-05 2020-08-01           NA
#> 33  2020-06-05 2020-09-01           NA
#> 34  2020-06-05 2020-10-01           NA
#> 35  2020-06-05 2020-11-01           NA
#> 36  2020-06-05 2020-12-01           NA
#> 37  2020-06-05 2021-01-01           NA
#> 38  2020-06-05 2021-02-01           NA
#> 39  2020-06-05 2021-03-01           NA
#> 40  2020-06-05 2021-04-01           NA
#> 41  2020-06-05 2021-05-01           NA
#> 42  2020-06-05 2021-06-01           NA
#> 43  2020-06-05 2021-07-01           NA
#> 44  2020-06-05 2021-08-01           NA
#> 45  2020-06-05 2021-09-01           NA
#> 46  2020-06-05 2021-10-01           NA
#> 47  2020-06-05 2021-11-01           NA
#> 48  2020-06-05 2021-12-01           NA
#> 49  2021-04-17 2020-01-01   5.70760677
#> 50  2021-04-17 2020-02-01   1.26734184
#> 51  2021-04-17 2020-03-01   0.27094446
#> 52  2021-04-17 2020-04-01  -2.42679961
#> 53  2021-04-17 2020-05-01  -1.15069595
#> 54  2021-04-17 2020-06-01   0.40886936
#> 55  2021-04-17 2020-07-01  -4.15845154
#> 56  2021-04-17 2020-08-01  -1.07923014
#> 57  2021-04-17 2020-09-01  -0.14033452
#> 58  2021-04-17 2020-10-01  -4.45247787
#> 59  2021-04-17 2020-11-01  -1.45017825
#> 60  2021-04-17 2020-12-01  -4.16763861
#> 61  2021-04-17 2021-01-01  -3.54907995
#> 62  2021-04-17 2021-02-01  -4.80310861
#> 63  2021-04-17 2021-03-01  -5.96435246
#> 64  2021-04-17 2021-04-01  -6.25374021
#> 65  2021-04-17 2021-05-01           NA
#> 66  2021-04-17 2021-06-01           NA
#> 67  2021-04-17 2021-07-01           NA
#> 68  2021-04-17 2021-08-01           NA
#> 69  2021-04-17 2021-09-01           NA
#> 70  2021-04-17 2021-10-01           NA
#> 71  2021-04-17 2021-11-01           NA
#> 72  2021-04-17 2021-12-01           NA
#> 73  2021-09-08 2020-01-01   5.32527323
#> 74  2021-09-08 2020-02-01   1.94451851
#> 75  2021-09-08 2020-03-01   1.04015019
#> 76  2021-09-08 2020-04-01  -2.54743277
#> 77  2021-09-08 2020-05-01  -1.72155735
#> 78  2021-09-08 2020-06-01   1.40429777
#> 79  2021-09-08 2020-07-01  -2.91260255
#> 80  2021-09-08 2020-08-01  -2.40462061
#> 81  2021-09-08 2020-09-01  -3.72196172
#> 82  2021-09-08 2020-10-01  -0.70556825
#> 83  2021-09-08 2020-11-01  -2.00886084
#> 84  2021-09-08 2020-12-01  -4.98947171
#> 85  2021-09-08 2021-01-01  -2.06110303
#> 86  2021-09-08 2021-02-01  -5.68647310
#> 87  2021-09-08 2021-03-01  -7.10308050
#> 88  2021-09-08 2021-04-01  -6.40717149
#> 89  2021-09-08 2021-05-01  -5.25217816
#> 90  2021-09-08 2021-06-01  -4.34524578
#> 91  2021-09-08 2021-07-01  -9.66640094
#> 92  2021-09-08 2021-08-01 -10.91058164
#> 93  2021-09-08 2021-09-01 -10.10901110
#> 94  2021-09-08 2021-10-01           NA
#> 95  2021-09-08 2021-11-01           NA
#> 96  2021-09-08 2021-12-01           NA
#> 97  2021-11-12 2020-01-01   5.27259936
#> 98  2021-11-12 2020-02-01   1.98641518
#> 99  2021-11-12 2020-03-01  -0.07925726
#> 100 2021-11-12 2020-04-01  -2.25861558
#> 101 2021-11-12 2020-05-01  -1.05262146
#> 102 2021-11-12 2020-06-01   1.55852225
#> 103 2021-11-12 2020-07-01  -1.65151266
#> 104 2021-11-12 2020-08-01  -1.82412472
#> 105 2021-11-12 2020-09-01  -2.05157390
#> 106 2021-11-12 2020-10-01  -2.06116932
#> 107 2021-11-12 2020-11-01  -2.11384122
#> 108 2021-11-12 2020-12-01  -3.04661266
#> 109 2021-11-12 2021-01-01  -2.77785883
#> 110 2021-11-12 2021-02-01  -6.27690886
#> 111 2021-11-12 2021-03-01  -6.16240078
#> 112 2021-11-12 2021-04-01  -5.08463642
#> 113 2021-11-12 2021-05-01  -3.91602380
#> 114 2021-11-12 2021-06-01  -3.24370385
#> 115 2021-11-12 2021-07-01  -8.04410157
#> 116 2021-11-12 2021-08-01  -9.16424539
#> 117 2021-11-12 2021-09-01  -6.67702091
#> 118 2021-11-12 2021-10-01 -13.29811621
#> 119 2021-11-12 2021-11-01 -15.44622901
#> 120 2021-11-12 2021-12-01           NA
#> 121 2021-11-17 2020-01-01   5.62781368
#> 122 2021-11-17 2020-02-01   1.86268555
#> 123 2021-11-17 2020-03-01   0.50033961
#> 124 2021-11-17 2020-04-01  -2.19881019
#> 125 2021-11-17 2020-05-01  -1.19660416
#> 126 2021-11-17 2020-06-01   1.44260617
#> 127 2021-11-17 2020-07-01  -1.88916911
#> 128 2021-11-17 2020-08-01  -2.02527869
#> 129 2021-11-17 2020-09-01  -1.98831076
#> 130 2021-11-17 2020-10-01  -3.29672209
#> 131 2021-11-17 2020-11-01  -2.21460363
#> 132 2021-11-17 2020-12-01  -2.65463448
#> 133 2021-11-17 2021-01-01  -2.33046569
#> 134 2021-11-17 2021-02-01  -5.46482405
#> 135 2021-11-17 2021-03-01  -5.79523993
#> 136 2021-11-17 2021-04-01  -4.59694631
#> 137 2021-11-17 2021-05-01  -5.72918812
#> 138 2021-11-17 2021-06-01  -2.05253969
#> 139 2021-11-17 2021-07-01  -6.86137581
#> 140 2021-11-17 2021-08-01  -8.55297530
#> 141 2021-11-17 2021-09-01  -7.90657210
#> 142 2021-11-17 2021-10-01 -13.37000920
#> 143 2021-11-17 2021-11-01 -12.86760251
#> 144 2021-11-17 2021-12-01           NA