Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01   7.0243885
#> 2   2020-02-17 2020-02-01   7.7974696
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01   6.3121098
#> 26  2020-06-05 2020-02-01   4.5245848
#> 27  2020-06-05 2020-03-01   1.6259654
#> 28  2020-06-05 2020-04-01   1.2382836
#> 29  2020-06-05 2020-05-01   6.1749680
#> 30  2020-06-05 2020-06-01   2.4706830
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01   6.3101801
#> 50  2021-04-17 2020-02-01   4.2647234
#> 51  2021-04-17 2020-03-01   2.2853171
#> 52  2021-04-17 2020-04-01  -1.3417630
#> 53  2021-04-17 2020-05-01   3.4887817
#> 54  2021-04-17 2020-06-01   2.6056797
#> 55  2021-04-17 2020-07-01  -1.3683380
#> 56  2021-04-17 2020-08-01   1.9401967
#> 57  2021-04-17 2020-09-01  -4.7072558
#> 58  2021-04-17 2020-10-01 -15.3035188
#> 59  2021-04-17 2020-11-01 -15.5432419
#> 60  2021-04-17 2020-12-01 -17.2904847
#> 61  2021-04-17 2021-01-01 -18.6732996
#> 62  2021-04-17 2021-02-01 -16.6596018
#> 63  2021-04-17 2021-03-01 -20.9430315
#> 64  2021-04-17 2021-04-01 -26.7916639
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01   5.9278466
#> 74  2021-09-08 2020-02-01   4.9419001
#> 75  2021-09-08 2020-03-01   3.0545228
#> 76  2021-09-08 2020-04-01  -1.4623961
#> 77  2021-09-08 2020-05-01   2.9179203
#> 78  2021-09-08 2020-06-01   3.6011081
#> 79  2021-09-08 2020-07-01  -0.1224890
#> 80  2021-09-08 2020-08-01   0.6148062
#> 81  2021-09-08 2020-09-01  -8.2888830
#> 82  2021-09-08 2020-10-01 -11.5566092
#> 83  2021-09-08 2020-11-01 -16.1019245
#> 84  2021-09-08 2020-12-01 -18.1123178
#> 85  2021-09-08 2021-01-01 -17.1853227
#> 86  2021-09-08 2021-02-01 -17.5429663
#> 87  2021-09-08 2021-03-01 -22.0817596
#> 88  2021-09-08 2021-04-01 -26.9450952
#> 89  2021-09-08 2021-05-01 -22.3013411
#> 90  2021-09-08 2021-06-01 -19.1336511
#> 91  2021-09-08 2021-07-01 -22.3216024
#> 92  2021-09-08 2021-08-01 -26.9068258
#> 93  2021-09-08 2021-09-01 -30.6787042
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01   5.8751727
#> 98  2021-11-12 2020-02-01   4.9837968
#> 99  2021-11-12 2020-03-01   1.9351154
#> 100 2021-11-12 2020-04-01  -1.1735789
#> 101 2021-11-12 2020-05-01   3.5868562
#> 102 2021-11-12 2020-06-01   3.7553326
#> 103 2021-11-12 2020-07-01   1.1386009
#> 104 2021-11-12 2020-08-01   1.1953021
#> 105 2021-11-12 2020-09-01  -6.6184952
#> 106 2021-11-12 2020-10-01 -12.9122103
#> 107 2021-11-12 2020-11-01 -16.2069049
#> 108 2021-11-12 2020-12-01 -16.1694588
#> 109 2021-11-12 2021-01-01 -17.9020785
#> 110 2021-11-12 2021-02-01 -18.1334021
#> 111 2021-11-12 2021-03-01 -21.1410798
#> 112 2021-11-12 2021-04-01 -25.6225601
#> 113 2021-11-12 2021-05-01 -20.9651867
#> 114 2021-11-12 2021-06-01 -18.0321092
#> 115 2021-11-12 2021-07-01 -20.6993031
#> 116 2021-11-12 2021-08-01 -25.1604896
#> 117 2021-11-12 2021-09-01 -27.2467141
#> 118 2021-11-12 2021-10-01 -30.4594282
#> 119 2021-11-12 2021-11-01 -29.4186946
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01   6.2303870
#> 122 2021-11-17 2020-02-01   4.8600671
#> 123 2021-11-17 2020-03-01   2.5147123
#> 124 2021-11-17 2020-04-01  -1.1137736
#> 125 2021-11-17 2020-05-01   3.4428735
#> 126 2021-11-17 2020-06-01   3.6394165
#> 127 2021-11-17 2020-07-01   0.9009444
#> 128 2021-11-17 2020-08-01   0.9941481
#> 129 2021-11-17 2020-09-01  -6.5552321
#> 130 2021-11-17 2020-10-01 -14.1477631
#> 131 2021-11-17 2020-11-01 -16.3076673
#> 132 2021-11-17 2020-12-01 -15.7774806
#> 133 2021-11-17 2021-01-01 -17.4546854
#> 134 2021-11-17 2021-02-01 -17.3213173
#> 135 2021-11-17 2021-03-01 -20.7739190
#> 136 2021-11-17 2021-04-01 -25.1348700
#> 137 2021-11-17 2021-05-01 -22.7783510
#> 138 2021-11-17 2021-06-01 -16.8409451
#> 139 2021-11-17 2021-07-01 -19.5165773
#> 140 2021-11-17 2021-08-01 -24.5492195
#> 141 2021-11-17 2021-09-01 -28.4762653
#> 142 2021-11-17 2021-10-01 -30.5313212
#> 143 2021-11-17 2021-11-01 -26.8400681
#> 144 2021-11-17 2021-12-01          NA