Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01   11.15473
#> 2   2020-02-17 2020-02-01   15.88589
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01   10.44245
#> 26  2020-06-05 2020-02-01   12.61301
#> 27  2020-06-05 2020-03-01   16.64620
#> 28  2020-06-05 2020-04-01   16.58526
#> 29  2020-06-05 2020-05-01   16.07196
#> 30  2020-06-05 2020-06-01   14.69198
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01   10.44052
#> 50  2021-04-17 2020-02-01   12.35315
#> 51  2021-04-17 2020-03-01   17.30555
#> 52  2021-04-17 2020-04-01   14.00521
#> 53  2021-04-17 2020-05-01   13.38577
#> 54  2021-04-17 2020-06-01   14.82698
#> 55  2021-04-17 2020-07-01   12.42638
#> 56  2021-04-17 2020-08-01   16.69154
#> 57  2021-04-17 2020-09-01   21.38487
#> 58  2021-04-17 2020-10-01   10.32588
#> 59  2021-04-17 2020-11-01   17.69456
#> 60  2021-04-17 2020-12-01   18.54227
#> 61  2021-04-17 2021-01-01   21.17446
#> 62  2021-04-17 2021-02-01   27.09769
#> 63  2021-04-17 2021-03-01   28.41397
#> 64  2021-04-17 2021-04-01   25.12469
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01   10.05819
#> 74  2021-09-08 2020-02-01   13.03033
#> 75  2021-09-08 2020-03-01   18.07475
#> 76  2021-09-08 2020-04-01   13.88458
#> 77  2021-09-08 2020-05-01   12.81491
#> 78  2021-09-08 2020-06-01   15.82240
#> 79  2021-09-08 2020-07-01   13.67223
#> 80  2021-09-08 2020-08-01   15.36615
#> 81  2021-09-08 2020-09-01   17.80324
#> 82  2021-09-08 2020-10-01   14.07279
#> 83  2021-09-08 2020-11-01   17.13588
#> 84  2021-09-08 2020-12-01   17.72044
#> 85  2021-09-08 2021-01-01   22.66244
#> 86  2021-09-08 2021-02-01   26.21432
#> 87  2021-09-08 2021-03-01   27.27524
#> 88  2021-09-08 2021-04-01   24.97126
#> 89  2021-09-08 2021-05-01   30.60700
#> 90  2021-09-08 2021-06-01   34.65269
#> 91  2021-09-08 2021-07-01   35.23749
#> 92  2021-09-08 2021-08-01   40.58526
#> 93  2021-09-08 2021-09-01   40.53592
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01   10.00552
#> 98  2021-11-12 2020-02-01   13.07222
#> 99  2021-11-12 2020-03-01   16.95535
#> 100 2021-11-12 2020-04-01   14.17339
#> 101 2021-11-12 2020-05-01   13.48385
#> 102 2021-11-12 2020-06-01   15.97663
#> 103 2021-11-12 2020-07-01   14.93332
#> 104 2021-11-12 2020-08-01   15.94665
#> 105 2021-11-12 2020-09-01   19.47363
#> 106 2021-11-12 2020-10-01   12.71719
#> 107 2021-11-12 2020-11-01   17.03090
#> 108 2021-11-12 2020-12-01   19.66330
#> 109 2021-11-12 2021-01-01   21.94568
#> 110 2021-11-12 2021-02-01   25.62388
#> 111 2021-11-12 2021-03-01   28.21592
#> 112 2021-11-12 2021-04-01   26.29379
#> 113 2021-11-12 2021-05-01   31.94316
#> 114 2021-11-12 2021-06-01   35.75424
#> 115 2021-11-12 2021-07-01   36.85979
#> 116 2021-11-12 2021-08-01   42.33159
#> 117 2021-11-12 2021-09-01   43.96791
#> 118 2021-11-12 2021-10-01   33.42048
#> 119 2021-11-12 2021-11-01   36.96172
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01   10.36073
#> 122 2021-11-17 2020-02-01   12.94849
#> 123 2021-11-17 2020-03-01   17.53494
#> 124 2021-11-17 2020-04-01   14.23320
#> 125 2021-11-17 2020-05-01   13.33987
#> 126 2021-11-17 2020-06-01   15.86071
#> 127 2021-11-17 2020-07-01   14.69566
#> 128 2021-11-17 2020-08-01   15.74549
#> 129 2021-11-17 2020-09-01   19.53689
#> 130 2021-11-17 2020-10-01   11.48164
#> 131 2021-11-17 2020-11-01   16.93013
#> 132 2021-11-17 2020-12-01   20.05528
#> 133 2021-11-17 2021-01-01   22.39307
#> 134 2021-11-17 2021-02-01   26.43597
#> 135 2021-11-17 2021-03-01   28.58308
#> 136 2021-11-17 2021-04-01   26.78148
#> 137 2021-11-17 2021-05-01   30.12999
#> 138 2021-11-17 2021-06-01   36.94540
#> 139 2021-11-17 2021-07-01   38.04251
#> 140 2021-11-17 2021-08-01   42.94286
#> 141 2021-11-17 2021-09-01   42.73836
#> 142 2021-11-17 2021-10-01   33.34859
#> 143 2021-11-17 2021-11-01   39.54034
#> 144 2021-11-17 2021-12-01         NA