Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time  obs_values
#> 1   2020-02-17 2020-01-01  8.06171222
#> 2   2020-02-17 2020-02-01  7.67977403
#> 3   2020-02-17 2020-03-01          NA
#> 4   2020-02-17 2020-04-01          NA
#> 5   2020-02-17 2020-05-01          NA
#> 6   2020-02-17 2020-06-01          NA
#> 7   2020-02-17 2020-07-01          NA
#> 8   2020-02-17 2020-08-01          NA
#> 9   2020-02-17 2020-09-01          NA
#> 10  2020-02-17 2020-10-01          NA
#> 11  2020-02-17 2020-11-01          NA
#> 12  2020-02-17 2020-12-01          NA
#> 13  2020-02-17 2021-01-01          NA
#> 14  2020-02-17 2021-02-01          NA
#> 15  2020-02-17 2021-03-01          NA
#> 16  2020-02-17 2021-04-01          NA
#> 17  2020-02-17 2021-05-01          NA
#> 18  2020-02-17 2021-06-01          NA
#> 19  2020-02-17 2021-07-01          NA
#> 20  2020-02-17 2021-08-01          NA
#> 21  2020-02-17 2021-09-01          NA
#> 22  2020-02-17 2021-10-01          NA
#> 23  2020-02-17 2021-11-01          NA
#> 24  2020-02-17 2021-12-01          NA
#> 25  2020-06-05 2020-01-01  7.34943358
#> 26  2020-06-05 2020-02-01  4.40688916
#> 27  2020-06-05 2020-03-01  1.50697910
#> 28  2020-06-05 2020-04-01  5.45196539
#> 29  2020-06-05 2020-05-01  7.78733102
#> 30  2020-06-05 2020-06-01  5.52611858
#> 31  2020-06-05 2020-07-01          NA
#> 32  2020-06-05 2020-08-01          NA
#> 33  2020-06-05 2020-09-01          NA
#> 34  2020-06-05 2020-10-01          NA
#> 35  2020-06-05 2020-11-01          NA
#> 36  2020-06-05 2020-12-01          NA
#> 37  2020-06-05 2021-01-01          NA
#> 38  2020-06-05 2021-02-01          NA
#> 39  2020-06-05 2021-03-01          NA
#> 40  2020-06-05 2021-04-01          NA
#> 41  2020-06-05 2021-05-01          NA
#> 42  2020-06-05 2021-06-01          NA
#> 43  2020-06-05 2021-07-01          NA
#> 44  2020-06-05 2021-08-01          NA
#> 45  2020-06-05 2021-09-01          NA
#> 46  2020-06-05 2021-10-01          NA
#> 47  2020-06-05 2021-11-01          NA
#> 48  2020-06-05 2021-12-01          NA
#> 49  2021-04-17 2020-01-01  7.34750384
#> 50  2021-04-17 2020-02-01  4.14702781
#> 51  2021-04-17 2020-03-01  2.16633084
#> 52  2021-04-17 2020-04-01  2.87191887
#> 53  2021-04-17 2020-05-01  5.10114475
#> 54  2021-04-17 2020-06-01  5.66111525
#> 55  2021-04-17 2020-07-01  0.28477765
#> 56  2021-04-17 2020-08-01  5.41569581
#> 57  2021-04-17 2020-09-01  2.55510572
#> 58  2021-04-17 2020-10-01  0.88379284
#> 59  2021-04-17 2020-11-01  2.89086883
#> 60  2021-04-17 2020-12-01  3.90780581
#> 61  2021-04-17 2021-01-01 -0.30373320
#> 62  2021-04-17 2021-02-01 -2.00388003
#> 63  2021-04-17 2021-03-01 -6.97870945
#> 64  2021-04-17 2021-04-01 -8.67226760
#> 65  2021-04-17 2021-05-01          NA
#> 66  2021-04-17 2021-06-01          NA
#> 67  2021-04-17 2021-07-01          NA
#> 68  2021-04-17 2021-08-01          NA
#> 69  2021-04-17 2021-09-01          NA
#> 70  2021-04-17 2021-10-01          NA
#> 71  2021-04-17 2021-11-01          NA
#> 72  2021-04-17 2021-12-01          NA
#> 73  2021-09-08 2020-01-01  6.96517029
#> 74  2021-09-08 2020-02-01  4.82420447
#> 75  2021-09-08 2020-03-01  2.93553657
#> 76  2021-09-08 2020-04-01  2.75128571
#> 77  2021-09-08 2020-05-01  4.53028335
#> 78  2021-09-08 2020-06-01  6.65654365
#> 79  2021-09-08 2020-07-01  1.53062664
#> 80  2021-09-08 2020-08-01  4.09030535
#> 81  2021-09-08 2020-09-01 -1.02652148
#> 82  2021-09-08 2020-10-01  4.63070245
#> 83  2021-09-08 2020-11-01  2.33218623
#> 84  2021-09-08 2020-12-01  3.08597271
#> 85  2021-09-08 2021-01-01  1.18424372
#> 86  2021-09-08 2021-02-01 -2.88724452
#> 87  2021-09-08 2021-03-01 -8.11743749
#> 88  2021-09-08 2021-04-01 -8.82569889
#> 89  2021-09-08 2021-05-01 -3.98175978
#> 90  2021-09-08 2021-06-01 -1.18988481
#> 91  2021-09-08 2021-07-01 -4.51823108
#> 92  2021-09-08 2021-08-01 -2.51342608
#> 93  2021-09-08 2021-09-01 -5.71453081
#> 94  2021-09-08 2021-10-01          NA
#> 95  2021-09-08 2021-11-01          NA
#> 96  2021-09-08 2021-12-01          NA
#> 97  2021-11-12 2020-01-01  6.91249643
#> 98  2021-11-12 2020-02-01  4.86610115
#> 99  2021-11-12 2020-03-01  1.81612912
#> 100 2021-11-12 2020-04-01  3.04010289
#> 101 2021-11-12 2020-05-01  5.19921924
#> 102 2021-11-12 2020-06-01  6.81076814
#> 103 2021-11-12 2020-07-01  2.79171653
#> 104 2021-11-12 2020-08-01  4.67080124
#> 105 2021-11-12 2020-09-01  0.64386634
#> 106 2021-11-12 2020-10-01  3.27510138
#> 107 2021-11-12 2020-11-01  2.22720585
#> 108 2021-11-12 2020-12-01  5.02883177
#> 109 2021-11-12 2021-01-01  0.46748792
#> 110 2021-11-12 2021-02-01 -3.47768028
#> 111 2021-11-12 2021-03-01 -7.17675778
#> 112 2021-11-12 2021-04-01 -7.50316381
#> 113 2021-11-12 2021-05-01 -2.64560542
#> 114 2021-11-12 2021-06-01 -0.08834287
#> 115 2021-11-12 2021-07-01 -2.89593171
#> 116 2021-11-12 2021-08-01 -0.76708982
#> 117 2021-11-12 2021-09-01 -2.28254063
#> 118 2021-11-12 2021-10-01 -0.99264342
#> 119 2021-11-12 2021-11-01 -4.57913103
#> 120 2021-11-12 2021-12-01          NA
#> 121 2021-11-17 2020-01-01  7.26771075
#> 122 2021-11-17 2020-02-01  4.74237151
#> 123 2021-11-17 2020-03-01  2.39572598
#> 124 2021-11-17 2020-04-01  3.09990828
#> 125 2021-11-17 2020-05-01  5.05523654
#> 126 2021-11-17 2020-06-01  6.69485206
#> 127 2021-11-17 2020-07-01  2.55406007
#> 128 2021-11-17 2020-08-01  4.46964727
#> 129 2021-11-17 2020-09-01  0.70712947
#> 130 2021-11-17 2020-10-01  2.03954861
#> 131 2021-11-17 2020-11-01  2.12644344
#> 132 2021-11-17 2020-12-01  5.42080994
#> 133 2021-11-17 2021-01-01  0.91488107
#> 134 2021-11-17 2021-02-01 -2.66559547
#> 135 2021-11-17 2021-03-01 -6.80959693
#> 136 2021-11-17 2021-04-01 -7.01547371
#> 137 2021-11-17 2021-05-01 -4.45876973
#> 138 2021-11-17 2021-06-01  1.10282129
#> 139 2021-11-17 2021-07-01 -1.71320595
#> 140 2021-11-17 2021-08-01 -0.15581974
#> 141 2021-11-17 2021-09-01 -3.51209181
#> 142 2021-11-17 2021-10-01 -1.06453642
#> 143 2021-11-17 2021-11-01 -2.00050454
#> 144 2021-11-17 2021-12-01          NA