Check long format

Check long format

Usage

check_long(x, date_format = "%Y-%m-%d")

Arguments

x

a formatted data.frame containing the input in the long format

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but with column and date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
check_long(long_format)
#>        revdate       time obs_values
#> 1   2020-02-17 2020-01-01  12.914333
#> 2   2020-02-17 2020-02-01  15.664932
#> 3   2020-02-17 2020-03-01         NA
#> 4   2020-02-17 2020-04-01         NA
#> 5   2020-02-17 2020-05-01         NA
#> 6   2020-02-17 2020-06-01         NA
#> 7   2020-02-17 2020-07-01         NA
#> 8   2020-02-17 2020-08-01         NA
#> 9   2020-02-17 2020-09-01         NA
#> 10  2020-02-17 2020-10-01         NA
#> 11  2020-02-17 2020-11-01         NA
#> 12  2020-02-17 2020-12-01         NA
#> 13  2020-02-17 2021-01-01         NA
#> 14  2020-02-17 2021-02-01         NA
#> 15  2020-02-17 2021-03-01         NA
#> 16  2020-02-17 2021-04-01         NA
#> 17  2020-02-17 2021-05-01         NA
#> 18  2020-02-17 2021-06-01         NA
#> 19  2020-02-17 2021-07-01         NA
#> 20  2020-02-17 2021-08-01         NA
#> 21  2020-02-17 2021-09-01         NA
#> 22  2020-02-17 2021-10-01         NA
#> 23  2020-02-17 2021-11-01         NA
#> 24  2020-02-17 2021-12-01         NA
#> 25  2020-06-05 2020-01-01  12.202054
#> 26  2020-06-05 2020-02-01  12.392048
#> 27  2020-06-05 2020-03-01  10.256745
#> 28  2020-06-05 2020-04-01  12.454618
#> 29  2020-06-05 2020-05-01  11.134856
#> 30  2020-06-05 2020-06-01   4.767894
#> 31  2020-06-05 2020-07-01         NA
#> 32  2020-06-05 2020-08-01         NA
#> 33  2020-06-05 2020-09-01         NA
#> 34  2020-06-05 2020-10-01         NA
#> 35  2020-06-05 2020-11-01         NA
#> 36  2020-06-05 2020-12-01         NA
#> 37  2020-06-05 2021-01-01         NA
#> 38  2020-06-05 2021-02-01         NA
#> 39  2020-06-05 2021-03-01         NA
#> 40  2020-06-05 2021-04-01         NA
#> 41  2020-06-05 2021-05-01         NA
#> 42  2020-06-05 2021-06-01         NA
#> 43  2020-06-05 2021-07-01         NA
#> 44  2020-06-05 2021-08-01         NA
#> 45  2020-06-05 2021-09-01         NA
#> 46  2020-06-05 2021-10-01         NA
#> 47  2020-06-05 2021-11-01         NA
#> 48  2020-06-05 2021-12-01         NA
#> 49  2021-04-17 2020-01-01  12.200125
#> 50  2021-04-17 2020-02-01  12.132186
#> 51  2021-04-17 2020-03-01  10.916097
#> 52  2021-04-17 2020-04-01   9.874572
#> 53  2021-04-17 2020-05-01   8.448669
#> 54  2021-04-17 2020-06-01   4.902891
#> 55  2021-04-17 2020-07-01   5.293044
#> 56  2021-04-17 2020-08-01   9.599482
#> 57  2021-04-17 2020-09-01  14.370306
#> 58  2021-04-17 2020-10-01  11.354043
#> 59  2021-04-17 2020-11-01  15.240121
#> 60  2021-04-17 2020-12-01  13.231272
#> 61  2021-04-17 2021-01-01  13.192875
#> 62  2021-04-17 2021-02-01  17.731806
#> 63  2021-04-17 2021-03-01  15.827864
#> 64  2021-04-17 2021-04-01  16.214866
#> 65  2021-04-17 2021-05-01         NA
#> 66  2021-04-17 2021-06-01         NA
#> 67  2021-04-17 2021-07-01         NA
#> 68  2021-04-17 2021-08-01         NA
#> 69  2021-04-17 2021-09-01         NA
#> 70  2021-04-17 2021-10-01         NA
#> 71  2021-04-17 2021-11-01         NA
#> 72  2021-04-17 2021-12-01         NA
#> 73  2021-09-08 2020-01-01  11.817791
#> 74  2021-09-08 2020-02-01  12.809363
#> 75  2021-09-08 2020-03-01  11.685302
#> 76  2021-09-08 2020-04-01   9.753938
#> 77  2021-09-08 2020-05-01   7.877808
#> 78  2021-09-08 2020-06-01   5.898319
#> 79  2021-09-08 2020-07-01   6.538893
#> 80  2021-09-08 2020-08-01   8.274092
#> 81  2021-09-08 2020-09-01  10.788679
#> 82  2021-09-08 2020-10-01  15.100953
#> 83  2021-09-08 2020-11-01  14.681438
#> 84  2021-09-08 2020-12-01  12.409438
#> 85  2021-09-08 2021-01-01  14.680852
#> 86  2021-09-08 2021-02-01  16.848441
#> 87  2021-09-08 2021-03-01  14.689136
#> 88  2021-09-08 2021-04-01  16.061435
#> 89  2021-09-08 2021-05-01  13.708009
#> 90  2021-09-08 2021-06-01  11.900877
#> 91  2021-09-08 2021-07-01  18.509510
#> 92  2021-09-08 2021-08-01  20.443692
#> 93  2021-09-08 2021-09-01  22.380538
#> 94  2021-09-08 2021-10-01         NA
#> 95  2021-09-08 2021-11-01         NA
#> 96  2021-09-08 2021-12-01         NA
#> 97  2021-11-12 2020-01-01  11.765117
#> 98  2021-11-12 2020-02-01  12.851260
#> 99  2021-11-12 2020-03-01  10.565895
#> 100 2021-11-12 2020-04-01  10.042756
#> 101 2021-11-12 2020-05-01   8.546744
#> 102 2021-11-12 2020-06-01   6.052544
#> 103 2021-11-12 2020-07-01   7.799983
#> 104 2021-11-12 2020-08-01   8.854588
#> 105 2021-11-12 2020-09-01  12.459067
#> 106 2021-11-12 2020-10-01  13.745352
#> 107 2021-11-12 2020-11-01  14.576458
#> 108 2021-11-12 2020-12-01  14.352297
#> 109 2021-11-12 2021-01-01  13.964096
#> 110 2021-11-12 2021-02-01  16.258006
#> 111 2021-11-12 2021-03-01  15.629815
#> 112 2021-11-12 2021-04-01  17.383970
#> 113 2021-11-12 2021-05-01  15.044163
#> 114 2021-11-12 2021-06-01  13.002419
#> 115 2021-11-12 2021-07-01  20.131809
#> 116 2021-11-12 2021-08-01  22.190028
#> 117 2021-11-12 2021-09-01  25.812529
#> 118 2021-11-12 2021-10-01  29.726740
#> 119 2021-11-12 2021-11-01  27.401219
#> 120 2021-11-12 2021-12-01         NA
#> 121 2021-11-17 2020-01-01  12.120332
#> 122 2021-11-17 2020-02-01  12.727530
#> 123 2021-11-17 2020-03-01  11.145492
#> 124 2021-11-17 2020-04-01  10.102561
#> 125 2021-11-17 2020-05-01   8.402761
#> 126 2021-11-17 2020-06-01   5.936628
#> 127 2021-11-17 2020-07-01   7.562326
#> 128 2021-11-17 2020-08-01   8.653434
#> 129 2021-11-17 2020-09-01  12.522330
#> 130 2021-11-17 2020-10-01  12.509799
#> 131 2021-11-17 2020-11-01  14.475695
#> 132 2021-11-17 2020-12-01  14.744276
#> 133 2021-11-17 2021-01-01  14.411489
#> 134 2021-11-17 2021-02-01  17.070090
#> 135 2021-11-17 2021-03-01  15.996976
#> 136 2021-11-17 2021-04-01  17.871660
#> 137 2021-11-17 2021-05-01  13.230999
#> 138 2021-11-17 2021-06-01  14.193583
#> 139 2021-11-17 2021-07-01  21.314535
#> 140 2021-11-17 2021-08-01  22.801298
#> 141 2021-11-17 2021-09-01  24.582977
#> 142 2021-11-17 2021-10-01  29.654847
#> 143 2021-11-17 2021-11-01  29.979846
#> 144 2021-11-17 2021-12-01         NA