Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04  2022-01-15
#> Jan 2020    9.41874   10.57558  10.482348  10.891238  11.261149 11.02738120
#> Feb 2020         NA   12.53513  11.475526  11.653165  12.320607 11.97428923
#> Mar 2020         NA   15.19634  18.052675  17.956860  17.836393 17.80020881
#> Apr 2020         NA   22.02160  17.266750  18.183574  18.294866 18.24248810
#> May 2020         NA   15.97667  20.213963  20.387971  20.333795 20.56654155
#> Jun 2020         NA   18.66569  15.057888  14.731666  14.271121 14.47467415
#> Jul 2020         NA   12.08845  14.489935  14.087796  14.018223 14.06778417
#> Aug 2020         NA         NA   9.320205  11.421038  11.910907 11.10726754
#> Sep 2020         NA         NA  10.921010  11.047438  10.764418 11.38936378
#> Oct 2020         NA         NA  13.422934   8.699095   9.506107 10.43331370
#> Nov 2020         NA         NA         NA  12.115476  12.979488 10.68914848
#> Dec 2020         NA         NA         NA   7.128602  10.334647  9.84831865
#> Jan 2021         NA         NA         NA   4.359173   3.561984  4.27046629
#> Feb 2021         NA         NA         NA   7.441437   5.377786  2.69881373
#> Mar 2021         NA         NA         NA  12.417635  12.328019 13.77703393
#> Apr 2021         NA         NA         NA  10.605844  14.637883 14.26560569
#> May 2021         NA         NA         NA  17.530844  14.914594 14.64334741
#> Jun 2021         NA         NA         NA  12.852278  13.105257 12.62668048
#> Jul 2021         NA         NA         NA         NA  12.672345  9.78930588
#> Aug 2021         NA         NA         NA         NA         NA 12.31553834
#> Sep 2021         NA         NA         NA         NA         NA  2.62284286
#> Oct 2021         NA         NA         NA         NA         NA  3.15546240
#> Nov 2021         NA         NA         NA         NA         NA  1.80942026
#> Dec 2021         NA         NA         NA         NA         NA -0.01451322