Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14  2020-07-25 2020-10-16 2021-06-16  2021-07-04 2022-01-15
#> Jan 2020   3.546197   4.7030323   4.609805   5.018694   5.3886054   5.154837
#> Feb 2020         NA  -0.9282904  -1.987893  -1.810254  -1.1428117  -1.489129
#> Mar 2020         NA  -9.9666772  -7.110338  -7.206153  -7.3266201  -7.362804
#> Apr 2020         NA  -1.8288070  -6.583654  -5.666829  -5.5555381  -5.607916
#> May 2020         NA -12.1136705  -7.876379  -7.702371  -7.7565469  -7.523800
#> Jun 2020         NA  -1.3055416  -4.913340  -5.239563  -5.7001076  -5.496554
#> Jul 2020         NA  -5.8559208  -3.454438  -3.856576  -3.9261490  -3.876588
#> Aug 2020         NA          NA  -7.239022  -5.138188  -4.6483194  -5.451959
#> Sep 2020         NA          NA  -7.119679  -6.993251  -7.2762709  -6.651325
#> Oct 2020         NA          NA  -4.396733  -9.120572  -8.3135593  -7.386353
#> Nov 2020         NA          NA         NA  -9.092025  -8.2280124 -10.518352
#> Dec 2020         NA          NA         NA  -8.448160  -5.2421150  -5.728444
#> Jan 2021         NA          NA         NA  -7.020772  -7.8179619  -7.109479
#> Feb 2021         NA          NA         NA -13.470142 -15.5337935 -18.212766
#> Mar 2021         NA          NA         NA -23.519447 -23.6090624 -22.160048
#> Apr 2021         NA          NA         NA -21.396217 -17.3641784 -17.736455
#> May 2021         NA          NA         NA -12.434962 -15.0512114 -15.322458
#> Jun 2021         NA          NA         NA  -6.629375  -6.3763963  -6.854973
#> Jul 2021         NA          NA         NA         NA   0.4435525  -2.439486
#> Aug 2021         NA          NA         NA         NA          NA   2.158903
#> Sep 2021         NA          NA         NA         NA          NA  -2.420863
#> Oct 2021         NA          NA         NA         NA          NA   4.072635
#> Nov 2021         NA          NA         NA         NA          NA   1.490674
#> Dec 2021         NA          NA         NA         NA          NA  -3.316186