Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12   2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   9.839960 11.343685455  11.268237 11.0559458 10.9435569 11.1755824
#> Feb 2020   6.336370  6.721652269   8.007823  7.8864918  8.1096169  7.9645664
#> Mar 2020   8.141116  6.195931875   6.006717  5.8499216  6.0556249  5.8779695
#> Apr 2020   3.472063  5.005785051   4.896924  5.2960017  4.7091070  4.8465153
#> May 2020         NA  0.233831147   1.987925  0.7368053  0.6156707  0.4891403
#> Jun 2020         NA -0.001971821   1.218041  2.1061110  1.5686791  1.9810139
#> Jul 2020         NA           NA   2.377944  3.7778399  4.3903192  4.9855319
#> Aug 2020         NA           NA   9.845726  6.0950060  7.4545869  7.6758082
#> Sep 2020         NA           NA         NA  6.4136708  9.3263779  9.4858092
#> Oct 2020         NA           NA         NA  9.2976041  7.8874741  6.5653638
#> Nov 2020         NA           NA         NA  1.5545844  3.0859705  2.7775755
#> Dec 2020         NA           NA         NA  1.7707522  4.9138978  3.9063030
#> Jan 2021         NA           NA         NA  8.4636310  7.7534316  6.5772521
#> Feb 2021         NA           NA         NA  8.7752108 11.6018959  9.5232034
#> Mar 2021         NA           NA         NA 17.0125281 17.7198208 16.5130934
#> Apr 2021         NA           NA         NA 17.9469867 16.4112346 18.6543226
#> May 2021         NA           NA         NA         NA  9.5338497  9.8989198
#> Jun 2021         NA           NA         NA         NA  8.5388510 11.4330316
#> Jul 2021         NA           NA         NA         NA  6.1766274  9.5175991
#> Aug 2021         NA           NA         NA         NA  8.0519869  9.4895953
#> Sep 2021         NA           NA         NA         NA         NA 13.1096954
#> Oct 2021         NA           NA         NA         NA         NA 13.6483014
#> Nov 2021         NA           NA         NA         NA         NA 10.5568718
#> Dec 2021         NA           NA         NA         NA         NA 10.4483482