Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16  2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   11.58166  13.085389 13.00994064  12.797649 12.6852601 12.9172856
#> Feb 2020   11.80459  12.189868 13.47603825  13.354708 13.5778327 13.4327822
#> Mar 2020   13.00756  11.062380 10.87316535  10.716370 10.9220735 10.7444181
#> Apr 2020   11.32771  12.861429 12.75256841  13.151646 12.5647514 12.7021597
#> May 2020         NA   9.427093 11.18118682   9.930067  9.8089323  9.6824019
#> Jun 2020         NA   5.993302  7.21331563   8.101385  7.5639534  7.9762882
#> Jul 2020         NA         NA -0.07227405   1.327622  1.9401009  2.5353136
#> Aug 2020         NA         NA  6.16282033   2.412100  3.7716812  3.9929025
#> Sep 2020         NA         NA          NA  -3.327017 -0.4143103 -0.2548789
#> Oct 2020         NA         NA          NA   1.172097 -0.2380326 -1.5601429
#> Nov 2020         NA         NA          NA  -3.117869 -1.5864828 -1.8948778
#> Dec 2020         NA         NA          NA  -2.291266  0.8518796 -0.1557151
#> Jan 2021         NA         NA          NA   8.186894  7.4766948  6.3005154
#> Feb 2021         NA         NA          NA   9.280449 12.1071344 10.0284419
#> Mar 2021         NA         NA          NA  12.566949 13.2742417 12.0675143
#> Apr 2021         NA         NA          NA  18.300975 16.7652227 19.0083106
#> May 2021         NA         NA          NA         NA 14.4518983 14.8169684
#> Jun 2021         NA         NA          NA         NA 12.7462031 15.6403837
#> Jul 2021         NA         NA          NA         NA  4.0752288  7.4162005
#> Aug 2021         NA         NA          NA         NA  5.4370166  6.8746250
#> Sep 2021         NA         NA          NA         NA         NA  3.8837198
#> Oct 2021         NA         NA          NA         NA         NA  5.5800677
#> Nov 2021         NA         NA          NA         NA         NA  6.6944687
#> Dec 2021         NA         NA          NA         NA         NA  1.8513882