Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02  2021-08-01   2021-12-08
#> Jan 2020  3.3913183  4.8950438  4.8195958  4.6073042   4.4949153   4.72694079
#> Feb 2020 -0.5439313 -0.1586488  1.1275214  1.0061907   1.2293158   1.08426534
#> Mar 2020  1.7225148 -0.2226692 -0.4118844 -0.5686796  -0.3629762  -0.54063164
#> Apr 2020 -5.7349016 -4.2011794 -4.3100404 -3.9109627  -4.4978574  -4.36044906
#> May 2020         NA -6.3991902 -4.6450961 -5.8962160  -6.0173506  -6.14388105
#> Jun 2020         NA -4.9957074 -3.7756942 -2.8876246  -3.4250564  -3.01272162
#> Jul 2020         NA         NA -3.9341484 -2.5342527  -1.9217734  -1.32656069
#> Aug 2020         NA         NA  0.7592494 -2.9914706  -1.6318897  -1.41066836
#> Sep 2020         NA         NA         NA -4.4524326  -1.5397256  -1.38029422
#> Oct 2020         NA         NA         NA  2.5300933   1.1199633  -0.20214695
#> Nov 2020         NA         NA         NA  0.8143451   2.3457311   2.03733612
#> Dec 2020         NA         NA         NA  5.5783971   8.7215426   7.71394790
#> Jan 2021         NA         NA         NA  3.1107054   2.4005060   1.22432652
#> Feb 2021         NA         NA         NA -5.6740852  -2.8474001  -4.92609263
#> Mar 2021         NA         NA         NA -9.6037572  -8.8964645 -10.10319193
#> Apr 2021         NA         NA         NA -7.6734061  -9.2091581  -6.96607022
#> May 2021         NA         NA         NA         NA -11.5979766 -11.23290656
#> Jun 2021         NA         NA         NA         NA  -8.9107024  -6.01652183
#> Jul 2021         NA         NA         NA         NA  -5.6990136  -2.35804191
#> Aug 2021         NA         NA         NA         NA  -5.8482028  -4.41059439
#> Sep 2021         NA         NA         NA         NA          NA  -4.48447464
#> Oct 2021         NA         NA         NA         NA          NA  -1.48093575
#> Nov 2021         NA         NA         NA         NA          NA  -0.02374917
#> Dec 2021         NA         NA         NA         NA          NA  -2.03477160