Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020  10.201743  11.705468  11.630020  11.417729  11.305340  11.537365
#> Feb 2020   5.816997   6.202279   7.488449   7.367119   7.590244   7.445193
#> Mar 2020  14.926696  12.981512  12.792297  12.635502  12.841205  12.663550
#> Apr 2020  15.901490  17.435212  17.326351  17.725429  17.138534  17.275942
#> May 2020         NA  21.913419  23.667513  22.416393  22.295258  22.168728
#> Jun 2020         NA  18.931219  20.151232  21.039302  20.501870  20.914205
#> Jul 2020         NA         NA  14.242757  15.642652  16.255132  16.850344
#> Aug 2020         NA         NA  17.820315  14.069595  15.429176  15.650398
#> Sep 2020         NA         NA         NA  20.784310  23.697018  23.856449
#> Oct 2020         NA         NA         NA  23.881768  22.471638  21.149527
#> Nov 2020         NA         NA         NA  21.284331  22.815717  22.507322
#> Dec 2020         NA         NA         NA  22.627899  25.771045  24.763450
#> Jan 2021         NA         NA         NA  33.314021  32.603821  31.427642
#> Feb 2021         NA         NA         NA  28.220275  31.046960  28.968267
#> Mar 2021         NA         NA         NA  33.073308  33.780601  32.573873
#> Apr 2021         NA         NA         NA  33.002857  31.467105  33.710193
#> May 2021         NA         NA         NA         NA  36.166637  36.531707
#> Jun 2021         NA         NA         NA         NA  36.651723  39.545903
#> Jul 2021         NA         NA         NA         NA  33.772505  37.113477
#> Aug 2021         NA         NA         NA         NA  36.967700  38.405308
#> Sep 2021         NA         NA         NA         NA         NA  48.828521
#> Oct 2021         NA         NA         NA         NA         NA  48.763595
#> Nov 2021         NA         NA         NA         NA         NA  55.314881
#> Dec 2021         NA         NA         NA         NA         NA  52.744232