Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   12.17564  13.332478  13.239251   13.64814   14.01805   13.78428
#> Feb 2020         NA  13.815682  12.756079   12.93372   13.60116   13.25484
#> Mar 2020         NA  11.519799  14.376138   14.28032   14.15986   14.12367
#> Apr 2020         NA  14.495811   9.740965   10.65779   10.76908   10.71670
#> May 2020         NA   6.323403  10.560695   10.73470   10.68053   10.91327
#> Jun 2020         NA  16.497241  12.889442   12.56322   12.10268   12.30623
#> Jul 2020         NA  10.884023  13.285506   12.88337   12.81379   12.86336
#> Aug 2020         NA         NA  12.944277   15.04511   15.53498   14.73134
#> Sep 2020         NA         NA  14.910313   15.03674   14.75372   15.37867
#> Oct 2020         NA         NA  21.200677   16.47684   17.28385   18.21106
#> Nov 2020         NA         NA         NA   20.95804   21.82205   19.53171
#> Dec 2020         NA         NA         NA   18.87544   22.08148   21.59515
#> Jan 2021         NA         NA         NA   31.20741   30.41022   31.11870
#> Feb 2021         NA         NA         NA   34.73017   32.66652   29.98754
#> Mar 2021         NA         NA         NA   30.68328   30.59367   32.04268
#> Apr 2021         NA         NA         NA   22.17858   26.21062   25.83834
#> May 2021         NA         NA         NA   32.75551   30.13926   29.86801
#> Jun 2021         NA         NA         NA   33.08790   33.34088   32.86231
#> Jul 2021         NA         NA         NA         NA   36.62523   33.74219
#> Aug 2021         NA         NA         NA         NA         NA   38.92539
#> Sep 2021         NA         NA         NA         NA         NA   27.25081
#> Oct 2021         NA         NA         NA         NA         NA   36.61781
#> Nov 2021         NA         NA         NA         NA         NA   41.25404
#> Dec 2021         NA         NA         NA         NA         NA   39.63625