Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   12.12419   13.28103   13.18780   13.59669   13.96660   13.73283
#> Feb 2020         NA   12.50306   11.44346   11.62110   12.28854   11.94223
#> Mar 2020         NA   12.11048   14.96682   14.87100   14.75054   14.71435
#> Apr 2020         NA   22.07423   17.31939   18.23621   18.34750   18.29512
#> May 2020         NA   16.40876   20.64605   20.82006   20.76588   20.99863
#> Jun 2020         NA   30.91019   27.30239   26.97617   26.51562   26.71917
#> Jul 2020         NA   29.40646   31.80795   31.40581   31.33624   31.38580
#> Aug 2020         NA         NA   27.48145   29.58228   30.07215   29.26851
#> Sep 2020         NA         NA   29.64192   29.76835   29.48533   30.11027
#> Oct 2020         NA         NA   30.60046   25.87663   26.68364   27.61084
#> Nov 2020         NA         NA         NA   26.37073   27.23474   24.94440
#> Dec 2020         NA         NA         NA   26.11687   29.32291   28.83658
#> Jan 2021         NA         NA         NA   27.73631   26.93912   27.64760
#> Feb 2021         NA         NA         NA   26.15150   24.08785   21.40888
#> Mar 2021         NA         NA         NA   25.58299   25.49337   26.94239
#> Apr 2021         NA         NA         NA   20.32852   24.36056   23.98828
#> May 2021         NA         NA         NA   27.90734   25.29109   25.01984
#> Jun 2021         NA         NA         NA   30.76354   31.01651   30.53794
#> Jul 2021         NA         NA         NA         NA   38.41678   35.53374
#> Aug 2021         NA         NA         NA         NA         NA   38.87094
#> Sep 2021         NA         NA         NA         NA         NA   30.95117
#> Oct 2021         NA         NA         NA         NA         NA   32.83800
#> Nov 2021         NA         NA         NA         NA         NA   34.67876
#> Dec 2021         NA         NA         NA         NA         NA   35.22372