Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   14.17745   15.68117   15.60572   15.39343   15.28104   15.51307
#> Feb 2020   14.58595   14.97123   16.25740   16.13607   16.35919   16.21414
#> Mar 2020   25.27858   23.33340   23.14418   22.98739   23.19309   23.01543
#> Apr 2020   19.73779   21.27151   21.16265   21.56173   20.97484   21.11224
#> May 2020         NA   24.40125   26.15534   24.90422   24.78309   24.65656
#> Jun 2020         NA   20.52122   21.74123   22.62930   22.09187   22.50421
#> Jul 2020         NA         NA   17.50722   18.90712   19.51960   20.11481
#> Aug 2020         NA         NA   19.51902   15.76830   17.12788   17.34910
#> Sep 2020         NA         NA         NA   15.84810   18.76081   18.92024
#> Oct 2020         NA         NA         NA   27.64707   26.23694   24.91483
#> Nov 2020         NA         NA         NA   18.22724   19.75862   19.45023
#> Dec 2020         NA         NA         NA   17.76525   20.90840   19.90080
#> Jan 2021         NA         NA         NA   27.58638   26.87618   25.70000
#> Feb 2021         NA         NA         NA   26.22952   29.05620   26.97751
#> Mar 2021         NA         NA         NA   34.88187   35.58916   34.38243
#> Apr 2021         NA         NA         NA   27.38814   25.85239   28.09548
#> May 2021         NA         NA         NA         NA   30.44152   30.80659
#> Jun 2021         NA         NA         NA         NA   34.62873   37.52291
#> Jul 2021         NA         NA         NA         NA   30.74987   34.09084
#> Aug 2021         NA         NA         NA         NA   31.27584   32.71344
#> Sep 2021         NA         NA         NA         NA         NA   36.72568
#> Oct 2021         NA         NA         NA         NA         NA   40.21934
#> Nov 2021         NA         NA         NA         NA         NA   36.14460
#> Dec 2021         NA         NA         NA         NA         NA   35.60645