Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020  13.148897  14.652623   14.57717   14.36488   14.25249   14.48452
#> Feb 2020   9.503406   9.888689   11.17486   11.05353   11.27665   11.13160
#> Mar 2020  17.738764  15.793580   15.60436   15.44757   15.65327   15.47562
#> Apr 2020  18.498265  20.031987   19.92313   20.32220   19.73531   19.87272
#> May 2020         NA  23.096213   24.85031   23.59919   23.47805   23.35152
#> Jun 2020         NA  25.555117   26.77513   27.66320   27.12577   27.53810
#> Jul 2020         NA         NA   27.11856   28.51846   29.13094   29.72615
#> Aug 2020         NA         NA   26.88829   23.13757   24.49715   24.71837
#> Sep 2020         NA         NA         NA   25.07053   27.98324   28.14267
#> Oct 2020         NA         NA         NA   30.37060   28.96047   27.63836
#> Nov 2020         NA         NA         NA   24.62519   26.15657   25.84818
#> Dec 2020         NA         NA         NA   29.22584   32.36898   31.36139
#> Jan 2021         NA         NA         NA   40.68349   39.97329   38.79711
#> Feb 2021         NA         NA         NA   33.09232   35.91900   33.84031
#> Mar 2021         NA         NA         NA   38.02252   38.72982   37.52309
#> Apr 2021         NA         NA         NA   40.96428   39.42853   41.67162
#> May 2021         NA         NA         NA         NA   44.10996   44.47503
#> Jun 2021         NA         NA         NA         NA   53.02176   55.91594
#> Jul 2021         NA         NA         NA         NA   51.90579   55.24676
#> Aug 2021         NA         NA         NA         NA   47.37037   48.80797
#> Sep 2021         NA         NA         NA         NA         NA   46.39536
#> Oct 2021         NA         NA         NA         NA         NA   41.42891
#> Nov 2021         NA         NA         NA         NA         NA   42.17105
#> Dec 2021         NA         NA         NA         NA         NA   39.44337