Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   7.317405  8.8211307   8.745683   8.533391   8.421002   8.653028
#> Feb 2020   5.105299  5.4905815   6.776752   6.655421   6.878546   6.733496
#> Mar 2020   9.220507  7.2753233   7.086108   6.929313   7.135016   6.957361
#> Apr 2020   9.100112 10.6338344  10.524973  10.924051  10.337156  10.474565
#> May 2020         NA  7.6153043   9.369398   8.118278   7.997144   7.870613
#> Jun 2020         NA  0.3145443   1.534557   2.422627   1.885195   2.297530
#> Jul 2020         NA         NA   3.908932   5.308827   5.921307   6.516519
#> Aug 2020         NA         NA  13.376368   9.625648  10.985229  11.206450
#> Sep 2020         NA         NA         NA  11.782104  14.694811  14.854242
#> Oct 2020         NA         NA         NA  20.260061  18.849931  17.527821
#> Nov 2020         NA         NA         NA  16.289206  17.820592  17.512197
#> Dec 2020         NA         NA         NA  15.110976  18.254122  17.246527
#> Jan 2021         NA         NA         NA  15.765467  15.055268  13.879088
#> Feb 2021         NA         NA         NA   9.313296  12.139981  10.061289
#> Mar 2021         NA         NA         NA  16.814159  17.521452  16.314725
#> Apr 2021         NA         NA         NA  17.731841  16.196089  18.439177
#> May 2021         NA         NA         NA         NA  14.278592  14.643662
#> Jun 2021         NA         NA         NA         NA  12.989061  15.883242
#> Jul 2021         NA         NA         NA         NA   8.941038  12.282009
#> Aug 2021         NA         NA         NA         NA  11.654986  13.092595
#> Sep 2021         NA         NA         NA         NA         NA  13.758839
#> Oct 2021         NA         NA         NA         NA         NA  14.492984
#> Nov 2021         NA         NA         NA         NA         NA  18.396986
#> Dec 2021         NA         NA         NA         NA         NA  15.158132