Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   12.24202  13.398856  13.305628  13.714518  14.084429  13.850661
#> Feb 2020         NA  15.747515  14.687913  14.865552  15.532994  15.186676
#> Mar 2020         NA   5.220924   8.077263   7.981448   7.860981   7.824797
#> Apr 2020         NA  12.382743   7.627896   8.544721   8.656012   8.603634
#> May 2020         NA   1.799698   6.036990   6.210998   6.156822   6.389568
#> Jun 2020         NA   8.692271   5.084472   4.758249   4.297705   4.501258
#> Jul 2020         NA   4.098902   6.500385   6.098246   6.028673   6.078234
#> Aug 2020         NA         NA  11.232294  13.333128  13.822997  13.019357
#> Sep 2020         NA         NA  16.664908  16.791336  16.508317  17.133262
#> Oct 2020         NA         NA  18.962191  14.238352  15.045364  15.972571
#> Nov 2020         NA         NA         NA  18.088929  18.952941  16.662602
#> Dec 2020         NA         NA         NA  14.606399  17.812444  17.326115
#> Jan 2021         NA         NA         NA  19.901929  19.104740  19.813222
#> Feb 2021         NA         NA         NA  22.834260  20.770608  18.091636
#> Mar 2021         NA         NA         NA   8.008435   7.918820   9.367835
#> Apr 2021         NA         NA         NA   6.710615  10.742654  10.370377
#> May 2021         NA         NA         NA   9.071113   6.454864   6.183617
#> Jun 2021         NA         NA         NA   2.495331   2.748310   2.269733
#> Jul 2021         NA         NA         NA         NA  11.450856   8.567817
#> Aug 2021         NA         NA         NA         NA         NA  21.868550
#> Sep 2021         NA         NA         NA         NA         NA  18.949795
#> Oct 2021         NA         NA         NA         NA         NA  23.746751
#> Nov 2021         NA         NA         NA         NA         NA  29.144788
#> Dec 2021         NA         NA         NA         NA         NA  25.004379