Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   12.04123   13.54496   13.46951   13.25722   13.14483   13.37686
#> Feb 2020   10.77055   11.15583   12.44200   12.32067   12.54380   12.39875
#> Mar 2020   15.56746   13.62228   13.43306   13.27627   13.48197   13.30431
#> Apr 2020   14.74076   16.27449   16.16562   16.56470   15.97781   16.11522
#> May 2020         NA   23.96976   25.72385   24.47273   24.35160   24.22507
#> Jun 2020         NA   27.87334   29.09335   29.98142   29.44399   29.85632
#> Jul 2020         NA         NA   25.57823   26.97812   27.59060   28.18582
#> Aug 2020         NA         NA   31.34199   27.59127   28.95086   29.17208
#> Sep 2020         NA         NA         NA   26.27687   29.18958   29.34901
#> Oct 2020         NA         NA         NA   32.00197   30.59184   29.26973
#> Nov 2020         NA         NA         NA   22.12594   23.65732   23.34893
#> Dec 2020         NA         NA         NA   13.60333   16.74648   15.73888
#> Jan 2021         NA         NA         NA   21.91544   21.20524   20.02906
#> Feb 2021         NA         NA         NA   18.51218   21.33886   19.26017
#> Mar 2021         NA         NA         NA   23.64078   24.34808   23.14135
#> Apr 2021         NA         NA         NA   30.96052   29.42477   31.66785
#> May 2021         NA         NA         NA         NA   40.20147   40.56654
#> Jun 2021         NA         NA         NA         NA   53.11191   56.00609
#> Jul 2021         NA         NA         NA         NA   50.77751   54.11848
#> Aug 2021         NA         NA         NA         NA   47.76777   49.20538
#> Sep 2021         NA         NA         NA         NA         NA   50.56734
#> Oct 2021         NA         NA         NA         NA         NA   50.97393
#> Nov 2021         NA         NA         NA         NA         NA   46.50113
#> Dec 2021         NA         NA         NA         NA         NA   32.37820