Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16  2021-07-04 2022-01-15
#> Jan 2020   8.444714   9.601550   9.508322  9.9172120 10.28712302 10.0533551
#> Feb 2020         NA  16.159917  15.100315 15.2779535 15.94539570 15.5990781
#> Mar 2020         NA  13.187903  16.044242 15.9484275 15.82796050 15.7917765
#> Apr 2020         NA  18.669150  13.914303 14.8311274 14.94241878 14.8900412
#> May 2020         NA   8.804480  13.041772 13.2157799 13.16160404 13.3943505
#> Jun 2020         NA  10.888314   7.280515  6.9542926  6.49374802  6.6973011
#> Jul 2020         NA   7.171023   9.572506  9.1703674  9.10079461  9.1503554
#> Aug 2020         NA         NA  13.720280 15.8211140 16.31098247 15.5073431
#> Sep 2020         NA         NA   5.780363  5.9067913  5.62377143  6.2487171
#> Oct 2020         NA         NA   7.521653  2.7978138  3.60482634  4.5320326
#> Nov 2020         NA         NA         NA  5.4698844  6.33389688  4.0435574
#> Dec 2020         NA         NA         NA  7.2332835 10.43932845  9.9529997
#> Jan 2021         NA         NA         NA 12.5567036 11.75951411 12.4679965
#> Feb 2021         NA         NA         NA 17.3243930 15.26074184 12.5817692
#> Mar 2021         NA         NA         NA 12.7902157 12.70060056 14.1496150
#> Apr 2021         NA         NA         NA 10.2262240 14.25826246 13.8859855
#> May 2021         NA         NA         NA 11.8036047  9.18735518  8.9161083
#> Jun 2021         NA         NA         NA -0.2785906 -0.02561144 -0.5041883
#> Jul 2021         NA         NA         NA         NA  0.15274551 -2.7302934
#> Aug 2021         NA         NA         NA         NA          NA  7.2455241
#> Sep 2021         NA         NA         NA         NA          NA -7.9889909
#> Oct 2021         NA         NA         NA         NA          NA -5.0205354
#> Nov 2021         NA         NA         NA         NA          NA  0.8588116
#> Dec 2021         NA         NA         NA         NA          NA  5.7400725