Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for mts
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for data.frame
check_vertical(x, ...)

# S3 method for matrix
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for default
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01  2021-12-08
#> Jan 2020   8.780928  10.284653  10.209205   9.996914   9.884525  10.1165503
#> Feb 2020   6.067713   6.452995   7.739165   7.617835   7.840960   7.6959094
#> Mar 2020   7.329348   5.384164   5.194949   5.038154   5.243857   5.0662015
#> Apr 2020   4.046265   5.579987   5.471126   5.870204   5.283309   5.4207175
#> May 2020         NA   7.962339   9.716433   8.465313   8.344178   8.2176476
#> Jun 2020         NA   5.783270   7.003283   7.891353   7.353921   7.7662559
#> Jul 2020         NA         NA  -4.561387  -3.161491  -2.549012  -1.9537990
#> Aug 2020         NA         NA  -7.123984 -10.874704  -9.515123  -9.2939018
#> Sep 2020         NA         NA         NA -14.393354 -11.480647 -11.3212156
#> Oct 2020         NA         NA         NA  -9.517586 -10.927716 -12.2498261
#> Nov 2020         NA         NA         NA -16.466435 -14.935049 -15.2434437
#> Dec 2020         NA         NA         NA -15.984011 -12.840866 -13.8484604
#> Jan 2021         NA         NA         NA -12.400739 -13.110938 -14.2871176
#> Feb 2021         NA         NA         NA -17.421320 -14.594635 -16.6733274
#> Mar 2021         NA         NA         NA -13.753936 -13.046643 -14.2533703
#> Apr 2021         NA         NA         NA  -9.644627 -11.180379  -8.9372913
#> May 2021         NA         NA         NA         NA  -8.461842  -8.0967716
#> Jun 2021         NA         NA         NA         NA  -3.542463  -0.6482821
#> Jul 2021         NA         NA         NA         NA -14.995964 -11.6549924
#> Aug 2021         NA         NA         NA         NA -21.233497 -19.7958890
#> Sep 2021         NA         NA         NA         NA         NA -22.0336378
#> Oct 2021         NA         NA         NA         NA         NA -24.6728350
#> Nov 2021         NA         NA         NA         NA         NA -26.7910475
#> Dec 2021         NA         NA         NA         NA         NA -31.7889390