Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14  2020-07-25 2020-10-16  2021-06-16  2021-07-04  2022-01-15
#> Jan 2020   4.180089  5.33692447   5.243697   5.6525865   6.0224976   5.7887297
#> Feb 2020         NA  5.95036616   4.890764   5.0684026   5.7358448   5.3895272
#> Mar 2020         NA  2.25312323   5.109462   5.0136474   4.8931804   4.8569964
#> Apr 2020         NA  5.87831085   1.123464   2.0402884   2.1515798   2.0992023
#> May 2020         NA -0.07770854   4.159583   4.3335909   4.2794151   4.5121615
#> Jun 2020         NA  6.59440949   2.986611   2.6603881   2.1998435   2.4033966
#> Jul 2020         NA  0.96046732   3.361951   2.9598119   2.8902391   2.9397999
#> Aug 2020         NA          NA   2.469553   4.5703867   5.0602552   4.2566158
#> Sep 2020         NA          NA   4.783914   4.9103425   4.6273227   5.2522684
#> Oct 2020         NA          NA   5.415797   0.6919576   1.4989702   2.4261764
#> Nov 2020         NA          NA         NA   2.4060456   3.2700580   0.9797185
#> Dec 2020         NA          NA         NA  -4.1843022  -0.9782572  -1.4645860
#> Jan 2021         NA          NA         NA  -5.8051445  -6.6023340  -5.8938516
#> Feb 2021         NA          NA         NA  -3.9263359  -5.9899871  -8.6689597
#> Mar 2021         NA          NA         NA -10.9607074 -11.0503225  -9.6013080
#> Apr 2021         NA          NA         NA -19.9829563 -15.9509178 -16.3231947
#> May 2021         NA          NA         NA -14.0418154 -16.6580649 -16.9293118
#> Jun 2021         NA          NA         NA  -8.9102589  -8.6572797  -9.1358566
#> Jul 2021         NA          NA         NA          NA  -7.4736888 -10.3567277
#> Aug 2021         NA          NA         NA          NA          NA  -7.5182539
#> Sep 2021         NA          NA         NA          NA          NA -10.1197383
#> Oct 2021         NA          NA         NA          NA          NA -10.9265825
#> Nov 2021         NA          NA         NA          NA          NA -10.0232426
#> Dec 2021         NA          NA         NA          NA          NA -15.6405283