Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   8.143223   9.646949   9.571501   9.359209   9.246820   9.478846
#> Feb 2020   5.122364   5.507647   6.793817   6.672486   6.895611   6.750561
#> Mar 2020   9.591418   7.646234   7.457019   7.300223   7.505927   7.328271
#> Apr 2020  12.935602  14.469325  14.360464  14.759541  14.172647  14.310055
#> May 2020         NA  18.828952  20.583046  19.331926  19.210792  19.084261
#> Jun 2020         NA  18.471366  19.691379  20.579449  20.042017  20.454352
#> Jul 2020         NA         NA  14.877400  16.277295  16.889775  17.484987
#> Aug 2020         NA         NA  23.243636  19.492916  20.852497  21.073718
#> Sep 2020         NA         NA         NA  12.341092  15.253799  15.413230
#> Oct 2020         NA         NA         NA  13.887608  12.477478  11.155367
#> Nov 2020         NA         NA         NA  11.414930  12.946316  12.637921
#> Dec 2020         NA         NA         NA  21.802166  24.945311  23.937716
#> Jan 2021         NA         NA         NA  30.695024  29.984824  28.808645
#> Feb 2021         NA         NA         NA  29.299038  32.125723  30.047030
#> Mar 2021         NA         NA         NA  36.363628  37.070921  35.864193
#> Apr 2021         NA         NA         NA  44.835138  43.299386  45.542474
#> May 2021         NA         NA         NA         NA  43.157311  43.522381
#> Jun 2021         NA         NA         NA         NA  42.232975  45.127155
#> Jul 2021         NA         NA         NA         NA  34.900664  38.241635
#> Aug 2021         NA         NA         NA         NA  37.201545  38.639154
#> Sep 2021         NA         NA         NA         NA         NA  36.418675
#> Oct 2021         NA         NA         NA         NA         NA  32.379958
#> Nov 2021         NA         NA         NA         NA         NA  35.651131
#> Dec 2021         NA         NA         NA         NA         NA  42.418230