Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20  2021-04-02  2021-08-01   2021-12-08
#> Jan 2020   7.935161   9.438887  9.3634385   9.1511470   9.0387580   9.27078352
#> Feb 2020   6.179583   6.564865  7.8510355   7.7297048   7.9528300   7.80777945
#> Mar 2020  12.060405  10.115221  9.9260061   9.7692109   9.9749142   9.79725881
#> Apr 2020   5.911456   7.445178  7.3363174   7.7353952   7.1485004   7.28590875
#> May 2020         NA   4.741011  6.4951053   5.2439854   5.1228508   4.99632035
#> Jun 2020         NA  -1.340101 -0.1200873   0.7679823   0.2305504   0.64288525
#> Jul 2020         NA         NA -4.0445265  -2.6446308  -2.0321516  -1.43693883
#> Aug 2020         NA         NA  0.2409430  -3.5097771  -2.1501962  -1.92897482
#> Sep 2020         NA         NA         NA  -0.8867421   2.0259650   2.18539635
#> Oct 2020         NA         NA         NA   6.5008007   5.0906708   3.76856050
#> Nov 2020         NA         NA         NA  -0.1771286   1.3542575   1.04586251
#> Dec 2020         NA         NA         NA  -2.1627487   0.9803969  -0.02719786
#> Jan 2021         NA         NA         NA  -3.1820111  -3.8922106  -5.06839001
#> Feb 2021         NA         NA         NA -10.1186972  -7.2920121  -9.37070462
#> Mar 2021         NA         NA         NA  -8.6891732  -7.9818805  -9.18860793
#> Apr 2021         NA         NA         NA -14.2465540 -15.7823060 -13.53921812
#> May 2021         NA         NA         NA          NA -17.3919667 -17.02689659
#> Jun 2021         NA         NA         NA          NA -22.7672065 -19.87302590
#> Jul 2021         NA         NA         NA          NA -26.8460098 -23.50503804
#> Aug 2021         NA         NA         NA          NA -27.8221943 -26.38458592
#> Sep 2021         NA         NA         NA          NA          NA -24.45586676
#> Oct 2021         NA         NA         NA          NA          NA -13.41027319
#> Nov 2021         NA         NA         NA          NA          NA  -8.47605653
#> Dec 2021         NA         NA         NA          NA          NA -14.52406690