Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20  2021-04-02  2021-08-01  2021-12-08
#> Jan 2020  3.3830250  4.8867506  4.8113026   4.5990110   4.4866221   4.7186476
#> Feb 2020  0.3245907  0.7098732  1.9960434   1.8747127   2.0978379   1.9527873
#> Mar 2020  7.1902191  5.2450351  5.0558199   4.8990248   5.1047281   4.9270727
#> Apr 2020  5.5465702  7.0802925  6.9714314   7.3705091   6.7836144   6.9210227
#> May 2020         NA  1.3869044  3.1409985   1.8898786   1.7687440   1.6422136
#> Jun 2020         NA  0.9524024  2.1724156   3.0604852   2.5230534   2.9353882
#> Jul 2020         NA         NA -2.1339478  -0.7340521  -0.1215729   0.4736399
#> Aug 2020         NA         NA -0.3984392  -4.1491593  -2.7895784  -2.5683570
#> Sep 2020         NA         NA         NA  -3.2751736  -0.3624666  -0.2030352
#> Oct 2020         NA         NA         NA   4.9853374   3.5752075   2.2530972
#> Nov 2020         NA         NA         NA  -3.0721765  -1.5407904  -1.8491854
#> Dec 2020         NA         NA         NA -10.7008660  -7.5577204  -8.5653151
#> Jan 2021         NA         NA         NA -13.1787507 -13.8889502 -15.0651296
#> Feb 2021         NA         NA         NA -18.5654475 -15.7387625 -17.8174550
#> Mar 2021         NA         NA         NA -11.4416818 -10.7343891 -11.9411165
#> Apr 2021         NA         NA         NA  -9.5432292 -11.0789812  -8.8358933
#> May 2021         NA         NA         NA          NA -12.7645619 -12.3994918
#> Jun 2021         NA         NA         NA          NA  -8.5858460  -5.6916654
#> Jul 2021         NA         NA         NA          NA -16.0439821 -12.7030104
#> Aug 2021         NA         NA         NA          NA -20.8005023 -19.3628939
#> Sep 2021         NA         NA         NA          NA          NA -16.1659867
#> Oct 2021         NA         NA         NA          NA          NA -16.0590215
#> Nov 2021         NA         NA         NA          NA          NA -21.8149872
#> Dec 2021         NA         NA         NA          NA          NA -34.6065780