Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   9.167418  10.671144  10.595696 10.3834043  10.271015 10.5030408
#> Feb 2020  11.058722  11.444005  12.730175 12.6088442  12.831969 12.6869188
#> Mar 2020  14.359505  12.414321  12.225105 12.0683102  12.274014 12.0963582
#> Apr 2020   8.835870  10.369593  10.260731 10.6598092  10.072914 10.2103228
#> May 2020         NA   9.415041  11.169135  9.9180155   9.796881  9.6703505
#> Jun 2020         NA   8.046183   9.266196 10.1542657   9.616834 10.0291687
#> Jul 2020         NA         NA   5.509295  6.9091905   7.521670  8.1168825
#> Aug 2020         NA         NA   5.970620  2.2199004   3.579481  3.8007027
#> Sep 2020         NA         NA         NA  2.1812220   5.093929  5.2533604
#> Oct 2020         NA         NA         NA  5.5403142   4.130184  2.8080740
#> Nov 2020         NA         NA         NA -0.2266044   1.304782  0.9963866
#> Dec 2020         NA         NA         NA  3.7866071   6.929753  5.9221580
#> Jan 2021         NA         NA         NA 12.8122679  12.102069 10.9258891
#> Feb 2021         NA         NA         NA 10.4644452  13.291130 11.2124377
#> Mar 2021         NA         NA         NA 10.1963375  10.903630  9.6969028
#> Apr 2021         NA         NA         NA  7.5376777   6.001926  8.2450136
#> May 2021         NA         NA         NA         NA   3.579631  3.9447010
#> Jun 2021         NA         NA         NA         NA   4.996554  7.8907344
#> Jul 2021         NA         NA         NA         NA   4.946248  8.2872194
#> Aug 2021         NA         NA         NA         NA   4.785433  6.2230417
#> Sep 2021         NA         NA         NA         NA         NA  7.1403592
#> Oct 2021         NA         NA         NA         NA         NA  1.0377263
#> Nov 2021         NA         NA         NA         NA         NA  0.9139089
#> Dec 2021         NA         NA         NA         NA         NA  1.1603366