Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for mts
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for data.frame
check_vertical(x, ...)

# S3 method for matrix
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for default
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   8.870904  10.374630  10.299182  10.086890   9.974501  10.206527
#> Feb 2020   7.719604   8.104886   9.391057   9.269726   9.492851   9.347801
#> Mar 2020   9.099342   7.154158   6.964943   6.808148   7.013851   6.836196
#> Apr 2020   4.786755   6.320477   6.211616   6.610694   6.023799   6.161208
#> May 2020         NA   6.217555   7.971649   6.720529   6.599394   6.472864
#> Jun 2020         NA   3.977328   5.197341   6.085410   5.547979   5.960313
#> Jul 2020         NA         NA   2.478108   3.878004   4.490483   5.085696
#> Aug 2020         NA         NA   6.418292   2.667572   4.027153   4.248374
#> Sep 2020         NA         NA         NA   5.993780   8.906487   9.065918
#> Oct 2020         NA         NA         NA  14.598788  13.188658  11.866548
#> Nov 2020         NA         NA         NA  10.515882  12.047268  11.738873
#> Dec 2020         NA         NA         NA  16.109083  19.252229  18.244634
#> Jan 2021         NA         NA         NA  17.268855  16.558656  15.382477
#> Feb 2021         NA         NA         NA  14.210052  17.036737  14.958044
#> Mar 2021         NA         NA         NA  11.897791  12.605084  11.398356
#> Apr 2021         NA         NA         NA  12.116797  10.581045  12.824133
#> May 2021         NA         NA         NA         NA   7.307387   7.672457
#> Jun 2021         NA         NA         NA         NA   5.601790   8.495971
#> Jul 2021         NA         NA         NA         NA   2.412472   5.753443
#> Aug 2021         NA         NA         NA         NA   1.350930   2.788539
#> Sep 2021         NA         NA         NA         NA         NA   9.312836
#> Oct 2021         NA         NA         NA         NA         NA  11.628625
#> Nov 2021         NA         NA         NA         NA         NA  15.420397
#> Dec 2021         NA         NA         NA         NA         NA  16.041991