Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16  2021-07-04  2022-01-15
#> Jan 2020   8.003176   9.160011  9.0667837   9.475673   9.8455845   9.6118166
#> Feb 2020         NA   5.247995  4.1883928   4.366032   5.0334738   4.6871562
#> Mar 2020         NA  -7.855611 -4.9992724  -5.095087  -5.2155542  -5.2517382
#> Apr 2020         NA  -3.633071 -8.3879176  -7.471093  -7.3598018  -7.4121794
#> May 2020         NA -10.838610 -6.6013176  -6.427310  -6.4814859  -6.2487395
#> Jun 2020         NA  -1.089559 -4.6973574  -5.023580  -5.4841248  -5.2805717
#> Jul 2020         NA  -3.312935 -0.9114517  -1.313590  -1.3831632  -1.3336024
#> Aug 2020         NA         NA -8.2930340  -6.192200  -5.7023317  -6.5059711
#> Sep 2020         NA         NA -6.0687889  -5.942361  -6.2253805  -5.6004349
#> Oct 2020         NA         NA -0.2592402  -4.983079  -4.1760668  -3.2488605
#> Nov 2020         NA         NA         NA  -4.109352  -3.2453395  -5.5356790
#> Dec 2020         NA         NA         NA  -3.709554  -0.5035094  -0.9898382
#> Jan 2021         NA         NA         NA  -2.410451  -3.2076404  -2.4991581
#> Feb 2021         NA         NA         NA  -7.129619  -9.1932706 -11.8722432
#> Mar 2021         NA         NA         NA -17.564126 -17.6537409 -16.2047264
#> Apr 2021         NA         NA         NA -10.436367  -6.4043282  -6.7766051
#> May 2021         NA         NA         NA  -2.659672  -5.2759216  -5.5471685
#> Jun 2021         NA         NA         NA  -5.090529  -4.8375497  -5.3161266
#> Jul 2021         NA         NA         NA         NA   5.3977617   2.5147227
#> Aug 2021         NA         NA         NA         NA          NA   5.0632975
#> Sep 2021         NA         NA         NA         NA          NA  -6.1075840
#> Oct 2021         NA         NA         NA         NA          NA  -0.1812540
#> Nov 2021         NA         NA         NA         NA          NA   8.7490182
#> Dec 2021         NA         NA         NA         NA          NA  13.9004494