Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   6.931273   8.434998   8.359550   8.147259   8.034870   8.266895
#> Feb 2020   7.355936   7.741218   9.027388   8.906058   9.129183   8.984132
#> Mar 2020   9.301450   7.356266   7.167051   7.010256   7.215959   7.038304
#> Apr 2020   6.609121   8.142843   8.033982   8.433060   7.846165   7.983573
#> May 2020         NA   7.299344   9.053438   7.802319   7.681184   7.554654
#> Jun 2020         NA   5.508355   6.728368   7.616438   7.079006   7.491341
#> Jul 2020         NA         NA   6.874316   8.274212   8.886691   9.481904
#> Aug 2020         NA         NA  11.960179   8.209459   9.569039   9.790261
#> Sep 2020         NA         NA         NA   8.281274  11.193981  11.353413
#> Oct 2020         NA         NA         NA  18.667092  17.256962  15.934852
#> Nov 2020         NA         NA         NA  19.982765  21.514151  21.205756
#> Dec 2020         NA         NA         NA  18.145827  21.288972  20.281378
#> Jan 2021         NA         NA         NA  21.850791  21.140592  19.964412
#> Feb 2021         NA         NA         NA  24.570508  27.397194  25.318501
#> Mar 2021         NA         NA         NA  22.775462  23.482755  22.276028
#> Apr 2021         NA         NA         NA  21.246691  19.710939  21.954027
#> May 2021         NA         NA         NA         NA  18.557812  18.922882
#> Jun 2021         NA         NA         NA         NA  22.830291  25.724472
#> Jul 2021         NA         NA         NA         NA  22.859937  26.200909
#> Aug 2021         NA         NA         NA         NA  24.549650  25.987258
#> Sep 2021         NA         NA         NA         NA         NA  31.402218
#> Oct 2021         NA         NA         NA         NA         NA  35.115226
#> Nov 2021         NA         NA         NA         NA         NA  44.369998
#> Dec 2021         NA         NA         NA         NA         NA  43.949473