Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   9.684901   11.18863   11.11318   10.90089   10.78850   11.02052
#> Feb 2020  10.479987   10.86527   12.15144   12.03011   12.25323   12.10818
#> Mar 2020  12.868872   10.92369   10.73447   10.57768   10.78338   10.60573
#> Apr 2020  11.375867   12.90959   12.80073   13.19981   12.61291   12.75032
#> May 2020         NA   12.09817   13.85226   12.60114   12.48001   12.35348
#> Jun 2020         NA   13.14748   14.36749   15.25556   14.71813   15.13047
#> Jul 2020         NA         NA   11.17496   12.57485   13.18733   13.78254
#> Aug 2020         NA         NA   16.78135   13.03063   14.39021   14.61143
#> Sep 2020         NA         NA         NA   13.92894   16.84165   17.00108
#> Oct 2020         NA         NA         NA   20.99431   19.58418   18.26207
#> Nov 2020         NA         NA         NA   20.86788   22.39927   22.09088
#> Dec 2020         NA         NA         NA   26.88812   30.03126   29.02367
#> Jan 2021         NA         NA         NA   24.82188   24.11168   22.93550
#> Feb 2021         NA         NA         NA   21.65073   24.47741   22.39872
#> Mar 2021         NA         NA         NA   22.97702   23.68431   22.47758
#> Apr 2021         NA         NA         NA   28.88571   27.34996   29.59305
#> May 2021         NA         NA         NA         NA   26.56284   26.92791
#> Jun 2021         NA         NA         NA         NA   31.71800   34.61218
#> Jul 2021         NA         NA         NA         NA   28.30163   31.64260
#> Aug 2021         NA         NA         NA         NA   32.69933   34.13694
#> Sep 2021         NA         NA         NA         NA         NA   36.61359
#> Oct 2021         NA         NA         NA         NA         NA   35.66437
#> Nov 2021         NA         NA         NA         NA         NA   38.84121
#> Dec 2021         NA         NA         NA         NA         NA   38.51530