Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020  2.4074529   3.911178  3.8357304   3.623439   3.511050   3.743075
#> Feb 2020  5.3040083   5.689291  6.9754611   6.854130   7.077255   6.932205
#> Mar 2020  0.1168606  -1.828323 -2.0175385  -2.174334  -1.968630  -2.146286
#> Apr 2020 -5.9657203  -4.431998 -4.5408591  -4.141781  -4.728676  -4.591268
#> May 2020         NA  -2.297048 -0.5429544  -1.794074  -1.915209  -2.041739
#> Jun 2020         NA  -6.986784 -5.7667712  -4.878702  -5.416133  -5.003799
#> Jul 2020         NA         NA -9.9304898  -8.530594  -7.918115  -7.322902
#> Aug 2020         NA         NA -1.0269397  -4.777660  -3.418079  -3.196857
#> Sep 2020         NA         NA         NA -12.499714  -9.587007  -9.427575
#> Oct 2020         NA         NA         NA  -7.308742  -8.718872 -10.040983
#> Nov 2020         NA         NA         NA -10.092716  -8.561330  -8.869725
#> Dec 2020         NA         NA         NA  -8.248691  -5.105546  -6.113140
#> Jan 2021         NA         NA         NA -11.980672 -12.690872 -13.867051
#> Feb 2021         NA         NA         NA -15.244503 -12.417818 -14.496511
#> Mar 2021         NA         NA         NA -16.973228 -16.265935 -17.472663
#> Apr 2021         NA         NA         NA -10.075345 -11.611097  -9.368010
#> May 2021         NA         NA         NA         NA   1.908016   2.273086
#> Jun 2021         NA         NA         NA         NA   6.800369   9.694549
#> Jul 2021         NA         NA         NA         NA   2.221423   5.562395
#> Aug 2021         NA         NA         NA         NA   9.214451  10.652059
#> Sep 2021         NA         NA         NA         NA         NA   7.253944
#> Oct 2021         NA         NA         NA         NA         NA   5.283264
#> Nov 2021         NA         NA         NA         NA         NA  11.515069
#> Dec 2021         NA         NA         NA         NA         NA   7.368243