Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   12.41368  13.917410  13.841962  13.629670   13.51728   13.74931
#> Feb 2020   19.98774  20.373020  21.659190  21.537859   21.76098   21.61593
#> Mar 2020   19.24367  17.298484  17.109269  16.952474   17.15818   16.98052
#> Apr 2020   13.50463  15.038348  14.929487  15.328565   14.74167   14.87908
#> May 2020         NA  12.781308  14.535402  13.284282   13.16315   13.03662
#> Jun 2020         NA   9.490566  10.710579  11.598649   11.06122   11.47355
#> Jul 2020         NA         NA   7.435395   8.835291    9.44777   10.04298
#> Aug 2020         NA         NA  13.314169   9.563449   10.92303   11.14425
#> Sep 2020         NA         NA         NA  11.696075   14.60878   14.76821
#> Oct 2020         NA         NA         NA  18.203224   16.79309   15.47098
#> Nov 2020         NA         NA         NA  21.566846   23.09823   22.78984
#> Dec 2020         NA         NA         NA  23.061255   26.20440   25.19681
#> Jan 2021         NA         NA         NA  32.706486   31.99629   30.82011
#> Feb 2021         NA         NA         NA  35.380326   38.20701   36.12832
#> Mar 2021         NA         NA         NA  33.854251   34.56154   33.35482
#> Apr 2021         NA         NA         NA  28.727968   27.19222   29.43530
#> May 2021         NA         NA         NA         NA   27.39147   27.75654
#> Jun 2021         NA         NA         NA         NA   28.92069   31.81488
#> Jul 2021         NA         NA         NA         NA   26.19663   29.53761
#> Aug 2021         NA         NA         NA         NA   29.27992   30.71753
#> Sep 2021         NA         NA         NA         NA         NA   35.87357
#> Oct 2021         NA         NA         NA         NA         NA   38.81011
#> Nov 2021         NA         NA         NA         NA         NA   45.38052
#> Dec 2021         NA         NA         NA         NA         NA   42.90617