Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16  2021-06-16 2021-07-04  2022-01-15
#> Jan 2020   11.63753  12.794366 12.7011379 13.11002763 13.4799386 13.24617074
#> Feb 2020         NA  18.421144 17.3615413 17.53918013 18.2066224 17.86030473
#> Mar 2020         NA  15.117459 17.9737983 17.87798358 17.7575166 17.72133254
#> Apr 2020         NA  24.468273 19.7134258 20.63025010 20.7415415 20.68916398
#> May 2020         NA  12.258472 16.4957640 16.66977150 16.6155957 16.84834210
#> Jun 2020         NA  13.100315  9.4925164  9.16629358  8.7057490  8.90930212
#> Jul 2020         NA   3.337463  5.7389462  5.33680750  5.2672347  5.31679552
#> Aug 2020         NA         NA  0.2178859  2.31871970  2.8085882  2.00494880
#> Sep 2020         NA         NA  2.3415472  2.46797545  2.1849556  2.80990128
#> Oct 2020         NA         NA  4.6767997 -0.04703947  0.7599731  1.68717940
#> Nov 2020         NA         NA         NA  5.61895235  6.4829648  4.19262532
#> Dec 2020         NA         NA         NA  0.84951059  4.0555556  3.56922680
#> Jan 2021         NA         NA         NA  6.25853194  5.4613425  6.16982489
#> Feb 2021         NA         NA         NA 15.43495401 13.3713028 10.69233024
#> Mar 2021         NA         NA         NA  9.66293608  9.5733209 11.02233540
#> Apr 2021         NA         NA         NA  6.64319137 10.6752298 10.30295292
#> May 2021         NA         NA         NA  8.41803562  5.8017861  5.53053921
#> Jun 2021         NA         NA         NA  2.99733262  3.2503118  2.77173488
#> Jul 2021         NA         NA         NA          NA  2.8176047 -0.06543427
#> Aug 2021         NA         NA         NA          NA         NA  4.96092358
#> Sep 2021         NA         NA         NA          NA         NA -1.15808184
#> Oct 2021         NA         NA         NA          NA         NA -1.36365700
#> Nov 2021         NA         NA         NA          NA         NA  2.23688041
#> Dec 2021         NA         NA         NA          NA         NA  0.36700032