Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02  2021-08-01   2021-12-08
#> Jan 2020   7.346113   8.849839  8.7743909  8.5620993  8.44971040   8.68173591
#> Feb 2020  10.650914  11.036197 12.3223669 12.2010362 12.42416135  12.27911085
#> Mar 2020  14.709683  12.764499 12.5752839 12.4184888 12.62419208  12.44653669
#> Apr 2020   6.239432   7.773154  7.6642934  8.0633711  7.47647639   7.61388473
#> May 2020         NA   2.771101  4.5251953  3.2740754  3.15294084   3.02641040
#> Jun 2020         NA  -1.371751 -0.1517382  0.7363314  0.19889953   0.61123434
#> Jul 2020         NA         NA -4.5869426 -3.1870469 -2.57456769  -1.97935495
#> Aug 2020         NA         NA  2.0526321 -1.6980879 -0.33850697  -0.11728564
#> Sep 2020         NA         NA         NA  2.7779821  5.69068917   5.85012052
#> Oct 2020         NA         NA         NA  3.7011654  2.29103539   0.96892514
#> Nov 2020         NA         NA         NA -4.6623496 -3.13096347  -3.43935850
#> Dec 2020         NA         NA         NA -8.3077385 -5.16459298  -6.17218771
#> Jan 2021         NA         NA         NA -3.8049618 -4.51516122  -5.69134066
#> Feb 2021         NA         NA         NA  1.1154087  3.94209379   1.86340125
#> Mar 2021         NA         NA         NA -0.6472064  0.06008627  -1.14664113
#> Apr 2021         NA         NA         NA -3.0971082 -4.63286027  -2.38977236
#> May 2021         NA         NA         NA         NA -4.33533929  -3.97026923
#> Jun 2021         NA         NA         NA         NA -2.93495430  -0.04077371
#> Jul 2021         NA         NA         NA         NA -4.98875994  -1.64778822
#> Aug 2021         NA         NA         NA         NA -4.09662697  -2.65901859
#> Sep 2021         NA         NA         NA         NA          NA  -1.59670350
#> Oct 2021         NA         NA         NA         NA          NA  -4.38372030
#> Nov 2021         NA         NA         NA         NA          NA  -5.07570956
#> Dec 2021         NA         NA         NA         NA          NA -11.76416274