Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   7.871664   9.375389   9.299941   9.087650   8.975261   9.207286
#> Feb 2020   6.725772   7.111055   8.397225   8.275894   8.499019   8.353969
#> Mar 2020  16.422966  14.477782  14.288566  14.131771  14.337475  14.159819
#> Apr 2020  14.172839  15.706561  15.597700  15.996778  15.409883  15.547291
#> May 2020         NA  15.998357  17.752451  16.501331  16.380196  16.253666
#> Jun 2020         NA  12.887100  14.107113  14.995183  14.457751  14.870086
#> Jul 2020         NA         NA  12.676640  14.076536  14.689015  15.284228
#> Aug 2020         NA         NA  15.417092  11.666371  13.025952  13.247174
#> Sep 2020         NA         NA         NA   6.878566   9.791273   9.950704
#> Oct 2020         NA         NA         NA  19.109172  17.699042  16.376932
#> Nov 2020         NA         NA         NA  19.107432  20.638818  20.330423
#> Dec 2020         NA         NA         NA  22.378165  25.521310  24.513715
#> Jan 2021         NA         NA         NA  25.692902  24.982702  23.806523
#> Feb 2021         NA         NA         NA  25.214591  28.041277  25.962584
#> Mar 2021         NA         NA         NA  30.826280  31.533573  30.326846
#> Apr 2021         NA         NA         NA  27.881405  26.345653  28.588741
#> May 2021         NA         NA         NA         NA  29.438379  29.803449
#> Jun 2021         NA         NA         NA         NA  26.486187  29.380368
#> Jul 2021         NA         NA         NA         NA  23.931452  27.272424
#> Aug 2021         NA         NA         NA         NA  24.689612  26.127221
#> Sep 2021         NA         NA         NA         NA         NA  22.813474
#> Oct 2021         NA         NA         NA         NA         NA  30.065369
#> Nov 2021         NA         NA         NA         NA         NA  37.098633
#> Dec 2021         NA         NA         NA         NA         NA  35.495335