Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01  2021-12-08
#> Jan 2020   9.071190  10.574916  10.499468 10.2871764 10.1747875 10.40681300
#> Feb 2020   6.918942   7.304225   8.590395  8.4690645  8.6921897  8.54713915
#> Mar 2020   9.482242   7.537058   7.347843  7.1910478  7.3967511  7.21909574
#> Apr 2020   3.782021   5.315743   5.206882  5.6059598  5.0190651  5.15647341
#> May 2020         NA   6.058267   7.812361  6.5612413  6.4401067  6.31357631
#> Jun 2020         NA   8.208964   9.428977 10.3170466  9.7796147 10.19194955
#> Jul 2020         NA         NA   2.969842  4.3697382  4.9822174  5.57743013
#> Aug 2020         NA         NA   6.865665  3.1149450  4.4745259  4.69574720
#> Sep 2020         NA         NA         NA  3.3615754  6.2742825  6.43371386
#> Oct 2020         NA         NA         NA  4.2791684  2.8690384  1.54692817
#> Nov 2020         NA         NA         NA -0.8525395  0.6788466  0.37045161
#> Dec 2020         NA         NA         NA  0.9728840  4.1160296  3.10843485
#> Jan 2021         NA         NA         NA  6.3932340  5.6830345  4.50685509
#> Feb 2021         NA         NA         NA  0.2917854  3.1184705  1.03977795
#> Mar 2021         NA         NA         NA  1.6444438  2.3517365  1.14500906
#> Apr 2021         NA         NA         NA -0.2506356 -1.7863876  0.45670030
#> May 2021         NA         NA         NA         NA -2.8947662 -2.52969615
#> Jun 2021         NA         NA         NA         NA -0.6497192  2.24446140
#> Jul 2021         NA         NA         NA         NA -3.3904364 -0.04946466
#> Aug 2021         NA         NA         NA         NA -1.0034528  0.43415556
#> Sep 2021         NA         NA         NA         NA         NA  3.10160065
#> Oct 2021         NA         NA         NA         NA         NA -3.90553666
#> Nov 2021         NA         NA         NA         NA         NA -2.41281974
#> Dec 2021         NA         NA         NA         NA         NA -3.84066517