Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for mts
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for data.frame
check_vertical(x, ...)

# S3 method for matrix
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for default
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12  2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   6.386263  7.88998885   7.814541  7.6022492  7.4898603  7.7218858
#> Feb 2020   5.855205  6.24048771   7.526658  7.4053272  7.6284523  7.4834018
#> Mar 2020  10.868186  8.92300247   8.733787  8.5769922  8.7826955  8.6050401
#> Apr 2020   2.314004  3.84772581   3.738865  4.1379425  3.5510478  3.6884561
#> May 2020         NA  5.98334135   7.737435  6.4863155  6.3651809  6.2386505
#> Jun 2020         NA -0.02982513   1.190188  2.0782576  1.5408258  1.9531606
#> Jul 2020         NA          NA  -2.183580 -0.7836844 -0.1712051  0.4240076
#> Aug 2020         NA          NA   8.219635  4.4689148  5.8284958  6.0497171
#> Sep 2020         NA          NA         NA  2.8900032  5.8027103  5.9621417
#> Oct 2020         NA          NA         NA 11.0487684  9.6386384  8.3165281
#> Nov 2020         NA          NA         NA  9.0682241 10.5996102 10.2912152
#> Dec 2020         NA          NA         NA 15.2192041 18.3623497 17.3547550
#> Jan 2021         NA          NA         NA 16.7935078 16.0833084 14.9071290
#> Feb 2021         NA          NA         NA 12.2503994 15.0770845 12.9983920
#> Mar 2021         NA          NA         NA 15.5611511 16.2684438 15.0617164
#> Apr 2021         NA          NA         NA  9.2331688  7.6974168  9.9405047
#> May 2021         NA          NA         NA         NA 11.9817506 12.3468206
#> Jun 2021         NA          NA         NA         NA  6.1322998  9.0264804
#> Jul 2021         NA          NA         NA         NA  0.5602359  3.9012077
#> Aug 2021         NA          NA         NA         NA  7.6581972  9.0958056
#> Sep 2021         NA          NA         NA         NA         NA 10.3530302
#> Oct 2021         NA          NA         NA         NA         NA 10.2399975
#> Nov 2021         NA          NA         NA         NA         NA 12.1155490
#> Dec 2021         NA          NA         NA         NA         NA 14.7996372