Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20  2021-04-02  2021-08-01 2021-12-08
#> Jan 2020   9.127333  10.631059  10.555611 10.34331932 10.23093038 10.4629559
#> Feb 2020   6.163523   6.548806   7.834976  7.71364533  7.93677048  7.7917200
#> Mar 2020   5.008226   3.063042   2.873827  2.71703201  2.92273531  2.7450799
#> Apr 2020   5.182946   6.716669   6.607808  7.00688537  6.41999063  6.5573990
#> May 2020         NA   1.710362   3.464456  2.21333605  2.09220147  1.9656710
#> Jun 2020         NA   5.822808   7.042822  7.93089125  7.39345942  7.8057942
#> Jul 2020         NA         NA   4.612785  6.01268058  6.62515982  7.2203726
#> Aug 2020         NA         NA   8.686177  4.93545667  6.29503758  6.5162589
#> Sep 2020         NA         NA         NA  7.09731422 10.01002129 10.1694526
#> Oct 2020         NA         NA         NA 12.67268037 11.26255040  9.9404402
#> Nov 2020         NA         NA         NA  4.48886173  6.02024782  5.7118528
#> Dec 2020         NA         NA         NA  6.78683941  9.92998496  8.9223902
#> Jan 2021         NA         NA         NA  7.22985879  6.51965935  5.3434799
#> Feb 2021         NA         NA         NA -0.02105406  2.80563104  0.7269385
#> Mar 2021         NA         NA         NA -3.33832562 -2.63103294 -3.8377603
#> Apr 2021         NA         NA         NA -2.16398764 -3.69973966 -1.4566517
#> May 2021         NA         NA         NA          NA -7.16557744 -6.8005074
#> Jun 2021         NA         NA         NA          NA -0.09845347  2.7957271
#> Jul 2021         NA         NA         NA          NA -5.23542368 -1.8944520
#> Aug 2021         NA         NA         NA          NA -7.22974723 -5.7921388
#> Sep 2021         NA         NA         NA          NA          NA -2.2541407
#> Oct 2021         NA         NA         NA          NA          NA -1.0285324
#> Nov 2021         NA         NA         NA          NA          NA  0.5032994
#> Dec 2021         NA         NA         NA          NA          NA  3.0862355