Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   9.865625   11.36935   11.29390  11.081611  10.969222  11.201248
#> Feb 2020  13.612739   13.99802   15.28419  15.162861  15.385986  15.240936
#> Mar 2020  22.873838   20.92865   20.73944  20.582643  20.788347  20.610691
#> Apr 2020  19.538716   21.07244   20.96358  21.362655  20.775760  20.913169
#> May 2020         NA   17.05579   18.80988  17.558763  17.437628  17.311098
#> Jun 2020         NA   11.07156   12.29158  13.179647  12.642215  13.054550
#> Jul 2020         NA         NA   10.76139  12.161285  12.773764  13.368977
#> Aug 2020         NA         NA   16.48027  12.729549  14.089129  14.310351
#> Sep 2020         NA         NA         NA   8.166466  11.079173  11.238604
#> Oct 2020         NA         NA         NA  13.565186  12.155056  10.832946
#> Nov 2020         NA         NA         NA   5.010428   6.541814   6.233419
#> Dec 2020         NA         NA         NA   6.849382   9.992528   8.984933
#> Jan 2021         NA         NA         NA  13.116600  12.406401  11.230222
#> Feb 2021         NA         NA         NA  14.205053  17.031738  14.953045
#> Mar 2021         NA         NA         NA  21.257142  21.964434  20.757707
#> Apr 2021         NA         NA         NA  22.061253  20.525501  22.768589
#> May 2021         NA         NA         NA         NA  19.227463  19.592533
#> Jun 2021         NA         NA         NA         NA  21.860864  24.755045
#> Jul 2021         NA         NA         NA         NA  21.123230  24.464202
#> Aug 2021         NA         NA         NA         NA  18.532248  19.969856
#> Sep 2021         NA         NA         NA         NA         NA  17.794118
#> Oct 2021         NA         NA         NA         NA         NA  23.631871
#> Nov 2021         NA         NA         NA         NA         NA  20.346554
#> Dec 2021         NA         NA         NA         NA         NA  16.567798