Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   9.332572   10.48941   10.39618   10.80507   11.17498   10.94121
#> Feb 2020         NA   14.99534   13.93574   14.11337   14.78082   14.43450
#> Mar 2020         NA   11.57995   14.43629   14.34048   14.22001   14.18382
#> Apr 2020         NA   18.88115   14.12630   15.04312   15.15442   15.10204
#> May 2020         NA   10.65814   14.89543   15.06944   15.01526   15.24801
#> Jun 2020         NA   20.20608   16.59828   16.27206   15.81152   16.01507
#> Jul 2020         NA   23.02381   25.42529   25.02315   24.95358   25.00314
#> Aug 2020         NA         NA   28.08792   30.18876   30.67863   29.87499
#> Sep 2020         NA         NA   36.62542   36.75185   36.46883   37.09377
#> Oct 2020         NA         NA   40.82816   36.10432   36.91134   37.83854
#> Nov 2020         NA         NA         NA   36.45634   37.32035   35.03001
#> Dec 2020         NA         NA         NA   34.07846   37.28451   36.79818
#> Jan 2021         NA         NA         NA   36.63315   35.83596   36.54444
#> Feb 2021         NA         NA         NA   47.29242   45.22877   42.54980
#> Mar 2021         NA         NA         NA   45.58885   45.49923   46.94825
#> Apr 2021         NA         NA         NA   45.97045   50.00249   49.63021
#> May 2021         NA         NA         NA   52.86771   50.25146   49.98021
#> Jun 2021         NA         NA         NA   47.09619   47.34917   46.87059
#> Jul 2021         NA         NA         NA         NA   53.68194   50.79891
#> Aug 2021         NA         NA         NA         NA         NA   59.03586
#> Sep 2021         NA         NA         NA         NA         NA   56.66598
#> Oct 2021         NA         NA         NA         NA         NA   60.84222
#> Nov 2021         NA         NA         NA         NA         NA   62.68419
#> Dec 2021         NA         NA         NA         NA         NA   61.69244