Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for mts
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for data.frame
check_vertical(x, ...)

# S3 method for matrix
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for default
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   8.478623   9.982348   9.906900   9.694609   9.582220  9.8142454
#> Feb 2020   5.715173   6.100455   7.386626   7.265295   7.488420  7.3433695
#> Mar 2020   9.784227   7.839043   7.649828   7.493033   7.698736  7.5210809
#> Apr 2020   4.836090   6.369812   6.260951   6.660029   6.073134  6.2105424
#> May 2020         NA   4.188676   5.942770   4.691650   4.570515  4.4439847
#> Jun 2020         NA   8.123717   9.343730  10.231799   9.694368 10.1067023
#> Jul 2020         NA         NA   6.263659   7.663554   8.276034  8.8712463
#> Aug 2020         NA         NA  14.687901  10.937181  12.296762 12.5179829
#> Sep 2020         NA         NA         NA   9.032814  11.945521 12.1049522
#> Oct 2020         NA         NA         NA  12.639373  11.229243  9.9071327
#> Nov 2020         NA         NA         NA   7.645069   9.176455  8.8680600
#> Dec 2020         NA         NA         NA   7.560264  10.703410  9.6958149
#> Jan 2021         NA         NA         NA   9.591825   8.881626  7.7054463
#> Feb 2021         NA         NA         NA  -1.290295   1.536390 -0.5423023
#> Mar 2021         NA         NA         NA  -4.311293  -3.604000 -4.8107276
#> Apr 2021         NA         NA         NA  -3.321196  -4.856948 -2.6138605
#> May 2021         NA         NA         NA         NA  -1.850671 -1.4856010
#> Jun 2021         NA         NA         NA         NA   5.308173  8.2023537
#> Jul 2021         NA         NA         NA         NA   1.176954  4.5179257
#> Aug 2021         NA         NA         NA         NA   3.776495  5.2141032
#> Sep 2021         NA         NA         NA         NA         NA  5.0783183
#> Oct 2021         NA         NA         NA         NA         NA  4.9944804
#> Nov 2021         NA         NA         NA         NA         NA 11.8830268
#> Dec 2021         NA         NA         NA         NA         NA  9.2593321