Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020  13.331710  14.835436  14.759988  14.547696  14.435308  14.667333
#> Feb 2020  12.030276  12.415559  13.701729  13.580398  13.803523  13.658473
#> Mar 2020  20.720703  18.775519  18.586304  18.429509  18.635212  18.457556
#> Apr 2020   8.496101  10.029824   9.920963  10.320040   9.733146   9.870554
#> May 2020         NA   8.550928  10.305022   9.053902   8.932767   8.806237
#> Jun 2020         NA   6.927299   8.147312   9.035381   8.497949   8.910284
#> Jul 2020         NA         NA   6.252404   7.652300   8.264779   8.859992
#> Aug 2020         NA         NA  12.928262   9.177542  10.537123  10.758344
#> Sep 2020         NA         NA         NA   7.730104  10.642811  10.802243
#> Oct 2020         NA         NA         NA  16.645686  15.235556  13.913446
#> Nov 2020         NA         NA         NA  10.980280  12.511666  12.203271
#> Dec 2020         NA         NA         NA  11.753986  14.897131  13.889536
#> Jan 2021         NA         NA         NA  24.249454  23.539254  22.363075
#> Feb 2021         NA         NA         NA  16.736334  19.563019  17.484327
#> Mar 2021         NA         NA         NA  17.499941  18.207234  17.000506
#> Apr 2021         NA         NA         NA  12.996977  11.461225  13.704313
#> May 2021         NA         NA         NA         NA  10.712998  11.078068
#> Jun 2021         NA         NA         NA         NA   5.589514   8.483694
#> Jul 2021         NA         NA         NA         NA   3.590430   6.931402
#> Aug 2021         NA         NA         NA         NA   8.555244   9.992852
#> Sep 2021         NA         NA         NA         NA         NA   9.043288
#> Oct 2021         NA         NA         NA         NA         NA  12.724900
#> Nov 2021         NA         NA         NA         NA         NA  14.075536
#> Dec 2021         NA         NA         NA         NA         NA   7.586890