Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14  2020-07-25 2020-10-16  2021-06-16  2021-07-04  2022-01-15
#> Jan 2020   10.92937  12.0862087 11.9929811  12.4018708  12.7717818  12.5380139
#> Feb 2020         NA  10.9208692  9.8612668  10.0389056  10.7063478  10.3600302
#> Mar 2020         NA   5.6302931  8.4866320   8.3908173   8.2703502   8.2341662
#> Apr 2020         NA  10.8769879  6.1221412   7.0389655   7.1502569   7.0978794
#> May 2020         NA  -4.6802967 -0.4430048  -0.2689973  -0.3231731  -0.0904267
#> Jun 2020         NA  -0.6791446 -4.2869432  -4.6131661  -5.0737106  -4.8701575
#> Jul 2020         NA -10.1321696 -7.7306863  -8.1328250  -8.2023978  -8.1528370
#> Aug 2020         NA          NA -9.5661207  -7.4652869  -6.9754184  -7.7790578
#> Sep 2020         NA          NA -0.8663152  -0.7398870  -1.0229068  -0.3979612
#> Oct 2020         NA          NA  0.3009215  -4.4229177  -3.6159051  -2.6886988
#> Nov 2020         NA          NA         NA  -3.5450584  -2.6810459  -4.9713854
#> Dec 2020         NA          NA         NA -12.9796610  -9.7736160 -10.2599448
#> Jan 2021         NA          NA         NA -12.0220116 -12.8192010 -12.1107186
#> Feb 2021         NA          NA         NA -10.0134325 -12.0770837 -14.7560563
#> Mar 2021         NA          NA         NA -16.7289517 -16.8185668 -15.3695524
#> Apr 2021         NA          NA         NA -20.4989943 -16.4669558 -16.8392327
#> May 2021         NA          NA         NA -18.6427081 -21.2589576 -21.5302045
#> Jun 2021         NA          NA         NA -20.2243628 -19.9713836 -20.4499605
#> Jul 2021         NA          NA         NA          NA -21.7396941 -24.6227330
#> Aug 2021         NA          NA         NA          NA          NA -15.6727415
#> Sep 2021         NA          NA         NA          NA          NA -15.9851505
#> Oct 2021         NA          NA         NA          NA          NA -17.3973375
#> Nov 2021         NA          NA         NA          NA          NA -12.0356463
#> Dec 2021         NA          NA         NA          NA          NA -11.6722246