Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-04 2021-06-13 2021-07-21 2021-07-23 2021-07-27 2021-09-26
#> Jan 2020   11.65607  13.801688  12.710104  11.343152  11.703850  11.656320
#> Feb 2020         NA  12.638333  11.212157  12.731503  12.571037  12.745760
#> Mar 2020         NA  10.532016  11.205995  11.159707  10.691536  11.103690
#> Apr 2020         NA   9.228794   9.928912   9.875749   9.110041   9.769733
#> May 2020         NA   8.004849   9.732271  12.037081  11.177072  11.318368
#> Jun 2020         NA  10.386425  12.095984  12.984044  12.804104  12.450895
#> Jul 2020         NA  11.205005  14.053487  14.100357  13.939640  14.554791
#> Aug 2020         NA  16.938964  16.236940  16.493993  16.145272  16.398185
#> Sep 2020         NA  14.848755  16.507493  17.694820  16.530321  16.587850
#> Oct 2020         NA  16.375106  14.319438  15.079264  15.162191  14.730937
#> Nov 2020         NA   9.366050  12.485158  13.060312  12.677990  12.294919
#> Dec 2020         NA   8.509013  10.930793  10.318171  10.867337  10.415796
#> Jan 2021         NA   3.416250   7.101651   7.887416   7.989170   8.287511
#> Feb 2021         NA   6.565632   6.924617   7.014272   7.305675   7.356182
#> Mar 2021         NA   4.933838   6.480647   6.033984   6.686270   6.517997
#> Apr 2021         NA   4.833797   5.312605   5.515787   5.427222   5.488383
#> May 2021         NA   3.382235   7.023817   7.263267   7.259365   6.914243
#> Jun 2021         NA   6.059657   3.761990   4.730963   4.900061   5.227554
#> Jul 2021         NA         NA   7.608055   6.904572   6.719220   5.990479
#> Aug 2021         NA         NA         NA         NA         NA   5.873273
#> Sep 2021         NA         NA         NA         NA         NA  11.558595
#> Oct 2021         NA         NA         NA         NA         NA         NA
#> Nov 2021         NA         NA         NA         NA         NA         NA
#> Dec 2021         NA         NA         NA         NA         NA         NA