Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   9.444528   10.60136   10.50814  10.917026  11.286937  11.053169
#> Feb 2020         NA   15.92978   14.87018  15.047819  15.715261  15.368944
#> Mar 2020         NA   12.94311   15.79945  15.703635  15.583168  15.546984
#> Apr 2020         NA   21.56822   16.81337  17.730195  17.841487  17.789109
#> May 2020         NA   14.09661   18.33391  18.507914  18.453738  18.686485
#> Jun 2020         NA   24.31097   20.70318  20.376952  19.916408  20.119961
#> Jul 2020         NA   16.77019   19.17167  18.769534  18.699961  18.749522
#> Aug 2020         NA         NA   14.61132  16.712149  17.202018  16.398378
#> Sep 2020         NA         NA   14.02207  14.148495  13.865475  14.490421
#> Oct 2020         NA         NA   15.34011  10.616274  11.423286  12.350492
#> Nov 2020         NA         NA         NA   6.192760   7.056773   4.766433
#> Dec 2020         NA         NA         NA   2.801819   6.007864   5.521535
#> Jan 2021         NA         NA         NA   4.711188   3.913999   4.622481
#> Feb 2021         NA         NA         NA  14.241793  12.178142   9.499170
#> Mar 2021         NA         NA         NA  11.944555  11.854939  13.303954
#> Apr 2021         NA         NA         NA  10.056532  14.088571  13.716294
#> May 2021         NA         NA         NA  17.692419  15.076169  14.804922
#> Jun 2021         NA         NA         NA  21.107672  21.360651  20.882074
#> Jul 2021         NA         NA         NA         NA  23.158885  20.275846
#> Aug 2021         NA         NA         NA         NA         NA  22.220893
#> Sep 2021         NA         NA         NA         NA         NA  13.778758
#> Oct 2021         NA         NA         NA         NA         NA  15.165788
#> Nov 2021         NA         NA         NA         NA         NA  12.735593
#> Dec 2021         NA         NA         NA         NA         NA  13.347655