Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-13 2020-05-16 2020-07-13 2021-03-25 2021-12-10 2022-01-23
#> Jan 2020   12.97336  10.191631   10.62886  10.580940  10.429539  10.326075
#> Feb 2020   10.69399   9.187389   10.46947   9.991256   9.661032   9.952713
#> Mar 2020   12.74122  15.451964   14.12639  14.834129  14.603072  14.740645
#> Apr 2020   17.77372  19.340934   20.84109  20.734829  21.066354  20.763636
#> May 2020         NA  18.606323   16.34094  16.833953  17.358698  17.471181
#> Jun 2020         NA         NA   19.45094  16.725712  16.632936  18.265496
#> Jul 2020         NA         NA   13.61532  12.899090  12.860125  13.287112
#> Aug 2020         NA         NA         NA  11.026981  11.993365  13.363852
#> Sep 2020         NA         NA         NA  13.531629  17.383816  17.399058
#> Oct 2020         NA         NA         NA  20.044125  18.546673  19.961477
#> Nov 2020         NA         NA         NA  24.638877  23.935124  24.923986
#> Dec 2020         NA         NA         NA  24.144151  24.696029  23.967542
#> Jan 2021         NA         NA         NA  33.527397  29.862472  29.663247
#> Feb 2021         NA         NA         NA  30.673317  34.233675  31.267486
#> Mar 2021         NA         NA         NA  40.271473  38.830843  37.213353
#> Apr 2021         NA         NA         NA         NA  47.426538  49.157798
#> May 2021         NA         NA         NA         NA  46.322563  45.869568
#> Jun 2021         NA         NA         NA         NA  43.129252  42.545171
#> Jul 2021         NA         NA         NA         NA  39.855688  38.261353
#> Aug 2021         NA         NA         NA         NA  36.720897  36.968627
#> Sep 2021         NA         NA         NA         NA  42.856542  42.493554
#> Oct 2021         NA         NA         NA         NA  48.765774  45.086857
#> Nov 2021         NA         NA         NA         NA  52.857985  48.304914
#> Dec 2021         NA         NA         NA         NA  53.072545  50.544823