Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   7.437084   8.593920   8.500692   8.909582   9.279493   9.045725
#> Feb 2020         NA   9.991845   8.932243   9.109881   9.777324   9.431006
#> Mar 2020         NA   8.081231  10.937570  10.841755  10.721288  10.685104
#> Apr 2020         NA  17.368300  12.613453  13.530278  13.641569  13.589192
#> May 2020         NA  12.512302  16.749594  16.923602  16.869426  17.102172
#> Jun 2020         NA  18.193032  14.585234  14.259011  13.798466  14.002019
#> Jul 2020         NA  14.311983  16.713466  16.311328  16.241755  16.291316
#> Aug 2020         NA         NA  16.244522  18.345356  18.835224  18.031585
#> Sep 2020         NA         NA  23.472614  23.599042  23.316023  23.940968
#> Oct 2020         NA         NA  34.052951  29.329111  30.136124  31.063330
#> Nov 2020         NA         NA         NA  30.644316  31.508329  29.217989
#> Dec 2020         NA         NA         NA  27.941993  31.148038  30.661710
#> Jan 2021         NA         NA         NA  29.353408  28.556218  29.264701
#> Feb 2021         NA         NA         NA  33.642888  31.579237  28.900264
#> Mar 2021         NA         NA         NA  27.864111  27.774496  29.223510
#> Apr 2021         NA         NA         NA  31.211665  35.243703  34.871426
#> May 2021         NA         NA         NA  41.958881  39.342632  39.071385
#> Jun 2021         NA         NA         NA  38.061385  38.314364  37.835787
#> Jul 2021         NA         NA         NA         NA  47.103294  44.220255
#> Aug 2021         NA         NA         NA         NA         NA  48.478083
#> Sep 2021         NA         NA         NA         NA         NA  45.930800
#> Oct 2021         NA         NA         NA         NA         NA  64.213784
#> Nov 2021         NA         NA         NA         NA         NA  67.128073
#> Dec 2021         NA         NA         NA         NA         NA  62.826741