Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020  11.528429  13.032154  12.956706  12.744415  12.632026  12.864051
#> Feb 2020   7.180635   7.565918   8.852088   8.730757   8.953882   8.808832
#> Mar 2020  13.698544  11.753360  11.564145  11.407350  11.613053  11.435398
#> Apr 2020  15.609740  17.143462  17.034601  17.433679  16.846784  16.984193
#> May 2020         NA  12.118938  13.873032  12.621912  12.500777  12.374247
#> Jun 2020         NA   9.970135  11.190148  12.078218  11.540786  11.953121
#> Jul 2020         NA         NA   9.782095  11.181991  11.794470  12.389683
#> Aug 2020         NA         NA   9.718034   5.967314   7.326895   7.548116
#> Sep 2020         NA         NA         NA   5.971397   8.884104   9.043536
#> Oct 2020         NA         NA         NA   9.951980   8.541850   7.219740
#> Nov 2020         NA         NA         NA   6.243464   7.774850   7.466455
#> Dec 2020         NA         NA         NA   3.214068   6.357213   5.349619
#> Jan 2021         NA         NA         NA  11.594274  10.884075   9.707895
#> Feb 2021         NA         NA         NA   5.241631   8.068316   5.989624
#> Mar 2021         NA         NA         NA   9.351259  10.058552   8.851824
#> Apr 2021         NA         NA         NA  10.682504   9.146752  11.389840
#> May 2021         NA         NA         NA         NA   2.186976   2.552046
#> Jun 2021         NA         NA         NA         NA   2.236436   5.130616
#> Jul 2021         NA         NA         NA         NA  -2.286628   1.054343
#> Aug 2021         NA         NA         NA         NA  -3.965672  -2.528063
#> Sep 2021         NA         NA         NA         NA         NA   3.569511
#> Oct 2021         NA         NA         NA         NA         NA   1.942092
#> Nov 2021         NA         NA         NA         NA         NA   5.266817
#> Dec 2021         NA         NA         NA         NA         NA  -1.134642