Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   13.38077   14.53761   14.44438   14.85327   15.22318   14.98941
#> Feb 2020         NA   16.25335   15.19374   15.37138   16.03882   15.69251
#> Mar 2020         NA   13.34310   16.19944   16.10363   15.98316   15.94697
#> Apr 2020         NA   21.57196   16.81711   17.73393   17.84522   17.79285
#> May 2020         NA   13.13500   17.37229   17.54629   17.49212   17.72487
#> Jun 2020         NA   21.73068   18.12288   17.79666   17.33611   17.53967
#> Jul 2020         NA   15.30272   17.70421   17.30207   17.23250   17.28206
#> Aug 2020         NA         NA   16.04495   18.14578   18.63565   17.83201
#> Sep 2020         NA         NA   14.95541   15.08184   14.79882   15.42376
#> Oct 2020         NA         NA   16.56617   11.84233   12.64935   13.57655
#> Nov 2020         NA         NA         NA   16.37010   17.23412   14.94378
#> Dec 2020         NA         NA         NA   14.92467   18.13072   17.64439
#> Jan 2021         NA         NA         NA   25.39539   24.59820   25.30668
#> Feb 2021         NA         NA         NA   29.61116   27.54751   24.86853
#> Mar 2021         NA         NA         NA   23.01084   22.92123   24.37024
#> Apr 2021         NA         NA         NA   21.60127   25.63330   25.26103
#> May 2021         NA         NA         NA   30.80244   28.18619   27.91495
#> Jun 2021         NA         NA         NA   30.85691   31.10989   30.63132
#> Jul 2021         NA         NA         NA         NA   29.75240   26.86937
#> Aug 2021         NA         NA         NA         NA         NA   32.05679
#> Sep 2021         NA         NA         NA         NA         NA   19.84824
#> Oct 2021         NA         NA         NA         NA         NA   16.81879
#> Nov 2021         NA         NA         NA         NA         NA   18.71982
#> Dec 2021         NA         NA         NA         NA         NA   22.12327