Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020  11.531086  13.034811  12.959363  12.747072  12.634683  12.866708
#> Feb 2020   7.263324   7.648607   8.934777   8.813446   9.036571   8.891521
#> Mar 2020  13.924047  11.978863  11.789648  11.632852  11.838556  11.660900
#> Apr 2020  15.836412  17.370134  17.261273  17.660351  17.073456  17.210865
#> May 2020         NA  21.830112  23.584206  22.333086  22.211952  22.085421
#> Jun 2020         NA  20.823381  22.043394  22.931464  22.394032  22.806366
#> Jul 2020         NA         NA  21.190910  22.590806  23.203285  23.798498
#> Aug 2020         NA         NA  27.540472  23.789752  25.149333  25.370554
#> Sep 2020         NA         NA         NA  23.275463  26.188170  26.347601
#> Oct 2020         NA         NA         NA  26.356365  24.946235  23.624125
#> Nov 2020         NA         NA         NA  16.724034  18.255421  17.947025
#> Dec 2020         NA         NA         NA  11.568002  14.711148  13.703553
#> Jan 2021         NA         NA         NA  16.207044  15.496845  14.320665
#> Feb 2021         NA         NA         NA   9.062385  11.889070   9.810378
#> Mar 2021         NA         NA         NA  10.770648  11.477941  10.271213
#> Apr 2021         NA         NA         NA  20.313771  18.778019  21.021107
#> May 2021         NA         NA         NA         NA  23.399641  23.764711
#> Jun 2021         NA         NA         NA         NA  22.299691  25.193871
#> Jul 2021         NA         NA         NA         NA  21.286360  24.627331
#> Aug 2021         NA         NA         NA         NA  25.887187  27.324795
#> Sep 2021         NA         NA         NA         NA         NA  35.572153
#> Oct 2021         NA         NA         NA         NA         NA  35.195954
#> Nov 2021         NA         NA         NA         NA         NA  36.631589
#> Dec 2021         NA         NA         NA         NA         NA  29.513778