Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   1.998816  3.5025415  3.4270935  3.2148019  3.1024130  3.3344385
#> Feb 2020  -0.962763 -0.5774805  0.7086897  0.5873590  0.8104841  0.6654336
#> Mar 2020  -0.411041 -2.3562250 -2.5454401 -2.7022353 -2.4965320 -2.6741874
#> Apr 2020  -2.789152 -1.2554298 -1.3642908 -0.9652131 -1.5521078 -1.4146995
#> May 2020         NA -2.7182233 -0.9641293 -2.2152492 -2.3363837 -2.4629142
#> Jun 2020         NA -3.2701469 -2.0501337 -1.1620641 -1.6994960 -1.2871611
#> Jul 2020         NA         NA -0.3648132  1.0350825  1.6475617  2.2427745
#> Aug 2020         NA         NA  3.6485610 -0.1021590  1.2574219  1.4786432
#> Sep 2020         NA         NA         NA -1.5518997  1.3608073  1.5202387
#> Oct 2020         NA         NA         NA 11.8060271 10.3958971  9.0737869
#> Nov 2020         NA         NA         NA  9.7761078 11.3074939 10.9990989
#> Dec 2020         NA         NA         NA  8.2410715 11.3842170 10.3766223
#> Jan 2021         NA         NA         NA  5.9923419  5.2821425  4.1059630
#> Feb 2021         NA         NA         NA  1.7274634  4.5541485  2.4754560
#> Mar 2021         NA         NA         NA -1.6321023 -0.9248096 -2.1315370
#> Apr 2021         NA         NA         NA -1.3537615 -2.8895135 -0.6464256
#> May 2021         NA         NA         NA         NA -0.9748963 -0.6098262
#> Jun 2021         NA         NA         NA         NA  2.3757808  5.2699614
#> Jul 2021         NA         NA         NA         NA  4.8368422  8.1778139
#> Aug 2021         NA         NA         NA         NA  3.4562889  4.8938973
#> Sep 2021         NA         NA         NA         NA         NA  6.5276275
#> Oct 2021         NA         NA         NA         NA         NA 17.0347229
#> Nov 2021         NA         NA         NA         NA         NA 22.6630308
#> Dec 2021         NA         NA         NA         NA         NA 18.2088800