Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   8.287687   9.444523   9.351295   9.760185  10.130096  9.8963280
#> Feb 2020         NA   7.191610   6.132007   6.309646   6.977088  6.6307705
#> Mar 2020         NA   7.856092  10.712431  10.616616  10.496149 10.4599648
#> Apr 2020         NA  16.472106  11.717259  12.634083  12.745375 12.6929974
#> May 2020         NA   9.620543  13.857835  14.031842  13.977666 14.2104127
#> Jun 2020         NA  17.256778  13.648979  13.322756  12.862212 13.0657649
#> Jul 2020         NA   7.637939  10.039422   9.637284   9.567711  9.6172717
#> Aug 2020         NA         NA   1.570594   3.671428   4.161296  3.3576570
#> Sep 2020         NA         NA   1.790762   1.917190   1.634170  2.2591158
#> Oct 2020         NA         NA   7.285543   2.561704   3.368717  4.2959232
#> Nov 2020         NA         NA         NA  12.603841  13.467854 11.1775141
#> Dec 2020         NA         NA         NA   6.330802   9.536847  9.0505178
#> Jan 2021         NA         NA         NA   8.744688   7.947499  8.6559810
#> Feb 2021         NA         NA         NA   4.869448   2.805797  0.1268241
#> Mar 2021         NA         NA         NA   3.198115   3.108500  4.5575144
#> Apr 2021         NA         NA         NA   2.782544   6.814583  6.4423058
#> May 2021         NA         NA         NA  10.809250   8.193001  7.9217538
#> Jun 2021         NA         NA         NA  10.974790  11.227769 10.7491923
#> Jul 2021         NA         NA         NA         NA  14.212679 11.3296401
#> Aug 2021         NA         NA         NA         NA         NA  8.9351813
#> Sep 2021         NA         NA         NA         NA         NA  0.2763846
#> Oct 2021         NA         NA         NA         NA         NA  5.0488699
#> Nov 2021         NA         NA         NA         NA         NA 16.8493696
#> Dec 2021         NA         NA         NA         NA         NA 14.2488181