Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25  2020-10-16  2021-06-16   2021-07-04  2022-01-15
#> Jan 2020   9.084563  10.241398 10.14817067  10.5570604  10.92697144  10.6932035
#> Feb 2020         NA   7.447548  6.38794547   6.5655843   7.23302650   6.8867089
#> Mar 2020         NA   1.516369  4.37270787   4.2768932   4.15642614   4.1202421
#> Apr 2020         NA   5.202715  0.44786843   1.3646927   1.47598416   1.4236066
#> May 2020         NA  -7.236099 -2.99880725  -2.8247997  -2.87897555  -2.6462291
#> Jun 2020         NA  -2.410011 -6.01780933  -6.3440322  -6.80457670  -6.6010236
#> Jul 2020         NA  -2.432829 -0.03134612  -0.4334848  -0.50305761  -0.4534968
#> Aug 2020         NA         NA  0.51254724   2.6133810   3.10324954   2.2996101
#> Sep 2020         NA         NA  0.11885643   0.2452847  -0.03773519   0.5872105
#> Oct 2020         NA         NA  0.86558606  -3.8582531  -3.05124056  -2.1240343
#> Nov 2020         NA         NA          NA   1.9322247   2.79623716   0.5058977
#> Dec 2020         NA         NA          NA   0.9554370   4.16148193   3.6751532
#> Jan 2021         NA         NA          NA   2.8958703   2.09868090   2.8071633
#> Feb 2021         NA         NA          NA  -1.7208098  -3.78446093  -6.4634335
#> Mar 2021         NA         NA          NA  -6.6149614  -6.70457657  -5.2555621
#> Apr 2021         NA         NA          NA -12.0463990  -8.01436052  -8.3866374
#> May 2021         NA         NA          NA  -8.7514911 -11.36774066 -11.6389875
#> Jun 2021         NA         NA          NA -17.6082106 -17.35523149 -17.8338084
#> Jul 2021         NA         NA          NA          NA  -5.31421149  -8.1972504
#> Aug 2021         NA         NA          NA          NA           NA   3.2505716
#> Sep 2021         NA         NA          NA          NA           NA  -3.8653732
#> Oct 2021         NA         NA          NA          NA           NA   0.3682568
#> Nov 2021         NA         NA          NA          NA           NA   6.4640384
#> Dec 2021         NA         NA          NA          NA           NA   4.6959315