Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for mts
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for data.frame
check_vertical(x, ...)

# S3 method for matrix
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for default
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   7.085207   8.588933   8.513485   8.301193   8.188804   8.420830
#> Feb 2020   4.406828   4.792110   6.078280   5.956950   6.180075   6.035024
#> Mar 2020  10.287699   8.342515   8.153300   7.996505   8.202208   8.024553
#> Apr 2020   9.579311  11.113033  11.004172  11.403250  10.816355  10.953763
#> May 2020         NA  12.942632  14.696726  13.445606  13.324471  13.197941
#> Jun 2020         NA  13.576282  14.796295  15.684365  15.146933  15.559268
#> Jul 2020         NA         NA  17.017445  18.417340  19.029820  19.625032
#> Aug 2020         NA         NA  20.444495  16.693775  18.053356  18.274577
#> Sep 2020         NA         NA         NA  10.511840  13.424547  13.583979
#> Oct 2020         NA         NA         NA  16.389107  14.978977  13.656867
#> Nov 2020         NA         NA         NA  11.092123  12.623509  12.315114
#> Dec 2020         NA         NA         NA   6.840491   9.983636   8.976041
#> Jan 2021         NA         NA         NA   6.102748   5.392549   4.216369
#> Feb 2021         NA         NA         NA   2.621082   5.447767   3.369074
#> Mar 2021         NA         NA         NA   4.849383   5.556676   4.349948
#> Apr 2021         NA         NA         NA   8.579349   7.043597   9.286685
#> May 2021         NA         NA         NA         NA  10.911924  11.276994
#> Jun 2021         NA         NA         NA         NA  14.208005  17.102186
#> Jul 2021         NA         NA         NA         NA  16.654434  19.995405
#> Aug 2021         NA         NA         NA         NA  14.537757  15.975365
#> Sep 2021         NA         NA         NA         NA         NA  13.483097
#> Oct 2021         NA         NA         NA         NA         NA  15.509188
#> Nov 2021         NA         NA         NA         NA         NA  13.044664
#> Dec 2021         NA         NA         NA         NA         NA   5.624456