Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   12.01865   13.52237   13.44693   13.23463   13.12225   13.35427
#> Feb 2020   12.86663   13.25191   14.53808   14.41675   14.63988   14.49482
#> Mar 2020   19.22306   17.27788   17.08866   16.93187   17.13757   16.95991
#> Apr 2020   18.67727   20.21099   20.10213   20.50121   19.91431   20.05172
#> May 2020         NA   17.93841   19.69250   18.44138   18.32025   18.19372
#> Jun 2020         NA   12.26604   13.48605   14.37412   13.83669   14.24903
#> Jul 2020         NA         NA   12.41436   13.81425   14.42673   15.02194
#> Aug 2020         NA         NA   22.97743   19.22671   20.58629   20.80751
#> Sep 2020         NA         NA         NA   23.24068   26.15338   26.31282
#> Oct 2020         NA         NA         NA   26.42645   25.01632   23.69421
#> Nov 2020         NA         NA         NA   14.29672   15.82811   15.51971
#> Dec 2020         NA         NA         NA   14.23367   17.37681   16.36922
#> Jan 2021         NA         NA         NA   22.33915   21.62895   20.45277
#> Feb 2021         NA         NA         NA   21.92111   24.74779   22.66910
#> Mar 2021         NA         NA         NA   23.74332   24.45062   23.24389
#> Apr 2021         NA         NA         NA   22.33855   20.80280   23.04589
#> May 2021         NA         NA         NA         NA   20.17394   20.53901
#> Jun 2021         NA         NA         NA         NA   18.57665   21.47083
#> Jul 2021         NA         NA         NA         NA   21.40620   24.74717
#> Aug 2021         NA         NA         NA         NA   31.85687   33.29448
#> Sep 2021         NA         NA         NA         NA         NA   40.14370
#> Oct 2021         NA         NA         NA         NA         NA   34.17571
#> Nov 2021         NA         NA         NA         NA         NA   25.12773
#> Dec 2021         NA         NA         NA         NA         NA   21.48470