Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   18.07435   19.57808   19.50263  19.290339 19.1779505  19.409976
#> Feb 2020   13.89195   14.27723   15.56340  15.442072 15.6651971  15.520147
#> Mar 2020   17.01691   15.07173   14.88252  14.725720 14.9314235  14.753768
#> Apr 2020   10.57378   12.10750   11.99864  12.397718 11.8108229  11.948231
#> May 2020         NA   15.83380   17.58789  16.336770 16.2156351  16.089105
#> Jun 2020         NA   15.14600   16.36602  17.254087 16.7166550  17.128990
#> Jul 2020         NA         NA   20.13195  21.531842 22.1443208  22.739534
#> Aug 2020         NA         NA   22.84117  19.090446 20.4500268  20.671248
#> Sep 2020         NA         NA         NA  18.697232 21.6099392  21.769371
#> Oct 2020         NA         NA         NA  20.987617 19.5774869  18.255377
#> Nov 2020         NA         NA         NA  15.847163 17.3785489  17.070154
#> Dec 2020         NA         NA         NA  12.782203 15.9253486  14.917754
#> Jan 2021         NA         NA         NA  22.404346 21.6941471  20.517968
#> Feb 2021         NA         NA         NA  11.364762 14.1914468  12.112754
#> Mar 2021         NA         NA         NA   5.907653  6.6149455   5.408218
#> Apr 2021         NA         NA         NA   1.401268 -0.1344843   2.108604
#> May 2021         NA         NA         NA         NA  1.7789242   2.143994
#> Jun 2021         NA         NA         NA         NA  3.8742236   6.768404
#> Jul 2021         NA         NA         NA         NA  7.2882971  10.629269
#> Aug 2021         NA         NA         NA         NA  6.9243224   8.361931
#> Sep 2021         NA         NA         NA         NA         NA  14.068565
#> Oct 2021         NA         NA         NA         NA         NA  12.898168
#> Nov 2021         NA         NA         NA         NA         NA  17.584720
#> Dec 2021         NA         NA         NA         NA         NA   9.924516