Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20  2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   4.431426   5.935152   5.859704   5.6474120   5.535023   5.767049
#> Feb 2020   1.866356   2.251638   3.537809   3.4164780   3.639603   3.494553
#> Mar 2020   8.181720   6.236536   6.047320   5.8905252   6.096229   5.918573
#> Apr 2020   2.140698   3.674421   3.565560   3.9646372   3.377742   3.515151
#> May 2020         NA  -4.557380  -2.803286  -4.0544056  -4.175540  -4.302071
#> Jun 2020         NA  -9.863611  -8.643597  -7.7555277  -8.292960  -7.880625
#> Jul 2020         NA         NA -12.630861 -11.2309651 -10.618486 -10.023273
#> Aug 2020         NA         NA  -4.440402  -8.1911224  -6.831541  -6.610320
#> Sep 2020         NA         NA         NA  -9.4934395  -6.580732  -6.421301
#> Oct 2020         NA         NA         NA  -0.3483632  -1.758493  -3.080603
#> Nov 2020         NA         NA         NA  -6.6375819  -5.106196  -5.414591
#> Dec 2020         NA         NA         NA -10.6044625  -7.461317  -8.468912
#> Jan 2021         NA         NA         NA  -6.8137295  -7.523929  -8.700108
#> Feb 2021         NA         NA         NA  -9.6169195  -6.790234  -8.868927
#> Mar 2021         NA         NA         NA  -7.5132124  -6.805920  -8.012647
#> Apr 2021         NA         NA         NA  -7.8213745  -9.357126  -7.114039
#> May 2021         NA         NA         NA          NA -15.781859 -15.416789
#> Jun 2021         NA         NA         NA          NA -15.597196 -12.703016
#> Jul 2021         NA         NA         NA          NA -17.761672 -14.420701
#> Aug 2021         NA         NA         NA          NA -19.515605 -18.077996
#> Sep 2021         NA         NA         NA          NA         NA -19.437066
#> Oct 2021         NA         NA         NA          NA         NA -14.093547
#> Nov 2021         NA         NA         NA          NA         NA -12.853078
#> Dec 2021         NA         NA         NA          NA         NA -19.441723