Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20  2021-04-02  2021-08-01  2021-12-08
#> Jan 2020  2.0594660  3.5631916  3.4877436   3.2754520   3.1630631   3.3950886
#> Feb 2020  3.5799429  3.9652255  5.2513957   5.1300650   5.3531901   5.2081396
#> Mar 2020  6.2280362  4.2828522  4.0936370   3.9368419   4.1425452   3.9648898
#> Apr 2020 -0.6724359  0.8612863  0.7524253   1.1515030   0.5646082   0.7020166
#> May 2020         NA -1.1340435  0.6200506  -0.6310693  -0.7522039  -0.8787343
#> Jun 2020         NA -2.7786480 -1.5586348  -0.6705652  -1.2079971  -0.7956623
#> Jul 2020         NA         NA  0.2013479   1.6012436   2.2137229   2.8089356
#> Aug 2020         NA         NA  2.7132684  -1.0374516   0.3221293   0.5433506
#> Sep 2020         NA         NA         NA  -5.1295175  -2.2168105  -2.0573791
#> Oct 2020         NA         NA         NA  -2.8606692  -4.2707991  -5.5929094
#> Nov 2020         NA         NA         NA -12.0560257 -10.5246396 -10.8330346
#> Dec 2020         NA         NA         NA  -7.8250623  -4.6819167  -5.6895114
#> Jan 2021         NA         NA         NA  -9.1836653  -9.8938647 -11.0700442
#> Feb 2021         NA         NA         NA  -9.3501251  -6.5234400  -8.6021325
#> Mar 2021         NA         NA         NA -11.2471067 -10.5398140 -11.7465414
#> Apr 2021         NA         NA         NA -12.5349722 -14.0707243 -11.8276363
#> May 2021         NA         NA         NA          NA -14.0812913 -13.7162212
#> Jun 2021         NA         NA         NA          NA -15.2848834 -12.3907029
#> Jul 2021         NA         NA         NA          NA -14.6189171 -11.2779454
#> Aug 2021         NA         NA         NA          NA -12.1452355 -10.7076271
#> Sep 2021         NA         NA         NA          NA          NA  -9.6361481
#> Oct 2021         NA         NA         NA          NA          NA -16.1494176
#> Nov 2021         NA         NA         NA          NA          NA -16.2022734
#> Dec 2021         NA         NA         NA          NA          NA -19.3167564