Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for mts
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for data.frame
check_vertical(x, ...)

# S3 method for matrix
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for default
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to `check_vertical` according to the class of the object `x`

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20   2021-04-02  2021-08-01  2021-12-08
#> Jan 2020  2.0967056  3.6004312  3.5249832   3.31269157   3.2003026   3.4323281
#> Feb 2020 -0.4934085 -0.1081259  1.1780443   1.05671354   1.2798387   1.1347882
#> Mar 2020  0.8971204 -1.0480635 -1.2372787  -1.39407386  -1.1883706  -1.3660259
#> Apr 2020 -1.0732646  0.4604576  0.3515966   0.75067429   0.1637796   0.3011879
#> May 2020         NA  2.5533202  4.3074142   3.05629436   2.9351598   2.8086293
#> Jun 2020         NA -1.0509597  0.1690535   1.05712307   0.5196913   0.9320261
#> Jul 2020         NA         NA -1.4337237  -0.03382797   0.5786513   1.1738640
#> Aug 2020         NA         NA  0.3411845  -3.40953556  -2.0499547  -1.8287333
#> Sep 2020         NA         NA         NA  -1.96899236   0.9437147   1.1031461
#> Oct 2020         NA         NA         NA   6.21154875   4.8014188   3.4793085
#> Nov 2020         NA         NA         NA   3.23174456   4.7631306   4.4547356
#> Dec 2020         NA         NA         NA   0.26915470   3.4123003   2.4047055
#> Jan 2021         NA         NA         NA  -3.68798402  -4.3981835  -5.5743629
#> Feb 2021         NA         NA         NA -13.30061844 -10.4739333 -12.5526259
#> Mar 2021         NA         NA         NA -12.98117787 -12.2738852 -13.4806126
#> Apr 2021         NA         NA         NA  -7.01711253  -8.5528645  -6.3097766
#> May 2021         NA         NA         NA           NA  -7.9755332  -7.6104631
#> Jun 2021         NA         NA         NA           NA -15.4220594 -12.5278788
#> Jul 2021         NA         NA         NA           NA -21.2448345 -17.9038628
#> Aug 2021         NA         NA         NA           NA -30.6214086 -29.1838002
#> Sep 2021         NA         NA         NA           NA          NA -23.9098656
#> Oct 2021         NA         NA         NA           NA          NA -24.8473892
#> Nov 2021         NA         NA         NA           NA          NA -25.4066773
#> Dec 2021         NA         NA         NA           NA          NA -32.5368021