Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02  2021-08-01  2021-12-08
#> Jan 2020   1.142592  2.6463176  2.5708696  2.3585780  2.24618904  2.47821455
#> Feb 2020   5.247765  5.6330472  6.9192175  6.7978867  7.02101189  6.87596138
#> Mar 2020   5.751753  3.8065693  3.6173542  3.4605590  3.66626231  3.48860692
#> Apr 2020  -1.299518  0.2342039  0.1253428  0.5244206 -0.06247419  0.07493415
#> May 2020         NA  2.1842864  3.9383804  2.6872605  2.56612594  2.43959550
#> Jun 2020         NA -0.2849527  0.9350605  1.8231300  1.28569823  1.69803303
#> Jul 2020         NA         NA  3.8340842  5.2339799  5.84645915  6.44167189
#> Aug 2020         NA         NA 12.8274060  9.0766860 10.43626692 10.65748825
#> Sep 2020         NA         NA         NA  8.0492952 10.96200227 11.12143362
#> Oct 2020         NA         NA         NA 12.4093997 10.99926977  9.67715952
#> Nov 2020         NA         NA         NA 10.4109477 11.94233379 11.63393877
#> Dec 2020         NA         NA         NA  8.5484644 11.69161000 10.68401528
#> Jan 2021         NA         NA         NA  4.4606493  3.75044991  2.57427046
#> Feb 2021         NA         NA         NA  4.8925584  7.71924345  5.64055091
#> Mar 2021         NA         NA         NA  4.8180064  5.52529904  4.31857164
#> Apr 2021         NA         NA         NA  3.0317624  1.49601035  3.73909827
#> May 2021         NA         NA         NA         NA  5.52970179  5.89477185
#> Jun 2021         NA         NA         NA         NA  6.87462278  9.76880337
#> Jul 2021         NA         NA         NA         NA  7.18895520 10.52992691
#> Aug 2021         NA         NA         NA         NA 12.54413851 13.98174690
#> Sep 2021         NA         NA         NA         NA          NA 16.13582852
#> Oct 2021         NA         NA         NA         NA          NA 15.03123597
#> Nov 2021         NA         NA         NA         NA          NA 21.67904513
#> Dec 2021         NA         NA         NA         NA          NA 19.19383942