Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   9.926965  11.430691   11.35524  11.142951  11.030562  11.262588
#> Feb 2020   9.423913   9.809196   11.09537  10.974035  11.197161  11.052110
#> Mar 2020  16.129566  14.184382   13.99517  13.838372  14.044075  13.866420
#> Apr 2020  16.539417  18.073139   17.96428  18.363356  17.776461  17.913869
#> May 2020         NA  18.438003   20.19210  18.940977  18.819842  18.693312
#> Jun 2020         NA  15.391233   16.61125  17.499315  16.961884  17.374218
#> Jul 2020         NA         NA   12.65119  14.051083  14.663562  15.258775
#> Aug 2020         NA         NA   16.65203  12.901313  14.260894  14.482115
#> Sep 2020         NA         NA         NA   9.785786  12.698493  12.857924
#> Oct 2020         NA         NA         NA  19.166345  17.756215  16.434104
#> Nov 2020         NA         NA         NA   9.293357  10.824743  10.516348
#> Dec 2020         NA         NA         NA   5.262553   8.405699   7.398104
#> Jan 2021         NA         NA         NA  11.040143  10.329944   9.153765
#> Feb 2021         NA         NA         NA   5.294320   8.121005   6.042312
#> Mar 2021         NA         NA         NA   5.809684   6.516977   5.310249
#> Apr 2021         NA         NA         NA  10.070103   8.534351  10.777439
#> May 2021         NA         NA         NA         NA  13.717975  14.083045
#> Jun 2021         NA         NA         NA         NA  14.628633  17.522813
#> Jul 2021         NA         NA         NA         NA   7.022489  10.363460
#> Aug 2021         NA         NA         NA         NA   5.727156   7.164764
#> Sep 2021         NA         NA         NA         NA         NA   4.945619
#> Oct 2021         NA         NA         NA         NA         NA  12.838039
#> Nov 2021         NA         NA         NA         NA         NA   8.805752
#> Dec 2021         NA         NA         NA         NA         NA   2.699482