Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

Integer. Periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-01-14 2020-07-25 2020-10-16 2021-06-16 2021-07-04 2022-01-15
#> Jan 2020   5.702682  6.8595181   6.766290   7.175180   7.545091   7.311323
#> Feb 2020         NA  2.0966591   1.037057   1.214695   1.882138   1.535820
#> Mar 2020         NA -0.7394578   2.116881   2.021066   1.900599   1.864415
#> Apr 2020         NA  6.8024401   2.047593   2.964418   3.075709   3.023331
#> May 2020         NA  1.7405286   5.977821   6.151828   6.097652   6.330399
#> Jun 2020         NA 14.9333616  11.325563  10.999340  10.538796  10.742349
#> Jul 2020         NA  6.2346527   8.636136   8.233997   8.164424   8.213985
#> Aug 2020         NA         NA   7.442052   9.542886  10.032754   9.229115
#> Sep 2020         NA         NA  12.052777  12.179205  11.896185  12.521131
#> Oct 2020         NA         NA  13.895474   9.171635   9.978648  10.905854
#> Nov 2020         NA         NA         NA  10.634912  11.498925   9.208585
#> Dec 2020         NA         NA         NA   6.143577   9.349622   8.863293
#> Jan 2021         NA         NA         NA   5.386084   4.588895   5.297377
#> Feb 2021         NA         NA         NA   4.614259   2.550608  -0.128365
#> Mar 2021         NA         NA         NA   2.154375   2.064760   3.513774
#> Apr 2021         NA         NA         NA  -5.125427  -1.093388  -1.465665
#> May 2021         NA         NA         NA   4.361726   1.745477   1.474230
#> Jun 2021         NA         NA         NA   1.511887   1.764866   1.286289
#> Jul 2021         NA         NA         NA         NA   0.121867  -2.761172
#> Aug 2021         NA         NA         NA         NA         NA   6.617023
#> Sep 2021         NA         NA         NA         NA         NA   1.337646
#> Oct 2021         NA         NA         NA         NA         NA   4.359635
#> Nov 2021         NA         NA         NA         NA         NA   4.706864
#> Dec 2021         NA         NA         NA         NA         NA  -1.310639