Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   8.013592   9.517317   9.441869   9.229578   9.117189   9.349214
#> Feb 2020   7.515210   7.900492   9.186663   9.065332   9.288457   9.143406
#> Mar 2020  11.928658   9.983474   9.794259   9.637464   9.843167   9.665512
#> Apr 2020   9.000685  10.534408  10.425547  10.824624  10.237730  10.375138
#> May 2020         NA   9.902062  11.656156  10.405036  10.283901  10.157371
#> Jun 2020         NA  11.050916  12.270929  13.158999  12.621567  13.033902
#> Jul 2020         NA         NA  13.044301  14.444196  15.056676  15.651888
#> Aug 2020         NA         NA  23.271310  19.520590  20.880171  21.101392
#> Sep 2020         NA         NA         NA  15.606595  18.519302  18.678733
#> Oct 2020         NA         NA         NA  16.134776  14.724646  13.402535
#> Nov 2020         NA         NA         NA  11.618373  13.149759  12.841364
#> Dec 2020         NA         NA         NA   9.375609  12.518755  11.511160
#> Jan 2021         NA         NA         NA   9.359352   8.649153   7.472973
#> Feb 2021         NA         NA         NA   6.389821   9.216506   7.137814
#> Mar 2021         NA         NA         NA   9.636219  10.343512   9.136785
#> Apr 2021         NA         NA         NA  11.349374   9.813622  12.056710
#> May 2021         NA         NA         NA         NA   9.913073  10.278143
#> Jun 2021         NA         NA         NA         NA  13.602318  16.496498
#> Jul 2021         NA         NA         NA         NA  14.797581  18.138553
#> Aug 2021         NA         NA         NA         NA  20.968670  22.406278
#> Sep 2021         NA         NA         NA         NA         NA  23.126703
#> Oct 2021         NA         NA         NA         NA         NA  21.776993
#> Nov 2021         NA         NA         NA         NA         NA  30.162575
#> Dec 2021         NA         NA         NA         NA         NA  22.593948