Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   8.339769   9.843495   9.768047   9.555755   9.443366   9.675392
#> Feb 2020  10.716164  11.101446  12.387617  12.266286  12.489411  12.344360
#> Mar 2020  12.183431  10.238247  10.049032   9.892237  10.097940   9.920285
#> Apr 2020   7.134461   8.668183   8.559322   8.958400   8.371505   8.508913
#> May 2020         NA   5.310611   7.064705   5.813585   5.692451   5.565920
#> Jun 2020         NA   8.909228  10.129241  11.017311  10.479879  10.892214
#> Jul 2020         NA         NA   8.997725  10.397620  11.010100  11.605312
#> Aug 2020         NA         NA  14.084853  10.334133  11.693713  11.914935
#> Sep 2020         NA         NA         NA  12.327045  15.239752  15.399183
#> Oct 2020         NA         NA         NA  14.658183  13.248053  11.925943
#> Nov 2020         NA         NA         NA  10.194304  11.725690  11.417295
#> Dec 2020         NA         NA         NA  14.754163  17.897309  16.889714
#> Jan 2021         NA         NA         NA  25.170372  24.460173  23.283993
#> Feb 2021         NA         NA         NA  29.298161  32.124846  30.046153
#> Mar 2021         NA         NA         NA  37.355400  38.062693  36.855966
#> Apr 2021         NA         NA         NA  40.266501  38.730749  40.973837
#> May 2021         NA         NA         NA         NA  32.847329  33.212400
#> Jun 2021         NA         NA         NA         NA  33.112756  36.006936
#> Jul 2021         NA         NA         NA         NA  27.338774  30.679746
#> Aug 2021         NA         NA         NA         NA  26.019642  27.457250
#> Sep 2021         NA         NA         NA         NA         NA  28.403025
#> Oct 2021         NA         NA         NA         NA         NA  21.840363
#> Nov 2021         NA         NA         NA         NA         NA  21.500311
#> Dec 2021         NA         NA         NA         NA         NA  16.152135