Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

# S3 method for class 'mts'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# S3 method for class 'data.frame'
check_vertical(x, ...)

# S3 method for class 'matrix'
check_vertical(x, periodicity, date_format = "%Y-%m-%d", ...)

# Default S3 method
check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

...

Arguments to be passed to check_vertical according to the class of the object x

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020   5.882732   7.386458   7.311010  7.0987183   6.986329  7.2183549
#> Feb 2020   4.982352   5.367634   6.653805  6.5324740   6.755599  6.6105486
#> Mar 2020   7.906751   5.961567   5.772352  5.6155566   5.821260  5.6436045
#> Apr 2020   4.726823   6.260546   6.151685  6.5507623   5.963868  6.1012759
#> May 2020         NA   6.756481   8.510575  7.2594551   7.138321  7.0117901
#> Jun 2020         NA   6.817953   8.037966  8.9260360   8.388604  8.8009390
#> Jul 2020         NA         NA   9.907391 11.3072870  11.919766 12.5149789
#> Aug 2020         NA         NA  13.191200  9.4404802  10.800061 11.0212824
#> Sep 2020         NA         NA         NA  8.0605140  10.973221 11.1326524
#> Oct 2020         NA         NA         NA 16.6948097  15.284680 13.9625695
#> Nov 2020         NA         NA         NA 11.0790439  12.610430 12.3020349
#> Dec 2020         NA         NA         NA  9.2082362  12.351382 11.3437871
#> Jan 2021         NA         NA         NA  5.5314298   4.821230  3.6450509
#> Feb 2021         NA         NA         NA  0.7997871   3.626472  1.5477796
#> Mar 2021         NA         NA         NA  1.2050866   1.912379  0.7056519
#> Apr 2021         NA         NA         NA  6.5556600   5.019908  7.2629959
#> May 2021         NA         NA         NA         NA   4.534881  4.8999515
#> Jun 2021         NA         NA         NA         NA   3.951776  6.8459565
#> Jul 2021         NA         NA         NA         NA   7.133717 10.4746883
#> Aug 2021         NA         NA         NA         NA  11.232174 12.6697826
#> Sep 2021         NA         NA         NA         NA         NA 18.1732418
#> Oct 2021         NA         NA         NA         NA         NA 23.1155894
#> Nov 2021         NA         NA         NA         NA         NA 28.0749805
#> Dec 2021         NA         NA         NA         NA         NA 21.6381729