Check vertical format

Check vertical format

Usage

check_vertical(x, ...)

Arguments

x

a formatted data.frame containing the input in the vertical format

periodicity

periodicity of the time period (12, 4 or 1 for resp. monthly, quarterly or annual data)

date_format

character string corresponding to the format used in the input data.frame for the revision dates.

Value

the same input but in a ts object and with revision date formatted

Examples

long_format <- rjd3revisions:::simulate_long(
    start_period = as.Date("2020-01-01"),
    n_period = 24,
    n_revision = 6,
    periodicity = 12L
)
vertical_format <- rjd3revisions:::from_long_to_vertical(long_format, periodicity = 12L)
check_vertical(vertical_format)
#>          2020-04-12 2020-06-16 2020-08-20 2021-04-02 2021-08-01 2021-12-08
#> Jan 2020  12.473735   13.97746  13.902013 13.6897210 13.5773321 13.8093576
#> Feb 2020  13.476297   13.86158  15.147749 15.0264186 15.2495437 15.1044932
#> Mar 2020  18.367279   16.42209  16.232880 16.0760847 16.2817880 16.1041326
#> Apr 2020   9.991821   11.52554  11.416683 11.8157603 11.2288655 11.3662739
#> May 2020         NA   10.93510  12.689195 11.4380748 11.3169402 11.1904098
#> Jun 2020         NA   14.37994  15.599954 16.4880238 15.9505920 16.3629268
#> Jul 2020         NA         NA   7.643382  9.0432774  9.6557566 10.2509693
#> Aug 2020         NA         NA  12.411641  8.6609206 10.0205015 10.2417228
#> Sep 2020         NA         NA         NA  0.8094656  3.7221727  3.8816040
#> Oct 2020         NA         NA         NA  3.1539099  1.7437800  0.4216697
#> Nov 2020         NA         NA         NA -3.3500768 -1.8186907 -2.1270857
#> Dec 2020         NA         NA         NA -2.6883950  0.4547505 -0.5528442
#> Jan 2021         NA         NA         NA 11.0794119 10.3692125  9.1930330
#> Feb 2021         NA         NA         NA  8.6536767 11.4803618  9.4016692
#> Mar 2021         NA         NA         NA 12.2691631 12.9764558 11.7697284
#> Apr 2021         NA         NA         NA  7.5820012  6.0462491  8.2893371
#> May 2021         NA         NA         NA         NA  8.6876574  9.0527275
#> Jun 2021         NA         NA         NA         NA 18.1501023 21.0442829
#> Jul 2021         NA         NA         NA         NA  6.9964488 10.3374205
#> Aug 2021         NA         NA         NA         NA  9.9540914 11.3916998
#> Sep 2021         NA         NA         NA         NA         NA  8.6637221
#> Oct 2021         NA         NA         NA         NA         NA  7.6964779
#> Nov 2021         NA         NA         NA         NA         NA  6.7048652
#> Dec 2021         NA         NA         NA         NA         NA  0.6762678