Efficiency Model 1

Linear regression model of the revisions (R) on a preliminary vintage (P)

Usage

efficiencyModel1(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
efficiencyModel1(vintages[["diagonal_view"]])
#>                            N         R2         F intercept.estimate
#> [Release[2]]-[Release[1]] 19 0.01283303 0.2209975          0.1363549
#> [Release[3]]-[Release[2]] 16 0.12520114 2.0036787         -0.6855472
#> [Release[4]]-[Release[3]] 10 0.43472636 6.1524377          1.1451832
#> [Release[5]]-[Release[4]]  9 0.17219775 1.4561258         -0.3737869
#>                           intercept.stderr intercept.pvalue slope.estimate
#> [Release[2]]-[Release[1]]        0.9824646       0.89124803   -0.008101707
#> [Release[3]]-[Release[2]]        0.5235022       0.21143558    0.016423023
#> [Release[4]]-[Release[3]]        0.5124609       0.05588792   -0.046375895
#> [Release[5]]-[Release[4]]        0.2737676       0.21439928    0.012803985
#>                           slope.stderr slope.pvalue   skewness   kurtosis
#> [Release[2]]-[Release[1]]   0.01723387   0.64425529  1.3094666  1.6961203
#> [Release[3]]-[Release[2]]   0.01160217   0.17877773  0.3411671 -0.9031256
#> [Release[4]]-[Release[3]]   0.01869686   0.03808726 -0.5984240 -0.8518502
#> [Release[5]]-[Release[4]]   0.01061074   0.26674003  0.1664578  0.7157798
#>                           JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> [Release[2]]-[Release[1]]        7.3017261        0.02596871     0.112040808
#> [Release[3]]-[Release[2]]        0.8007599        0.67006542     0.089780980
#> [Release[4]]-[Release[3]]        0.8092852        0.66721522     0.083332126
#> [Release[5]]-[Release[4]]        0.2077245        0.90134944     0.001005461
#>                           BreuschPagan.value BreuschPagan.pvalue   White.R2
#> [Release[2]]-[Release[1]]        2.145023963           0.1612812 0.27543108
#> [Release[3]]-[Release[2]]        1.380913494           0.2595452 0.14994759
#> [Release[4]]-[Release[3]]        0.727261233           0.4185708 0.46234276
#> [Release[5]]-[Release[4]]        0.007045312           0.9354570 0.00263113
#>                           White.value White.pvalue      arch.R2   arch.value
#> [Release[2]]-[Release[1]]  5.23319045   0.07305116 1.027382e-02 1.849288e-01
#> [Release[3]]-[Release[2]]  2.39916142   0.30132053 8.206081e-03 1.230912e-01
#> [Release[4]]-[Release[3]]  4.62342756   0.09909129 9.108695e-07 8.197826e-06
#> [Release[5]]-[Release[4]]  0.02368017   0.98822973 1.547083e-01 1.237666e+00
#>                           arch.pvalue
#> [Release[2]]-[Release[1]]   0.6671713
#> [Release[3]]-[Release[2]]   0.7257057
#> [Release[4]]-[Release[3]]   0.9977155
#> [Release[5]]-[Release[4]]   0.2659213