Efficiency Model 1

Linear regression model of the revisions (R) on a preliminary vintage (P)

Usage

efficiencyModel1(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
efficiencyModel1(vintages[["diagonal_view"]])
#>                            N          R2         F intercept.estimate
#> [Release[2]]-[Release[1]] 31 0.035071535 1.0540414         0.99408661
#> [Release[3]]-[Release[2]] 25 0.107211493 2.7619804         0.62104373
#> [Release[4]]-[Release[3]] 23 0.004813013 0.1015621        -0.30201799
#> [Release[5]]-[Release[4]] 11 0.515788675 9.5869259         0.08987556
#>                           intercept.stderr intercept.pvalue slope.estimate
#> [Release[2]]-[Release[1]]       0.74080275       0.19003997     0.01828467
#> [Release[3]]-[Release[2]]       0.27004595       0.03087481     0.01294279
#> [Release[4]]-[Release[3]]       0.14598617       0.05109623    -0.00145085
#> [Release[5]]-[Release[4]]       0.09209355       0.35460629     0.02894538
#>                           slope.stderr slope.pvalue    skewness   kurtosis
#> [Release[2]]-[Release[1]]  0.017809772   0.31306594 -0.03338609  0.2895871
#> [Release[3]]-[Release[2]]  0.007787850   0.11009970 -0.51175393  0.3549411
#> [Release[4]]-[Release[3]]  0.004552572   0.75311105 -0.21635863 -0.3850843
#> [Release[5]]-[Release[4]]  0.009348450   0.01279914 -0.54166333  0.9299452
#>                           JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> [Release[2]]-[Release[1]]        0.1103990         0.9462963      0.07359144
#> [Release[3]]-[Release[2]]        1.1735515         0.5561175      0.04080190
#> [Release[4]]-[Release[3]]        0.3075730         0.8574551      0.05567029
#> [Release[5]]-[Release[4]]        0.8493311         0.6539885      0.14397228
#>                           BreuschPagan.value BreuschPagan.pvalue   White.R2
#> [Release[2]]-[Release[1]]          2.3036832           0.1398948 0.08720619
#> [Release[3]]-[Release[2]]          0.9783627           0.3328957 0.07888948
#> [Release[4]]-[Release[3]]          1.2379956           0.2784328 0.05577587
#> [Release[5]]-[Release[4]]          1.5136782           0.2497655 0.16098348
#>                           White.value White.pvalue      arch.R2   arch.value
#> [Release[2]]-[Release[1]]    2.703392    0.2588010 2.974097e-05 0.0008922291
#> [Release[3]]-[Release[2]]    1.972237    0.3730218 7.377195e-03 0.1770526697
#> [Release[4]]-[Release[3]]    1.282845    0.5265429 5.976352e-02 1.3147973425
#> [Release[5]]-[Release[4]]    1.770818    0.4125454 9.675632e-03 0.0967563208
#>                           arch.pvalue
#> [Release[2]]-[Release[1]]   0.9761706
#> [Release[3]]-[Release[2]]   0.6739185
#> [Release[4]]-[Release[3]]   0.2515279
#> [Release[5]]-[Release[4]]   0.7557575