Efficiency Model 2

Linear regression model of R_v on R_{v-1}

Usage

efficiencyModel2(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
period_range <- seq(as.Date('2011-01-01'),as.Date('2020-10-01'),by='quarter')
qtr <- (as.numeric(substr(period_range,6,7))+2)/3
time_period <- rep(paste0(format(period_range, "%Y"), "Q", qtr),5)
np <- length(period_range)
rev_date <- c(rep("2021-06-30",np), rep("2021-12-31",np), rep("2022-06-30",np),
            rep("2022-12-31",np), rep("2023-06-30",np))
set.seed(1)
xt <- cumsum(sample(rnorm(1000,0,1), np, TRUE))
rev <- rnorm(np*4,0,.1)
obs_values <- xt
for(i in 1:4) {
  xt <- xt+rev[(1+(i-1)*np):(i*np)]
  obs_values <- c(obs_values,xt)
}
df <- data.frame(rev_date, time_period, obs_values)

## Create vintage and test
vintages <- create_vintages(df, periodicity = 4)
efficiencyModel2(vintages$diagonal_view)
#>                            N         R2         F intercept.estimate
#> [Release[3]]-[Release[2]] 40 0.03414641 1.3434373        -0.01502902
#> [Release[4]]-[Release[3]] 40 0.01248758 0.4805288        -0.02512668
#> [Release[5]]-[Release[4]] 40 0.03583727 1.4124340        -0.02279074
#>                           intercept.stderr intercept.pvalue slope.estimate
#> [Release[3]]-[Release[2]]       0.01769005       0.40088255     0.22683721
#> [Release[4]]-[Release[3]]       0.01476059       0.09687119    -0.09476903
#> [Release[5]]-[Release[4]]       0.01557162       0.15152475    -0.19619317
#>                           slope.stderr slope.pvalue   skewness    kurtosis
#> [Release[3]]-[Release[2]]    0.1957067    0.2536606 -0.2663346 -0.70594162
#> [Release[4]]-[Release[3]]    0.1367120    0.4923983  0.1428890  0.05980705
#> [Release[5]]-[Release[4]]    0.1650820    0.2420273 -0.3799207  0.48908872
#>                           JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> [Release[3]]-[Release[2]]        1.2708964         0.5296980    0.0008479461
#> [Release[4]]-[Release[3]]        0.1385246         0.9330819    0.0127461791
#> [Release[5]]-[Release[4]]        1.3269210         0.5150659    0.0759441218
#>                           BreuschPagan.value BreuschPagan.pvalue   White.R2
#> [Release[3]]-[Release[2]]          0.0322493          0.85843578 0.06507212
#> [Release[4]]-[Release[3]]          0.4906082          0.48792282 0.01724923
#> [Release[5]]-[Release[4]]          3.1230542          0.08522265 0.09138218
#>                           White.value White.pvalue      arch.R2 arch.value
#> [Release[3]]-[Release[2]]   2.6028850    0.2721390 1.675820e-03 0.06535699
#> [Release[4]]-[Release[3]]   0.6899692    0.7082313 7.696581e-02 3.00166646
#> [Release[5]]-[Release[4]]   3.6552873    0.1607920 8.745591e-05 0.00341078
#>                           arch.pvalue
#> [Release[3]]-[Release[2]]  0.79822100
#> [Release[4]]-[Release[3]]  0.08317892
#> [Release[5]]-[Release[4]]  0.95342851