Generate report on Revision Analysis

Generate report on Revision Analysis

Usage

render_report(
  rslt,
  output_file,
  output_dir,
  output_format = c("html_document", "pdf_document"),
  open_report = TRUE,
  ...
)

Arguments

rslt

an object of class "rjd3rev_vintages" which is the output of the function revision_analysis()

output_file

path or name of the output file containing the report

output_dir

path of the dir containing the output file (Optional)

output_format

either an HTML document (default) or a PDF document

open_report

Boolean. Default is TRUE meaning that the report will open automatically after being generated.

...

Arguments to be passed to rmarkdown::render(), for example:

output_options List of output options that can override the options specified in metadata

...

See also

Examples

## Simulated data
period_range <- seq(as.Date('2011-01-01'),as.Date('2020-10-01'),by='quarter')
qtr <- (as.numeric(substr(period_range,6,7))+2)/3
time_period <- rep(paste0(format(period_range, "%Y"), "Q", qtr),5)
np <- length(period_range)
rev_date <- c(rep("2021-06-30",np), rep("2021-12-31",np), rep("2022-06-30",np),
            rep("2022-12-31",np), rep("2023-06-30",np))
set.seed(1)
xt <- cumsum(sample(rnorm(1000,0,1), np, TRUE))
rev <- rnorm(np*4,0,.1)
obs_values <- xt
for(i in 1:4) {
  xt <- xt+rev[(1+(i-1)*np):(i*np)]
  obs_values <- c(obs_values,xt)
}
df <- data.frame(rev_date, time_period, obs_values)

## Make analysis and generate the report

vintages <- create_vintages(df, periodicity = 4L, type = "long")
rslt <- revision_analysis(vintages, view = "diagonal")

if (FALSE) { # \dontrun{
render_report(
    rslt,
    output_file = "my_report",
    output_dir = "C:/Users/xxx",
    output_format = "pdf_document"
)
} # }