Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9979537  13167.82         0.25228074        0.6905622
#> Release[3] 26 0.9994990  47881.39        -0.16654551        0.3306236
#> Release[4] 17 0.9996625  44430.27         0.17204128        0.2240514
#> Release[5] 14 0.9998892 108314.66        -0.03367712        0.1108606
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.7177140      0.9992982  0.008708394    0.9363591 -0.3843051
#> Release[3]        0.6190428      1.0029813  0.004583629    0.5215906 -0.6700807
#> Release[4]        0.4544861      1.0031840  0.004759279    0.5136543 -0.2978260
#> Release[5]        0.7664990      0.9995360  0.003037070    0.8811139 -0.4277637
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.7655336        1.3729373         0.5033505     0.002445143
#> Release[3] -0.4463501        2.0783969         0.3537381     0.006462237
#> Release[4] -1.1212910        1.0747299         0.5842859     0.063120266
#> Release[5]  0.1710056        0.4123003         0.8137109     0.109984532
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]         0.06618069           0.7989316 0.04450725    1.290710
#> Release[3]         0.15610245           0.6962603 0.06738670    1.752054
#> Release[4]         1.01059289           0.3307031 0.07891887    1.341621
#> Release[5]         1.48291173           0.2467232 0.11828197    1.655948
#>            White.pvalue      arch.R2 arch.value arch.pvalue
#> Release[2]    0.5244762 0.0736680203 2.06270457   0.1509421
#> Release[3]    0.4164341 0.0355235067 0.88808767   0.3459961
#> Release[4]    0.5112941 0.0006707363 0.01073178   0.9174913
#> Release[5]    0.4369337 0.0205587456 0.26726369   0.6051732