Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9848098  1750.463         0.05716094       0.63558544
#> Release[3] 26 0.9969149  7755.312        -0.20385749       0.30034720
#> Release[4] 17 0.9987769 12248.547         0.16324585       0.18646666
#> Release[5] 14 0.9996517 34440.468        -0.04043166       0.08913221
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.9290033      0.9888680  0.023635340    0.6414317 -0.3296497
#> Release[3]        0.5037960      1.0068171  0.011432756    0.5565742 -0.6511112
#> Release[4]        0.3951216      1.0079857  0.009107771    0.3944193 -0.3028167
#> Release[5]        0.6582004      0.9978907  0.005377104    0.7017388 -0.3260379
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  1.0353155        1.7576462         0.4152714      0.01563830
#> Release[3] -0.4503938        1.9777475         0.3719954      0.01460366
#> Release[4] -1.1412031        1.1127575         0.5732813      0.04384944
#> Release[5] -0.1963940        0.2512107         0.8819629      0.14023427
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]          0.4289420           0.5180498 0.01660638   0.4815851
#> Release[3]          0.3556820           0.5564982 0.03552142   0.9235569
#> Release[4]          0.6879058           0.4198813 0.12140761   2.0639294
#> Release[5]          1.9572904           0.1871161 0.23347619   3.2686666
#>            White.pvalue     arch.R2 arch.value arch.pvalue
#> Release[2]    0.7860047 0.070956480 1.98678145   0.1586778
#> Release[3]    0.6301619 0.048389020 1.20972549   0.2713865
#> Release[4]    0.3563062 0.003690859 0.05905375   0.8079978
#> Release[5]    0.1950824 0.010959873 0.14247835   0.7058295