Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9800414 1325.797         0.61356115       0.41842206
#> Release[3] 26 0.9946557 4466.805        -0.27999049       0.20832741
#> Release[4] 17 0.9952903 3169.884         0.11011840       0.12810672
#> Release[5] 14 0.9987038 9245.945        -0.03186357       0.07765351
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.1540989      0.9605742   0.02638106   0.14664790 -0.0850249
#> Release[3]        0.1915192      0.9898111   0.01480996   0.49806191 -0.7519888
#> Release[4]        0.4035478      1.0408045   0.01848619   0.04328832 -0.2306685
#> Release[5]        0.6887952      0.9951683   0.01034954   0.64897042 -0.3213415
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2] -0.1721676       0.06831839         0.9664176     0.073639022
#> Release[3] -0.2887216       2.44303018         0.2947832     0.003923309
#> Release[4] -0.4852551       0.29886956         0.8611946     0.002123927
#> Release[5] -0.1198281       0.23150842         0.8906941     0.115993347
#>            BreuschPagan.value BreuschPagan.pvalue    White.R2 White.value
#> Release[2]         2.14630544           0.1544620 0.115892782   3.3608907
#> Release[3]         0.09453030           0.7611471 0.004928838   0.1281498
#> Release[4]         0.03192671           0.8605796 0.013918467   0.2366139
#> Release[5]         1.57455846           0.2334350 0.116680259   1.6335236
#>            White.pvalue    arch.R2 arch.value arch.pvalue
#> Release[2]    0.1862910 0.12398604  3.4716091  0.06243055
#> Release[3]    0.9379348 0.02465222  0.6163055  0.43242350
#> Release[4]    0.8884233 0.01747473  0.2795957  0.59696635
#> Release[5]    0.4418602 0.01800530  0.2340690  0.62852304