Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2          F intercept.estimate intercept.stderr
#> Release[2] 29 0.9673622   800.2609        -0.05024430       0.49972382
#> Release[3] 26 0.9934013  3613.0849        -0.20741016       0.23863861
#> Release[4] 17 0.9975699  6157.6368         0.09638763       0.15471222
#> Release[5] 14 0.9992632 16273.6670        -0.02415571       0.07536499
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.9206552      0.9646644  0.034100477    0.3092913 -0.3201805
#> Release[3]        0.3933802      1.0153181  0.016891298    0.3734955 -0.6204077
#> Release[4]        0.5426334      1.0085757  0.012852910    0.5147672 -0.3461895
#> Release[5]        0.7540896      0.9977879  0.007821599    0.7821391 -0.4237412
#>               kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  1.65425844        3.6710722         0.1595280     0.005020713
#> Release[3] -0.54736518        1.9158664         0.3836851     0.017143671
#> Release[4] -1.12674266        1.1659584         0.5582328     0.030389940
#> Release[5]  0.02378062        0.3893457         0.8231039     0.045686235
#>            BreuschPagan.value BreuschPagan.pvalue    White.R2 White.value
#> Release[2]          0.1362433           0.7149235 0.005063403   0.1468387
#> Release[3]          0.4186249           0.5237646 0.022056451   0.5734677
#> Release[4]          0.4701365           0.5033846 0.175903458   2.9903588
#> Release[5]          0.5744807           0.4631095 0.062790198   0.8790628
#>            White.pvalue      arch.R2   arch.value arch.pvalue
#> Release[2]    0.9292111 4.112498e-02 1.1514993912   0.2832355
#> Release[3]    0.7507115 2.457650e-02 0.6144125850   0.4331312
#> Release[4]    0.2242084 2.726273e-05 0.0004362036   0.9833370
#> Release[5]    0.6443383 1.239053e-02 0.1610769357   0.6881669