Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9272174  343.968        0.209127824       0.37986768
#> Release[3] 26 0.9799851 1175.108       -0.345308145       0.18836043
#> Release[4] 17 0.9935751 2319.673       -0.015782265       0.13962565
#> Release[5] 14 0.9985254 8125.916        0.004230246       0.07948182
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]       0.58648375      0.9482927   0.05113089    0.3208626 -0.3524067
#> Release[3]       0.07920174      0.9965547   0.02907115    0.9066470 -0.6869940
#> Release[4]       0.91150352      1.0361169   0.02151272    0.1138831 -0.4193529
#> Release[5]       0.95843016      0.9911573   0.01099528    0.4369066 -0.2451359
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  1.2755037        2.4776169         0.2897292    3.606825e-05
#> Release[3] -0.3580369        2.1000349         0.3499316    2.645708e-04
#> Release[4] -0.7491451        0.8430973         0.6560301    3.094769e-04
#> Release[5] -0.6263742        0.3427184         0.8425189    1.188472e-01
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]       0.0009738778           0.9753340 0.02222356   0.6444833
#> Release[3]       0.0063513807           0.9371402 0.03693333   0.9602666
#> Release[4]       0.0046435902           0.9465712 0.02646833   0.4499616
#> Release[5]       1.6185229083           0.2273982 0.13123426   1.8372797
#>            White.pvalue      arch.R2   arch.value arch.pvalue
#> Release[2]    0.7245231 8.485672e-02 2.3759882848   0.1232136
#> Release[3]    0.6187009 3.740978e-02 0.9352444991   0.3335038
#> Release[4]    0.7985315 1.534489e-02 0.2455182782   0.6202486
#> Release[5]    0.3990615 4.196366e-05 0.0005455275   0.9813659