Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9904743 2807.441         0.76057122       0.46806635
#> Release[3] 26 0.9957930 5680.754        -0.29554305       0.24516592
#> Release[4] 17 0.9900832 1497.589        -0.01249996       0.16143013
#> Release[5] 14 0.9970278 4025.362        -0.02745502       0.07809968
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.1157972      0.9716309   0.01833775    0.1334977 -0.2413972
#> Release[3]        0.2397687      0.9959287   0.01321372    0.7606537 -0.7078987
#> Release[4]        0.9393028      1.0232879   0.02644244    0.3923716 -0.3303287
#> Release[5]        0.7312801      1.0043245   0.01582966    0.7893536 -0.5699779
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2] -0.2482000        0.3438093         0.8420595     0.020157182
#> Release[3] -0.3648034        2.2266290         0.3284684     0.003240577
#> Release[4] -0.7919401        0.7090915         0.7014920     0.015493548
#> Release[5]  0.9079396        1.1504208         0.5625865     0.033553608
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]          0.5554400           0.4625423 0.02287585   0.6633997
#> Release[3]          0.0780267           0.7823827 0.05714120   1.4856712
#> Release[4]          0.2360606           0.6340868 0.09003378   1.5305743
#> Release[5]          0.4166225           0.5307756 0.04627781   0.6478894
#>            White.pvalue    arch.R2 arch.value arch.pvalue
#> Release[2]    0.7177027 0.07811523  2.1872266   0.1391597
#> Release[3]    0.4757629 0.03693263  0.9233158   0.3366059
#> Release[4]    0.4652003 0.03940289  0.6304462   0.4271917
#> Release[5]    0.7232902 0.02950276  0.3835359   0.5357169