Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9983422 16259.40         0.45280116       0.58038339
#> Release[3] 26 0.9994894 46981.91        -0.28807967       0.28472110
#> Release[4] 17 0.9994408 26811.26         0.14559977       0.19483988
#> Release[5] 14 0.9998117 63719.74        -0.02617075       0.09380221
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.4420735      1.0027335  0.007863815    0.7308347 -0.3905903
#> Release[3]        0.3217282      1.0012098  0.004619125    0.7956269 -0.6924725
#> Release[4]        0.4664488      1.0041186  0.006132343    0.5120389 -0.2674100
#> Release[5]        0.7849954      0.9996502  0.003960145    0.9310690 -0.4593251
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.4698915        0.9695482         0.6158363    7.371649e-05
#> Release[3] -0.3898231        2.1562863         0.3402267    9.461569e-06
#> Release[4] -1.0900583        0.9828397         0.6117572    8.037288e-02
#> Release[5]  0.3051149        0.5075490         0.7758668    7.698298e-02
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]       0.0019904919           0.9647426 0.01788877   0.5187745
#> Release[3]       0.0002270798           0.9881016 0.09540603   2.4805569
#> Release[4]       1.3109586444           0.2701621 0.08170954   1.3890621
#> Release[5]       1.0008437288           0.3368532 0.08333814   1.1667340
#>            White.pvalue     arch.R2 arch.value arch.pvalue
#> Release[2]    0.7715242 0.075573472 2.11605722   0.1457614
#> Release[3]    0.2893037 0.041931137 1.04827843   0.3059039
#> Release[4]    0.4993085 0.001451135 0.02321816   0.8788911
#> Release[5]    0.5580164 0.023919080 0.31094804   0.5770990