Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9910462  2988.473         0.34747345       0.45661101
#> Release[3] 26 0.9978152 10960.995        -0.32974329       0.21453571
#> Release[4] 17 0.9963338  4076.471         0.03749383       0.14534167
#> Release[5] 14 0.9987798  9822.242         0.01271726       0.08683094
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.4532622      1.0034383  0.018355491    0.8528123 -0.3817457
#> Release[3]        0.1373715      1.0013312  0.009564291    0.8904623 -0.6922641
#> Release[4]        0.7999356      1.0107104  0.015830132    0.5089713 -0.3056872
#> Release[5]        0.8859900      0.9927126  0.010016552    0.4808548 -0.2164920
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.5646643        1.0520590         0.5909467     0.001337359
#> Release[3] -0.3772209        2.1450144         0.3421496     0.001104567
#> Release[4] -1.1348405        1.1077610         0.5747153     0.046313424
#> Release[5] -0.4943103        0.2339013         0.8896291     0.097963770
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]         0.03615704           0.8506140 0.07798147    2.261463
#> Release[3]         0.02653892           0.8719551 0.08471049    2.202473
#> Release[4]         0.72843782           0.4068134 0.09824311    1.670133
#> Release[5]         1.30323506           0.2758922 0.25938561    3.631399
#>            White.pvalue     arch.R2 arch.value arch.pvalue
#> Release[2]    0.3227971 0.078060270  2.1856875   0.1392988
#> Release[3]    0.3324598 0.045030649  1.1257662   0.2886802
#> Release[4]    0.4338457 0.010815954  0.1730553   0.6774107
#> Release[5]    0.1627241 0.008132143  0.1057179   0.7450730