Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages[["diagonal_view"]])
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 38 0.9391994  556.0994        -0.06568268        1.2909579
#> Release[3] 34 0.9843329 2010.4958         0.24997827        0.6488602
#> Release[4] 19 0.9803355  847.5016        -0.08068656        0.8230157
#> Release[5]  8 0.9914649  696.9773         0.86260328        0.8061034
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue
#> Release[2]        0.9597032      0.9890335   0.04194062    0.7952150
#> Release[3]        0.7025957      0.9894999   0.02206806    0.6374460
#> Release[4]        0.9230490      1.0014081   0.03439861    0.9678248
#> Release[5]        0.3257257      0.9604211   0.03637914    0.3183743
#>               skewness   kurtosis JarqueBera.value JarqueBera.pvalue
#> Release[2]  0.15605397 -0.6334815       0.76884488         0.6808438
#> Release[3] -0.39550546 -0.9955188       2.22303950         0.3290585
#> Release[4] -0.03394303 -0.7621633       0.43912604         0.8028696
#> Release[5]  0.06076384 -0.1652076       0.01226824         0.9938847
#>            BreuschPagan.R2 BreuschPagan.value BreuschPagan.pvalue   White.R2
#> Release[2]     0.097312987         3.88093269          0.05656759 0.10018015
#> Release[3]     0.010771053         0.34842663          0.55915092 0.02862400
#> Release[4]     0.002025215         0.03449852          0.85484856 0.09595106
#> Release[5]     0.464049947         5.19507305          0.06286266 0.66098032
#>            White.value White.pvalue      arch.R2 arch.value arch.pvalue
#> Release[2]   3.8068457   0.14905754 0.0004645339 0.01718776   0.8956946
#> Release[3]   0.9732159   0.61470800 0.0219271593 0.72359626   0.3949668
#> Release[4]   1.8230701   0.40190681 0.0094828160 0.17069069   0.6794987
#> Release[5]   5.2878426   0.07108199 0.0470527504 0.32936925   0.5660307