Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9582194  619.2325         0.02239269       0.48147068
#> Release[3] 26 0.9860235 1693.1638        -0.26275777       0.23163725
#> Release[4] 17 0.9938404 2420.2169         0.11596702       0.14720340
#> Release[5] 14 0.9986236 8706.1058        -0.02639480       0.07536153
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.9632469      0.9650841   0.03878272    0.3759230 -0.3409252
#> Release[3]        0.2678510      1.0146423   0.02465832    0.5581979 -0.6867388
#> Release[4]        0.4430848      1.0247360   0.02082979    0.2534843 -0.3330940
#> Release[5]        0.7322276      0.9955205   0.01066936    0.6820171 -0.4100232
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  1.2932913        2.4937764         0.2873977     0.001706483
#> Release[3] -0.4624466        2.1878101         0.3349061     0.001704986
#> Release[4] -0.9771695        0.9324445         0.6273678     0.012010733
#> Release[5] -0.1474985        0.3760422         0.8285972     0.045716144
#>            BreuschPagan.value BreuschPagan.pvalue    White.R2 White.value
#> Release[2]         0.04615381           0.8315109 0.003151524  0.09139419
#> Release[3]         0.04098955           0.8412646 0.042962763  1.11703185
#> Release[4]         0.18235117           0.6754265 0.057692184  0.98076713
#> Release[5]         0.57487479           0.4629597 0.093393151  1.30750411
#>            White.pvalue     arch.R2 arch.value arch.pvalue
#> Release[2]    0.9553313 0.051637846 1.44585968   0.2291935
#> Release[3]    0.5720574 0.044430209 1.11075521   0.2919178
#> Release[4]    0.6123915 0.011467362 0.18347779   0.6684014
#> Release[5]    0.5200907 0.005566283 0.07236168   0.7879287