Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9988152 22762.28         0.71845323        0.6579143
#> Release[3] 26 0.9996480 68167.34        -0.15941075        0.3315227
#> Release[4] 17 0.9996476 42547.65         0.17550823        0.2463335
#> Release[5] 14 0.9998592 85228.62        -0.07874324        0.1307114
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.2844747      0.9949011  0.006594350    0.4461038 -0.4005897
#> Release[3]        0.6349818      0.9974263  0.003820258    0.5069476 -0.6819607
#> Release[4]        0.4871019      0.9971298  0.004834081    0.5615351 -0.3227043
#> Release[5]        0.5581065      1.0018334  0.003431648    0.6029128 -0.3022480
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.1870358        0.7896825         0.6737870    0.0003062929
#> Release[3] -0.4159870        2.1180486         0.3467940    0.0038736773
#> Release[4] -1.1443459        1.1507199         0.5625024    0.0409721248
#> Release[5] -0.2069105        0.2211232         0.8953312    0.0887133337
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]        0.008272443           0.9282014 0.07663052    2.222285
#> Release[3]        0.093329786           0.7626200 0.12441912    3.234897
#> Release[4]        0.640838382           0.4359037 0.09910705    1.684820
#> Release[5]        1.168194427           0.3010159 0.17334556    2.426838
#>            White.pvalue    arch.R2 arch.value arch.pvalue
#> Release[2]    0.3291826 0.08148589 2.28160490   0.1309163
#> Release[3]    0.1984043 0.03577054 0.89426356   0.3443245
#> Release[4]    0.4306714 0.00281621 0.04505937   0.8318949
#> Release[5]    0.2971795 0.01154329 0.15006280   0.6984753