Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9806685 1369.6875         1.26372859        0.7621188
#> Release[3] 26 0.9927148 3270.3349        -0.17250004        0.4422107
#> Release[4] 17 0.9797289  724.9683         0.18461250        0.6022899
#> Release[5] 14 0.9909045 1307.3343        -0.06710649        0.4097601
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.1088564      0.9625664   0.02600878    0.1615684 -0.3205854
#> Release[3]        0.6999138      0.9925633   0.01735650    0.6721334 -0.7090328
#> Release[4]        0.7634241      0.9919254   0.03683999    0.8294646 -0.3452907
#> Release[5]        0.8726380      1.0031631   0.02774458    0.9111176 -0.4396152
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2] -0.1205197        0.4965625         0.7801405    0.0075889612
#> Release[3] -0.3853660        2.2493928         0.3247510    0.0006597965
#> Release[4] -1.1179182        1.1510958         0.5623967    0.0453929755
#> Release[5]  0.2365084        0.4490322         0.7989027    0.1235266530
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]         0.20646884           0.6531813 0.02519912   0.7307744
#> Release[3]         0.01584557           0.9008761 0.13086051   3.4023732
#> Release[4]         0.71327218           0.4116278 0.14586926   2.4797775
#> Release[5]         1.69123207           0.2178600 0.12934753   1.8108654
#>            White.pvalue     arch.R2 arch.value arch.pvalue
#> Release[2]    0.6939279 0.094065454  2.6338327   0.1046089
#> Release[3]    0.1824669 0.035544826  0.8886206   0.3458514
#> Release[4]    0.2894164 0.009922423  0.1587588   0.6903019
#> Release[5]    0.4043669 0.022749862  0.2957482   0.5865603