Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9502374  515.5757        0.009651361        0.3580953
#> Release[3] 26 0.9836032 1439.6995       -0.312307784        0.2161779
#> Release[4] 17 0.9905912 1579.2480        0.269660095        0.2511123
#> Release[5] 14 0.9981832 6592.9551       -0.042017456        0.1200570
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.9786963      0.9006606   0.03966568   0.01860586  0.1501636
#> Release[3]        0.1614782      1.0078901   0.02656301   0.76899742 -0.6684999
#> Release[4]        0.2998563      1.0264129   0.02582837   0.32269240 -0.3431709
#> Release[5]        0.7324202      0.9972761   0.01228218   0.82821498 -0.4263190
#>               kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2] -0.33548990        0.2365416         0.8884554     0.074647936
#> Release[3] -0.39797480        2.0270336         0.3629403     0.009872645
#> Release[4] -1.03964455        1.0346172         0.5961228     0.003183667
#> Release[5]  0.06808993        0.3962984         0.8202475     0.070289287
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]         2.17808374           0.1515567 0.07712476   2.2366182
#> Release[3]         0.23930607           0.6291488 0.02718033   0.7066887
#> Release[4]         0.04790753           0.8296954 0.08350249   1.4195424
#> Release[5]         0.90724075           0.3596399 0.08784916   1.2298883
#>            White.pvalue     arch.R2 arch.value arch.pvalue
#> Release[2]    0.3268320 0.081066222 2.26985421   0.1319123
#> Release[3]    0.7023353 0.042342169 1.05855422   0.3035452
#> Release[4]    0.4917567 0.002339482 0.03743171   0.8465887
#> Release[5]    0.5406711 0.017725405 0.23043027   0.6312050