Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9710539  905.7677         1.17383183        0.6765187
#> Release[3] 26 0.9931630 3486.3341        -0.12280816        0.3440016
#> Release[4] 17 0.9873597 1171.6836         0.29380346        0.3156599
#> Release[5] 14 0.9956096 2721.2306         0.05492447        0.1780500
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]       0.09412491      0.9516507   0.03162053    0.1378853 -0.2732869
#> Release[3]       0.72421261      0.9872167   0.01671968    0.4519813 -0.6908495
#> Release[4]       0.36671676      0.9760172   0.02851361    0.4134991 -0.3973136
#> Release[5]       0.76300826      0.9910673   0.01899855    0.6466567 -0.4806674
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2] -0.1744794        0.3840503         0.8252861     0.017960597
#> Release[3] -0.3870077        2.1446531         0.3422114     0.001437701
#> Release[4] -1.0750333        1.1914193         0.5511713     0.008291797
#> Release[5]  0.5342579        0.6551979         0.7206520     0.002762073
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]         0.49380515           0.4882498 0.25837483    7.492870
#> Release[3]         0.03455451           0.8540944 0.04002519    1.040655
#> Release[4]         0.12541688           0.7281611 0.19443899    3.305463
#> Release[5]         0.03323667           0.8583836 0.09818978    1.374657
#>            White.pvalue    arch.R2 arch.value arch.pvalue
#> Release[2]   0.02360174 0.12461947  3.4893452  0.06176501
#> Release[3]   0.59432586 0.04832469  1.2081173  0.27170530
#> Release[4]   0.19152606 0.01076164  0.1721863  0.67817607
#> Release[5]   0.50291784 0.04760841  0.6189093  0.43145289