Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9241806  329.1094          1.5627810        1.4795960
#> Release[3] 26 0.9689267  748.3667          1.2626136        0.9038471
#> Release[4] 17 0.9875397 1188.8278          0.4866647        0.6965017
#> Release[5] 14 0.9972222 4307.9970          0.2006928        0.3544707
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.3002267      0.9536182   0.05256594   0.38537395 -0.4717124
#> Release[3]        0.1752179      0.9378608   0.03428318   0.08243421 -0.6081212
#> Release[4]        0.4954142      0.9820257   0.02848153   0.53748191 -0.4032438
#> Release[5]        0.5817079      0.9903452   0.01508860   0.53428117 -0.6252678
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.3400771         1.173320         0.5561819     0.005161192
#> Release[3] -0.6870209         2.032545         0.3619416     0.012935522
#> Release[4] -1.1066091         1.250004         0.5352603     0.001110412
#> Release[5]  1.6643926         2.347606         0.3091889     0.012430517
#>            BreuschPagan.value BreuschPagan.pvalue    White.R2 White.value
#> Release[2]         0.14007514           0.7111296 0.005609891   0.1626869
#> Release[3]         0.31452103           0.5801178 0.023104440   0.6007154
#> Release[4]         0.01667469           0.8989699 0.327759522   5.5719119
#> Release[5]         0.15104376           0.7043474 0.024160917   0.3382528
#>            White.pvalue      arch.R2  arch.value arch.pvalue
#> Release[2]   0.92187704 0.0822234932 2.302257811   0.1291861
#> Release[3]   0.74055327 0.0039630538 0.099076345   0.7529409
#> Release[4]   0.06167011 0.0002858218 0.004573149   0.9460841
#> Release[5]   0.84440215 0.0534068950 0.694289635   0.4047091