Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages[["diagonal_view"]])
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 30 0.9582096  642.0096         2.38012100        1.0109255
#> Release[3] 29 0.9877298 2173.4588         0.09727184        0.6110874
#> Release[4] 27 0.9970934 8575.9841        -0.09147092        0.3134805
#> Release[5]  7 0.9987433 3973.7494        -0.07546477        0.2618083
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]       0.02579832      0.9158434   0.03614518   0.02734305 -0.1103667
#> Release[3]       0.87471316      1.0058199   0.02157469   0.78939966 -0.7141803
#> Release[4]       0.77285373      0.9983005   0.01078001   0.87600007 -0.1302344
#> Release[5]       0.78472634      0.9983556   0.01583744   0.92134195  0.3942945
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2] -0.2758592        0.1508260         0.9273604     0.005063969
#> Release[3]  0.3052164        2.4889330         0.2880946     0.022796382
#> Release[4] -0.9212259        0.9928762         0.6086949     0.078338705
#> Release[5] -2.0722253        1.2289976         0.5409119     0.006901294
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]         0.14251281           0.7086421 0.11607410  3.48222286
#> Release[3]         0.62986087           0.4343262 0.09507711  2.75723624
#> Release[4]         2.12493206           0.1573676 0.09411348  2.54106406
#> Release[5]         0.03474626           0.8594553 0.01384836  0.09693853
#>            White.pvalue     arch.R2 arch.value arch.pvalue
#> Release[2]    0.1753254 0.034419904  0.9981772   0.3177520
#> Release[3]    0.2519264 0.003325625  0.0931175   0.7602507
#> Release[4]    0.2806823 0.045597249  1.1855285   0.2762334
#> Release[5]    0.9526866 0.236467682  1.4188061   0.2336000