Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the `create_vintages()` output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
period_range<-seq(as.Date('2011-01-01'),as.Date('2020-10-01'),by='quarter')
qtr<-(as.numeric(substr(period_range,6,7))+2)/3
time_period<-rep(paste0(format(period_range, "%Y"), "Q", qtr),5)
np<-length(period_range)
rev_date<-c(rep("2021-06-30",np), rep("2021-12-31",np), rep("2022-06-30",np),
            rep("2022-12-31",np), rep("2023-06-30",np))
set.seed(1)
xt<-cumsum(sample(rnorm(1000,0,1), np, TRUE))
rev<-rnorm(np*4,0,.1)
obs_values<-xt
for(i in 1:4) {
  xt<-xt+rev[(1+(i-1)*np):(i*np)]
  obs_values<-c(obs_values,xt)
}
df<-data.frame(rev_date, time_period, obs_values)

## Create vintage and test
vintages<-create_vintages(df, periodicity = 4)
slope_and_drift(vintages$diagonal_view)
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 40 0.9995199 79105.04         0.03955525       0.01737525
#> Release[3] 40 0.9992716 52129.28         0.01253707       0.02135381
#> Release[4] 40 0.9994474 68733.73        -0.03276902       0.01868319
#> Release[5] 40 0.9994009 63395.10        -0.01101970       0.01933166
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]       0.02853379      0.9942853  0.003535159   0.11425292  0.2597218
#> Release[3]       0.56060100      0.9925766  0.004347335   0.09587371 -0.2938120
#> Release[4]       0.08750526      1.0028430  0.003825146   0.46190246  0.2012318
#> Release[5]       0.57200970      0.9976545  0.003962345   0.55739632 -0.7122024
#>               kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2] -0.43395506        0.7444754        0.68919041    1.218545e-02
#> Release[3] -0.06368272        0.5677057        0.75287742    3.525564e-05
#> Release[4] -0.05408149        0.2679654        0.87460520    4.268766e-02
#> Release[5]  1.53571113        7.1294240        0.02830514    1.671692e-02
#>            BreuschPagan.value BreuschPagan.pvalue    White.R2 White.value
#> Release[2]        0.468759236           0.4977135 0.012285119  0.49140475
#> Release[3]        0.001339762           0.9709934 0.000453343  0.01813372
#> Release[4]        1.694463842           0.2008499 0.054232001  2.16928004
#> Release[5]        0.646042725           0.4265307 0.017604610  0.70418439
#>            White.pvalue      arch.R2 arch.value arch.pvalue
#> Release[2]    0.7821550 0.0456815355 1.78157988  0.18195545
#> Release[3]    0.9909741 0.0033041245 0.12886085  0.71961608
#> Release[4]    0.3380234 0.0853776151 3.32972699  0.06803817
#> Release[5]    0.7032153 0.0005330063 0.02078725  0.88536005