Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages[["diagonal_view"]])
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9974182 10430.76         0.65600685        0.6725764
#> Release[3] 26 0.9992234 30880.90        -0.20444515        0.3414613
#> Release[4] 17 0.9990922 16508.74         0.18839593        0.2680672
#> Release[5] 14 0.9996566 34937.01        -0.01799946        0.1447275
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.3380381      0.9937888  0.009730520    0.5286393 -0.3941035
#> Release[3]        0.5549587      0.9972245  0.005674766    0.6292188 -0.6919391
#> Release[4]        0.4929481      0.9954896  0.007747826    0.5691160 -0.3216747
#> Release[5]        0.9030830      0.9998628  0.005349307    0.9799564 -0.5008900
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.2795882        0.8160133         0.6649745    0.0001009154
#> Release[3] -0.4043082        2.1651916         0.3387151    0.0007185730
#> Release[4] -1.1150441        1.1048145         0.5755626    0.0377598728
#> Release[5]  0.4223382        0.6402136         0.7260715    0.0857117055
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]        0.002724992           0.9587525 0.02350995   0.6817887
#> Release[3]        0.017258152           0.8965775 0.10057245   2.6148836
#> Release[4]        0.588624478           0.4548589 0.03775991   0.6419185
#> Release[5]        1.124962959           0.3097338 0.08586313   1.2020838
#>            White.pvalue     arch.R2 arch.value arch.pvalue
#> Release[2]    0.7111340 0.076217818 2.13409889   0.1440551
#> Release[3]    0.2705112 0.037812529 0.94531324   0.3309151
#> Release[4]    0.7254528 0.001769649 0.02831438   0.8663718
#> Release[5]    0.5482401 0.027466461 0.35706399   0.5501414