Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2          F intercept.estimate intercept.stderr
#> Release[2] 29 0.9669368   789.6175        -0.49785220        1.1274022
#> Release[3] 26 0.9921593  3036.9521         0.25288063        0.5754476
#> Release[4] 17 0.9976762  6439.9543         0.48369239        0.3850624
#> Release[5] 14 0.9993956 19843.7170        -0.03074795        0.2043766
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.6623007      1.0273769  0.036561283    0.4604547 -0.1564402
#> Release[3]        0.6642668      0.9805232  0.017792578    0.2845297 -0.4935667
#> Release[4]        0.2282779      0.9854004  0.012279236    0.2529365 -0.1728264
#> Release[5]        0.8829109      1.0003574  0.007101395    0.9606874 -0.4706846
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.9265080        1.1156964         0.5724395      0.12926430
#> Release[3] -0.5312663        1.3090378         0.5196920      0.09031852
#> Release[4] -0.9207616        0.6448519         0.7243896      0.08433568
#> Release[5]  0.3311169        0.5393995         0.7636087      0.09425557
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]           4.008261          0.05541698 0.13167136    3.818469
#> Release[3]           2.382861          0.13575715 0.09942855    2.585142
#> Release[4]           1.381549          0.25815781 0.12970603    2.205003
#> Release[5]           1.248770          0.28566152 0.10088147    1.412341
#>            White.pvalue      arch.R2 arch.value arch.pvalue
#> Release[2]    0.1481937 0.0623053877  1.7445509   0.1865632
#> Release[3]    0.2745639 0.0234072917  0.5851823   0.4442882
#> Release[4]    0.3320395 0.0005472188  0.0087555   0.9254501
#> Release[5]    0.4935307 0.0248718268  0.3233337   0.5696110