Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages[["diagonal_view"]])
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9979591 13202.29         0.48765376       0.54223483
#> Release[3] 26 0.9993142 34972.32        -0.26789751       0.26612732
#> Release[4] 17 0.9985204 10122.96         0.03966474       0.19399697
#> Release[5] 14 0.9991216 13648.54        -0.02350095       0.09837807
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.3764225      1.0042074  0.008739744    0.6341030 -0.3687473
#> Release[3]        0.3241443      1.0021627  0.005358905    0.6901023 -0.7128799
#> Release[4]        0.8407419      0.9987099  0.009926260    0.8983152 -0.3687026
#> Release[5]        0.8152265      0.9996883  0.008556999    0.9715459 -0.4804566
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.3863789        0.8087180         0.6674045     0.002245877
#> Release[3] -0.3651786        2.2564025         0.3236148     0.002422574
#> Release[4] -1.0480374        1.0947659         0.5784617     0.006784973
#> Release[5]  0.3638676        0.5718665         0.7513127     0.047197381
#>            BreuschPagan.value BreuschPagan.pvalue    White.R2 White.value
#> Release[2]         0.06077517           0.8071396 0.005375382   0.1558861
#> Release[3]         0.05828298           0.8112814 0.064203392   1.6692882
#> Release[4]         0.10246985           0.7533012 0.188347384   3.2019055
#> Release[5]         0.59442382           0.4556266 0.054720011   0.7660801
#>            White.pvalue    arch.R2 arch.value arch.pvalue
#> Release[2]    0.9250171 0.08037962  2.2506294   0.1335601
#> Release[3]    0.4340289 0.04003033  1.0007582   0.3171271
#> Release[4]    0.2017042 0.02480827  0.3969323   0.5286778
#> Release[5]    0.6817856 0.02547433  0.3311663   0.5649731