Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages[["diagonal_view"]])
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9983266 16107.98         0.47255746       0.49662674
#> Release[3] 26 0.9994256 41757.05        -0.28533741       0.24005019
#> Release[4] 17 0.9990485 15750.18         0.07135634       0.15384725
#> Release[5] 14 0.9996165 31282.24        -0.02209951       0.07452462
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.3497741      1.0042086  0.007912312    0.5991444 -0.3752511
#> Release[3]        0.2462036      1.0019232  0.004903087    0.6983343 -0.7100992
#> Release[4]        0.6494411      1.0022404  0.007985998    0.7828988 -0.3762428
#> Release[5]        0.7718878      0.9979873  0.005642559    0.7275078 -0.3536372
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.3548203        0.8040096         0.6689775    0.0008441087
#> Release[3] -0.3592511        2.2354427         0.3270241    0.0022059424
#> Release[4] -1.1096911        1.1984326         0.5492419    0.0013492123
#> Release[5]  0.1063617        0.2770895         0.8706243    0.0945525718
#>            BreuschPagan.value BreuschPagan.pvalue    White.R2 White.value
#> Release[2]         0.02281019           0.8810745 0.008073709   0.2341376
#> Release[3]         0.05305966           0.8197744 0.070150728   1.8239189
#> Release[4]         0.02026553           0.8886925 0.169325345   2.8785309
#> Release[5]         1.25311622           0.2848646 0.127038015   1.7785322
#>            White.pvalue    arch.R2 arch.value arch.pvalue
#> Release[2]    0.8895240 0.08218466  2.3011706   0.1292765
#> Release[3]    0.4017363 0.04024112  1.0060281   0.3158563
#> Release[4]    0.2371019 0.01192467  0.1907947   0.6622560
#> Release[5]    0.4109572 0.01848273  0.2402755   0.6240072