Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9902774  2750.024          0.8908468        1.1170455
#> Release[3] 26 0.9976945 10385.966          0.1109613        0.5513961
#> Release[4] 17 0.9983577  9118.788          0.3358855        0.4628145
#> Release[5] 14 0.9994718 22705.351          0.1083409        0.2630604
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.4321172      0.9893903  0.018866851    0.5785193 -0.3996248
#> Release[3]        0.8422091      0.9914755  0.009728783    0.3895991 -0.6703437
#> Release[4]        0.4791661      0.9933968  0.010402894    0.5351534 -0.2877108
#> Release[5]        0.6877122      0.9966064  0.006613929    0.6171915 -0.6411404
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.2871175        0.8414423         0.6565732    0.0001184677
#> Release[3] -0.4736728        2.1060505         0.3488807    0.0082433327
#> Release[4] -1.1356423        1.0805290         0.5825941    0.0755188237
#> Release[5]  1.1794827        1.6441878         0.4395104    0.0209410361
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]        0.003199008           0.9553124 0.22143543    6.421627
#> Release[3]        0.199484402           0.6591416 0.05800748    1.508195
#> Release[4]        1.225316841           0.2857561 0.12147063    2.065001
#> Release[5]        0.256667313           0.6215988 0.10246736    1.434543
#>            White.pvalue      arch.R2  arch.value arch.pvalue
#> Release[2]   0.04032379 0.0816395745 2.285908086   0.1305537
#> Release[3]   0.47043510 0.0234050285 0.585125712   0.4443102
#> Release[4]   0.35611543 0.0005664737 0.009063579   0.9241537
#> Release[5]   0.48808215 0.0453422833 0.589449683   0.4426320