Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.8937200  227.0460        -0.56520181        0.3959332
#> Release[3] 26 0.9616838  602.3666        -0.39815586        0.2412155
#> Release[4] 17 0.9847824  970.6984         0.12680576        0.2177065
#> Release[5] 14 0.9970260 4022.9160         0.01003821        0.1006509
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.1648950      0.8101652   0.05376710  0.001509423  0.6341641
#> Release[3]        0.1118430      0.9856798   0.04016109  0.724531115 -0.6701992
#> Release[4]        0.5689083      1.0140132   0.03254629  0.672909211 -0.3013040
#> Release[5]        0.9222032      1.0072922   0.01588126  0.654316593 -0.6530458
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.7826654         2.591425         0.2737028     0.074279154
#> Release[3] -0.3652685         2.010510         0.3659514     0.006522574
#> Release[4] -1.0914718         1.036298         0.5956220     0.076137251
#> Release[5]  1.1528929         1.643979         0.4395564     0.033722051
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]          2.1664600           0.1526117 0.07783796    2.257301
#> Release[3]          0.1575695           0.6949121 0.02658658    0.691251
#> Release[4]          1.2361780           0.2837121 0.11675839    1.984893
#> Release[5]          0.4187870           0.5297280 0.24927741    3.489884
#>            White.pvalue      arch.R2 arch.value arch.pvalue
#> Release[2]    0.3234695 0.0002896607 0.00811050   0.9282409
#> Release[3]    0.7077775 0.0647201425 1.61800356   0.2033704
#> Release[4]    0.3706688 0.0053948281 0.08631725   0.7689123
#> Release[5]    0.1746551 0.0417007377 0.54210959   0.4615598