Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9689649  842.9822         0.23963576        0.3346597
#> Release[3] 26 0.9849456 1570.2217        -0.30341430        0.1998332
#> Release[4] 17 0.9933935 2255.4952         0.03621974        0.1564925
#> Release[5] 14 0.9970190 4013.5012        -0.02053889        0.0751047
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.4801033      0.9149771   0.03151382   0.01187847  0.1254149
#> Release[3]        0.1419933      1.0147155   0.02560730   0.57086982 -0.6252826
#> Release[4]        0.8200950      0.9930282   0.02090935   0.74342503 -0.3919279
#> Release[5]        0.7891418      1.0016047   0.01581010   0.92083145 -0.5496236
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.2601121        0.1523362         0.9266604     0.097490673
#> Release[3] -0.4331580        1.8245200         0.4016155     0.026984555
#> Release[4] -0.9135668        0.9660228         0.6169228     0.000822599
#> Release[5]  0.5679368        0.8292357         0.6605927     0.182150513
#>            BreuschPagan.value BreuschPagan.pvalue     White.R2 White.value
#> Release[2]         2.91658833          0.09914818 0.1052397620  3.05195310
#> Release[3]         0.66558997          0.42261694 0.0294994586  0.76698592
#> Release[4]         0.01234914          0.91298920 0.0008271737  0.01406195
#> Release[5]         2.67262643          0.12802929 0.2861332066  4.00586489
#>            White.pvalue    arch.R2 arch.value arch.pvalue
#> Release[2]    0.2174086 0.03216809  0.9007066   0.3425923
#> Release[3]    0.6814769 0.02455595  0.6138988   0.4333236
#> Release[4]    0.9929937 0.02911387  0.4658219   0.4949156
#> Release[5]    0.1349390 0.02746914  0.3570988   0.5501219