Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2        F intercept.estimate intercept.stderr
#> Release[2] 29 0.9974143 10415.16         0.83881633        0.6627249
#> Release[3] 26 0.9992197 30732.29        -0.17046178        0.3386074
#> Release[4] 17 0.9988860 13449.52         0.22929514        0.2841058
#> Release[5] 14 0.9995171 24839.09        -0.08119401        0.1639039
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.2164309      0.9904736  0.009705321    0.3350288 -0.3630866
#> Release[3]        0.6192583      0.9965059  0.005684370    0.5445422 -0.6890327
#> Release[4]        0.4322342      0.9939738  0.008570803    0.4927604 -0.3256826
#> Release[5]        0.6292817      1.0026139  0.006361593    0.6884006 -0.4017573
#>               kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.03889366        0.6169804         0.7345552    0.0007177393
#> Release[3] -0.42138438        2.1631553         0.3390602    0.0020119577
#> Release[4] -1.12629895        1.1285506         0.5687722    0.0468641335
#> Release[5]  0.03013751        0.3502113         0.8393683    0.0397841171
#>            BreuschPagan.value BreuschPagan.pvalue   White.R2 White.value
#> Release[2]         0.01939288           0.8902796 0.10834792   3.1420896
#> Release[3]         0.04838433           0.8277591 0.09503013   2.4707833
#> Release[4]         0.73752550           0.4039701 0.16100350   2.7370595
#> Release[5]         0.49718966           0.4941973 0.05248734   0.7348228
#>            White.pvalue      arch.R2 arch.value arch.pvalue
#> Release[2]    0.2078279 0.0836991961 2.34357749   0.1258004
#> Release[3]    0.2907209 0.0372478104 0.93119526   0.3345525
#> Release[4]    0.2544808 0.0000956531 0.00153045   0.9687939
#> Release[5]    0.6925247 0.0112919791 0.14679573   0.7016165