Slope and Drift

Linear regression model of a latter vintage (L) on a preliminary vintage (P)

Usage

slope_and_drift(vintages.view, gap = 1, na.zero = FALSE)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

gap

Integer. Gap to consider between each vintages. Default is 1 which means that revisions are calculated and tested for each vintages consecutively.

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

See also

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
slope_and_drift(vintages$diagonal_view)
#>             N        R2         F intercept.estimate intercept.stderr
#> Release[2] 29 0.9950684  5447.874         0.44504754       0.44599024
#> Release[3] 26 0.9981662 13063.584        -0.29639236       0.21187513
#> Release[4] 17 0.9973975  5748.798         0.04999196       0.14499657
#> Release[5] 14 0.9987115  9301.294         0.04722080       0.08795311
#>            intercept.pvalue slope.estimate slope.stderr slope.pvalue   skewness
#> Release[2]        0.3271942      1.0082021  0.013659478    0.5531992 -0.3591072
#> Release[3]        0.1746337      1.0042286  0.008786206    0.6346822 -0.6967400
#> Release[4]        0.7350462      1.0051688  0.013257159    0.7021004 -0.3941925
#> Release[5]        0.6011591      0.9868031  0.010231959    0.2214320 -0.2867853
#>              kurtosis JarqueBera.value JarqueBera.pvalue BreuschPagan.R2
#> Release[2]  0.3093391        0.7134431         0.6999674    2.875448e-03
#> Release[3] -0.3966026        2.1865418         0.3351186    8.673001e-05
#> Release[4] -1.1178435        1.2474167         0.5359533    7.623687e-06
#> Release[5] -0.6337590        0.3957599         0.8204684    3.684791e-02
#>            BreuschPagan.value BreuschPagan.pvalue    White.R2 White.value
#> Release[2]       0.0778609780           0.7823436 0.006939144   0.2012352
#> Release[3]       0.0020817007           0.9639861 0.038710633   1.0064765
#> Release[4]       0.0001143562           0.9916087 0.143339323   2.4367685
#> Release[5]       0.4590915257           0.5109015 0.053015148   0.7422121
#>            White.pvalue     arch.R2 arch.value arch.pvalue
#> Release[2]    0.9042788 0.081562164 2.28374060   0.1307362
#> Release[3]    0.6045698 0.042631212 1.06578030   0.3019005
#> Release[4]    0.2957076 0.007598763 0.12158021   0.7273269
#> Release[5]    0.6899708 0.001983922 0.02579099   0.8724118