Vector error correction model (VECM)

Can lead to a better understanding of the nature of any nonstationary process among the different component series.

Usage

vecm(
  vintages.view,
  lag = 2,
  model = c("none", "cnt", "trend"),
  na.zero = FALSE
)

Arguments

vintages.view

mts object. Vertical or diagonal view of the create_vintages() output

lag

Number of lags

model

Character. Must be "none" (the default), "cnt" or "trend".

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

Examples

## Simulated data
df_long <- simulate_long(
    n_period = 10L * 4L,
    n_revision = 5L,
    periodicity = 4L,
    start_period = as.Date("2010-01-01")
)

## Create vintage and test
vintages <- create_vintages(df_long, periodicity = 4L)
vecm(vintages[["diagonal_view"]])
#>                            trace(2)  trace(1)    max(2)    max(1)
#> [Release[1]]_[Release[2]] 6.6909224 19.041196 6.6909224 12.350273
#> [Release[1]]_[Release[3]] 1.3597393 12.174505 1.3597393 10.814766
#> [Release[1]]_[Release[4]] 3.3051576 12.162728 3.3051576  8.857570
#> [Release[1]]_[Release[5]] 1.0324828  9.401064 1.0324828  8.368581
#> [Release[2]]_[Release[3]] 3.9989639 17.162561 3.9989639 13.163597
#> [Release[2]]_[Release[4]] 4.7018434 17.827830 4.7018434 13.125986
#> [Release[2]]_[Release[5]] 1.1038851 14.569560 1.1038851 13.465675
#> [Release[3]]_[Release[4]] 3.2701793 13.843195 3.2701793 10.573016
#> [Release[3]]_[Release[5]] 1.0924880 15.748867 1.0924880 14.656379
#> [Release[4]]_[Release[5]] 0.7348546 11.173550 0.7348546 10.438696