Vector error correction model (VECM)

Can lead to a better understanding of the nature of any nonstationary process among the different component series.

Usage

vecm(
  vintages.view,
  lag = 2,
  model = c("none", "cnt", "trend"),
  na.zero = FALSE
)

Arguments

vintages.view

mts object. Vertical or diagonal view of the `create_vintages()` output

lag

Number of lags

model

Character. Must be "none" (the default), "cnt" or "trend".

na.zero

Boolean whether missing values should be considered as 0 or rather as data not (yet) available (the default).

Examples

## Simulated data
period_range<-seq(as.Date('2011-01-01'),as.Date('2020-10-01'),by='quarter')
qtr<-(as.numeric(substr(period_range,6,7))+2)/3
time_period<-rep(paste0(format(period_range, "%Y"), "Q", qtr),5)
np<-length(period_range)
rev_date<-c(rep("2021-06-30",np), rep("2021-12-31",np), rep("2022-06-30",np),
            rep("2022-12-31",np), rep("2023-06-30",np))
set.seed(1)
xt<-cumsum(sample(rnorm(1000,0,1), np, TRUE))
rev<-rnorm(np*4,0,.1)
obs_values<-xt
for(i in 1:4) {
  xt<-xt+rev[(1+(i-1)*np):(i*np)]
  obs_values<-c(obs_values,xt)
}
df<-data.frame(rev_date, time_period, obs_values)

## Create vintage and test
vintages<-create_vintages(df, periodicity = 4)
vecm(vintages$diagonal_view)
#>                           trace(2) trace(1)   max(2)   max(1)
#> [Release[1]]_[Release[2]] 1.657778 24.16726 1.657778 22.50948
#> [Release[1]]_[Release[3]] 1.671109 30.62985 1.671109 28.95874
#> [Release[1]]_[Release[4]] 1.640661 23.37093 1.640661 21.73027
#> [Release[1]]_[Release[5]] 1.760839 21.22042 1.760839 19.45958
#> [Release[2]]_[Release[3]] 1.902635 24.26282 1.902635 22.36019
#> [Release[2]]_[Release[4]] 1.852304 17.28764 1.852304 15.43534
#> [Release[2]]_[Release[5]] 2.022543 17.67682 2.022543 15.65428
#> [Release[3]]_[Release[4]] 2.039084 21.89822 2.039084 19.85913
#> [Release[3]]_[Release[5]] 2.233745 22.39236 2.233745 20.15862
#> [Release[4]]_[Release[5]] 2.106855 23.91079 2.106855 21.80393